Retrieve Acoustic data from the LifeWatch project

Retrieve Acoustic data from the LifeWatch project

getAcousticData(startdate, stopdate, minband, maxband, by, params = FALSE, ...)

Arguments

startdate

Starting date for the query in UTC. Must be a character date of the form "2020-02-26T10:35:00.00Z"

stopdate

Stopping date for the query in UTC. Must be a character of the form "2020-02-26T11:35:00.00Z"

minband

lower frequency band from which the sound value will be retrieved

maxband

upper frequency band from which the sound value will be retrieved

by

Sample period, one of "1 min", "1 hour" or "1 day"

params

If TRUE, returns a list with the dataset and the query parameters applied in the server side. IF FALSE returns only the data.

...

Reserved for internal use.

Value

Dataframe with the acoustic data within the specified daterange and location.

Examples

getAcousticData("2021-03-26T10:35:00.00Z", "2021-03-26T10:45:00.00Z",
                minband = 9, maxband = 11, by = "1 min")
#> - Query mode: Post request to OpenCPU server
#> No encoding supplied: defaulting to UTF-8.
#> - Query parameters: 
#>    - type: acoustic data
#>    - binSize: 1 min
#>    - daterange: c("2021-03-26T10:35:00.00Z", "2021-03-26T10:45:00.00Z")
#>    - bands: c(9, 11)
#>                   Time frequency    value upper_frequency lower_frequency
#> 1  2021-03-26 10:35:00         9 109.8838             9.5             8.5
#> 2  2021-03-26 10:35:00        10 107.1073            10.5             9.5
#> 3  2021-03-26 10:35:00        11 111.0535            11.5            10.5
#> 4  2021-03-26 10:36:00         9 107.9184             9.5             8.5
#> 5  2021-03-26 10:36:00        10 105.5102            10.5             9.5
#> 6  2021-03-26 10:36:00        11 109.9490            11.5            10.5
#> 7  2021-03-26 10:37:00         9 105.3126             9.5             8.5
#> 8  2021-03-26 10:37:00        10 101.9817            10.5             9.5
#> 9  2021-03-26 10:37:00        11 106.2558            11.5            10.5
#> 10 2021-03-26 10:38:00         9 111.9692             9.5             8.5
#> 11 2021-03-26 10:38:00        10 109.0077            10.5             9.5
#> 12 2021-03-26 10:38:00        11 112.8820            11.5            10.5
#> 13 2021-03-26 10:39:00         9 112.2446             9.5             8.5
#> 14 2021-03-26 10:39:00        10 110.5370            10.5             9.5
#> 15 2021-03-26 10:39:00        11 112.6930            11.5            10.5
#> 16 2021-03-26 10:40:00         9 111.6737             9.5             8.5
#> 17 2021-03-26 10:40:00        10 109.5911            10.5             9.5
#> 18 2021-03-26 10:40:00        11 111.7508            11.5            10.5
#> 19 2021-03-26 10:41:00         9 106.9139             9.5             8.5
#> 20 2021-03-26 10:41:00        10 104.0844            10.5             9.5
#> 21 2021-03-26 10:41:00        11 106.2386            11.5            10.5
#> 22 2021-03-26 10:42:00         9 108.2736             9.5             8.5
#> 23 2021-03-26 10:42:00        10 105.8012            10.5             9.5
#> 24 2021-03-26 10:42:00        11 107.9502            11.5            10.5
#> 25 2021-03-26 10:43:00         9 111.7977             9.5             8.5
#> 26 2021-03-26 10:43:00        10 109.7071            10.5             9.5
#> 27 2021-03-26 10:43:00        11 113.2867            11.5            10.5
#> 28 2021-03-26 10:44:00         9 111.0933             9.5             8.5
#> 29 2021-03-26 10:44:00        10 109.3659            10.5             9.5
#> 30 2021-03-26 10:44:00        11 112.1425            11.5            10.5
#>              station_name longitude latitude etnID
#> 1         bpns-Birkenfels    2.5362  51.6483 12206
#> 2         bpns-Birkenfels    2.5362  51.6483 12206
#> 3         bpns-Birkenfels    2.5362  51.6483 12206
#> 4         bpns-Birkenfels    2.5362  51.6483 12206
#> 5         bpns-Birkenfels    2.5362  51.6483 12206
#> 6         bpns-Birkenfels    2.5362  51.6483 12206
#> 7         bpns-Birkenfels    2.5362  51.6483 12206
#> 8         bpns-Birkenfels    2.5362  51.6483 12206
#> 9         bpns-Birkenfels    2.5362  51.6483 12206
#> 10        bpns-Birkenfels    2.5362  51.6483 12206
#> 11        bpns-Birkenfels    2.5362  51.6483 12206
#> 12        bpns-Birkenfels    2.5362  51.6483 12206
#> 13        bpns-Birkenfels    2.5362  51.6483 12206
#> 14        bpns-Birkenfels    2.5362  51.6483 12206
#> 15        bpns-Birkenfels    2.5362  51.6483 12206
#> 16        bpns-Birkenfels    2.5362  51.6483 12206
#> 17        bpns-Birkenfels    2.5362  51.6483 12206
#> 18        bpns-Birkenfels    2.5362  51.6483 12206
#> 19        bpns-Birkenfels    2.5362  51.6483 12206
#> 20        bpns-Birkenfels    2.5362  51.6483 12206
#> 21        bpns-Birkenfels    2.5362  51.6483 12206
#> 22 bpns-Reefballs Belwind    2.8126  51.7044  9586
#> 23 bpns-Reefballs Belwind    2.8126  51.7044  9586
#> 24 bpns-Reefballs Belwind    2.8126  51.7044  9586
#> 25 bpns-Reefballs Belwind    2.8126  51.7044  9586
#> 26 bpns-Reefballs Belwind    2.8126  51.7044  9586
#> 27 bpns-Reefballs Belwind    2.8126  51.7044  9586
#> 28 bpns-Reefballs Belwind    2.8126  51.7044  9586
#> 29 bpns-Reefballs Belwind    2.8126  51.7044  9586
#> 30 bpns-Reefballs Belwind    2.8126  51.7044  9586