Retrieve C-POD data from the LifeWatch project

Retrieves the C-POD (Cetacean Acoustic Hydrophone Network) data from the LifeWatch project. Need valid authentication to access the entire data.To get an account, register via the Lifewatch RShiny registration webpage.

getCpodData(
  startdate,
  stopdate,
  processing,
  quality = c("Hi", "Mod", "Lo"),
  by,
  usr = NULL,
  pwd = NULL,
  params = FALSE,
  ...
)

Arguments

startdate

Starting date for the query

stopdate

Stopping date for the query

processing

One of ('Validated','Raw'). If "Validated", the quality parameter is ignored.

quality

One or more of ("Hi","Mod", "Lo"). This parameter is ignored if processing = "Validated"

by

Sample period, one of ("1 day", "1 week", "60 min","10 min", "1 min")

usr

Username to connect to ETN database

pwd

Password to connect to ETN database

params

If TRUE, returns a list with the dataset and the query parameters applied in the server side. IF FALSE returns only the data.

...

Params to be passed to lw_check_lwdataserver().

Value

Dataframe with the specified C-POD data.

Examples

getCpodData("2020-04-19", "2020-04-21", processing = "Validated", by = "1 week")
#> - Query mode: Post request to OpenCPU server
#> No encoding supplied: defaulting to UTF-8.
#> - Query parameters: 
#>    - daterange: c("2020-04-19", "2020-04-21")
#>    - type: CPOD data
#>    - binSize: 1 week
#>    - processing: Validated
#>    - quality: c("Hi", "Mod", "Lo")
#>    deployment_fk       time  species milliseconds number_clicks_filtered
#> 1           6168 2020-04-13 Dolphins            0                      0
#> 2           6168 2020-04-13     NBHF        24970                   1064
#> 3           6168 2020-04-13    sonar         7841                     77
#> 4           6170 2020-04-13     NBHF         5035                    166
#> 5           6170 2020-04-13    sonar         3752                     53
#> 6           6171 2020-04-13 Dolphins            0                      0
#> 7           6171 2020-04-13     NBHF         2686                    411
#> 8           6171 2020-04-13    sonar        22027                    281
#> 9           9100 2020-04-13 Dolphins            0                      0
#> 10          9100 2020-04-13     NBHF        76376                   3323
#> 11          9100 2020-04-13    sonar        17761                     69
#> 12          9103 2020-04-13 Dolphins            0                      0
#> 13          9103 2020-04-13     NBHF        99311                   4818
#> 14          6170 2020-04-20     NBHF         6450                    359
#> 15          6170 2020-04-20    sonar       423814                   2070
#> 16          6171 2020-04-20 Dolphins            0                      0
#> 17          6171 2020-04-20     NBHF        21977                    923
#> 18          6171 2020-04-20    sonar       210067                   1091
#> 19          9100 2020-04-20 Dolphins            0                      0
#> 20          9100 2020-04-20     NBHF        57559                   2137
#> 21          9100 2020-04-20    sonar         9000                     95
#> 22          9103 2020-04-20 Dolphins            0                      0
#> 23          9103 2020-04-20     NBHF       340232                  21053
#>    number_clicks_total lost_minutes recorded dpm dph dpd dpw quality
#> 1               134240            0      900   0   0   0   0       4
#> 2               134240            0      900  26   9   1   1       4
#> 3               134240            0      900   2   1   1   1       4
#> 4                 6768            0     1440   7   3   1   1       4
#> 5                 6768            0     1440   2   1   1   1       4
#> 6                42406            0     1440   0   0   0   0       4
#> 7                42406            0     1440   7   3   1   1       4
#> 8                42406            0     1440   2   1   1   1       4
#> 9                59464            0     1440   0   0   0   0       4
#> 10               59464            0     1440 121  14   1   1       4
#> 11               59464            0     1440   3   3   1   1       4
#> 12              173744            0     1380   0   0   0   0       4
#> 13              173744            0     1380 140  21   1   1       4
#> 14               21838            0     1500  10   5   1   1       4
#> 15               21838            0     1500  22   5   1   1       4
#> 16               37954            0     1500   0   0   0   0       4
#> 17               37954            0     1500  32  10   2   1       4
#> 18               37954            0     1500  24   3   1   1       4
#> 19               34882            0     1500   0   0   0   0       4
#> 20               34882            0     1500  79  18   2   1       4
#> 21               34882            0     1500   3   1   1   1       4
#> 22              340646            0     1320   0   0   0   0       4
#> 23              340646            0     1320 293  20   2   1       4
#>    projectname          station latitude longitude   mooring_type receiver
#> 1    Lifewatch bpns-Buitenratel  51.2403    2.5047 bottom-mooring POD-3416
#> 2    Lifewatch bpns-Buitenratel  51.2403    2.5047 bottom-mooring POD-3416
#> 3    Lifewatch bpns-Buitenratel  51.2403    2.5047 bottom-mooring POD-3416
#> 4    Lifewatch bpns-Nautica Ena  51.3007    2.8814 bottom-mooring POD-3401
#> 5    Lifewatch bpns-Nautica Ena  51.3007    2.8814 bottom-mooring POD-3401
#> 6    Lifewatch bpns-Nautica Ena  51.3007    2.8809 bottom-mooring POD-2977
#> 7    Lifewatch bpns-Nautica Ena  51.3007    2.8809 bottom-mooring POD-2977
#> 8    Lifewatch bpns-Nautica Ena  51.3007    2.8809 bottom-mooring POD-2977
#> 9    Lifewatch     bpns-Grafton  51.4065    2.8190 bottom-mooring POD-3331
#> 10   Lifewatch     bpns-Grafton  51.4065    2.8190 bottom-mooring POD-3331
#> 11   Lifewatch     bpns-Grafton  51.4065    2.8190 bottom-mooring POD-3331
#> 12   Lifewatch   bpns-Faulbaums  51.3345    2.6150 bottom-mooring POD-3219
#> 13   Lifewatch   bpns-Faulbaums  51.3345    2.6150 bottom-mooring POD-3219
#> 14   Lifewatch bpns-Nautica Ena  51.3007    2.8814 bottom-mooring POD-3401
#> 15   Lifewatch bpns-Nautica Ena  51.3007    2.8814 bottom-mooring POD-3401
#> 16   Lifewatch bpns-Nautica Ena  51.3007    2.8809 bottom-mooring POD-2977
#> 17   Lifewatch bpns-Nautica Ena  51.3007    2.8809 bottom-mooring POD-2977
#> 18   Lifewatch bpns-Nautica Ena  51.3007    2.8809 bottom-mooring POD-2977
#> 19   Lifewatch     bpns-Grafton  51.4065    2.8190 bottom-mooring POD-3331
#> 20   Lifewatch     bpns-Grafton  51.4065    2.8190 bottom-mooring POD-3331
#> 21   Lifewatch     bpns-Grafton  51.4065    2.8190 bottom-mooring POD-3331
#> 22   Lifewatch   bpns-Faulbaums  51.3345    2.6150 bottom-mooring POD-3219
#> 23   Lifewatch   bpns-Faulbaums  51.3345    2.6150 bottom-mooring POD-3219
#>       deploy_date_time   recover_date_time valid_data_until_datetime
#> 1  2019-10-24 09:15:00 2020-05-04 14:00:00       2020-05-04 10:27:00
#> 2  2019-10-24 09:15:00 2020-05-04 14:00:00       2020-05-04 10:27:00
#> 3  2019-10-24 09:15:00 2020-05-04 14:00:00       2020-05-04 10:27:00
#> 4  2019-10-24 14:42:00 2020-05-05 10:29:00       2020-05-01 18:24:00
#> 5  2019-10-24 14:42:00 2020-05-05 10:29:00       2020-05-01 18:24:00
#> 6  2020-03-04 15:30:00 2020-07-17 09:00:00       2020-05-10 20:15:00
#> 7  2020-03-04 15:30:00 2020-07-17 09:00:00       2020-05-10 20:15:00
#> 8  2020-03-04 15:30:00 2020-07-17 09:00:00       2020-05-10 20:15:00
#> 9  2020-03-04 10:00:00 2020-07-13 13:25:00       2020-07-13 13:25:00
#> 10 2020-03-04 10:00:00 2020-07-13 13:25:00       2020-07-13 13:25:00
#> 11 2020-03-04 10:00:00 2020-07-13 13:25:00       2020-07-13 13:25:00
#> 12 2020-03-04 13:35:00 2020-07-13 10:30:00       2020-07-13 10:30:00
#> 13 2020-03-04 13:35:00 2020-07-13 10:30:00       2020-07-13 10:30:00
#> 14 2019-10-24 14:42:00 2020-05-05 10:29:00       2020-05-01 18:24:00
#> 15 2019-10-24 14:42:00 2020-05-05 10:29:00       2020-05-01 18:24:00
#> 16 2020-03-04 15:30:00 2020-07-17 09:00:00       2020-05-10 20:15:00
#> 17 2020-03-04 15:30:00 2020-07-17 09:00:00       2020-05-10 20:15:00
#> 18 2020-03-04 15:30:00 2020-07-17 09:00:00       2020-05-10 20:15:00
#> 19 2020-03-04 10:00:00 2020-07-13 13:25:00       2020-07-13 13:25:00
#> 20 2020-03-04 10:00:00 2020-07-13 13:25:00       2020-07-13 13:25:00
#> 21 2020-03-04 10:00:00 2020-07-13 13:25:00       2020-07-13 13:25:00
#> 22 2020-03-04 13:35:00 2020-07-13 10:30:00       2020-07-13 10:30:00
#> 23 2020-03-04 13:35:00 2020-07-13 10:30:00       2020-07-13 10:30:00
getCpodData("2020-04-19", "2020-04-21", processing = "Raw",
quality = c("Hi", "Lo"), by = "1 day", params = TRUE)
#> - Query mode: Post request to OpenCPU server
#> No encoding supplied: defaulting to UTF-8.
#> - Query parameters: 
#>    - daterange: c("2020-04-19", "2020-04-21")
#>    - type: CPOD data
#>    - binSize: 1 day
#>    - processing: Raw
#>    - quality: c("Hi", "Lo")
#> $mdf
#>    deployment_fk       Time  species milliseconds number_clicks_filtered
#> 1           6171 2020-04-20 Dolphins            0                      0
#> 2           6171 2020-04-21    sonar            0                      0
#> 3           6171 2020-04-21     NBHF         6268                    197
#> 4           6171 2020-04-21 Dolphins            0                      0
#> 5           6171 2020-04-20    sonar       210067                   1091
#> 6           6171 2020-04-19 Dolphins            0                      0
#> 7           6171 2020-04-19     NBHF         6101                    514
#> 8           6171 2020-04-19    sonar        22027                    281
#> 9           6171 2020-04-20     NBHF        30300                   1012
#> 10          9103 2020-04-21     NBHF            0                      0
#> 11          9103 2020-04-19 Dolphins            0                      0
#> 12          9103 2020-04-19     NBHF       333027                   9287
#> 13          9103 2020-04-20 Dolphins            0                      0
#> 14          9103 2020-04-20     NBHF       743165                  27716
#> 15          9103 2020-04-21 Dolphins            0                      0
#> 16          9100 2020-04-21    sonar            0                      0
#> 17          9100 2020-04-20 Dolphins            0                      0
#> 18          9100 2020-04-19    sonar        17761                     69
#> 19          9100 2020-04-19     NBHF       136226                   4700
#> 20          9100 2020-04-19 Dolphins            0                      0
#> 21          9100 2020-04-21 Dolphins            0                      0
#> 22          9100 2020-04-21     NBHF            0                      0
#> 23          9100 2020-04-20     NBHF        91828                   3016
#> 24          9100 2020-04-20    sonar         9000                     95
#> 25          6170 2020-04-21     NBHF            0                      0
#> 26          6170 2020-04-19     NBHF         8231                    201
#> 27          6170 2020-04-19    sonar         3752                     53
#> 28          6170 2020-04-20     NBHF        15134                    480
#> 29          6170 2020-04-20    sonar       423814                   2070
#> 30          6170 2020-04-21    sonar            0                      0
#> 31          6168 2020-04-20     NBHF          164                     16
#> 32          6168 2020-04-19 Dolphins            0                      0
#> 33          6168 2020-04-21 Dolphins            0                      0
#> 34          6168 2020-04-21     NBHF            0                      0
#> 35          6168 2020-04-21    sonar            0                      0
#> 36          6168 2020-04-19    sonar         7841                     77
#> 37          6168 2020-04-19     NBHF        52050                   1487
#> 38          6168 2020-04-20 Dolphins            0                      0
#> 39          6168 2020-04-20    sonar            0                      0
#>    number_clicks_total lost_minutes recorded dpm dph dpd quality projectname
#> 1                17983            0     1440   0   0   0       3   Lifewatch
#> 2                  572            0       10   0   0   0       3   Lifewatch
#> 3                  572            0       10   2   1   1       3   Lifewatch
#> 4                  572            0       10   0   0   0       3   Lifewatch
#> 5                17983            0     1440  24   3   1       3   Lifewatch
#> 6                21203            0     1440   0   0   0       3   Lifewatch
#> 7                21203            0     1440   9   3   1       3   Lifewatch
#> 8                21203            0     1440   2   1   1       3   Lifewatch
#> 9                17983            0     1440  36   9   1       3   Lifewatch
#> 10                 218            0       10   0   0   0       3   Lifewatch
#> 11              109383            3     1440   0   0   0       3   Lifewatch
#> 12              109383            3     1440 261  22   1       3   Lifewatch
#> 13              329189           18     1440   0   0   0       3   Lifewatch
#> 14              329189           18     1440 410  21   1       3   Lifewatch
#> 15                 218            0       10   0   0   0       3   Lifewatch
#> 16                  56            0       10   0   0   0       3   Lifewatch
#> 17               16731            0     1440   0   0   0       3   Lifewatch
#> 18               29732            0     1440   3   3   1       3   Lifewatch
#> 19               29732            0     1440 164  15   1       3   Lifewatch
#> 20               29732            0     1440   0   0   0       3   Lifewatch
#> 21                  56            0       10   0   0   0       3   Lifewatch
#> 22                  56            0       10   0   0   0       3   Lifewatch
#> 23               16731            0     1440 113  17   1       3   Lifewatch
#> 24               16731            0     1440   3   1   1       3   Lifewatch
#> 25                  22            0       10   0   0   0       3   Lifewatch
#> 26                3384            0     1440   8   3   1       3   Lifewatch
#> 27                3384            0     1440   2   1   1       3   Lifewatch
#> 28               10858            0     1440  14   5   1       3   Lifewatch
#> 29               10858            0     1440  22   5   1       3   Lifewatch
#> 30                  22            0       10   0   0   0       3   Lifewatch
#> 31             5394317         1175     1440   3   3   1       3   Lifewatch
#> 32             1630852          214     1440   0   0   0       3   Lifewatch
#> 33               40950           10       10   0   0   0       3   Lifewatch
#> 34               40950           10       10   0   0   0       3   Lifewatch
#> 35               40950           10       10   0   0   0       3   Lifewatch
#> 36             1630852          214     1440   2   1   1       3   Lifewatch
#> 37             1630852          214     1440  38  12   1       3   Lifewatch
#> 38             5394317         1175     1440   0   0   0       3   Lifewatch
#> 39             5394317         1175     1440   0   0   0       3   Lifewatch
#>             station latitude longitude   mooring_type receiver
#> 1  bpns-Nautica Ena  51.3007    2.8809 bottom-mooring POD-2977
#> 2  bpns-Nautica Ena  51.3007    2.8809 bottom-mooring POD-2977
#> 3  bpns-Nautica Ena  51.3007    2.8809 bottom-mooring POD-2977
#> 4  bpns-Nautica Ena  51.3007    2.8809 bottom-mooring POD-2977
#> 5  bpns-Nautica Ena  51.3007    2.8809 bottom-mooring POD-2977
#> 6  bpns-Nautica Ena  51.3007    2.8809 bottom-mooring POD-2977
#> 7  bpns-Nautica Ena  51.3007    2.8809 bottom-mooring POD-2977
#> 8  bpns-Nautica Ena  51.3007    2.8809 bottom-mooring POD-2977
#> 9  bpns-Nautica Ena  51.3007    2.8809 bottom-mooring POD-2977
#> 10   bpns-Faulbaums  51.3345    2.6150 bottom-mooring POD-3219
#> 11   bpns-Faulbaums  51.3345    2.6150 bottom-mooring POD-3219
#> 12   bpns-Faulbaums  51.3345    2.6150 bottom-mooring POD-3219
#> 13   bpns-Faulbaums  51.3345    2.6150 bottom-mooring POD-3219
#> 14   bpns-Faulbaums  51.3345    2.6150 bottom-mooring POD-3219
#> 15   bpns-Faulbaums  51.3345    2.6150 bottom-mooring POD-3219
#> 16     bpns-Grafton  51.4065    2.8190 bottom-mooring POD-3331
#> 17     bpns-Grafton  51.4065    2.8190 bottom-mooring POD-3331
#> 18     bpns-Grafton  51.4065    2.8190 bottom-mooring POD-3331
#> 19     bpns-Grafton  51.4065    2.8190 bottom-mooring POD-3331
#> 20     bpns-Grafton  51.4065    2.8190 bottom-mooring POD-3331
#> 21     bpns-Grafton  51.4065    2.8190 bottom-mooring POD-3331
#> 22     bpns-Grafton  51.4065    2.8190 bottom-mooring POD-3331
#> 23     bpns-Grafton  51.4065    2.8190 bottom-mooring POD-3331
#> 24     bpns-Grafton  51.4065    2.8190 bottom-mooring POD-3331
#> 25 bpns-Nautica Ena  51.3007    2.8814 bottom-mooring POD-3401
#> 26 bpns-Nautica Ena  51.3007    2.8814 bottom-mooring POD-3401
#> 27 bpns-Nautica Ena  51.3007    2.8814 bottom-mooring POD-3401
#> 28 bpns-Nautica Ena  51.3007    2.8814 bottom-mooring POD-3401
#> 29 bpns-Nautica Ena  51.3007    2.8814 bottom-mooring POD-3401
#> 30 bpns-Nautica Ena  51.3007    2.8814 bottom-mooring POD-3401
#> 31 bpns-Buitenratel  51.2403    2.5047 bottom-mooring POD-3416
#> 32 bpns-Buitenratel  51.2403    2.5047 bottom-mooring POD-3416
#> 33 bpns-Buitenratel  51.2403    2.5047 bottom-mooring POD-3416
#> 34 bpns-Buitenratel  51.2403    2.5047 bottom-mooring POD-3416
#> 35 bpns-Buitenratel  51.2403    2.5047 bottom-mooring POD-3416
#> 36 bpns-Buitenratel  51.2403    2.5047 bottom-mooring POD-3416
#> 37 bpns-Buitenratel  51.2403    2.5047 bottom-mooring POD-3416
#> 38 bpns-Buitenratel  51.2403    2.5047 bottom-mooring POD-3416
#> 39 bpns-Buitenratel  51.2403    2.5047 bottom-mooring POD-3416
#>       deploy_date_time   recover_date_time valid_data_until_datetime
#> 1  2020-03-04 15:30:00 2020-07-17 09:00:00       2020-05-10 20:15:00
#> 2  2020-03-04 15:30:00 2020-07-17 09:00:00       2020-05-10 20:15:00
#> 3  2020-03-04 15:30:00 2020-07-17 09:00:00       2020-05-10 20:15:00
#> 4  2020-03-04 15:30:00 2020-07-17 09:00:00       2020-05-10 20:15:00
#> 5  2020-03-04 15:30:00 2020-07-17 09:00:00       2020-05-10 20:15:00
#> 6  2020-03-04 15:30:00 2020-07-17 09:00:00       2020-05-10 20:15:00
#> 7  2020-03-04 15:30:00 2020-07-17 09:00:00       2020-05-10 20:15:00
#> 8  2020-03-04 15:30:00 2020-07-17 09:00:00       2020-05-10 20:15:00
#> 9  2020-03-04 15:30:00 2020-07-17 09:00:00       2020-05-10 20:15:00
#> 10 2020-03-04 13:35:00 2020-07-13 10:30:00       2020-07-13 10:30:00
#> 11 2020-03-04 13:35:00 2020-07-13 10:30:00       2020-07-13 10:30:00
#> 12 2020-03-04 13:35:00 2020-07-13 10:30:00       2020-07-13 10:30:00
#> 13 2020-03-04 13:35:00 2020-07-13 10:30:00       2020-07-13 10:30:00
#> 14 2020-03-04 13:35:00 2020-07-13 10:30:00       2020-07-13 10:30:00
#> 15 2020-03-04 13:35:00 2020-07-13 10:30:00       2020-07-13 10:30:00
#> 16 2020-03-04 10:00:00 2020-07-13 13:25:00       2020-07-13 13:25:00
#> 17 2020-03-04 10:00:00 2020-07-13 13:25:00       2020-07-13 13:25:00
#> 18 2020-03-04 10:00:00 2020-07-13 13:25:00       2020-07-13 13:25:00
#> 19 2020-03-04 10:00:00 2020-07-13 13:25:00       2020-07-13 13:25:00
#> 20 2020-03-04 10:00:00 2020-07-13 13:25:00       2020-07-13 13:25:00
#> 21 2020-03-04 10:00:00 2020-07-13 13:25:00       2020-07-13 13:25:00
#> 22 2020-03-04 10:00:00 2020-07-13 13:25:00       2020-07-13 13:25:00
#> 23 2020-03-04 10:00:00 2020-07-13 13:25:00       2020-07-13 13:25:00
#> 24 2020-03-04 10:00:00 2020-07-13 13:25:00       2020-07-13 13:25:00
#> 25 2019-10-24 14:42:00 2020-05-05 10:29:00       2020-05-01 18:24:00
#> 26 2019-10-24 14:42:00 2020-05-05 10:29:00       2020-05-01 18:24:00
#> 27 2019-10-24 14:42:00 2020-05-05 10:29:00       2020-05-01 18:24:00
#> 28 2019-10-24 14:42:00 2020-05-05 10:29:00       2020-05-01 18:24:00
#> 29 2019-10-24 14:42:00 2020-05-05 10:29:00       2020-05-01 18:24:00
#> 30 2019-10-24 14:42:00 2020-05-05 10:29:00       2020-05-01 18:24:00
#> 31 2019-10-24 09:15:00 2020-05-04 14:00:00       2020-05-04 10:27:00
#> 32 2019-10-24 09:15:00 2020-05-04 14:00:00       2020-05-04 10:27:00
#> 33 2019-10-24 09:15:00 2020-05-04 14:00:00       2020-05-04 10:27:00
#> 34 2019-10-24 09:15:00 2020-05-04 14:00:00       2020-05-04 10:27:00
#> 35 2019-10-24 09:15:00 2020-05-04 14:00:00       2020-05-04 10:27:00
#> 36 2019-10-24 09:15:00 2020-05-04 14:00:00       2020-05-04 10:27:00
#> 37 2019-10-24 09:15:00 2020-05-04 14:00:00       2020-05-04 10:27:00
#> 38 2019-10-24 09:15:00 2020-05-04 14:00:00       2020-05-04 10:27:00
#> 39 2019-10-24 09:15:00 2020-05-04 14:00:00       2020-05-04 10:27:00
#> 
#> $par
#> $par$daterange
#> [1] "2020-04-19" "2020-04-21"
#> 
#> $par$type
#> [1] "CPOD data"
#> 
#> $par$binSize
#> [1] "1 day"
#> 
#> $par$processing
#> [1] "Raw"
#> 
#> $par$quality
#> [1] "Hi" "Lo"
#> 
#>