Retrieve MVB-data from the LifeWatch project

Retrieves the Meetnet Vlaams Banken (MVB) data from the LifeWatch project. Without valid credentials you are only allowed to view the "Tide TAW" parameter for the last 30 days, grouped by day/hour. To get an account, register via the Lifewatch RShiny registration webpage.

getMvbData(
  startdate,
  stopdate,
  parameters,
  stations = NULL,
  by,
  calc,
  usr = NULL,
  pwd = NULL,
  params = FALSE,
  ...
)

Arguments

startdate

Starting date for the query. Without a login, this is restricted to the last month

stopdate

Stopping date for the query. Without a login, this is restricted to the last month

parameters

One (or list) of the $parameter in listMvbStations.

stations

List of stations to be included in the query. See listMvbStations. Use stations="All" or stations=NULL to get all stations.

by

Time grouping for calculation, one of ('day','hour', '10min' ,'none')

calc

Calculation to perform given time grouping, one of ('avg', 'max', 'min', 'none').

usr

Username to connect to database

pwd

Password to connect to database

params

If TRUE, returns a list with the dataset and the query parameters applied in the server side. IF FALSE returns only the data.

...

Reserved for internal use.

Value

Dataframe with the specified MVB data.

Examples

getMvbData(Sys.Date() - 30, Sys.Date() + 1, parameters = 'Tide TAW',
stations = "Nieuwpoort", by = "day", calc = "avg")
#> - Query mode: Post request to OpenCPU server
#> No encoding supplied: defaulting to UTF-8.
#> - Your query:
#>    - daterange: c("2026-03-08", "2026-04-08")
#>    - type: MVB data
#>    - stations: Nieuwpoort
#>    - calc: avg
#>    - binSize: day
#>    - parameters: Tide TAW
#> - Server query:
#>    - daterange: c("2026-03-08", "2026-04-08")
#>    - type: MVB data
#>    - stations: Nieuwpoort
#>    - calc: avg
#>    - binSize: day
#>    - parameters: Tide (cmTAW)
#> Warning: The query applied on the server differ from the parameters you used. 
#>  Hint: You may need an account to fully access the data under moratorium. 
#>  Hint: Request access at: https://rshiny.lifewatch.be/account?p=register
#>          time    station latitude longitude Tide (cmTAW)
#> 1  2026-03-08 Nieuwpoort  51.1506    2.7281     231.9549
#> 2  2026-03-09 Nieuwpoort  51.1506    2.7281     233.6667
#> 3  2026-03-10 Nieuwpoort  51.1506    2.7281     236.2292
#> 4  2026-03-11 Nieuwpoort  51.1506    2.7281     254.2674
#> 5  2026-03-12 Nieuwpoort  51.1506    2.7281     222.2361
#> 6  2026-03-13 Nieuwpoort  51.1506    2.7281     285.2118
#> 7  2026-03-14 Nieuwpoort  51.1506    2.7281     254.4167
#> 8  2026-03-15 Nieuwpoort  51.1506    2.7281     222.2708
#> 9  2026-03-16 Nieuwpoort  51.1506    2.7281     260.3889
#> 10 2026-03-17 Nieuwpoort  51.1506    2.7281     242.6640
#> 11 2026-03-18 Nieuwpoort  51.1506    2.7281     234.2153
#> 12 2026-03-19 Nieuwpoort  51.1506    2.7281     218.6562
#> 13 2026-03-20 Nieuwpoort  51.1506    2.7281     223.9688
#> 14 2026-03-21 Nieuwpoort  51.1506    2.7281     231.1042
#> 15 2026-03-22 Nieuwpoort  51.1506    2.7281     279.5505
#> 16 2026-03-23 Nieuwpoort  51.1506    2.7281     193.2287
#> 17 2026-03-24 Nieuwpoort  51.1506    2.7281     244.7465
#> 18 2026-03-25 Nieuwpoort  51.1506    2.7281     303.0035
#> 19 2026-03-26 Nieuwpoort  51.1506    2.7281     262.4028
#> 20 2026-03-27 Nieuwpoort  51.1506    2.7281     227.3090
#> 21 2026-03-28 Nieuwpoort  51.1506    2.7281     260.0278
#> 22 2026-03-29 Nieuwpoort  51.1506    2.7281     210.3577
#> 23 2026-03-30 Nieuwpoort  51.1506    2.7281     274.2361
#> 24 2026-03-31 Nieuwpoort  51.1506    2.7281     205.3419
#> 25 2026-04-01 Nieuwpoort  51.1506    2.7281     245.8377
#> 26 2026-04-02 Nieuwpoort  51.1506    2.7281     235.0868
#> 27 2026-04-03 Nieuwpoort  51.1506    2.7281     218.4266
#> 28 2026-04-04 Nieuwpoort  51.1506    2.7281     258.9416
#> 29 2026-04-07 Nieuwpoort  51.1506    2.7281     195.6273
getMvbData(Sys.Date() - 30, Sys.Date() + 1, parameters = 'Tide TAW',
stations = "Blankenberge", by = "hour", calc = "max", params = TRUE)
#> - Query mode: Post request to OpenCPU server
#> No encoding supplied: defaulting to UTF-8.
#> - Your query:
#>    - daterange: c("2026-03-08", "2026-04-08")
#>    - type: MVB data
#>    - stations: Blankenberge
#>    - calc: max
#>    - binSize: hour
#>    - parameters: Tide TAW
#> - Server query:
#>    - daterange: c("2026-03-08", "2026-04-08")
#>    - type: MVB data
#>    - stations: Blankenberge
#>    - calc: max
#>    - binSize: hour
#>    - parameters: Tide (cmTAW)
#> Warning: The query applied on the server differ from the parameters you used. 
#>  Hint: You may need an account to fully access the data under moratorium. 
#>  Hint: Request access at: https://rshiny.lifewatch.be/account?p=register
#> $mdf
#>                    time      station latitude longitude Tide (cmTAW)
#> 1   2026-03-08 00:00:00 Blankenberge  51.3106    3.1164          230
#> 2   2026-03-08 01:00:00 Blankenberge  51.3106    3.1164          377
#> 3   2026-03-08 02:00:00 Blankenberge  51.3106    3.1164          449
#> 4   2026-03-08 03:00:00 Blankenberge  51.3106    3.1164          452
#> 5   2026-03-08 04:00:00 Blankenberge  51.3106    3.1164          427
#> 6   2026-03-08 05:00:00 Blankenberge  51.3106    3.1164          369
#> 7   2026-03-08 06:00:00 Blankenberge  51.3106    3.1164          283
#> 8   2026-03-08 07:00:00 Blankenberge  51.3106    3.1164          186
#> 9   2026-03-08 08:00:00 Blankenberge  51.3106    3.1164           94
#> 10  2026-03-08 09:00:00 Blankenberge  51.3106    3.1164           36
#> 11  2026-03-08 10:00:00 Blankenberge  51.3106    3.1164           75
#> 12  2026-03-08 11:00:00 Blankenberge  51.3106    3.1164          119
#> 13  2026-03-08 12:00:00 Blankenberge  51.3106    3.1164          189
#> 14  2026-03-08 13:00:00 Blankenberge  51.3106    3.1164          328
#> 15  2026-03-08 14:00:00 Blankenberge  51.3106    3.1164          429
#> 16  2026-03-08 15:00:00 Blankenberge  51.3106    3.1164          439
#> 17  2026-03-08 16:00:00 Blankenberge  51.3106    3.1164          425
#> 18  2026-03-08 17:00:00 Blankenberge  51.3106    3.1164          379
#> 19  2026-03-08 18:00:00 Blankenberge  51.3106    3.1164          304
#> 20  2026-03-08 19:00:00 Blankenberge  51.3106    3.1164          214
#> 21  2026-03-08 20:00:00 Blankenberge  51.3106    3.1164          126
#> 22  2026-03-08 21:00:00 Blankenberge  51.3106    3.1164           66
#> 23  2026-03-08 22:00:00 Blankenberge  51.3106    3.1164           89
#> 24  2026-03-08 23:00:00 Blankenberge  51.3106    3.1164          128
#> 25  2026-03-09 00:00:00 Blankenberge  51.3106    3.1164          188
#> 26  2026-03-09 01:00:00 Blankenberge  51.3106    3.1164          298
#> 27  2026-03-09 02:00:00 Blankenberge  51.3106    3.1164          416
#> 28  2026-03-09 03:00:00 Blankenberge  51.3106    3.1164          433
#> 29  2026-03-09 04:00:00 Blankenberge  51.3106    3.1164          426
#> 30  2026-03-09 05:00:00 Blankenberge  51.3106    3.1164          386
#> 31  2026-03-09 06:00:00 Blankenberge  51.3106    3.1164          321
#> 32  2026-03-09 07:00:00 Blankenberge  51.3106    3.1164          233
#> 33  2026-03-09 08:00:00 Blankenberge  51.3106    3.1164          145
#> 34  2026-03-09 09:00:00 Blankenberge  51.3106    3.1164           77
#> 35  2026-03-09 10:00:00 Blankenberge  51.3106    3.1164           64
#> 36  2026-03-09 11:00:00 Blankenberge  51.3106    3.1164          106
#> 37  2026-03-09 12:00:00 Blankenberge  51.3106    3.1164          159
#> 38  2026-03-09 13:00:00 Blankenberge  51.3106    3.1164          258
#> 39  2026-03-09 14:00:00 Blankenberge  51.3106    3.1164          386
#> 40  2026-03-09 15:00:00 Blankenberge  51.3106    3.1164          421
#> 41  2026-03-09 16:00:00 Blankenberge  51.3106    3.1164          420
#> 42  2026-03-09 17:00:00 Blankenberge  51.3106    3.1164          394
#> 43  2026-03-09 18:00:00 Blankenberge  51.3106    3.1164          337
#> 44  2026-03-09 19:00:00 Blankenberge  51.3106    3.1164          259
#> 45  2026-03-09 20:00:00 Blankenberge  51.3106    3.1164          176
#> 46  2026-03-09 21:00:00 Blankenberge  51.3106    3.1164          102
#> 47  2026-03-09 22:00:00 Blankenberge  51.3106    3.1164           77
#> 48  2026-03-09 23:00:00 Blankenberge  51.3106    3.1164          113
#> 49  2026-03-10 00:00:00 Blankenberge  51.3106    3.1164          164
#> 50  2026-03-10 01:00:00 Blankenberge  51.3106    3.1164          241
#> 51  2026-03-10 02:00:00 Blankenberge  51.3106    3.1164          362
#> 52  2026-03-10 03:00:00 Blankenberge  51.3106    3.1164          413
#> 53  2026-03-10 04:00:00 Blankenberge  51.3106    3.1164          415
#> 54  2026-03-10 05:00:00 Blankenberge  51.3106    3.1164          398
#> 55  2026-03-10 06:00:00 Blankenberge  51.3106    3.1164          349
#> 56  2026-03-10 07:00:00 Blankenberge  51.3106    3.1164          279
#> 57  2026-03-10 08:00:00 Blankenberge  51.3106    3.1164          198
#> 58  2026-03-10 09:00:00 Blankenberge  51.3106    3.1164          119
#> 59  2026-03-10 10:00:00 Blankenberge  51.3106    3.1164           73
#> 60  2026-03-10 11:00:00 Blankenberge  51.3106    3.1164           96
#> 61  2026-03-10 12:00:00 Blankenberge  51.3106    3.1164          137
#> 62  2026-03-10 13:00:00 Blankenberge  51.3106    3.1164          208
#> 63  2026-03-10 14:00:00 Blankenberge  51.3106    3.1164          320
#> 64  2026-03-10 15:00:00 Blankenberge  51.3106    3.1164          392
#> 65  2026-03-10 16:00:00 Blankenberge  51.3106    3.1164          400
#> 66  2026-03-10 17:00:00 Blankenberge  51.3106    3.1164          392
#> 67  2026-03-10 18:00:00 Blankenberge  51.3106    3.1164          357
#> 68  2026-03-10 19:00:00 Blankenberge  51.3106    3.1164          296
#> 69  2026-03-10 20:00:00 Blankenberge  51.3106    3.1164          221
#> 70  2026-03-10 21:00:00 Blankenberge  51.3106    3.1164          153
#> 71  2026-03-10 22:00:00 Blankenberge  51.3106    3.1164          100
#> 72  2026-03-10 23:00:00 Blankenberge  51.3106    3.1164           98
#> 73  2026-03-11 00:00:00 Blankenberge  51.3106    3.1164          135
#> 74  2026-03-11 01:00:00 Blankenberge  51.3106    3.1164          201
#> 75  2026-03-11 02:00:00 Blankenberge  51.3106    3.1164          299
#> 76  2026-03-11 03:00:00 Blankenberge  51.3106    3.1164          371
#> 77  2026-03-11 04:00:00 Blankenberge  51.3106    3.1164          382
#> 78  2026-03-11 05:00:00 Blankenberge  51.3106    3.1164          375
#> 79  2026-03-11 06:00:00 Blankenberge  51.3106    3.1164          345
#> 80  2026-03-11 07:00:00 Blankenberge  51.3106    3.1164          294
#> 81  2026-03-11 08:00:00 Blankenberge  51.3106    3.1164          236
#> 82  2026-03-11 09:00:00 Blankenberge  51.3106    3.1164          177
#> 83  2026-03-11 10:00:00 Blankenberge  51.3106    3.1164          132
#> 84  2026-03-11 11:00:00 Blankenberge  51.3106    3.1164          106
#> 85  2026-03-11 12:00:00 Blankenberge  51.3106    3.1164          142
#> 86  2026-03-11 13:00:00 Blankenberge  51.3106    3.1164          199
#> 87  2026-03-11 14:00:00 Blankenberge  51.3106    3.1164          280
#> 88  2026-03-11 15:00:00 Blankenberge  51.3106    3.1164          301
#> 89  2026-03-12 06:00:00 Blankenberge  51.3106    3.1164          352
#> 90  2026-03-12 07:00:00 Blankenberge  51.3106    3.1164          328
#> 91  2026-03-12 08:00:00 Blankenberge  51.3106    3.1164          276
#> 92  2026-03-12 09:00:00 Blankenberge  51.3106    3.1164          218
#> 93  2026-03-12 10:00:00 Blankenberge  51.3106    3.1164          160
#> 94  2026-03-12 11:00:00 Blankenberge  51.3106    3.1164          118
#> 95  2026-03-12 12:00:00 Blankenberge  51.3106    3.1164           94
#> 96  2026-03-12 13:00:00 Blankenberge  51.3106    3.1164           92
#> 97  2026-03-12 14:00:00 Blankenberge  51.3106    3.1164          144
#> 98  2026-03-12 15:00:00 Blankenberge  51.3106    3.1164          199
#> 99  2026-03-12 16:00:00 Blankenberge  51.3106    3.1164          242
#> 100 2026-03-12 17:00:00 Blankenberge  51.3106    3.1164          263
#> 101 2026-03-12 18:00:00 Blankenberge  51.3106    3.1164          265
#> 102 2026-03-12 19:00:00 Blankenberge  51.3106    3.1164          254
#> 103 2026-03-12 20:00:00 Blankenberge  51.3106    3.1164          234
#> 104 2026-03-12 21:00:00 Blankenberge  51.3106    3.1164          216
#> 105 2026-03-12 22:00:00 Blankenberge  51.3106    3.1164          188
#> 106 2026-03-12 23:00:00 Blankenberge  51.3106    3.1164          165
#> 107 2026-03-13 00:00:00 Blankenberge  51.3106    3.1164          163
#> 108 2026-03-13 01:00:00 Blankenberge  51.3106    3.1164          182
#> 109 2026-03-13 02:00:00 Blankenberge  51.3106    3.1164          214
#> 110 2026-03-13 03:00:00 Blankenberge  51.3106    3.1164          268
#> 111 2026-03-13 04:00:00 Blankenberge  51.3106    3.1164          328
#> 112 2026-03-13 05:00:00 Blankenberge  51.3106    3.1164          359
#> 113 2026-03-13 06:00:00 Blankenberge  51.3106    3.1164          413
#> 114 2026-03-13 07:00:00 Blankenberge  51.3106    3.1164          410
#> 115 2026-03-13 08:00:00 Blankenberge  51.3106    3.1164          387
#> 116 2026-03-13 09:00:00 Blankenberge  51.3106    3.1164          365
#> 117 2026-03-13 10:00:00 Blankenberge  51.3106    3.1164          348
#> 118 2026-03-13 11:00:00 Blankenberge  51.3106    3.1164          292
#> 119 2026-03-13 12:00:00 Blankenberge  51.3106    3.1164          264
#> 120 2026-03-13 13:00:00 Blankenberge  51.3106    3.1164          211
#> 121 2026-03-13 14:00:00 Blankenberge  51.3106    3.1164          201
#> 122 2026-03-13 15:00:00 Blankenberge  51.3106    3.1164          218
#> 123 2026-03-13 16:00:00 Blankenberge  51.3106    3.1164          250
#> 124 2026-03-13 17:00:00 Blankenberge  51.3106    3.1164          294
#> 125 2026-03-13 18:00:00 Blankenberge  51.3106    3.1164          328
#> 126 2026-03-13 19:00:00 Blankenberge  51.3106    3.1164          356
#> 127 2026-03-13 20:00:00 Blankenberge  51.3106    3.1164          361
#> 128 2026-03-13 21:00:00 Blankenberge  51.3106    3.1164          347
#> 129 2026-03-13 22:00:00 Blankenberge  51.3106    3.1164          329
#> 130 2026-03-13 23:00:00 Blankenberge  51.3106    3.1164          288
#> 131 2026-03-14 00:00:00 Blankenberge  51.3106    3.1164          245
#> 132 2026-03-14 01:00:00 Blankenberge  51.3106    3.1164          202
#> 133 2026-03-14 02:00:00 Blankenberge  51.3106    3.1164          173
#> 134 2026-03-14 03:00:00 Blankenberge  51.3106    3.1164          177
#> 135 2026-03-14 04:00:00 Blankenberge  51.3106    3.1164          205
#> 136 2026-03-14 05:00:00 Blankenberge  51.3106    3.1164          246
#> 137 2026-03-14 06:00:00 Blankenberge  51.3106    3.1164          302
#> 138 2026-03-14 07:00:00 Blankenberge  51.3106    3.1164          353
#> 139 2026-03-14 08:00:00 Blankenberge  51.3106    3.1164          368
#> 140 2026-03-14 09:00:00 Blankenberge  51.3106    3.1164          371
#> 141 2026-03-14 10:00:00 Blankenberge  51.3106    3.1164          361
#> 142 2026-03-14 11:00:00 Blankenberge  51.3106    3.1164          334
#> 143 2026-03-14 12:00:00 Blankenberge  51.3106    3.1164          284
#> 144 2026-03-14 13:00:00 Blankenberge  51.3106    3.1164          228
#> 145 2026-03-14 14:00:00 Blankenberge  51.3106    3.1164          178
#> 146 2026-03-14 15:00:00 Blankenberge  51.3106    3.1164          147
#> 147 2026-03-14 16:00:00 Blankenberge  51.3106    3.1164          158
#> 148 2026-03-14 17:00:00 Blankenberge  51.3106    3.1164          195
#> 149 2026-03-14 18:00:00 Blankenberge  51.3106    3.1164          232
#> 150 2026-03-14 19:00:00 Blankenberge  51.3106    3.1164          280
#> 151 2026-03-14 20:00:00 Blankenberge  51.3106    3.1164          328
#> 152 2026-03-14 21:00:00 Blankenberge  51.3106    3.1164          342
#> 153 2026-03-14 22:00:00 Blankenberge  51.3106    3.1164          340
#> 154 2026-03-14 23:00:00 Blankenberge  51.3106    3.1164          319
#> 155 2026-03-15 00:00:00 Blankenberge  51.3106    3.1164          280
#> 156 2026-03-15 01:00:00 Blankenberge  51.3106    3.1164          225
#> 157 2026-03-15 02:00:00 Blankenberge  51.3106    3.1164          169
#> 158 2026-03-15 03:00:00 Blankenberge  51.3106    3.1164          119
#> 159 2026-03-15 04:00:00 Blankenberge  51.3106    3.1164           90
#> 160 2026-03-15 05:00:00 Blankenberge  51.3106    3.1164          122
#> 161 2026-03-15 06:00:00 Blankenberge  51.3106    3.1164          173
#> 162 2026-03-15 07:00:00 Blankenberge  51.3106    3.1164          243
#> 163 2026-03-15 08:00:00 Blankenberge  51.3106    3.1164          328
#> 164 2026-03-15 09:00:00 Blankenberge  51.3106    3.1164          371
#> 165 2026-03-15 10:00:00 Blankenberge  51.3106    3.1164          376
#> 166 2026-03-15 11:00:00 Blankenberge  51.3106    3.1164          366
#> 167 2026-03-15 12:00:00 Blankenberge  51.3106    3.1164          331
#> 168 2026-03-15 13:00:00 Blankenberge  51.3106    3.1164          272
#> 169 2026-03-15 14:00:00 Blankenberge  51.3106    3.1164          206
#> 170 2026-03-15 15:00:00 Blankenberge  51.3106    3.1164          134
#> 171 2026-03-15 16:00:00 Blankenberge  51.3106    3.1164           88
#> 172 2026-03-15 17:00:00 Blankenberge  51.3106    3.1164           89
#> 173 2026-03-15 18:00:00 Blankenberge  51.3106    3.1164          122
#> 174 2026-03-15 19:00:00 Blankenberge  51.3106    3.1164          193
#> 175 2026-03-15 20:00:00 Blankenberge  51.3106    3.1164          297
#> 176 2026-03-15 21:00:00 Blankenberge  51.3106    3.1164          366
#> 177 2026-03-15 22:00:00 Blankenberge  51.3106    3.1164          389
#> 178 2026-03-15 23:00:00 Blankenberge  51.3106    3.1164          388
#> 179 2026-03-16 00:00:00 Blankenberge  51.3106    3.1164          362
#> 180 2026-03-16 01:00:00 Blankenberge  51.3106    3.1164          313
#> 181 2026-03-16 02:00:00 Blankenberge  51.3106    3.1164          262
#> 182 2026-03-16 03:00:00 Blankenberge  51.3106    3.1164          204
#> 183 2026-03-16 04:00:00 Blankenberge  51.3106    3.1164          158
#> 184 2026-03-16 05:00:00 Blankenberge  51.3106    3.1164          155
#> 185 2026-03-16 06:00:00 Blankenberge  51.3106    3.1164          188
#> 186 2026-03-16 07:00:00 Blankenberge  51.3106    3.1164          245
#> 187 2026-03-16 08:00:00 Blankenberge  51.3106    3.1164          324
#> 188 2026-03-16 09:00:00 Blankenberge  51.3106    3.1164          423
#> 189 2026-03-16 10:00:00 Blankenberge  51.3106    3.1164          450
#> 190 2026-03-16 11:00:00 Blankenberge  51.3106    3.1164          452
#> 191 2026-03-16 12:00:00 Blankenberge  51.3106    3.1164          426
#> 192 2026-03-16 13:00:00 Blankenberge  51.3106    3.1164          377
#> 193 2026-03-16 14:00:00 Blankenberge  51.3106    3.1164          304
#> 194 2026-03-16 15:00:00 Blankenberge  51.3106    3.1164          218
#> 195 2026-03-16 16:00:00 Blankenberge  51.3106    3.1164          143
#> 196 2026-03-16 17:00:00 Blankenberge  51.3106    3.1164           95
#> 197 2026-03-16 18:00:00 Blankenberge  51.3106    3.1164          125
#> 198 2026-03-16 19:00:00 Blankenberge  51.3106    3.1164          165
#> 199 2026-03-16 20:00:00 Blankenberge  51.3106    3.1164          232
#> 200 2026-03-16 21:00:00 Blankenberge  51.3106    3.1164          339
#> 201 2026-03-16 22:00:00 Blankenberge  51.3106    3.1164          405
#> 202 2026-03-16 23:00:00 Blankenberge  51.3106    3.1164          411
#> 203 2026-03-17 00:00:00 Blankenberge  51.3106    3.1164          393
#> 204 2026-03-17 01:00:00 Blankenberge  51.3106    3.1164          347
#> 205 2026-03-17 02:00:00 Blankenberge  51.3106    3.1164          284
#> 206 2026-03-17 03:00:00 Blankenberge  51.3106    3.1164          203
#> 207 2026-03-17 04:00:00 Blankenberge  51.3106    3.1164          123
#> 208 2026-03-17 05:00:00 Blankenberge  51.3106    3.1164           61
#> 209 2026-03-17 06:00:00 Blankenberge  51.3106    3.1164           68
#> 210 2026-03-17 07:00:00 Blankenberge  51.3106    3.1164          111
#> 211 2026-03-17 08:00:00 Blankenberge  51.3106    3.1164          186
#> 212 2026-03-17 09:00:00 Blankenberge  51.3106    3.1164          311
#> 213 2026-03-17 10:00:00 Blankenberge  51.3106    3.1164          412
#> 214 2026-03-17 11:00:00 Blankenberge  51.3106    3.1164          431
#> 215 2026-03-17 12:00:00 Blankenberge  51.3106    3.1164          425
#> 216 2026-03-17 13:00:00 Blankenberge  51.3106    3.1164          385
#> 217 2026-03-17 14:00:00 Blankenberge  51.3106    3.1164          321
#> 218 2026-03-17 15:00:00 Blankenberge  51.3106    3.1164          238
#> 219 2026-03-17 16:00:00 Blankenberge  51.3106    3.1164          151
#> 220 2026-03-17 17:00:00 Blankenberge  51.3106    3.1164           74
#> 221 2026-03-17 20:00:00 Blankenberge  51.3106    3.1164          155
#> 222 2026-03-17 21:00:00 Blankenberge  51.3106    3.1164          262
#> 223 2026-03-17 22:00:00 Blankenberge  51.3106    3.1164          399
#> 224 2026-03-17 23:00:00 Blankenberge  51.3106    3.1164          435
#> 225 2026-03-18 00:00:00 Blankenberge  51.3106    3.1164          433
#> 226 2026-03-18 01:00:00 Blankenberge  51.3106    3.1164          398
#> 227 2026-03-18 02:00:00 Blankenberge  51.3106    3.1164          344
#> 228 2026-03-18 03:00:00 Blankenberge  51.3106    3.1164          261
#> 229 2026-03-18 04:00:00 Blankenberge  51.3106    3.1164          169
#> 230 2026-03-18 05:00:00 Blankenberge  51.3106    3.1164           83
#> 231 2026-03-18 06:00:00 Blankenberge  51.3106    3.1164           41
#> 232 2026-03-18 07:00:00 Blankenberge  51.3106    3.1164           89
#> 233 2026-03-18 08:00:00 Blankenberge  51.3106    3.1164          141
#> 234 2026-03-18 09:00:00 Blankenberge  51.3106    3.1164          226
#> 235 2026-03-18 10:00:00 Blankenberge  51.3106    3.1164          386
#> 236 2026-03-18 11:00:00 Blankenberge  51.3106    3.1164          462
#> 237 2026-03-18 12:00:00 Blankenberge  51.3106    3.1164          467
#> 238 2026-03-18 13:00:00 Blankenberge  51.3106    3.1164          449
#> 239 2026-03-18 14:00:00 Blankenberge  51.3106    3.1164          395
#> 240 2026-03-18 15:00:00 Blankenberge  51.3106    3.1164          313
#> 241 2026-03-18 16:00:00 Blankenberge  51.3106    3.1164          212
#> 242 2026-03-18 17:00:00 Blankenberge  51.3106    3.1164          119
#> 243 2026-03-18 18:00:00 Blankenberge  51.3106    3.1164           69
#> 244 2026-03-18 19:00:00 Blankenberge  51.3106    3.1164          121
#> 245 2026-03-18 20:00:00 Blankenberge  51.3106    3.1164          159
#> 246 2026-03-18 21:00:00 Blankenberge  51.3106    3.1164          210
#> 247 2026-03-18 22:00:00 Blankenberge  51.3106    3.1164          315
#> 248 2026-03-18 23:00:00 Blankenberge  51.3106    3.1164          448
#> 249 2026-03-19 00:00:00 Blankenberge  51.3106    3.1164          456
#> 250 2026-03-19 01:00:00 Blankenberge  51.3106    3.1164          444
#> 251 2026-03-19 02:00:00 Blankenberge  51.3106    3.1164          393
#> 252 2026-03-19 03:00:00 Blankenberge  51.3106    3.1164          316
#> 253 2026-03-19 04:00:00 Blankenberge  51.3106    3.1164          216
#> 254 2026-03-19 05:00:00 Blankenberge  51.3106    3.1164          111
#> 255 2026-03-19 06:00:00 Blankenberge  51.3106    3.1164           32
#> 256 2026-03-19 07:00:00 Blankenberge  51.3106    3.1164           25
#> 257 2026-03-19 08:00:00 Blankenberge  51.3106    3.1164           82
#> 258 2026-03-19 09:00:00 Blankenberge  51.3106    3.1164          137
#> 259 2026-03-19 10:00:00 Blankenberge  51.3106    3.1164          240
#> 260 2026-03-19 11:00:00 Blankenberge  51.3106    3.1164          422
#> 261 2026-03-19 12:00:00 Blankenberge  51.3106    3.1164          460
#> 262 2026-03-19 13:00:00 Blankenberge  51.3106    3.1164          460
#> 263 2026-03-19 14:00:00 Blankenberge  51.3106    3.1164          424
#> 264 2026-03-19 15:00:00 Blankenberge  51.3106    3.1164          356
#> 265 2026-03-19 16:00:00 Blankenberge  51.3106    3.1164          258
#> 266 2026-03-19 17:00:00 Blankenberge  51.3106    3.1164          154
#> 267 2026-03-19 18:00:00 Blankenberge  51.3106    3.1164           62
#> 268 2026-03-19 19:00:00 Blankenberge  51.3106    3.1164           60
#> 269 2026-03-19 20:00:00 Blankenberge  51.3106    3.1164          106
#> 270 2026-03-19 21:00:00 Blankenberge  51.3106    3.1164          156
#> 271 2026-03-19 22:00:00 Blankenberge  51.3106    3.1164          227
#> 272 2026-03-19 23:00:00 Blankenberge  51.3106    3.1164          394
#> 273 2026-03-20 00:00:00 Blankenberge  51.3106    3.1164          467
#> 274 2026-03-20 01:00:00 Blankenberge  51.3106    3.1164          466
#> 275 2026-03-20 02:00:00 Blankenberge  51.3106    3.1164          434
#> 276 2026-03-20 03:00:00 Blankenberge  51.3106    3.1164          374
#> 277 2026-03-20 04:00:00 Blankenberge  51.3106    3.1164          283
#> 278 2026-03-20 05:00:00 Blankenberge  51.3106    3.1164          173
#> 279 2026-03-20 06:00:00 Blankenberge  51.3106    3.1164           76
#> 280 2026-03-20 07:00:00 Blankenberge  51.3106    3.1164            4
#> 281 2026-03-20 08:00:00 Blankenberge  51.3106    3.1164           57
#> 282 2026-03-20 09:00:00 Blankenberge  51.3106    3.1164          102
#> 283 2026-03-20 10:00:00 Blankenberge  51.3106    3.1164          177
#> 284 2026-03-20 11:00:00 Blankenberge  51.3106    3.1164          335
#> 285 2026-03-20 12:00:00 Blankenberge  51.3106    3.1164          466
#> 286 2026-03-20 13:00:00 Blankenberge  51.3106    3.1164          476
#> 287 2026-03-20 14:00:00 Blankenberge  51.3106    3.1164          464
#> 288 2026-03-20 15:00:00 Blankenberge  51.3106    3.1164          413
#> 289 2026-03-20 16:00:00 Blankenberge  51.3106    3.1164          327
#> 290 2026-03-20 17:00:00 Blankenberge  51.3106    3.1164          222
#> 291 2026-03-20 18:00:00 Blankenberge  51.3106    3.1164          117
#> 292 2026-03-20 19:00:00 Blankenberge  51.3106    3.1164           40
#> 293 2026-03-20 20:00:00 Blankenberge  51.3106    3.1164           85
#> 294 2026-03-20 21:00:00 Blankenberge  51.3106    3.1164          126
#> 295 2026-03-20 22:00:00 Blankenberge  51.3106    3.1164          182
#> 296 2026-03-20 23:00:00 Blankenberge  51.3106    3.1164          285
#> 297 2026-03-21 00:00:00 Blankenberge  51.3106    3.1164          459
#> 298 2026-03-21 01:00:00 Blankenberge  51.3106    3.1164          475
#> 299 2026-03-21 02:00:00 Blankenberge  51.3106    3.1164          470
#> 300 2026-03-21 03:00:00 Blankenberge  51.3106    3.1164          421
#> 301 2026-03-21 04:00:00 Blankenberge  51.3106    3.1164          345
#> 302 2026-03-21 05:00:00 Blankenberge  51.3106    3.1164          239
#> 303 2026-03-21 06:00:00 Blankenberge  51.3106    3.1164          133
#> 304 2026-03-21 07:00:00 Blankenberge  51.3106    3.1164           43
#> 305 2026-03-21 08:00:00 Blankenberge  51.3106    3.1164           27
#> 306 2026-03-21 09:00:00 Blankenberge  51.3106    3.1164           86
#> 307 2026-03-21 10:00:00 Blankenberge  51.3106    3.1164          138
#> 308 2026-03-21 11:00:00 Blankenberge  51.3106    3.1164          225
#> 309 2026-03-21 12:00:00 Blankenberge  51.3106    3.1164          424
#> 310 2026-03-21 13:00:00 Blankenberge  51.3106    3.1164          481
#> 311 2026-03-21 14:00:00 Blankenberge  51.3106    3.1164          482
#> 312 2026-03-21 15:00:00 Blankenberge  51.3106    3.1164          451
#> 313 2026-03-21 16:00:00 Blankenberge  51.3106    3.1164          388
#> 314 2026-03-21 17:00:00 Blankenberge  51.3106    3.1164          291
#> 315 2026-03-21 18:00:00 Blankenberge  51.3106    3.1164          183
#> 316 2026-03-21 19:00:00 Blankenberge  51.3106    3.1164           83
#> 317 2026-03-21 20:00:00 Blankenberge  51.3106    3.1164           42
#> 318 2026-03-21 21:00:00 Blankenberge  51.3106    3.1164          103
#> 319 2026-03-21 22:00:00 Blankenberge  51.3106    3.1164          151
#> 320 2026-03-21 23:00:00 Blankenberge  51.3106    3.1164          211
#> 321 2026-03-22 00:00:00 Blankenberge  51.3106    3.1164          359
#> 322 2026-03-22 01:00:00 Blankenberge  51.3106    3.1164          472
#> 323 2026-03-22 02:00:00 Blankenberge  51.3106    3.1164          476
#> 324 2026-03-22 03:00:00 Blankenberge  51.3106    3.1164          454
#> 325 2026-03-22 04:00:00 Blankenberge  51.3106    3.1164          397
#> 326 2026-03-22 05:00:00 Blankenberge  51.3106    3.1164          303
#> 327 2026-03-22 06:00:00 Blankenberge  51.3106    3.1164          199
#> 328 2026-03-22 07:00:00 Blankenberge  51.3106    3.1164           94
#> 329 2026-03-22 08:00:00 Blankenberge  51.3106    3.1164           15
#> 330 2026-03-22 09:00:00 Blankenberge  51.3106    3.1164           38
#> 331 2026-03-22 10:00:00 Blankenberge  51.3106    3.1164           89
#> 332 2026-03-22 11:00:00 Blankenberge  51.3106    3.1164          152
#> 333 2026-03-22 12:00:00 Blankenberge  51.3106    3.1164          284
#> 334 2026-03-22 13:00:00 Blankenberge  51.3106    3.1164          452
#> 335 2026-03-22 14:00:00 Blankenberge  51.3106    3.1164          473
#> 336 2026-03-22 15:00:00 Blankenberge  51.3106    3.1164          469
#> 337 2026-03-22 16:00:00 Blankenberge  51.3106    3.1164          427
#> 338 2026-03-22 17:00:00 Blankenberge  51.3106    3.1164          351
#> 339 2026-03-22 18:00:00 Blankenberge  51.3106    3.1164          252
#> 340 2026-03-23 08:00:00 Blankenberge  51.3106    3.1164           86
#> 341 2026-03-23 09:00:00 Blankenberge  51.3106    3.1164           34
#> 342 2026-03-23 10:00:00 Blankenberge  51.3106    3.1164           90
#> 343 2026-03-23 11:00:00 Blankenberge  51.3106    3.1164          134
#> 344 2026-03-23 12:00:00 Blankenberge  51.3106    3.1164          203
#> 345 2026-03-23 13:00:00 Blankenberge  51.3106    3.1164          357
#> 346 2026-03-23 14:00:00 Blankenberge  51.3106    3.1164          456
#> 347 2026-03-23 15:00:00 Blankenberge  51.3106    3.1164          465
#> 348 2026-03-23 16:00:00 Blankenberge  51.3106    3.1164          447
#> 349 2026-03-23 17:00:00 Blankenberge  51.3106    3.1164          394
#> 350 2026-03-23 18:00:00 Blankenberge  51.3106    3.1164          308
#> 351 2026-03-23 19:00:00 Blankenberge  51.3106    3.1164          211
#> 352 2026-03-23 20:00:00 Blankenberge  51.3106    3.1164          112
#> 353 2026-03-23 21:00:00 Blankenberge  51.3106    3.1164           44
#> 354 2026-03-23 22:00:00 Blankenberge  51.3106    3.1164           76
#> 355 2026-03-23 23:00:00 Blankenberge  51.3106    3.1164          125
#> 356 2026-03-24 00:00:00 Blankenberge  51.3106    3.1164          183
#> 357 2026-03-24 01:00:00 Blankenberge  51.3106    3.1164          298
#> 358 2026-03-24 02:00:00 Blankenberge  51.3106    3.1164          438
#> 359 2026-03-24 03:00:00 Blankenberge  51.3106    3.1164          449
#> 360 2026-03-24 04:00:00 Blankenberge  51.3106    3.1164          441
#> 361 2026-03-24 05:00:00 Blankenberge  51.3106    3.1164          385
#> 362 2026-03-24 06:00:00 Blankenberge  51.3106    3.1164          325
#> 363 2026-03-24 07:00:00 Blankenberge  51.3106    3.1164          223
#> 364 2026-03-24 08:00:00 Blankenberge  51.3106    3.1164          129
#> 365 2026-03-24 09:00:00 Blankenberge  51.3106    3.1164           61
#> 366 2026-03-24 10:00:00 Blankenberge  51.3106    3.1164           46
#> 367 2026-03-24 11:00:00 Blankenberge  51.3106    3.1164           91
#> 368 2026-03-24 12:00:00 Blankenberge  51.3106    3.1164          154
#> 369 2026-03-24 13:00:00 Blankenberge  51.3106    3.1164          251
#> 370 2026-03-24 14:00:00 Blankenberge  51.3106    3.1164          405
#> 371 2026-03-24 15:00:00 Blankenberge  51.3106    3.1164          453
#> 372 2026-03-24 16:00:00 Blankenberge  51.3106    3.1164          458
#> 373 2026-03-24 17:00:00 Blankenberge  51.3106    3.1164          438
#> 374 2026-03-24 18:00:00 Blankenberge  51.3106    3.1164          390
#> 375 2026-03-24 19:00:00 Blankenberge  51.3106    3.1164          307
#> 376 2026-03-24 20:00:00 Blankenberge  51.3106    3.1164          232
#> 377 2026-03-24 21:00:00 Blankenberge  51.3106    3.1164          153
#> 378 2026-03-24 22:00:00 Blankenberge  51.3106    3.1164          119
#> 379 2026-03-24 23:00:00 Blankenberge  51.3106    3.1164          161
#> 380 2026-03-25 00:00:00 Blankenberge  51.3106    3.1164          232
#> 381 2026-03-25 01:00:00 Blankenberge  51.3106    3.1164          306
#> 382 2026-03-25 02:00:00 Blankenberge  51.3106    3.1164          429
#> 383 2026-03-25 03:00:00 Blankenberge  51.3106    3.1164          489
#> 384 2026-03-25 04:00:00 Blankenberge  51.3106    3.1164          492
#> 385 2026-03-25 05:00:00 Blankenberge  51.3106    3.1164          486
#> 386 2026-03-25 06:00:00 Blankenberge  51.3106    3.1164          444
#> 387 2026-03-25 07:00:00 Blankenberge  51.3106    3.1164          381
#> 388 2026-03-25 08:00:00 Blankenberge  51.3106    3.1164          295
#> 389 2026-03-25 09:00:00 Blankenberge  51.3106    3.1164          226
#> 390 2026-03-25 10:00:00 Blankenberge  51.3106    3.1164          165
#> 391 2026-03-25 11:00:00 Blankenberge  51.3106    3.1164          169
#> 392 2026-03-25 12:00:00 Blankenberge  51.3106    3.1164          210
#> 393 2026-03-25 13:00:00 Blankenberge  51.3106    3.1164          268
#> 394 2026-03-25 14:00:00 Blankenberge  51.3106    3.1164          335
#> 395 2026-03-25 15:00:00 Blankenberge  51.3106    3.1164          433
#> 396 2026-03-25 16:00:00 Blankenberge  51.3106    3.1164          454
#> 397 2026-03-25 17:00:00 Blankenberge  51.3106    3.1164          457
#> 398 2026-03-25 18:00:00 Blankenberge  51.3106    3.1164          437
#> 399 2026-03-25 19:00:00 Blankenberge  51.3106    3.1164          387
#> 400 2026-03-25 20:00:00 Blankenberge  51.3106    3.1164          306
#> 401 2026-03-25 21:00:00 Blankenberge  51.3106    3.1164          236
#> 402 2026-03-25 22:00:00 Blankenberge  51.3106    3.1164          177
#> 403 2026-03-26 06:00:00 Blankenberge  51.3106    3.1164          429
#> 404 2026-03-26 07:00:00 Blankenberge  51.3106    3.1164          427
#> 405 2026-03-26 08:00:00 Blankenberge  51.3106    3.1164          343
#> 406 2026-03-26 09:00:00 Blankenberge  51.3106    3.1164          263
#> 407 2026-03-26 10:00:00 Blankenberge  51.3106    3.1164          198
#> 408 2026-03-26 11:00:00 Blankenberge  51.3106    3.1164          154
#> 409 2026-03-26 12:00:00 Blankenberge  51.3106    3.1164          136
#> 410 2026-03-26 13:00:00 Blankenberge  51.3106    3.1164          162
#> 411 2026-03-26 14:00:00 Blankenberge  51.3106    3.1164          214
#> 412 2026-03-26 15:00:00 Blankenberge  51.3106    3.1164          254
#> 413 2026-03-26 16:00:00 Blankenberge  51.3106    3.1164          332
#> 414 2026-03-26 17:00:00 Blankenberge  51.3106    3.1164          372
#> 415 2026-03-26 18:00:00 Blankenberge  51.3106    3.1164          382
#> 416 2026-03-26 19:00:00 Blankenberge  51.3106    3.1164          352
#> 417 2026-03-26 20:00:00 Blankenberge  51.3106    3.1164          314
#> 418 2026-03-26 21:00:00 Blankenberge  51.3106    3.1164          249
#> 419 2026-03-26 22:00:00 Blankenberge  51.3106    3.1164          184
#> 420 2026-03-26 23:00:00 Blankenberge  51.3106    3.1164          119
#> 421 2026-03-27 00:00:00 Blankenberge  51.3106    3.1164           71
#> 422 2026-03-27 01:00:00 Blankenberge  51.3106    3.1164           65
#> 423 2026-03-27 02:00:00 Blankenberge  51.3106    3.1164          109
#> 424 2026-03-27 03:00:00 Blankenberge  51.3106    3.1164          182
#> 425 2026-03-27 04:00:00 Blankenberge  51.3106    3.1164          275
#> 426 2026-03-27 05:00:00 Blankenberge  51.3106    3.1164          341
#> 427 2026-03-27 06:00:00 Blankenberge  51.3106    3.1164          368
#> 428 2026-03-27 07:00:00 Blankenberge  51.3106    3.1164          365
#> 429 2026-03-27 08:00:00 Blankenberge  51.3106    3.1164          344
#> 430 2026-03-27 09:00:00 Blankenberge  51.3106    3.1164          301
#> 431 2026-03-27 10:00:00 Blankenberge  51.3106    3.1164          243
#> 432 2026-03-27 11:00:00 Blankenberge  51.3106    3.1164          177
#> 433 2026-03-27 12:00:00 Blankenberge  51.3106    3.1164          125
#> 434 2026-03-27 13:00:00 Blankenberge  51.3106    3.1164           88
#> 435 2026-03-27 14:00:00 Blankenberge  51.3106    3.1164          108
#> 436 2026-03-27 15:00:00 Blankenberge  51.3106    3.1164          157
#> 437 2026-03-27 16:00:00 Blankenberge  51.3106    3.1164          221
#> 438 2026-03-27 17:00:00 Blankenberge  51.3106    3.1164          302
#> 439 2026-03-27 18:00:00 Blankenberge  51.3106    3.1164          359
#> 440 2026-03-27 19:00:00 Blankenberge  51.3106    3.1164          371
#> 441 2026-03-27 20:00:00 Blankenberge  51.3106    3.1164          372
#> 442 2026-03-27 21:00:00 Blankenberge  51.3106    3.1164          351
#> 443 2026-03-27 22:00:00 Blankenberge  51.3106    3.1164          311
#> 444 2026-03-27 23:00:00 Blankenberge  51.3106    3.1164          261
#> 445 2026-03-28 00:00:00 Blankenberge  51.3106    3.1164          204
#> 446 2026-03-28 01:00:00 Blankenberge  51.3106    3.1164          155
#> 447 2026-03-28 02:00:00 Blankenberge  51.3106    3.1164          121
#> 448 2026-03-28 03:00:00 Blankenberge  51.3106    3.1164          161
#> 449 2026-03-28 04:00:00 Blankenberge  51.3106    3.1164          218
#> 450 2026-03-28 05:00:00 Blankenberge  51.3106    3.1164          283
#> 451 2026-03-28 06:00:00 Blankenberge  51.3106    3.1164          362
#> 452 2026-03-28 07:00:00 Blankenberge  51.3106    3.1164          407
#> 453 2026-03-28 08:00:00 Blankenberge  51.3106    3.1164          420
#> 454 2026-03-28 09:00:00 Blankenberge  51.3106    3.1164          414
#> 455 2026-03-28 10:00:00 Blankenberge  51.3106    3.1164          389
#> 456 2026-03-28 11:00:00 Blankenberge  51.3106    3.1164          335
#> 457 2026-03-28 12:00:00 Blankenberge  51.3106    3.1164          271
#> 458 2026-03-28 13:00:00 Blankenberge  51.3106    3.1164          196
#> 459 2026-03-28 14:00:00 Blankenberge  51.3106    3.1164          141
#> 460 2026-03-28 15:00:00 Blankenberge  51.3106    3.1164          129
#> 461 2026-03-28 16:00:00 Blankenberge  51.3106    3.1164          174
#> 462 2026-03-28 17:00:00 Blankenberge  51.3106    3.1164          200
#> 463 2026-03-28 18:00:00 Blankenberge  51.3106    3.1164          259
#> 464 2026-03-28 19:00:00 Blankenberge  51.3106    3.1164          337
#> 465 2026-03-28 20:00:00 Blankenberge  51.3106    3.1164          391
#> 466 2026-03-28 21:00:00 Blankenberge  51.3106    3.1164          404
#> 467 2026-03-28 22:00:00 Blankenberge  51.3106    3.1164          387
#> 468 2026-03-28 23:00:00 Blankenberge  51.3106    3.1164          342
#> 469 2026-03-29 00:00:00 Blankenberge  51.3106    3.1164          297
#> 470 2026-03-29 01:00:00 Blankenberge  51.3106    3.1164          221
#> 471 2026-03-29 02:00:00 Blankenberge  51.3106    3.1164          143
#> 472 2026-03-29 03:00:00 Blankenberge  51.3106    3.1164           85
#> 473 2026-03-29 04:00:00 Blankenberge  51.3106    3.1164           88
#> 474 2026-03-29 05:00:00 Blankenberge  51.3106    3.1164          131
#> 475 2026-03-29 06:00:00 Blankenberge  51.3106    3.1164          179
#> 476 2026-03-29 07:00:00 Blankenberge  51.3106    3.1164          260
#> 477 2026-03-29 08:00:00 Blankenberge  51.3106    3.1164          359
#> 478 2026-03-29 09:00:00 Blankenberge  51.3106    3.1164          394
#> 479 2026-03-29 10:00:00 Blankenberge  51.3106    3.1164          394
#> 480 2026-03-29 11:00:00 Blankenberge  51.3106    3.1164          371
#> 481 2026-03-29 12:00:00 Blankenberge  51.3106    3.1164          319
#> 482 2026-03-29 13:00:00 Blankenberge  51.3106    3.1164          248
#> 483 2026-03-29 14:00:00 Blankenberge  51.3106    3.1164          169
#> 484 2026-03-29 15:00:00 Blankenberge  51.3106    3.1164           92
#> 485 2026-03-29 16:00:00 Blankenberge  51.3106    3.1164           36
#> 486 2026-03-29 17:00:00 Blankenberge  51.3106    3.1164           52
#> 487 2026-03-29 18:00:00 Blankenberge  51.3106    3.1164           95
#> 488 2026-03-29 19:00:00 Blankenberge  51.3106    3.1164          167
#> 489 2026-03-29 20:00:00 Blankenberge  51.3106    3.1164          283
#> 490 2026-03-29 21:00:00 Blankenberge  51.3106    3.1164          366
#> 491 2026-03-29 22:00:00 Blankenberge  51.3106    3.1164          374
#> 492 2026-03-29 23:00:00 Blankenberge  51.3106    3.1164          381
#> 493 2026-03-30 00:00:00 Blankenberge  51.3106    3.1164          355
#> 494 2026-03-30 01:00:00 Blankenberge  51.3106    3.1164          318
#> 495 2026-03-30 02:00:00 Blankenberge  51.3106    3.1164          255
#> 496 2026-03-30 03:00:00 Blankenberge  51.3106    3.1164          193
#> 497 2026-03-30 04:00:00 Blankenberge  51.3106    3.1164          155
#> 498 2026-03-30 05:00:00 Blankenberge  51.3106    3.1164          178
#> 499 2026-03-30 06:00:00 Blankenberge  51.3106    3.1164          211
#> 500 2026-03-30 07:00:00 Blankenberge  51.3106    3.1164          269
#> 501 2026-03-30 08:00:00 Blankenberge  51.3106    3.1164          361
#> 502 2026-03-30 09:00:00 Blankenberge  51.3106    3.1164          460
#> 503 2026-03-30 10:00:00 Blankenberge  51.3106    3.1164          492
#> 504 2026-03-30 11:00:00 Blankenberge  51.3106    3.1164          488
#> 505 2026-03-30 12:00:00 Blankenberge  51.3106    3.1164          463
#> 506 2026-03-30 13:00:00 Blankenberge  51.3106    3.1164          397
#> 507 2026-03-30 14:00:00 Blankenberge  51.3106    3.1164          318
#> 508 2026-03-30 15:00:00 Blankenberge  51.3106    3.1164          232
#> 509 2026-03-30 16:00:00 Blankenberge  51.3106    3.1164          141
#> 510 2026-03-30 17:00:00 Blankenberge  51.3106    3.1164          120
#> 511 2026-03-30 18:00:00 Blankenberge  51.3106    3.1164          160
#> 512 2026-03-30 19:00:00 Blankenberge  51.3106    3.1164          197
#> 513 2026-03-30 20:00:00 Blankenberge  51.3106    3.1164          245
#> 514 2026-03-30 21:00:00 Blankenberge  51.3106    3.1164          341
#> 515 2026-03-30 22:00:00 Blankenberge  51.3106    3.1164          430
#> 516 2026-03-30 23:00:00 Blankenberge  51.3106    3.1164          436
#> 517 2026-03-31 00:00:00 Blankenberge  51.3106    3.1164          416
#> 518 2026-03-31 01:00:00 Blankenberge  51.3106    3.1164          367
#> 519 2026-03-31 02:00:00 Blankenberge  51.3106    3.1164          294
#> 520 2026-03-31 03:00:00 Blankenberge  51.3106    3.1164          201
#> 521 2026-03-31 04:00:00 Blankenberge  51.3106    3.1164          113
#> 522 2026-03-31 05:00:00 Blankenberge  51.3106    3.1164           52
#> 523 2026-03-31 06:00:00 Blankenberge  51.3106    3.1164           78
#> 524 2026-03-31 07:00:00 Blankenberge  51.3106    3.1164          124
#> 525 2026-03-31 08:00:00 Blankenberge  51.3106    3.1164          183
#> 526 2026-03-31 09:00:00 Blankenberge  51.3106    3.1164          300
#> 527 2026-03-31 10:00:00 Blankenberge  51.3106    3.1164          422
#> 528 2026-03-31 11:00:00 Blankenberge  51.3106    3.1164          440
#> 529 2026-03-31 12:00:00 Blankenberge  51.3106    3.1164          431
#> 530 2026-03-31 13:00:00 Blankenberge  51.3106    3.1164          392
#> 531 2026-03-31 14:00:00 Blankenberge  51.3106    3.1164          319
#> 532 2026-03-31 15:00:00 Blankenberge  51.3106    3.1164          231
#> 533 2026-03-31 16:00:00 Blankenberge  51.3106    3.1164          136
#> 534 2026-03-31 17:00:00 Blankenberge  51.3106    3.1164           59
#> 535 2026-03-31 18:00:00 Blankenberge  51.3106    3.1164           64
#> 536 2026-03-31 19:00:00 Blankenberge  51.3106    3.1164          107
#> 537 2026-03-31 20:00:00 Blankenberge  51.3106    3.1164          156
#> 538 2026-03-31 21:00:00 Blankenberge  51.3106    3.1164          243
#> 539 2026-03-31 22:00:00 Blankenberge  51.3106    3.1164          400
#> 540 2026-03-31 23:00:00 Blankenberge  51.3106    3.1164          442
#> 541 2026-04-01 00:00:00 Blankenberge  51.3106    3.1164          439
#> 542 2026-04-01 01:00:00 Blankenberge  51.3106    3.1164          403
#> 543 2026-04-01 02:00:00 Blankenberge  51.3106    3.1164          344
#> 544 2026-04-01 03:00:00 Blankenberge  51.3106    3.1164          253
#> 545 2026-04-01 04:00:00 Blankenberge  51.3106    3.1164          157
#> 546 2026-04-01 05:00:00 Blankenberge  51.3106    3.1164           67
#> 547 2026-04-01 06:00:00 Blankenberge  51.3106    3.1164           12
#> 548 2026-04-01 07:00:00 Blankenberge  51.3106    3.1164           70
#> 549 2026-04-01 08:00:00 Blankenberge  51.3106    3.1164          118
#> 550 2026-04-01 09:00:00 Blankenberge  51.3106    3.1164          204
#> 551 2026-04-01 10:00:00 Blankenberge  51.3106    3.1164          372
#> 552 2026-04-01 11:00:00 Blankenberge  51.3106    3.1164          450
#> 553 2026-04-01 12:00:00 Blankenberge  51.3106    3.1164          455
#> 554 2026-04-01 13:00:00 Blankenberge  51.3106    3.1164          436
#> 555 2026-04-01 14:00:00 Blankenberge  51.3106    3.1164          387
#> 556 2026-04-01 15:00:00 Blankenberge  51.3106    3.1164          303
#> 557 2026-04-01 16:00:00 Blankenberge  51.3106    3.1164          207
#> 558 2026-04-01 17:00:00 Blankenberge  51.3106    3.1164          112
#> 559 2026-04-01 18:00:00 Blankenberge  51.3106    3.1164           48
#> 560 2026-04-01 19:00:00 Blankenberge  51.3106    3.1164           95
#> 561 2026-04-01 20:00:00 Blankenberge  51.3106    3.1164          141
#> 562 2026-04-01 21:00:00 Blankenberge  51.3106    3.1164          199
#> 563 2026-04-01 22:00:00 Blankenberge  51.3106    3.1164          316
#> 564 2026-04-01 23:00:00 Blankenberge  51.3106    3.1164          452
#> 565 2026-04-02 00:00:00 Blankenberge  51.3106    3.1164          461
#> 566 2026-04-02 01:00:00 Blankenberge  51.3106    3.1164          450
#> 567 2026-04-02 02:00:00 Blankenberge  51.3106    3.1164          403
#> 568 2026-04-02 03:00:00 Blankenberge  51.3106    3.1164          328
#> 569 2026-04-02 04:00:00 Blankenberge  51.3106    3.1164          231
#> 570 2026-04-02 05:00:00 Blankenberge  51.3106    3.1164          133
#> 571 2026-04-02 06:00:00 Blankenberge  51.3106    3.1164           55
#> 572 2026-04-02 07:00:00 Blankenberge  51.3106    3.1164           72
#> 573 2026-04-02 08:00:00 Blankenberge  51.3106    3.1164          124
#> 574 2026-04-02 09:00:00 Blankenberge  51.3106    3.1164          182
#> 575 2026-04-02 10:00:00 Blankenberge  51.3106    3.1164          283
#> 576 2026-04-02 11:00:00 Blankenberge  51.3106    3.1164          447
#> 577 2026-04-02 12:00:00 Blankenberge  51.3106    3.1164          481
#> 578 2026-04-02 13:00:00 Blankenberge  51.3106    3.1164          479
#> 579 2026-04-02 14:00:00 Blankenberge  51.3106    3.1164          443
#> 580 2026-04-02 15:00:00 Blankenberge  51.3106    3.1164          377
#> 581 2026-04-02 16:00:00 Blankenberge  51.3106    3.1164          279
#> 582 2026-04-02 17:00:00 Blankenberge  51.3106    3.1164          173
#> 583 2026-04-02 18:00:00 Blankenberge  51.3106    3.1164           84
#> 584 2026-04-02 19:00:00 Blankenberge  51.3106    3.1164           67
#> 585 2026-04-02 20:00:00 Blankenberge  51.3106    3.1164          114
#> 586 2026-04-02 21:00:00 Blankenberge  51.3106    3.1164          159
#> 587 2026-04-02 22:00:00 Blankenberge  51.3106    3.1164          227
#> 588 2026-04-02 23:00:00 Blankenberge  51.3106    3.1164          387
#> 589 2026-04-03 00:00:00 Blankenberge  51.3106    3.1164          455
#> 590 2026-04-03 01:00:00 Blankenberge  51.3106    3.1164          455
#> 591 2026-04-03 02:00:00 Blankenberge  51.3106    3.1164          417
#> 592 2026-04-03 03:00:00 Blankenberge  51.3106    3.1164          361
#> 593 2026-04-03 04:00:00 Blankenberge  51.3106    3.1164          267
#> 594 2026-04-03 05:00:00 Blankenberge  51.3106    3.1164          166
#> 595 2026-04-03 06:00:00 Blankenberge  51.3106    3.1164           72
#> 596 2026-04-03 07:00:00 Blankenberge  51.3106    3.1164            3
#> 597 2026-04-03 08:00:00 Blankenberge  51.3106    3.1164           33
#> 598 2026-04-03 09:00:00 Blankenberge  51.3106    3.1164           85
#> 599 2026-04-03 10:00:00 Blankenberge  51.3106    3.1164          164
#> 600 2026-04-03 11:00:00 Blankenberge  51.3106    3.1164          345
#> 601 2026-04-03 12:00:00 Blankenberge  51.3106    3.1164          429
#> 602 2026-04-03 13:00:00 Blankenberge  51.3106    3.1164          433
#> 603 2026-04-03 14:00:00 Blankenberge  51.3106    3.1164          414
#> 604 2026-04-03 15:00:00 Blankenberge  51.3106    3.1164          364
#> 605 2026-04-03 16:00:00 Blankenberge  51.3106    3.1164          289
#> 606 2026-04-03 17:00:00 Blankenberge  51.3106    3.1164          200
#> 607 2026-04-03 18:00:00 Blankenberge  51.3106    3.1164          117
#> 608 2026-04-03 19:00:00 Blankenberge  51.3106    3.1164           50
#> 609 2026-04-03 20:00:00 Blankenberge  51.3106    3.1164           96
#> 610 2026-04-03 21:00:00 Blankenberge  51.3106    3.1164          146
#> 611 2026-04-03 22:00:00 Blankenberge  51.3106    3.1164          218
#> 612 2026-04-03 23:00:00 Blankenberge  51.3106    3.1164          355
#> 613 2026-04-04 00:00:00 Blankenberge  51.3106    3.1164          477
#> 614 2026-04-04 01:00:00 Blankenberge  51.3106    3.1164          496
#> 615 2026-04-04 02:00:00 Blankenberge  51.3106    3.1164          485
#> 616 2026-04-04 03:00:00 Blankenberge  51.3106    3.1164          443
#> 617 2026-04-04 04:00:00 Blankenberge  51.3106    3.1164          368
#> 618 2026-04-04 05:00:00 Blankenberge  51.3106    3.1164          270
#> 619 2026-04-04 06:00:00 Blankenberge  51.3106    3.1164          172
#> 620 2026-04-04 07:00:00 Blankenberge  51.3106    3.1164           87
#> 621 2026-04-04 08:00:00 Blankenberge  51.3106    3.1164           85
#> 622 2026-04-04 09:00:00 Blankenberge  51.3106    3.1164          134
#> 623 2026-04-04 10:00:00 Blankenberge  51.3106    3.1164          190
#> 624 2026-04-04 11:00:00 Blankenberge  51.3106    3.1164          286
#> 625 2026-04-04 12:00:00 Blankenberge  51.3106    3.1164          446
#> 626 2026-04-04 13:00:00 Blankenberge  51.3106    3.1164          477
#> 627 2026-04-04 14:00:00 Blankenberge  51.3106    3.1164          478
#> 628 2026-04-04 15:00:00 Blankenberge  51.3106    3.1164          432
#> 629 2026-04-04 16:00:00 Blankenberge  51.3106    3.1164          364
#> 630 2026-04-04 17:00:00 Blankenberge  51.3106    3.1164          269
#> 631 2026-04-04 18:00:00 Blankenberge  51.3106    3.1164          169
#> 632 2026-04-04 19:00:00 Blankenberge  51.3106    3.1164           79
#> 633 2026-04-04 20:00:00 Blankenberge  51.3106    3.1164           40
#> 634 2026-04-04 21:00:00 Blankenberge  51.3106    3.1164           84
#> 635 2026-04-04 22:00:00 Blankenberge  51.3106    3.1164          136
#> 636 2026-04-04 23:00:00 Blankenberge  51.3106    3.1164          232
#> 637 2026-04-05 00:00:00 Blankenberge  51.3106    3.1164          393
#> 638 2026-04-05 01:00:00 Blankenberge  51.3106    3.1164          440
#> 639 2026-04-05 02:00:00 Blankenberge  51.3106    3.1164          438
#> 640 2026-04-05 03:00:00 Blankenberge  51.3106    3.1164          404
#> 641 2026-04-05 04:00:00 Blankenberge  51.3106    3.1164          348
#> 642 2026-04-05 05:00:00 Blankenberge  51.3106    3.1164          266
#> 643 2026-04-05 06:00:00 Blankenberge  51.3106    3.1164          186
#> 644 2026-04-05 07:00:00 Blankenberge  51.3106    3.1164          105
#> 645 2026-04-05 08:00:00 Blankenberge  51.3106    3.1164           42
#> 646 2026-04-05 09:00:00 Blankenberge  51.3106    3.1164           91
#> 647 2026-04-05 10:00:00 Blankenberge  51.3106    3.1164          142
#> 648 2026-04-05 11:00:00 Blankenberge  51.3106    3.1164          236
#> 649 2026-04-05 12:00:00 Blankenberge  51.3106    3.1164          385
#> 650 2026-04-05 13:00:00 Blankenberge  51.3106    3.1164          473
#> 651 2026-04-05 14:00:00 Blankenberge  51.3106    3.1164          482
#> 652 2026-04-05 15:00:00 Blankenberge  51.3106    3.1164          463
#> 653 2026-04-05 16:00:00 Blankenberge  51.3106    3.1164          419
#> 654 2026-04-05 17:00:00 Blankenberge  51.3106    3.1164          341
#> 655 2026-04-05 18:00:00 Blankenberge  51.3106    3.1164          249
#> 656 2026-04-05 19:00:00 Blankenberge  51.3106    3.1164          165
#> 657 2026-04-05 20:00:00 Blankenberge  51.3106    3.1164          103
#> 658 2026-04-05 21:00:00 Blankenberge  51.3106    3.1164          141
#> 659 2026-04-05 22:00:00 Blankenberge  51.3106    3.1164          180
#> 660 2026-04-05 23:00:00 Blankenberge  51.3106    3.1164          243
#> 661 2026-04-06 00:00:00 Blankenberge  51.3106    3.1164          348
#> 662 2026-04-06 01:00:00 Blankenberge  51.3106    3.1164          462
#> 663 2026-04-06 02:00:00 Blankenberge  51.3106    3.1164          472
#> 664 2026-04-06 03:00:00 Blankenberge  51.3106    3.1164          468
#> 665 2026-04-06 04:00:00 Blankenberge  51.3106    3.1164          421
#> 666 2026-04-06 05:00:00 Blankenberge  51.3106    3.1164          344
#> 667 2026-04-06 06:00:00 Blankenberge  51.3106    3.1164          245
#> 668 2026-04-06 07:00:00 Blankenberge  51.3106    3.1164          150
#> 669 2026-04-06 08:00:00 Blankenberge  51.3106    3.1164           76
#> 670 2026-04-06 09:00:00 Blankenberge  51.3106    3.1164           78
#> 671 2026-04-06 10:00:00 Blankenberge  51.3106    3.1164          119
#> 672 2026-04-06 11:00:00 Blankenberge  51.3106    3.1164          176
#> 673 2026-04-06 12:00:00 Blankenberge  51.3106    3.1164          259
#> 674 2026-04-06 13:00:00 Blankenberge  51.3106    3.1164          400
#> 675 2026-04-06 14:00:00 Blankenberge  51.3106    3.1164          436
#> 676 2026-04-06 15:00:00 Blankenberge  51.3106    3.1164          433
#> 677 2026-04-06 16:00:00 Blankenberge  51.3106    3.1164          402
#> 678 2026-04-06 17:00:00 Blankenberge  51.3106    3.1164          339
#> 679 2026-04-06 18:00:00 Blankenberge  51.3106    3.1164          247
#> 680 2026-04-06 19:00:00 Blankenberge  51.3106    3.1164          153
#> 681 2026-04-06 20:00:00 Blankenberge  51.3106    3.1164           76
#> 682 2026-04-06 21:00:00 Blankenberge  51.3106    3.1164           43
#> 683 2026-04-06 22:00:00 Blankenberge  51.3106    3.1164           82
#> 684 2026-04-06 23:00:00 Blankenberge  51.3106    3.1164          132
#> 685 2026-04-07 00:00:00 Blankenberge  51.3106    3.1164          214
#> 686 2026-04-07 01:00:00 Blankenberge  51.3106    3.1164          356
#> 687 2026-04-07 02:00:00 Blankenberge  51.3106    3.1164          411
#> 688 2026-04-07 03:00:00 Blankenberge  51.3106    3.1164          412
#> 689 2026-04-07 04:00:00 Blankenberge  51.3106    3.1164          390
#> 690 2026-04-07 05:00:00 Blankenberge  51.3106    3.1164          336
#> 691 2026-04-07 06:00:00 Blankenberge  51.3106    3.1164          256
#> 692 2026-04-07 07:00:00 Blankenberge  51.3106    3.1164          172
#> 693 2026-04-07 08:00:00 Blankenberge  51.3106    3.1164           92
#> 694 2026-04-07 09:00:00 Blankenberge  51.3106    3.1164           41
#> 695 2026-04-07 10:00:00 Blankenberge  51.3106    3.1164           86
#> 696 2026-04-07 11:00:00 Blankenberge  51.3106    3.1164          127
#> 697 2026-04-07 12:00:00 Blankenberge  51.3106    3.1164          198
#> 698 2026-04-07 13:00:00 Blankenberge  51.3106    3.1164          321
#> 699 2026-04-07 14:00:00 Blankenberge  51.3106    3.1164          403
#> 700 2026-04-07 15:00:00 Blankenberge  51.3106    3.1164          414
#> 701 2026-04-07 16:00:00 Blankenberge  51.3106    3.1164          401
#> 
#> $par
#> $par$daterange
#> [1] "2026-03-08" "2026-04-08"
#> 
#> $par$type
#> [1] "MVB data"
#> 
#> $par$stations
#> [1] "Blankenberge"
#> 
#> $par$calc
#> [1] "max"
#> 
#> $par$binSize
#> [1] "hour"
#> 
#> $par$parameters
#> [1] "Tide (cmTAW)"
#> 
#>