Retrieve UVA-bird-data from the LifeWatch project

Retrieves the University of Amsterdam (UvA) Bird Tracking System data from the LifeWatch project. Need valid authentication to access the entire data. To get an account, register via the Lifewatch RShiny registration webpage. Without login you can only access data collected at least two years ago.

getUvaBirdData(
  startdate,
  stopdate,
  tagcodes,
  by,
  usr = NULL,
  pwd = NULL,
  params = FALSE,
  ...
)

Arguments

startdate

Starting date for the query

stopdate

Stopping date for the query

tagcodes

List of tag numbers from $tagnr in listUvaTags. Can also use "All HG" "All LBB" or "All MH" (MH only when logged in).

by

Time grouping for aggregation, one of ("1 day","60 min","10 min","1 min","1 week")

usr

Username to connect to database

pwd

Password to connect to database

params

If TRUE, returns a list with the dataset and the query parameters applied in the server side. IF FALSE returns only the data.

...

Params to be passed to lw_check_lwdataserver().

Value

Dataframe with the specified bird tracking data.

Examples

getUvaBirdData("2013-06-10", "2013-06-14", tagcodes = c("719","6013","610"), by = "1 day")
#> - Query mode: Post request to OpenCPU server
#> No encoding supplied: defaulting to UTF-8.
#> - Query parameters: 
#>    - daterange: c("2013-06-10", "2013-06-14")
#>    - type: UVA bird data
#>    - tags: c("719", "6013", "610")
#>    - binSize: 1 day
#>             code                time species longitude latitude tracklength
#> 1 LBB:Harry[719] 2013-06-13 00:01:40     LBB    3.2018  51.2750       34830
#> 2 LBB:Harry[719] 2013-06-12 00:22:26     LBB    3.2003  51.2887       35887
#> 3 LBB:Harry[719] 2013-06-11 00:25:45     LBB    3.2250  51.0115      156938
#> 4 LBB:Harry[719] 2013-06-10 00:13:41     LBB    3.1783  51.1361      205916
#>   maxdist tracktime counts logcounts
#> 1   12476   23.8306    315    5.7526
#> 2   12232   23.0139    320    5.7683
#> 3   67524   23.3367    704    6.5568
#> 4   67587   23.5953    706    6.5596
getUvaBirdData("2013-06-10", "2013-06-14", tagcodes = "All HG", by = "1 week", params = TRUE)
#> - Query mode: Post request to OpenCPU server
#> No encoding supplied: defaulting to UTF-8.
#> - Query parameters: 
#>    - daterange: c("2013-06-10", "2013-06-14")
#>    - type: UVA bird data
#>    - tags: All HG
#>    - binSize: 1 week
#> $mdf
#>             code                time species longitude latitude tracklength
#> 1  HG:Wilma[783] 2013-06-10 00:28:10      HG    2.9759  51.2335      491575
#> 2   HG:Anne[786] 2013-06-10 00:01:10      HG    2.8989  51.2160      469280
#> 3    HG:Luc[799] 2013-06-10 00:59:13      HG    2.9355  51.1385      334647
#> 4 HG:Jurgen[801] 2013-06-10 00:27:42      HG    2.9073  51.2545      770960
#>   maxdist tracktime counts logcounts
#> 1   15461  167.4911   2560    7.8478
#> 2   13157  167.4853   3217    8.0762
#> 3   17729  166.0297   1158    7.0544
#> 4   17785  167.0494   2837    7.9505
#> 
#> $par
#> $par$daterange
#> [1] "2013-06-10" "2013-06-14"
#> 
#> $par$type
#> [1] "UVA bird data"
#> 
#> $par$tags
#> [1] "All HG"
#> 
#> $par$binSize
#> [1] "1 week"
#> 
#>