Retrieve an overview of the ETN data.

Retrieves the name of network/animals? To get an account, register via the Lifewatch RShiny registration webpage.

listEtnProjects(usr = NULL, pwd = NULL, ...)

Arguments

usr

Username to connect to ETN database

pwd

Password to connect to ETN database

...

Reserved for internal use.

Value

Dataframe with name and type of networks.

Examples

listEtnProjects()
#> - Query mode: Post request to OpenCPU server
#> No encoding supplied: defaulting to UTF-8.
#>                                                                                 name
#> 1                                                                  2011_Bovenschelde
#> 2                                                                         2011_Loire
#> 3                                                                         2013_Foyle
#> 4                                                                          2013_Maas
#> 5                                                                          2014_Nene
#> 6                                                           2015_PhD_Gutmann_Roberts
#> 7                                                               2016_Diaccia_Botrona
#> 8                                                                        2017_Fremur
#> 9                                                                    2020_PhD_Winter
#> 10                                                                     Aberdeen 2017
#> 11                                       Acousic receiver array for the MOVE project
#> 12                     Acoustic array in the Ria de Vigo (Galicia, Spain) since 2024
#> 13   Acoustic data between 2026 and 2028 for rheophilic fish in weir-regulated Meuse
#> 14                                      Acoustic Ecology of Pelagic Fish Communities
#> 15                             Acoustic Monitoring of Bluefin Tuna in the Aegean Sea
#> 16              Acoustic receiver array in the Gudena River, Denmark, from 2004-2006
#> 17                          Acoustic receiver array in the River Hevring (2023-2028)
#> 18              Acoustic receiver array in the Zuidelijke Randmeren, the Netherlands
#> 19    Acoustic telemetry array active since 2019 to track species in Bristol Channel
#> 20    Acoustic telemetry array data for elasmobranch in Northern Ireland (2024-2030)
#> 21   Acoustic telemetry array in the river Oude IJssel (Germany and the Netherlands)
#> 22                      Acoustic telemetry array on the east cost of Corsica, France
#> 23    Acoustic telemetry data for Atlantic cod at an oil and gas platform in Denmark
#> 24         Acoustic telemetry data for blue crab in the S'Ena Arrubia lagoon (Italy)
#> 25          Acoustic telemetry data for coastal fishes in Marintique Island (France)
#> 26    Acoustic telemetry data for European eel (Anguilla anguilla) in the River Test
#> 27       Acoustic telemetry data for Lionfish distribution in Gökova Bay, Aegean Sea
#> 28    Acoustic telemetry data for Sakarya Telemetry Array in the Black Sea (Türkiye)
#> 29    Acoustic telemetry data from the Clew Bay and Achill region in west of Ireland
#> 30  Acoustic telemetry data of marine species monitored 6 French offshore wind farms
#> 31                     Acoustic telemetry data of the Super Smolt project, the Rhine
#> 32   Acoustic telemetry data on Allis shad migrating in the Oise River (Seine basin)
#> 33                                                   Acoustic telemetry in lake Osby
#> 34           Acoustic telemetry network in the Ecowende wind farm in the Netherlands
#> 35              Acoustic telemetry network in the harbours of IJmuiden and Rotterdam
#> 36     Acoustic telemetry network in the non-tidal River Scheldt and its tributaries
#> 37   Acoustic telemetry network in the North Sea and the Eastern Scheldt (2025-2028)
#> 38   Acoustic telemetry network in the western River Rhine and hinterlands 2025–2026
#> 39      Acoustic telemetry on Atlantic cod in the Horns Rev 2 offshore windmill farm
#> 40   Acoustic telemetry receiver array in Galway Bay and Dunkellin River (2015-2019)
#> 41                         Acoustic telemetry receiver array in Murcia Region, Spain
#> 42            Acoustic telemetry receiver array in the harinxma channel in Harlingen
#> 43               Acoustic telemetry receiver array in the Nederrijn-Lek, Netherlands
#> 44                 Acoustic telemetry receiver array in the North Sea from 2017-2018
#> 45                  Acoustic telemetry receiver array in the Orbetello lagoon, Italy
#> 46    Acoustic telemetry receiver array in the River Warnow (Germany) from 2011-2012
#> 47         Acoustic telemetry receiver array in the Waal river between 2025 and 2027
#> 48               Acoustic telemetry study in the Meuse river basin at Lith and Linne
#> 49                                                                      Albertkanaal
#> 50                                                              Albertkanaal_VPS_Ham
#> 51                                                          Albertkanaal_VPS_Hasselt
#> 52           An acoustic receiver network operating at 69 kHz in and around Blekinge
#> 53                                                               Apelafico_acoustics
#> 54                                                                        ARAISOLA02
#> 55                                                                        ARAISOLA03
#> 56                                                                               Ars
#> 57                                                                          Artevigo
#> 58             Artisanal fisheries & wetland management: eel and mullet in SE Iberia
#> 59                                                 Azorean acoustic receiver network
#> 60         Balearic strategy for climate change monitoring and prediction in the Med
#> 61                                                         Balearic Tracking Network
#> 62                   Balearic Tracking Network - Spanish Deep Water Receiver Network
#> 63                                                       Baltic Sea Tracking Network
#> 64                                                            Basking_shark_Hebrides
#> 65                                                                            BECORV
#> 66    Behaviour of Cancer pagurus in proximity to a fish farm in Storfjorden, Norway
#> 67                           Belwind temporary acoustic receiver network 2020 - 2021
#> 68                                           Bergen's Ocean Acoustic Tracking System
#> 69                              BlueConnect: innovative monitoring at Vlaamse banken
#> 70  Blue crab telemetry monitoring in three lagoons, channels and coast of Occitanie
#> 71                                                                       Boddenhecht
#> 72                                                                              BPNS
#> 73                                                           Brasem Ijsselmeergebied
#> 74                                      British and Irish Tracking array (permanent)
#> 75                                               Brown trout in the Falkland Islands
#> 76                                                                    BTN_LOST_ARRAY
#> 77                                                      Canaries Angel Shark Project
#> 78                                                          Catalan Tracking Network
#> 79                                                   CMRC European Sea Bass Tracking
#> 80                                                            CODEVCO fish detectors
#> 81                          Connectivity of coastal fish in the French Mediterranean
#> 82                                                                        Conon 2017
#> 83                                                                       cpodnetwork
#> 84                                                                     DAbecasis PhD
#> 85                                                                     DAK 2018-2019
#> 86                                                          Dalalven Smolt Migration
#> 87                                                        Danish_southeastern_strait
#> 88                                            Danish Straits Acoustic Infrastructure
#> 89                                                                  Danube_Sturgeons
#> 90                                                                Delfzijl 2014-2015
#> 91                                                                             Demer
#> 92    Deployment of acoustic receivers in Carlingford Lough for the NorTrack project
#> 93                             Deployment of acoustic receivers in inner Lough Foyle
#> 94                             Deployment of acoustic receivers in outer Lough Foyle
#> 95             Deployment of acoustic receivers in rivers within the Foyle Catchment
#> 96      Deployment of an acoustic array in the North Channel for the STRAITS project
#> 97                                                                      Deveron 2016
#> 98                                                                             Dijle
#> 99                                                                         Dijle_VPS
#> 100                                                                       DTU-Skjern
#> 101   Dusky grouper´s resettlement - Acoustic telemetry in Maro-Cerro Gordo´s cliffs
#> 102   Duurzame economische dragers in Vlaamse kustvisserijgemeenschappen post-Brexit
#> 103                                                 Eel migration from source to sea
#> 104                                                                     Eemskanaal I
#> 105                                                                    Eemskanaal II
#> 106                                                                   Eemskanaal III
#> 107                                                                             EMMN
#> 108                                                                             ESGL
#> 109                                                      ETN_network_project_group_1
#> 110                                                      ETN_network_project_group_2
#> 111                                                      ETN_network_project_group_3
#> 112                                                      ETN_network_project_group_4
#> 113                                                      ETN_network_project_group_5
#> 114                                                      ETN_network_project_group_6
#> 115                                                      ETN_network_project_group_7
#> 116                                                      ETN_network_project_group_8
#> 117                                                      ETN_network_project_group_9
#> 118                                                          Euskal Tracking Network
#> 119   Evaluation of the effectiveness of the Balearic Islands Marine Reserve Network
#> 120                                                                FarmTrack_Spain22
#> 121     Fine scale acoustic telemetry array at Hinkley Point C water intakes in 2025
#> 122     Fine scale acoustic telemetry array at Hinkley Point C water intakes in 2026
#> 123  Fine-scale multispecies acoustic telemetry positioning system data at Helgoland
#> 124                                           Fisheries Industry Science Partnership
#> 125               Fisheries Innovation for sustainable SHared INTerchannEL resources
#> 126                                          Fish migration from source to Waddensea
#> 127                                                             FISHOWF_Array_France
#> 128                                                              Friesland 2011-2012
#> 129                                    Galician (NW Spain) acoustic Tracking Network
#> 130                                                                   Gudena_network
#> 131                              Haringvlietdam - Rhine-Meuse Delta Acoustic Network
#> 132    High-frequency acoustic telemetry array in the northwestern Mediterranean sea
#> 133  HR2 tech and 3D tracking of European eel in landlocked drinking water reservoir
#> 134 ICOD: array to monitor interaction between cod spawning grounds and salmon farms
#> 135                                                                  IG II Fish Pass
#> 136                                                                        Iller_VPS
#> 137                                        Immature bass acoustic stock surveillance
#> 138                                                                          IMR_OWF
#> 139                                                                    Inforbiomares
#> 140                                                       Infragravity waves project
#> 141                                                                  INTERREG_CANAPE
#> 142         Intra-lagoon ecological connectivity: migration surveys of European eels
#> 143                                                     Jersey Coastal Fish Tracking
#> 144                                                              JSATS-PalmaBay-2019
#> 145                                                   Kalmar Baltic Tracking Network
#> 146                                                           Kattegat Islands Array
#> 147                                                      Kiel Bight Acoustic Network
#> 148                                                                   Kornwerderzand
#> 149                                                                Lake_Anundsjo_VPS
#> 150                       Land-use change and human activity on anadromous salmonids
#> 151                                                                      Lauwersmeer
#> 152                                                                    Leopoldkanaal
#> 153        Le Tréport harbour receiver array in the Bresle estuary, northwest France
#> 154                                                                        Lifewatch
#> 155                                                   Lifewatch: additional stations
#> 156                                             Lifewatch: Porpoise Performance test
#> 157                                                                         Limfjord
#> 158                                             Ljusne Bay Salmon Spawning Migration
#> 159 Long-term array MOnitoring BEhaviour in Illas Atlánticas National Park, NW Spain
#> 160                                                                       LSTSTJ_MPA
#> 161                                                                          MacFish
#> 162       Mapping critical habitats of threatened shark species in the Algarve coast
#> 163                                                          Marble vs rainbow trout
#> 164    MARE-WIND finescale setup acoustic deployments in Belgian Offshore Wind Farms
#> 165                                                            MariagerFjord_network
#> 166        Marine Ecological connectivity and conseRvation in the MOZambique Channel
#> 167                                                                       MBA_Massmo
#> 168                                                                      MBA_Wavehub
#> 169                                                                     MBA_Whitsand
#> 170                                                                           MEPNSW
#> 171                                                          Middle_Danube_sturgeons
#> 172   Migration of pikeperch within the Lake Peipus complex (Estonia) from 2021-2023
#> 173                                                                      MIGRATOEBRE
#> 174                                                            MMERMAID_Array_France
#> 175                                                                      Mobula Imar
#> 176                             Monitoring Gibraltar Strait Using Acoustic Telemetry
#> 177                                                                 Moray Firth 2016
#> 178      Movement and habitat connectivity of Atlantic cod in a coastal fjord system
#> 179     Movement, Range Size and Habitat Utilization of magafish in Turkey 2019-2021
#> 180   Movements of cod, plaice, flounder and turbot in the western German Baltic Sea
#> 181   MPAs' effect on Balearic chondrichthyes' biodiversity using acoustic telemetry
#> 182                                                                         mrc_vliz
#> 183                                                                         NARVAEEL
#> 184                                                          Northern_Norfolk_Broads
#> 185                                                                 North_sea_wrecks
#> 186                                                                    NTNU-Gaulosen
#> 187                                                       NTNU River Nidelva Project
#> 188   Observing the effect of group dynamics on individual behavior in basking shark
#> 189                                                Open Protocols scientific testing
#> 190                                                                       Orsted Cod
#> 191                                                                  OTN-Hemnfjorden
#> 192                                                          OTN Loan - Beiarfjorden
#> 193                                                             OTN-Skjerstadfjorden
#> 194                                                     OTN Strait of Gibraltar Line
#> 195                                                                 OTN-Tosenfjorden
#> 196                                                           OTN VR2W Loan - Norway
#> 197                                                                    paintedcomber
#> 198                                         PAM Harbour porpoise Borssele, 2019-2024
#> 199                                        Passive Acoustic Monitoring (PAM) STRAITS
#> 200                                                                             PC4C
#> 201             Permanent acoustic receiver array in the River Grote Nete since 2019
#> 202                                                                PhD Barbara Koeck
#> 203                                                                PhD Jeremy Pastor
#> 204                                                                        PhD_Nolan
#> 205                                                                         PhysFish
#> 206                       Pilot permanent deployment acoustic array Gibraltar Strait
#> 207                                   Porbeagle Tracking Network in Brittany, France
#> 208                          Predators and prey around renewable energy developments
#> 209                                                                        PTN-MARSW
#> 210                                                                  PTN_PROTECT2012
#> 211                                                                  PTN_PROTECT2013
#> 212                                                           PTN-Silver eel-Mondego
#> 213                                                          Purewind fish detectors
#> 214                                                                     RECCRU_array
#> 215              Receiver network in Killary Harbour and River Erriff from 2014-2024
#> 216  Redes de Sensores submarinos autonomos para la medida de indicadores biologicos
#> 217                                                                         Reelease
#> 218                 Research on the impact of management actions on beaver behaviour
#> 219                                                                           ResMed
#> 220   Restoration of Baltic sturgeon stock in Oder river & in Mecklenburg-Vorpommern
#> 221                                                River_lamprey_River Trent_2020_21
#> 222                                             River_Lamprey_Yorkshire Ouse_2018-20
#> 223                                             River Usk receiver network 2021-2024
#> 224                                                                rt2020_zeeschelde
#> 225                                                             Ruim_baan_voor_vis_2
#> 226                                                                      Running_eel
#> 227                                                                       Saeftinghe
#> 228                                               Salmon_migration_eastern_Greenland
#> 229                                                                          SAMARCH
#> 230                                                                       SARTELARAM
#> 231                                                                         SARTELTG
#> 232                                                                    SARTELZINGARO
#> 233                                                                           SCHUFI
#> 234                                                                         Scytrack
#> 235                                                                       SeaMonitor
#> 236           SeaMonitor - Costs and Pathways of Marine Migration in Atlantic salmon
#> 237                                                        SeaMonitor - GMIT Project
#> 238                                                        sea trout_2018-9_gullmarn
#> 239 SEAWave: small cetacean presence at the EMEC Orkney Billia Croo wave energy site
#> 240           SIARC network of scientific moorings in Pen Llŷn a’r Sarnau SAC, Wales
#> 241                                                                 Siganid_East_Med
#> 242                                                               Siganid_Gulf_Aqaba
#> 243              Silver Eel downstream migration in the Walloon Meuse - 2021 to 2024
#> 244       Silver European Eel escapement success and migration patterns in River Ems
#> 245                                                         Skagerrak-NorthSea-array
#> 246                                                 Skagerrak_temeletry_array_Norway
#> 247                                                                        Skye 2017
#> 248                                                             SMOLTRACK-I-Engeland
#> 249                                                            SMOLTRACK-II-Engeland
#> 250         Spatial Ecology of spurdog (Squalus acanthias) in a Scottish loch system
#> 251 Spawning migration of pike in MatsaluBay and RiverKasari (Baltic Sea, 2018-2020)
#> 252  Spawning migration of pike in Teorehe-Sauemeri floodplain (BalticSea 2017-2019)
#> 253                                                                        SPAWNSEIS
#> 254                                                                 SPIDER_GNB_array
#> 255                                                                S_rivulatus_Aqaba
#> 256                                                S_rivulatus_Eastern_Mediterranean
#> 257                                                                      STR_MOV_ANG
#> 258                                                     Sturgeon_reintroduction_Ebre
#> 259                                                  Sud_le_Impuls_voor_Diadrome_Vis
#> 260                                                    Sud_le_Innovatieve_monitoring
#> 261                                          survival_movement_elasmobranch_Adriatic
#> 262                                                                  SVNL-FISH-WATCH
#> 263                                                            Swedish eel migration
#> 264                                                                     SWIMWAY_2021
#> 265                                                           Swimway Rijnland Array
#> 266                                                                   Swimway Vechte
#> 267                                                 Swiss Acoustic Telemetry Network
#> 268   Tagging and tracking butterfly ray with acoustic telemetry in the Balearic Sea
#> 269                                  Temporary acoustic network in the Sines harbour
#> 270    Temporary acoustic network in the SW Alentejo and Vicentine Coast Marine Park
#> 271                                  Temporary acoustic receiver network in Bråviken
#> 272                     Temporary receiver array in Lithuanian rivers from 2014-2015
#> 273                                                                             TeRi
#> 274                                                                       Test VR2AR
#> 275                                                     The Bergen Telemetry Network
#> 276         The National Programme for Monitoring the Aquatic Environment and Nature
#> 277                                                                     Thorntonbank
#> 278                                                                  Trent¨_Tracking
#> 279                                                   Turkish Straits Detection Data
#> 280                         Turkish Telemetry Array in the eastern Mediterranean sea
#> 281       Understanding the movement of dolphinfish in the Mediterranean FAD fishery
#> 282                      Understand movement and life history of trout: Isle of Skye
#> 283    Underwater acoustic data from the Belgian ARMS-MBON hard-substrate monitoring
#> 284                                                             Unlocking the Severn
#> 285                  Unravelling the migrations of the invasive blue crab in Algarve
#> 286                                                                 VFM_BalticProper
#> 287                                                                 VFM_GulfOfBotnia
#> 288                                                                    VFM_Hjalmaren
#> 289                                                                      VFM_Malaren
#> 290                                                                    VFM_RiverGota
#> 291                                                                       VFM_Siljan
#> 292                                                                      VFM_Vattern
#> 293                                                              VFM_WestCoastCostal
#> 294                                                           Vissen_voor_verbinding
#> 295                                       West Coast Tracking Project Receiver Array
#> 296                                                                  Westerschelde 1
#> 297                                                                  Westerschelde 2
#> 298                                                                  Westerschelde 3
#> 299                                                       Winde Tjeukemeer 2023-2026
#> 300                                      Yellow eel movement behaviour in the Azores
#> 301                                                                       Zeeschelde
#> 302                                                                 2010_phd_reubens
#> 303                                                            2010_phd_reubens_sync
#> 304                                                                       2011_Loire
#> 305                                                                  2011 Rivierprik
#> 306                                                               2012 Leopoldkanaal
#> 307                                                                2013 Albertkanaal
#> 308                                                                       2014 Demer
#> 309                                                                       2014_Frome
#> 310                                                                        2014_Nene
#> 311                                                        2015_Albertkanaal_VPS_Ham
#> 312                                                                       2015 Dijle
#> 313                                                                     2015_Homarus
#> 314                                                            2015 PhD Verhelst_cod
#> 315                                                            2015 PhD Verhelst_eel
#> 316                                                        2016_Albertkanaal_VPS_Ham
#> 317                                                               2016 PhD Vergeynst
#> 318                                                        2017_Albertkanaal_VPS_Ham
#> 319                                                                   2019_Grotenete
#> 320                                                                   2021_Dijle_VPS
#> 321                                                         2021_Sturgeons_Danube_BG
#> 322                                                    2022_Albertkanaal_VPS_Hasselt
#> 323   Acoustic telemetry data for Atlantic cod at an oil and gas platform in Denmark
#> 324        Acoustic telemetry data for blue crab in the S'Ena Arrubia lagoon (Italy)
#> 325         Acoustic telemetry data for coastal fishes in Marintique Island (France)
#> 326    Acoustic telemetry data for Dicentrarchus labrax in the Gulf of Lion (France)
#> 327 Acoustic telemetry data for elasmobranch species in Northern Ireland (2024-2030)
#> 328    Acoustic telemetry data for garfish (Belone belone), Western Baltic (Denmark)
#> 329  Acoustic telemetry data for porbeagle sharks (Lamna nasus) tagged across Europe
#> 330          Acoustic telemetry data for Spicara smaris on the east coast of Corsica
#> 331 Acoustic telemetry data of Atlantic salmon and sea trout in Galway and Dunkellin
#> 332              Acoustic telemetry data of Carp in lake Markermeer, the Netherlands
#> 333    Acoustic telemetry for Atlantic mackerel in the German Bight of the North Sea
#> 334 Acoustic Telemetry on blue sharks (Prionace glauca) in the Basque country, Spain
#> 335                                                                       ADST-Shark
#> 336                                                                           AFICHE
#> 337                                                                       ARAISOLA02
#> 338                                                                       ARAISOLA03
#> 339                                                                              Ars
#> 340                                                                         Artevigo
#> 341        Balearic strategy for climate change monitoring and prediction in the Med
#> 342                                                        Balearic Tracking Network
#> 343                  Balearic Tracking Network - Spanish Deep Water Receiver Network
#> 344                                                                         CodNoise
#> 345      Comparative study of marble and rainbow trout behaviour in Kanomljica River
#> 346 Conventional and electronic telemetry data for starry smooth-hound in NEAtlantic
#> 347                                                                    DAbecasis PhD
#> 348                                                                    DAK 2018-2019
#> 349                                                                 Danube_Sturgeons
#> 350                                                         Delfzijl Drie Delfzijlen
#> 351   Dusky grouper´s resettlement - Acoustic telemetry in Maro-Cerro Gordo´s cliffs
#> 352                                                                             EMMN
#> 353                                                                             ESGL
#> 354                                                       ETN_animal_project_group_1
#> 355                                                       ETN_animal_project_group_2
#> 356                                                       ETN_animal_project_group_3
#> 357                                                       ETN_animal_project_group_4
#> 358                                                       ETN_animal_project_group_5
#> 359                                                       ETN_animal_project_group_6
#> 360                                                       ETN_animal_project_group_7
#> 361                                                       ETN_animal_project_group_8
#> 362                                                       ETN_animal_project_group_9
#> 363   Evaluation of the effectiveness of the Balearic Islands Marine Reserve Network
#> 364  High-frequency data for rocky reef fishes in the northwestern Mediterranean sea
#> 365                                                                        Hunze_eel
#> 366                                                              JSATS-PalmaBay-2019
#> 367                                                                         Limfjord
#> 368                                                          Marble vs rainbow trout
#> 369                                                                       MBA_Massmo
#> 370                                                                      MBA_Wavehub
#> 371                                                                     MBA_Whitsand
#> 372                                                          Middle_Danube_sturgeons
#> 373          Migration of common bream in the Zuidelijke Randmeren, the Netherlands.
#> 374                                                                      Mobula Imar
#> 375      Movement and habitat connectivity of Atlantic cod in a coastal fjord system
#> 376   MPAs' effect on Balearic chondrichthyes' biodiversity using acoustic telemetry
#> 377  Multispecies acoustic telemetry data from fine-scale positioning system MarGate
#> 378                                                         Noordzeekanaal 2017-2018
#> 379                                                   North Aegean Anguilla Tracking
#> 380                                                                    NTNU-Gaulosen
#> 381                                                       NTNU River Nidelva Project
#> 382                                                Open Protocols scientific testing
#> 383                                                                       Orsted Cod
#> 384                                                                  OTN-Hemnfjorden
#> 385                                                                 OTN-Tosenfjorden
#> 386                                                           OTN VR2W Loan - Norway
#> 387                                                                PhD Barbara Koeck
#> 388                                                                     PhD_Goossens
#> 389                                                                PhD Jeremy Pastor
#> 390                                                                  PhD_JolienBuyse
#> 391                                                                         PhysFish
#> 392                                                           PTN-Silver eel-Mondego
#> 393                                                                         Reelease
#> 394                                                                   Rijke_Noordzee
#> 395                                                                       SARTELARAM
#> 396                                                                         SARTELTG
#> 397                                                                    SARTELZINGARO
#> 398                                                                         Scytrack
#> 399                                                        sea trout_2018-9_gullmarn
#> 400                                                                 Siganid_East_Med
#> 401                                                               Siganid_Gulf_Aqaba
#> 402                                Silver eel migration in the Gudena River, Denmark
#> 403                                     Silver eel migration in the Warnow catchment
#> 404          Silver eel migration patterns in lowland rivers and lagoon in Lithuania
#> 405       Silver European Eel escapement success and migration patterns in River Ems
#> 406         Spatial Ecology of spurdog (Squalus acanthias) in a Scottish loch system
#> 407                                                S_rivulatus_Eastern_Mediterranean
#> 408                                                     Sturgeon_reintroduction_Ebre
#> 409   Tagging and tracking butterfly ray with acoustic telemetry in the Balearic Sea
#> 410                                                                             TeRi
#> 411                      Tracking anadromous brown trout at sea by data storage tags
#> 412  Tracking movements of 15 Dusky groupers released in the harbor of Sines in 2021
#> 413                          Tracking twaite shads at sea by using data storage tags
#> 414                 Trap and transport of silver eels from a reservoir to an estuary
#> 415       Understanding the movement of dolphinfish in the Mediterranean FAD fishery
#> 416                                                                    VPS Duurswold
#> 417                                                                  VPS Garmerwolde
#> 418                                      Yellow eel movement behaviour in the Azores
#>        type
#> 1   network
#> 2   network
#> 3   network
#> 4   network
#> 5   network
#> 6   network
#> 7   network
#> 8   network
#> 9   network
#> 10  network
#> 11  network
#> 12  network
#> 13  network
#> 14  network
#> 15  network
#> 16  network
#> 17  network
#> 18  network
#> 19  network
#> 20  network
#> 21  network
#> 22  network
#> 23  network
#> 24  network
#> 25  network
#> 26  network
#> 27  network
#> 28  network
#> 29  network
#> 30  network
#> 31  network
#> 32  network
#> 33  network
#> 34  network
#> 35  network
#> 36  network
#> 37  network
#> 38  network
#> 39  network
#> 40  network
#> 41  network
#> 42  network
#> 43  network
#> 44  network
#> 45  network
#> 46  network
#> 47  network
#> 48  network
#> 49  network
#> 50  network
#> 51  network
#> 52  network
#> 53  network
#> 54  network
#> 55  network
#> 56  network
#> 57  network
#> 58  network
#> 59  network
#> 60  network
#> 61  network
#> 62  network
#> 63  network
#> 64  network
#> 65  network
#> 66  network
#> 67  network
#> 68  network
#> 69  network
#> 70  network
#> 71  network
#> 72  network
#> 73  network
#> 74  network
#> 75  network
#> 76  network
#> 77  network
#> 78  network
#> 79  network
#> 80  network
#> 81  network
#> 82  network
#> 83  network
#> 84  network
#> 85  network
#> 86  network
#> 87  network
#> 88  network
#> 89  network
#> 90  network
#> 91  network
#> 92  network
#> 93  network
#> 94  network
#> 95  network
#> 96  network
#> 97  network
#> 98  network
#> 99  network
#> 100 network
#> 101 network
#> 102 network
#> 103 network
#> 104 network
#> 105 network
#> 106 network
#> 107 network
#> 108 network
#> 109 network
#> 110 network
#> 111 network
#> 112 network
#> 113 network
#> 114 network
#> 115 network
#> 116 network
#> 117 network
#> 118 network
#> 119 network
#> 120 network
#> 121 network
#> 122 network
#> 123 network
#> 124 network
#> 125 network
#> 126 network
#> 127 network
#> 128 network
#> 129 network
#> 130 network
#> 131 network
#> 132 network
#> 133 network
#> 134 network
#> 135 network
#> 136 network
#> 137 network
#> 138 network
#> 139 network
#> 140 network
#> 141 network
#> 142 network
#> 143 network
#> 144 network
#> 145 network
#> 146 network
#> 147 network
#> 148 network
#> 149 network
#> 150 network
#> 151 network
#> 152 network
#> 153 network
#> 154 network
#> 155 network
#> 156 network
#> 157 network
#> 158 network
#> 159 network
#> 160 network
#> 161 network
#> 162 network
#> 163 network
#> 164 network
#> 165 network
#> 166 network
#> 167 network
#> 168 network
#> 169 network
#> 170 network
#> 171 network
#> 172 network
#> 173 network
#> 174 network
#> 175 network
#> 176 network
#> 177 network
#> 178 network
#> 179 network
#> 180 network
#> 181 network
#> 182 network
#> 183 network
#> 184 network
#> 185 network
#> 186 network
#> 187 network
#> 188 network
#> 189 network
#> 190 network
#> 191 network
#> 192 network
#> 193 network
#> 194 network
#> 195 network
#> 196 network
#> 197 network
#> 198 network
#> 199 network
#> 200 network
#> 201 network
#> 202 network
#> 203 network
#> 204 network
#> 205 network
#> 206 network
#> 207 network
#> 208 network
#> 209 network
#> 210 network
#> 211 network
#> 212 network
#> 213 network
#> 214 network
#> 215 network
#> 216 network
#> 217 network
#> 218 network
#> 219 network
#> 220 network
#> 221 network
#> 222 network
#> 223 network
#> 224 network
#> 225 network
#> 226 network
#> 227 network
#> 228 network
#> 229 network
#> 230 network
#> 231 network
#> 232 network
#> 233 network
#> 234 network
#> 235 network
#> 236 network
#> 237 network
#> 238 network
#> 239 network
#> 240 network
#> 241 network
#> 242 network
#> 243 network
#> 244 network
#> 245 network
#> 246 network
#> 247 network
#> 248 network
#> 249 network
#> 250 network
#> 251 network
#> 252 network
#> 253 network
#> 254 network
#> 255 network
#> 256 network
#> 257 network
#> 258 network
#> 259 network
#> 260 network
#> 261 network
#> 262 network
#> 263 network
#> 264 network
#> 265 network
#> 266 network
#> 267 network
#> 268 network
#> 269 network
#> 270 network
#> 271 network
#> 272 network
#> 273 network
#> 274 network
#> 275 network
#> 276 network
#> 277 network
#> 278 network
#> 279 network
#> 280 network
#> 281 network
#> 282 network
#> 283 network
#> 284 network
#> 285 network
#> 286 network
#> 287 network
#> 288 network
#> 289 network
#> 290 network
#> 291 network
#> 292 network
#> 293 network
#> 294 network
#> 295 network
#> 296 network
#> 297 network
#> 298 network
#> 299 network
#> 300 network
#> 301 network
#> 302  animal
#> 303  animal
#> 304  animal
#> 305  animal
#> 306  animal
#> 307  animal
#> 308  animal
#> 309  animal
#> 310  animal
#> 311  animal
#> 312  animal
#> 313  animal
#> 314  animal
#> 315  animal
#> 316  animal
#> 317  animal
#> 318  animal
#> 319  animal
#> 320  animal
#> 321  animal
#> 322  animal
#> 323  animal
#> 324  animal
#> 325  animal
#> 326  animal
#> 327  animal
#> 328  animal
#> 329  animal
#> 330  animal
#> 331  animal
#> 332  animal
#> 333  animal
#> 334  animal
#> 335  animal
#> 336  animal
#> 337  animal
#> 338  animal
#> 339  animal
#> 340  animal
#> 341  animal
#> 342  animal
#> 343  animal
#> 344  animal
#> 345  animal
#> 346  animal
#> 347  animal
#> 348  animal
#> 349  animal
#> 350  animal
#> 351  animal
#> 352  animal
#> 353  animal
#> 354  animal
#> 355  animal
#> 356  animal
#> 357  animal
#> 358  animal
#> 359  animal
#> 360  animal
#> 361  animal
#> 362  animal
#> 363  animal
#> 364  animal
#> 365  animal
#> 366  animal
#> 367  animal
#> 368  animal
#> 369  animal
#> 370  animal
#> 371  animal
#> 372  animal
#> 373  animal
#> 374  animal
#> 375  animal
#> 376  animal
#> 377  animal
#> 378  animal
#> 379  animal
#> 380  animal
#> 381  animal
#> 382  animal
#> 383  animal
#> 384  animal
#> 385  animal
#> 386  animal
#> 387  animal
#> 388  animal
#> 389  animal
#> 390  animal
#> 391  animal
#> 392  animal
#> 393  animal
#> 394  animal
#> 395  animal
#> 396  animal
#> 397  animal
#> 398  animal
#> 399  animal
#> 400  animal
#> 401  animal
#> 402  animal
#> 403  animal
#> 404  animal
#> 405  animal
#> 406  animal
#> 407  animal
#> 408  animal
#> 409  animal
#> 410  animal
#> 411  animal
#> 412  animal
#> 413  animal
#> 414  animal
#> 415  animal
#> 416  animal
#> 417  animal
#> 418  animal