Function to retrieve data from the LifeWatch project

Large wrapper function that retrieves the LifeWatch data. Use datatype specific functions for clearer use.

lwdata(datatype = "Buoy data", from = as.character(Sys.Date() - 90),
  to = as.character(Sys.Date()), stations = c("Buoy at C-Power"),
  binSize = "60 min", calc = "Time bins", UrlPar = NULL,
  code = NULL, posres = 2, logged = FALSE, projectlist = NULL,
  tagprojectlist = NULL, loggedInUserPostgresUsername = NULL,
  loggedInUserPostgresPwd = NULL, phylasp = "#2#51#",
  taxranks = c(Species = 220), qualities = c("Hi"),
  processing = "Validated", minband = 10, maxband = 10,
  params = FALSE, force_opencpu = FALSE)

Arguments

datatype

Type of data to request

from

Starting date for the query

to

Stopping date for the query

stations

(Buoy, MVB, Station) list of stations to be included in the query

binSize

(Bats, ETN, CPOD, MVB, UvaBird, Underway) Sample period.

calc

(MVB) Calculation to perform given time grouping, one of ('avg', 'max', 'min', 'none')

UrlPar

.

code

.

posres

.

logged

(CPOD, MVB, ETN, UvaBird) Request data under moratory if you have an account

projectlist

.

tagprojectlist

.

loggedInUserPostgresUsername

(CPOD, MVB, ETN, UvaBird) Postgres username

loggedInUserPostgresPwd

(CPOD, MVB, ETN, UvaBird) Postgres password

phylasp

.

taxranks

.

qualities

.

processing

(CPOD) One of ('Validated','Raw')

minband

lower frequency band from which the sound value will be retrieved

maxband

upper frequency band from which the sound value will be retrieved

params

If TRUE, returns a list with the dataset and the query parameters applied in the server side. IF FALSE returns only the data.

force_opencpu

boolean. If true, the query will be forced to run through the OpenCPU server.

Value

The sum of x and y

Examples

if (FALSE) {
lwdata()
lwdata('zooscan data')
lwdata('flowcam data')
lwdata('MVB')
lwdata('station data')
lwdata('cpod data')
lwdata('etn data')
}