For use in the LifeWatch Rshiny data-explorer (LWDE). Reads input and USER arguments

For use in the LifeWatch Rshiny data-explorer (LWDE). Reads input and USER arguments

lwdata2(type, input = NULL, USER = NULL)

Arguments

type
Datatype as defined in LWDE

input
LWDE input element, as a list. Can be obtained by using shiny::reactiveValuesToList()

USER
LWDE USER element, as a list. Can be obtained by using shiny::reactiveValuesToList()

Value

Returns dataframe of the requested datatype

Examples

lwdata2("listETNprojects")

#> No encoding supplied: defaulting to UTF-8.

#>                                                                                name
#> 1                                                                       2004_Gudena
#> 2                                                                        2011_Loire
#> 3                                                                       2011_Warnow
#> 4                                                                        2013_Foyle
#> 5                                                                         2014_Nene
#> 6                                                          2015_PhD_Gutmann_Roberts
#> 7                                                              2016_Diaccia_Botrona
#> 8                                                                       2017_Fremur
#> 9                                                                    2019_Grotenete
#> 10                                                                  2020_PhD_Winter
#> 11                                                                    Aberdeen 2017
#> 12                                     Acoustic Ecology of Pelagic Fish Communities
#> 13                                                                     Albertkanaal
#> 14                                                             Albertkanaal_VPS_Ham
#> 15                                                                        Apelafico
#> 16                                                                       ARAISOLA02
#> 17                                                                       ARAISOLA03
#> 18                                                                              Ars
#> 19                                                                         Artevigo
#> 20                                                Azorean acoustic receiver network
#> 21                                                                           BECORV
#> 22                          Belwind temporary acoustic receiver network 2020 - 2021
#> 23                                                                     Bovenschelde
#> 24                                                                             BPNS
#> 25                                                          Brasem Ijsselmeergebied
#> 26                         Connectivity of coastal fish in the French Mediterranean
#> 27                                                                       Conon 2017
#> 28                                                                      cpodnetwork
#> 29                                                                    DAbecasis PhD
#> 30                                                                    DAK 2018-2019
#> 31                                                         Dalalven Smolt Migration
#> 32                                                                 Danube_Sturgeons
#> 33                                                               Delfzijl 2014-2015
#> 34                                                                            Demer
#> 35                                                                     Deveron 2016
#> 36                                                                            Dijle
#> 37                                                                       DTU-Skjern
#> 38                                                 Eel migration from source to sea
#> 39                                                                     Eemskanaal I
#> 40                                                                    Eemskanaal II
#> 41                                                                   Eemskanaal III
#> 42                                                                             EMMN
#> 43                                                                             ESGL
#> 44                                          Fish migration from source to Waddensea
#> 45                                                              Friesland 2011-2012
#> 46                                                                  IG II Fish Pass
#> 47                                                                            Iller
#> 48                                                                    Inforbiomares
#> 49                                                                  INTERREG_CANAPE
#> 50                                                     Jersey Coastal Fish Tracking
#> 51                                                              JSATS-PalmaBay-2019
#> 52                                                   Kalmar Baltic Tracking Network
#> 53                                                                   Kornwerderzand
#> 54                                                                      Lauwersmeer
#> 55                                                                    Leopoldkanaal
#> 56                                                                        Lifewatch
#> 57                                                                         Limfjord
#> 58                                             Ljusne Bay Salmon Spawning Migration
#> 59                                                                             Maas
#> 60                                                                          MacFish
#> 61                                                          Marble vs rainbow trout
#> 62                                                                       MBA_Massmo
#> 63                                                                      MBA_Wavehub
#> 64                                                                     MBA_Whitsand
#> 65                                                                           MEPNSW
#> 66                                                          Middle_Danube_sturgeons
#> 67                                                                      MIGRATOEBRE
#> 68                                                                           MOBEIA
#> 69                                                                      Mobula Imar
#> 70                                                                 Moray Firth 2016
#> 71                                                                         NARVAEEL
#> 72                                                         Noordzeekanaal 2017-2018
#> 73                                                          Northern_Norfolk_Broads
#> 74                                                                 North_sea_wrecks
#> 75                                                                    NTNU-Gaulosen
#> 76                                                                       Orsted Cod
#> 77                                                                  OTN-Hemnfjorden
#> 78                                                                 OTN-Tosenfjorden
#> 79                                                                    paintedcomber
#> 80                                                                             PC4C
#> 81                                                                PhD Barbara Koeck
#> 82                                                                PhD Jeremy Pastor
#> 83                                                                        PhD_Nolan
#> 84                                                                         PhysFish
#> 85                                                                  PTN/PROTECT2012
#> 86                                                                  PTN/PROTECT2013
#> 87                                                           PTN-Silver eel-Mondego
#> 88  Redes de Sensores submarinos autonomos para la medida de indicadores biologicos
#> 89                                                                         Reelease
#> 90                                                                           ResMed
#> 91                                             River Usk receiver network 2021-2024
#> 92                                                             Ruim_baan_voor_vis_2
#> 93                                                                      Running_eel
#> 94                                                                       Saeftinghe
#> 95                                                                          SAMARCH
#> 96                                                                       SARTELARAM
#> 97                                                                         SARTELTG
#> 98                                                                    SARTELZINGARO
#> 99                                                                         Scytrack
#> 100                                                                      SeaMonitor
#> 101                                                                  Semp Lithuania
#> 102                                                                Siganid_East_Med
#> 103                                                              Siganid_Gulf_Aqaba
#> 104                                                                       Skye 2017
#> 105                                                            SMOLTRACK-I-Engeland
#> 106                                                           SMOLTRACK-II-Engeland
#> 107                                                                       SPAWNSEIS
#> 108                                                 Sud_le_Impuls_voor_Diadrome_Vis
#> 109                                                   Sud_le_Innovatieve_monitoring
#> 110                                                                 SVNL-FISH-WATCH
#> 111                                                                  Swimway Vechte
#> 112                                                                            TeRi
#> 113                                                                    Thorntonbank
#> 114                                                            Unlocking the Severn
#> 115                                                          Vissen_voor_verbinding
#> 116                                                                 Westerschelde 1
#> 117                                                                 Westerschelde 2
#> 118                                                                 Westerschelde 3
#> 119                                                                      Zeeschelde
#> 120                                                                2010_phd_reubens
#> 121                                                           2010_phd_reubens_sync
#> 122                                                                 2011 Rivierprik
#> 123                                                              2012 Leopoldkanaal
#> 124                                                               2013 Albertkanaal
#> 125                                                                      2014 Demer
#> 126                                                       2015_Albertkanaal_VPS_Ham
#> 127                                                                      2015 Dijle
#> 128                                                                    2015_Homarus
#> 129                                                           2015 PhD Verhelst_cod
#> 130                                                           2015 PhD Verhelst_eel
#> 131                                                       2016_Albertkanaal_VPS_Ham
#> 132                                                              2016 PhD Vergeynst
#> 133                                                       2017_Albertkanaal_VPS_Ham
#> 134                                                                      ARAISOLA02
#> 135                                                                      ARAISOLA03
#> 136                                                                             Ars
#> 137                                                                        Artevigo
#> 138     Comparative study of marble and rainbow trout behaviour in Kanomljica River
#> 139                                                                   DAbecasis PhD
#> 140                                                                Danube_Sturgeons
#> 141                                                                            EMMN
#> 142                                                                            ESGL
#> 143                                                             JSATS-PalmaBay-2019
#> 144                                                                        Limfjord
#> 145                                                         Marble vs rainbow trout
#> 146                                                                      MBA_Massmo
#> 147                                                                     MBA_Wavehub
#> 148                                                                    MBA_Whitsand
#> 149                                                         Middle_Danube_sturgeons
#> 150                                                                     Mobula Imar
#> 151                                                  North Aegean Anguilla Tracking
#> 152                                                                   NTNU-Gaulosen
#> 153                                                          Ocean Tracking Network
#> 154                                                                      Orsted Cod
#> 155                                                                 OTN-Hemnfjorden
#> 156                                                                OTN-Tosenfjorden
#> 157                                                               PhD Barbara Koeck
#> 158                                                               PhD Jeremy Pastor
#> 159                                                                        PhysFish
#> 160                                                                      SARTELARAM
#> 161                                                                        SARTELTG
#> 162                                                                   SARTELZINGARO
#> 163                                                                        Scytrack
#> 164                                                                  Semp Lithuania
#> 165                                                              Siganid_Gulf_Aqaba
#> 166                                                                            TeRi
#>        type
#> 1   network
#> 2   network
#> 3   network
#> 4   network
#> 5   network
#> 6   network
#> 7   network
#> 8   network
#> 9   network
#> 10  network
#> 11  network
#> 12  network
#> 13  network
#> 14  network
#> 15  network
#> 16  network
#> 17  network
#> 18  network
#> 19  network
#> 20  network
#> 21  network
#> 22  network
#> 23  network
#> 24  network
#> 25  network
#> 26  network
#> 27  network
#> 28  network
#> 29  network
#> 30  network
#> 31  network
#> 32  network
#> 33  network
#> 34  network
#> 35  network
#> 36  network
#> 37  network
#> 38  network
#> 39  network
#> 40  network
#> 41  network
#> 42  network
#> 43  network
#> 44  network
#> 45  network
#> 46  network
#> 47  network
#> 48  network
#> 49  network
#> 50  network
#> 51  network
#> 52  network
#> 53  network
#> 54  network
#> 55  network
#> 56  network
#> 57  network
#> 58  network
#> 59  network
#> 60  network
#> 61  network
#> 62  network
#> 63  network
#> 64  network
#> 65  network
#> 66  network
#> 67  network
#> 68  network
#> 69  network
#> 70  network
#> 71  network
#> 72  network
#> 73  network
#> 74  network
#> 75  network
#> 76  network
#> 77  network
#> 78  network
#> 79  network
#> 80  network
#> 81  network
#> 82  network
#> 83  network
#> 84  network
#> 85  network
#> 86  network
#> 87  network
#> 88  network
#> 89  network
#> 90  network
#> 91  network
#> 92  network
#> 93  network
#> 94  network
#> 95  network
#> 96  network
#> 97  network
#> 98  network
#> 99  network
#> 100 network
#> 101 network
#> 102 network
#> 103 network
#> 104 network
#> 105 network
#> 106 network
#> 107 network
#> 108 network
#> 109 network
#> 110 network
#> 111 network
#> 112 network
#> 113 network
#> 114 network
#> 115 network
#> 116 network
#> 117 network
#> 118 network
#> 119 network
#> 120  animal
#> 121  animal
#> 122  animal
#> 123  animal
#> 124  animal
#> 125  animal
#> 126  animal
#> 127  animal
#> 128  animal
#> 129  animal
#> 130  animal
#> 131  animal
#> 132  animal
#> 133  animal
#> 134  animal
#> 135  animal
#> 136  animal
#> 137  animal
#> 138  animal
#> 139  animal
#> 140  animal
#> 141  animal
#> 142  animal
#> 143  animal
#> 144  animal
#> 145  animal
#> 146  animal
#> 147  animal
#> 148  animal
#> 149  animal
#> 150  animal
#> 151  animal
#> 152  animal
#> 153  animal
#> 154  animal
#> 155  animal
#> 156  animal
#> 157  animal
#> 158  animal
#> 159  animal
#> 160  animal
#> 161  animal
#> 162  animal
#> 163  animal
#> 164  animal
#> 165  animal
#> 166  animal

input=c()
input$daterange = c(Sys.Date()-365*2, Sys.Date())
lwdata2("ZooScan data", input)

#> No encoding supplied: defaulting to UTF-8.

#> $mdf
#> [1] "No data"
#> 
#> $par
#> $par$daterange
#> [1] "2019-08-12" "2021-08-11"
#> 
#> $par$type
#> [1] "ZooScan data"
#> 
#> 

lwdata2("flow", input)

#> No encoding supplied: defaulting to UTF-8.

#> $mdf
#>    Tripaction                Time Station Longitude Latitude
#> 1      111335 2019-09-26 02:27:35     215    2.6120  51.2768
#> 2      111495 2019-10-22 07:58:30     230    2.8518  51.3071
#> 3      111337 2019-09-26 03:15:47    ZG02    2.4969  51.3340
#> 4      111521 2019-10-22 13:11:42     700    3.2199  51.3766
#> 5      111357 2019-09-26 06:50:41     710    3.1380  51.4407
#> 6      111082 2019-08-20 06:28:26     120    2.7001  51.1844
#> 7      111323 2019-09-25 23:34:17     130    2.9072  51.2729
#> 8      111568 2019-10-23 14:36:12     120    2.6984  51.1839
#> 9      111328 2019-09-26 01:12:47     120    2.7007  51.1852
#> 10     111327 2019-09-25 23:58:47     230    2.8500  51.3087
#> 11     111148 2019-08-21 08:25:04     710    3.1403  51.4404
#> 12     111114 2019-08-20 11:39:22     230    2.8519  51.3068
#> 13     111105 2019-08-20 09:24:16    ZG02    2.4994  51.3329
#> 14     111347 2019-09-26 06:13:38     780        NA       NA
#> 15     111090 2019-08-20 07:58:04     215    2.6083  51.2758
#> 16     111169 2019-08-21 10:26:28     330    2.8073  51.4323
#> 17     111155 2019-08-21 09:14:22     780    3.0571  51.4707
#> 18     111126 2019-08-21 05:00:04     130    2.9041  51.2713
#> 19     111366 2019-09-26 07:35:35     700    3.2218  51.3780
#> 20     111534 2019-10-22 14:44:12     710    3.1394  51.4408
#> 21     111342 2019-09-26 04:58:35     330    2.8091  51.4332
#>    Fluid Volume Imaged (ml) Sample Volume (ml) Volume Filtered (l)
#> 1                    7.7331                172                  50
#> 2                    6.9636                225                  50
#> 3                    7.6095                214                  50
#> 4                    6.9991                210                  50
#> 5                    5.1393                285                  50
#> 6                    6.2953                200                  50
#> 7                    7.5130                 80                  50
#> 8                    6.6888                210                  50
#> 9                    5.9443                222                  50
#> 10                   7.7913                210                  50
#> 11                   5.0255                245                  50
#> 12                   7.1956                205                  50
#> 13                   6.3662                198                  50
#> 14                   7.7675                300                  50
#> 15                   7.4233                235                  50
#> 16                   6.5970                320                  50
#> 17                   6.5961                295                  50
#> 18                   7.4226                200                  50
#> 19                   7.6172                255                  50
#> 20                   7.4667                330                  50
#> 21                   7.0058                208                  50
#>    Sample Dilution Factor Count Total Count Density (cell/l)
#> 1                    0.30  2115        2115        106007.61
#> 2                    1.00  2897        2897         36979.60
#> 3                    0.30  1998        1998         81796.43
#> 4                    0.30  1706        1706         77379.56
#> 5                    0.10  1455        1455        198675.43
#> 6                    0.15  1524        1524        161390.24
#> 7                    0.20  2863        2863        476341.01
#> 8                    1.00  3134        3134         44623.28
#> 9                    1.00  3074        3074         46588.65
#> 10                   0.20  1757        1757        107384.73
#> 11                   0.20  1009        1009         81949.40
#> 12                   0.15  1661        1661        150136.92
#> 13                   0.10  3619        3619        574213.12
#> 14                   0.20   460         460         19740.37
#> 15                   0.10  2190        2190        251078.33
#> 16                   0.30  2889        2889         91234.65
#> 17                   0.15  2169        2169        148624.03
#> 18                   0.15  2203        2203        197864.18
#> 19                   0.10  2294        2294        236204.34
#> 20                   0.30  1362        1362         36850.48
#> 21                   0.20  1756        1756        120504.55
#>             ActionType   Project
#> 1  Planktonnet Apstein Lifewatch
#> 2  Planktonnet Apstein Lifewatch
#> 3  Planktonnet Apstein Lifewatch
#> 4  Planktonnet Apstein Lifewatch
#> 5  Planktonnet Apstein Lifewatch
#> 6  Planktonnet Apstein Lifewatch
#> 7  Planktonnet Apstein Lifewatch
#> 8  Planktonnet Apstein Lifewatch
#> 9  Planktonnet Apstein Lifewatch
#> 10 Planktonnet Apstein Lifewatch
#> 11 Planktonnet Apstein Lifewatch
#> 12 Planktonnet Apstein Lifewatch
#> 13 Planktonnet Apstein Lifewatch
#> 14 Planktonnet Apstein Lifewatch
#> 15 Planktonnet Apstein Lifewatch
#> 16 Planktonnet Apstein Lifewatch
#> 17 Planktonnet Apstein Lifewatch
#> 18 Planktonnet Apstein Lifewatch
#> 19 Planktonnet Apstein Lifewatch
#> 20 Planktonnet Apstein Lifewatch
#> 21 Planktonnet Apstein Lifewatch
#> 
#> $par
#> $par$daterange
#> [1] "2019-08-12" "2021-08-11"
#> 
#> $par$type
#> [1] "flow"
#> 
#>