Retrieve CTD data from the LifeWatch project

Retrieve CTD data from the LifeWatch project

getCTDData(startdate, stopdate, stations = "All", by, params = FALSE, ...)

Arguments

startdate

Starting date for the query in UTC. Must be a character date of the form "YYYY-MM-DD".

stopdate

Stopping date for the query in UTC. Must be a character date of the form "YYYY-MM-DD".

stations

List of stations to be included in the query. Use stations = "All" to get all stations. See stations list in details section.

by

Aggregation criteria. One of c("Per meter", "Per cast")

params

If TRUE, returns a list with the dataset and the query parameters applied in the server side. IF FALSE returns only the data.

...

Reserved for internal use.

Value

Dataframe with the CTD data within the specified daterange and location.

Details

List of stations: c("120", "130", "230", "215", "330", "421", "435", "710", "700", "780", "ZG02", "W09", "W08", "LW01", "W10", "LW02", "W07bis")

Examples

getCTDData("2023-01-15", "2023-02-15", stations = c("120", "W09"), by = "Per meter")
#> - Query mode: Post request to OpenCPU server
#> No encoding supplied: defaulting to UTF-8.
#> Warning: No data returned
getCTDData("2023-01-15", "2023-02-15", stations = "All", by = "Per cast", params = TRUE)
#> - Query mode: Post request to OpenCPU server
#> No encoding supplied: defaulting to UTF-8.
#> - Query parameters: 
#>    - type: ctd
#>    - bintype: Per cast
#>    - daterange: c("2023-01-15", "2023-02-15")
#>    - stationlist: c("120", "130", "230", "215", "330", "421", "435", "710", "700", "780", "ZG02", "W09", "W08", "LW01", "W10", "LW02", "W07bis")
#> $mdf
#>    Station castId Latitude Longitude                Time  Depth   obs3+    altM
#> 1      130   6196  51.2708    2.9068 2023-01-19 08:34:23 4.7611 14.6267  2.3650
#> 2      130   6252  51.2712    2.9052 2023-02-14 08:33:29 5.0910 19.9545 13.9735
#> 3      130   6253  51.2727    2.9052 2023-02-14 08:16:45 3.9039 20.7619 13.1296
#> 4      230   6192  51.3082    2.8497 2023-01-18 13:18:12 3.5980  3.2552  2.6275
#> 5      230   6197  51.3092    2.8565 2023-01-19 09:26:43 4.7582 10.2913  2.1567
#> 6      230   6254  51.3077    2.8495 2023-02-14 09:06:13 6.4903 20.4068  9.0758
#> 7      230   6255  51.3102    2.8477 2023-02-14 09:17:09 3.4394 21.6769  6.2602
#> 8      330   6193  51.4335    2.8080 2023-01-18 12:18:42 7.6749  1.5191  2.6722
#> 9      330   6198  51.4333    2.8093 2023-01-19 14:16:04 9.2345  1.6487  2.4778
#> 10     330   6256  51.4337    2.8108 2023-02-14 14:04:27 2.6596 22.1543  2.1606
#> 11     330   6257  51.4335    2.8085 2023-02-14 13:55:45 9.0525 24.1070  1.2774
#> 12     700   6199  51.3755    3.2208 2023-01-19 10:58:04 3.9539 41.5902  1.7888
#> 13     700   6258  51.3765    3.2195 2023-02-14 10:57:51 5.2882 22.9542 43.9897
#> 14     700   6259  51.3778    3.2163 2023-02-14 11:17:57 3.0080 21.5800 35.3752
#> 15     710   6200  51.4420    3.1397 2023-01-19 12:17:15 4.1490  7.9523  2.3970
#> 16     710   6260  51.4408    3.1373 2023-02-14 12:11:53 3.3953 21.1194  6.2488
#> 17     710   6261  51.4417    3.1393 2023-02-14 12:03:51 4.1134 22.9171  7.3001
#> 18     780   6201  51.4705    3.0588 2023-01-19 12:55:58 6.2757  3.4518  2.3660
#> 19     780   6262  51.4687    3.0498 2023-02-14 12:57:45 2.7708 22.3813  3.4645
#> 20     780   6263  51.4708    3.0547 2023-02-14 12:46:29 8.8920 23.2441  6.8361
#> 21  W07bis   6195  51.5868    3.0117 2023-01-18 10:39:29 8.0705  1.4171  8.4079
#>         par flECO-AFL oxigenvoltage sbeox0Mg/L sbeox0PS   prdM Temperature
#> 1    4.7619    1.2472        2.7853     8.3389  87.4570 4.8029      7.9547
#> 2   79.1521    1.6465        0.1753    -1.2989 -12.8322 5.1357      6.0584
#> 3   82.8675    1.6406        0.1334    -0.9902  -9.7954 3.9381      6.0666
#> 4  145.9285    0.7387        2.8456     8.4033  89.6550 3.6296      8.5304
#> 5    9.2331    0.9855        2.7830     8.2889  87.3987 4.7999      8.0553
#> 6   86.3170    1.7009        0.1116    -1.5225 -15.1501 6.5474      6.2616
#> 7   26.6101    1.5663        0.0653    -1.5712 -15.6094 3.4696      6.2222
#> 8  195.5979    0.7616        2.9221     8.4850  92.4587 7.7425      9.3490
#> 9   70.9728    1.0181        2.9255     8.5484  92.6640 9.3159      9.1196
#> 10  43.6063    1.9758        0.0090    -1.6129 -16.4495 2.6830      6.9433
#> 11 136.2572    3.5722        0.0123    -1.8411 -18.8157 9.1323      6.9538
#> 12   8.8902    1.7804        2.8090     8.5828  88.4432 3.9887      7.5922
#> 13  98.1906    2.4285        0.5711     0.1978   1.9729 5.3347      6.4699
#> 14  33.8922    2.3394        0.4060    -0.2621  -2.6267 3.0345      6.7261
#> 15  50.2606    0.9460        2.7262     8.0660  85.8742 4.1855      8.4354
#> 16  35.0141    4.7073        0.0689    -1.6772 -16.7893 3.4252      6.5149
#> 17  90.1166    1.7750        0.0930    -1.4858 -14.8691 4.1496      6.4925
#> 18 157.1863    0.9643        2.8941     8.5369  91.5507 6.3310      8.6866
#> 19  52.7276    3.2070        0.0296    -1.8089 -18.2467 2.7952      6.7563
#> 20 103.3015    2.1109        0.0844    -1.6002 -16.1163 8.9704      6.6488
#> 21  38.1960    0.9147        2.9161     8.4712  92.4474 8.1417      9.3999
#>     C0uS/cm Salinity soundspeed  density descSpeed        flag
#> 1  34754.53  33.4787   1480.497 1026.108    0.0141 -8.2872e-29
#> 2  30937.39  31.0982   1470.138 1024.488   -0.0019 -5.7086e-29
#> 3  31042.64  31.2052   1470.285 1024.566   -0.0009 -9.9900e-29
#> 4  35888.63  34.1314   1483.450 1026.528    0.0006 -8.1509e-29
#> 5  35295.89  33.9609   1481.475 1026.472    0.0081 -7.7720e-29
#> 6  31429.80  31.4595   1471.426 1024.755   -0.0017 -5.5816e-29
#> 7  31107.41  31.1414   1470.820 1024.495   -0.0038 -7.2602e-29
#> 8  37088.35  34.5945   1487.107 1026.779    0.0045 -5.2615e-29
#> 9  36851.20  34.5692   1486.262 1026.804    0.0012 -6.6879e-29
#> 10 32858.65  32.4197   1475.257 1025.408   -0.0042 -4.9950e-29
#> 11 33450.62  33.0293   1476.163 1025.915   -0.0011 -4.5906e-29
#> 12 33102.78  32.0496   1477.325 1025.033    0.0075 -8.8850e-29
#> 13 30811.27  30.5850   1471.136 1024.034   -0.0015 -6.3843e-29
#> 14 31264.11  30.8498   1472.441 1024.200   -0.0097 -6.9602e-29
#> 15 35801.26  34.1306   1483.103 1026.544    0.0068 -7.5057e-29
#> 16 31618.86  31.4418   1472.359 1024.695   -0.0021 -7.5480e-29
#> 17 31473.75  31.2973   1472.100 1024.587   -0.0031 -7.6796e-29
#> 18 36275.08  34.3868   1484.390 1026.716    0.0082 -6.8943e-29
#> 19 32096.15  31.7509   1473.689 1024.906   -0.0047 -6.3646e-29
#> 20 32140.55  31.8802   1473.526 1025.049   -0.0014 -4.3945e-29
#> 21 37194.74  34.6553   1487.374 1026.820    0.0138 -5.8006e-29
#> 
#> $par
#> $par$type
#> [1] "ctd"
#> 
#> $par$bintype
#> [1] "Per cast"
#> 
#> $par$daterange
#> [1] "2023-01-15" "2023-02-15"
#> 
#> $par$stationlist
#>  [1] "120"    "130"    "230"    "215"    "330"    "421"    "435"    "710"   
#>  [9] "700"    "780"    "ZG02"   "W09"    "W08"    "LW01"   "W10"    "LW02"  
#> [17] "W07bis"
#> 
#>