Retrieve MVB-data from the LifeWatch project

Retrieves the Meetnet Vlaams Banken (MVB) data from the LifeWatch project. Without valid credentials you are only allowed to view the "Tide TAW" parameter for the last 30 days, grouped by day/hour. To get an account, register via the Lifewatch RShiny registration webpage.

getMvbData(
  startdate,
  stopdate,
  parameters,
  stations = NULL,
  by,
  calc,
  usr = NULL,
  pwd = NULL,
  params = FALSE,
  ...
)

Arguments

startdate

Starting date for the query. Without a login, this is restricted to the last month

stopdate

Stopping date for the query. Without a login, this is restricted to the last month

parameters

One (or list) of the $parameter in listMvbStations.

stations

List of stations to be included in the query. See listMvbStations. Use stations="All" or stations=NULL to get all stations.

by

Time grouping for calculation, one of ('day','hour', '10min' ,'none')

calc

Calculation to perform given time grouping, one of ('avg', 'max', 'min', 'none').

usr

Username to connect to database

pwd

Password to connect to database

params

If TRUE, returns a list with the dataset and the query parameters applied in the server side. IF FALSE returns only the data.

...

Reserved for internal use.

Value

Dataframe with the specified MVB data.

Examples

getMvbData(Sys.Date() - 30, Sys.Date() + 1, parameters = 'Tide TAW',
stations = "Nieuwpoort", by = "day", calc = "avg")
#> - Query mode: Post request to OpenCPU server
#> No encoding supplied: defaulting to UTF-8.
#> - Your query:
#>    - daterange: c("2023-04-03", "2023-05-04")
#>    - type: MVB data
#>    - stations: Nieuwpoort
#>    - calc: avg
#>    - binSize: day
#>    - parameters: Tide TAW
#> - Server query:
#>    - daterange: c("2023-04-03", "2023-05-04")
#>    - type: MVB data
#>    - stations: Nieuwpoort
#>    - calc: avg
#>    - binSize: day
#>    - parameters: Tide (cmTAW)
#> Warning: The query applied on the server differ from the parameters you used. 
#>  Hint: You may need an account to fully access the data under moratorium. 
#>  Hint: Request access at: https://rshiny.lifewatch.be/account?p=register
#>          time    station latitude longitude Tide (cmTAW)
#> 1  2023-04-03 Nieuwpoort  51.1506    2.7281     207.4062
#> 2  2023-04-04 Nieuwpoort  51.1506    2.7281     214.6736
#> 3  2023-04-05 Nieuwpoort  51.1506    2.7281     221.3403
#> 4  2023-04-06 Nieuwpoort  51.1506    2.7281     233.7188
#> 5  2023-04-07 Nieuwpoort  51.1506    2.7281     238.9410
#> 6  2023-04-08 Nieuwpoort  51.1506    2.7281     224.7847
#> 7  2023-04-09 Nieuwpoort  51.1506    2.7281     225.6319
#> 8  2023-04-10 Nieuwpoort  51.1506    2.7281     227.2917
#> 9  2023-04-11 Nieuwpoort  51.1506    2.7281     254.1840
#> 10 2023-04-12 Nieuwpoort  51.1506    2.7281     245.6969
#> 11 2023-04-13 Nieuwpoort  51.1506    2.7281     262.0208
#> 12 2023-04-14 Nieuwpoort  51.1506    2.7281     239.6063
#> 13 2023-04-15 Nieuwpoort  51.1506    2.7281     246.4583
#> 14 2023-04-16 Nieuwpoort  51.1506    2.7281     224.9444
#> 15 2023-04-17 Nieuwpoort  51.1506    2.7281     221.7083
#> 16 2023-04-18 Nieuwpoort  51.1506    2.7281     212.0244
#> 17 2023-04-19 Nieuwpoort  51.1506    2.7281     213.3125
#> 18 2023-04-20 Nieuwpoort  51.1506    2.7281     218.0799
#> 19 2023-04-21 Nieuwpoort  51.1506    2.7281     225.9113
#> 20 2023-04-22 Nieuwpoort  51.1506    2.7281     236.2153
#> 21 2023-04-23 Nieuwpoort  51.1506    2.7281     245.9444
#> 22 2023-04-24 Nieuwpoort  51.1506    2.7281     274.5243
#> 23 2023-04-25 Nieuwpoort  51.1506    2.7281     254.0868
#> 24 2023-04-26 Nieuwpoort  51.1506    2.7281     238.6840
#> 25 2023-04-27 Nieuwpoort  51.1506    2.7281     236.6944
#> 26 2023-04-28 Nieuwpoort  51.1506    2.7281     250.7526
#> 27 2023-04-29 Nieuwpoort  51.1506    2.7281     244.2648
#> 28 2023-04-30 Nieuwpoort  51.1506    2.7281     226.3750
#> 29 2023-05-01 Nieuwpoort  51.1506    2.7281     240.7604
#> 30 2023-05-02 Nieuwpoort  51.1506    2.7281     234.7022
#> 31 2023-05-03 Nieuwpoort  51.1506    2.7281     133.1538
getMvbData(Sys.Date() - 30, Sys.Date() + 1, parameters = 'Tide TAW',
stations = "Blankenberge", by = "hour", calc = "max", params = TRUE)
#> - Query mode: Post request to OpenCPU server
#> No encoding supplied: defaulting to UTF-8.
#> - Your query:
#>    - daterange: c("2023-04-03", "2023-05-04")
#>    - type: MVB data
#>    - stations: Blankenberge
#>    - calc: max
#>    - binSize: hour
#>    - parameters: Tide TAW
#> - Server query:
#>    - daterange: c("2023-04-03", "2023-05-04")
#>    - type: MVB data
#>    - stations: Blankenberge
#>    - calc: max
#>    - binSize: hour
#>    - parameters: Tide (cmTAW)
#> Warning: The query applied on the server differ from the parameters you used. 
#>  Hint: You may need an account to fully access the data under moratorium. 
#>  Hint: Request access at: https://rshiny.lifewatch.be/account?p=register
#> $mdf
#>                    time      station latitude longitude Tide (cmTAW)
#> 1   2023-04-03 00:00:00 Blankenberge  51.3106    3.1164          357
#> 2   2023-04-03 01:00:00 Blankenberge  51.3106    3.1164          309
#> 3   2023-04-03 02:00:00 Blankenberge  51.3106    3.1164          236
#> 4   2023-04-03 03:00:00 Blankenberge  51.3106    3.1164          159
#> 5   2023-04-03 04:00:00 Blankenberge  51.3106    3.1164           82
#> 6   2023-04-03 05:00:00 Blankenberge  51.3106    3.1164           37
#> 7   2023-04-03 06:00:00 Blankenberge  51.3106    3.1164           75
#> 8   2023-04-03 07:00:00 Blankenberge  51.3106    3.1164          121
#> 9   2023-04-03 08:00:00 Blankenberge  51.3106    3.1164          204
#> 10  2023-04-03 09:00:00 Blankenberge  51.3106    3.1164          331
#> 11  2023-04-03 10:00:00 Blankenberge  51.3106    3.1164          392
#> 12  2023-04-03 11:00:00 Blankenberge  51.3106    3.1164          395
#> 13  2023-04-03 12:00:00 Blankenberge  51.3106    3.1164          384
#> 14  2023-04-03 13:00:00 Blankenberge  51.3106    3.1164          337
#> 15  2023-04-03 14:00:00 Blankenberge  51.3106    3.1164          275
#> 16  2023-04-03 15:00:00 Blankenberge  51.3106    3.1164          193
#> 17  2023-04-03 16:00:00 Blankenberge  51.3106    3.1164          109
#> 18  2023-04-03 17:00:00 Blankenberge  51.3106    3.1164           58
#> 19  2023-04-03 18:00:00 Blankenberge  51.3106    3.1164           75
#> 20  2023-04-03 19:00:00 Blankenberge  51.3106    3.1164          117
#> 21  2023-04-03 20:00:00 Blankenberge  51.3106    3.1164          168
#> 22  2023-04-03 21:00:00 Blankenberge  51.3106    3.1164          275
#> 23  2023-04-03 22:00:00 Blankenberge  51.3106    3.1164          375
#> 24  2023-04-03 23:00:00 Blankenberge  51.3106    3.1164          388
#> 25  2023-04-04 00:00:00 Blankenberge  51.3106    3.1164          378
#> 26  2023-04-04 01:00:00 Blankenberge  51.3106    3.1164          342
#> 27  2023-04-04 02:00:00 Blankenberge  51.3106    3.1164          282
#> 28  2023-04-04 03:00:00 Blankenberge  51.3106    3.1164          195
#> 29  2023-04-04 04:00:00 Blankenberge  51.3106    3.1164          112
#> 30  2023-04-04 05:00:00 Blankenberge  51.3106    3.1164           46
#> 31  2023-04-04 06:00:00 Blankenberge  51.3106    3.1164           44
#> 32  2023-04-04 07:00:00 Blankenberge  51.3106    3.1164           88
#> 33  2023-04-04 08:00:00 Blankenberge  51.3106    3.1164          151
#> 34  2023-04-04 09:00:00 Blankenberge  51.3106    3.1164          268
#> 35  2023-04-04 10:00:00 Blankenberge  51.3106    3.1164          394
#> 36  2023-04-04 11:00:00 Blankenberge  51.3106    3.1164          423
#> 37  2023-04-04 12:00:00 Blankenberge  51.3106    3.1164          420
#> 38  2023-04-04 13:00:00 Blankenberge  51.3106    3.1164          392
#> 39  2023-04-04 14:00:00 Blankenberge  51.3106    3.1164          332
#> 40  2023-04-04 15:00:00 Blankenberge  51.3106    3.1164          252
#> 41  2023-04-04 16:00:00 Blankenberge  51.3106    3.1164          166
#> 42  2023-04-04 17:00:00 Blankenberge  51.3106    3.1164           88
#> 43  2023-04-04 18:00:00 Blankenberge  51.3106    3.1164           69
#> 44  2023-04-04 19:00:00 Blankenberge  51.3106    3.1164          107
#> 45  2023-04-04 20:00:00 Blankenberge  51.3106    3.1164          156
#> 46  2023-04-04 21:00:00 Blankenberge  51.3106    3.1164          229
#> 47  2023-04-04 22:00:00 Blankenberge  51.3106    3.1164          371
#> 48  2023-04-04 23:00:00 Blankenberge  51.3106    3.1164          422
#> 49  2023-04-05 00:00:00 Blankenberge  51.3106    3.1164          423
#> 50  2023-04-05 01:00:00 Blankenberge  51.3106    3.1164          395
#> 51  2023-04-05 02:00:00 Blankenberge  51.3106    3.1164          341
#> 52  2023-04-05 03:00:00 Blankenberge  51.3106    3.1164          262
#> 53  2023-04-05 04:00:00 Blankenberge  51.3106    3.1164          172
#> 54  2023-04-05 05:00:00 Blankenberge  51.3106    3.1164           84
#> 55  2023-04-05 06:00:00 Blankenberge  51.3106    3.1164           27
#> 56  2023-04-05 07:00:00 Blankenberge  51.3106    3.1164           76
#> 57  2023-04-05 08:00:00 Blankenberge  51.3106    3.1164          121
#> 58  2023-04-05 09:00:00 Blankenberge  51.3106    3.1164          200
#> 59  2023-04-05 10:00:00 Blankenberge  51.3106    3.1164          354
#> 60  2023-04-05 11:00:00 Blankenberge  51.3106    3.1164          436
#> 61  2023-04-05 12:00:00 Blankenberge  51.3106    3.1164          441
#> 62  2023-04-05 13:00:00 Blankenberge  51.3106    3.1164          426
#> 63  2023-04-05 14:00:00 Blankenberge  51.3106    3.1164          376
#> 64  2023-04-05 15:00:00 Blankenberge  51.3106    3.1164          301
#> 65  2023-04-05 16:00:00 Blankenberge  51.3106    3.1164          208
#> 66  2023-04-05 17:00:00 Blankenberge  51.3106    3.1164          118
#> 67  2023-04-05 18:00:00 Blankenberge  51.3106    3.1164           54
#> 68  2023-04-05 19:00:00 Blankenberge  51.3106    3.1164           80
#> 69  2023-04-05 20:00:00 Blankenberge  51.3106    3.1164          123
#> 70  2023-04-05 21:00:00 Blankenberge  51.3106    3.1164          183
#> 71  2023-04-05 22:00:00 Blankenberge  51.3106    3.1164          306
#> 72  2023-04-05 23:00:00 Blankenberge  51.3106    3.1164          437
#> 73  2023-04-06 00:00:00 Blankenberge  51.3106    3.1164          446
#> 74  2023-04-06 01:00:00 Blankenberge  51.3106    3.1164          438
#> 75  2023-04-06 02:00:00 Blankenberge  51.3106    3.1164          390
#> 76  2023-04-06 03:00:00 Blankenberge  51.3106    3.1164          320
#> 77  2023-04-06 04:00:00 Blankenberge  51.3106    3.1164          228
#> 78  2023-04-06 05:00:00 Blankenberge  51.3106    3.1164          135
#> 79  2023-04-06 06:00:00 Blankenberge  51.3106    3.1164           53
#> 80  2023-04-06 07:00:00 Blankenberge  51.3106    3.1164           51
#> 81  2023-04-06 08:00:00 Blankenberge  51.3106    3.1164           97
#> 82  2023-04-06 09:00:00 Blankenberge  51.3106    3.1164          163
#> 83  2023-04-06 10:00:00 Blankenberge  51.3106    3.1164          282
#> 84  2023-04-06 11:00:00 Blankenberge  51.3106    3.1164          436
#> 85  2023-04-06 12:00:00 Blankenberge  51.3106    3.1164          464
#> 86  2023-04-06 13:00:00 Blankenberge  51.3106    3.1164          458
#> 87  2023-04-06 14:00:00 Blankenberge  51.3106    3.1164          426
#> 88  2023-04-06 15:00:00 Blankenberge  51.3106    3.1164          361
#> 89  2023-04-06 16:00:00 Blankenberge  51.3106    3.1164          272
#> 90  2023-04-06 17:00:00 Blankenberge  51.3106    3.1164          177
#> 91  2023-04-06 18:00:00 Blankenberge  51.3106    3.1164           92
#> 92  2023-04-06 19:00:00 Blankenberge  51.3106    3.1164           77
#> 93  2023-04-06 20:00:00 Blankenberge  51.3106    3.1164          120
#> 94  2023-04-06 21:00:00 Blankenberge  51.3106    3.1164          174
#> 95  2023-04-06 22:00:00 Blankenberge  51.3106    3.1164          246
#> 96  2023-04-06 23:00:00 Blankenberge  51.3106    3.1164          414
#> 97  2023-04-07 00:00:00 Blankenberge  51.3106    3.1164          468
#> 98  2023-04-07 01:00:00 Blankenberge  51.3106    3.1164          468
#> 99  2023-04-07 02:00:00 Blankenberge  51.3106    3.1164          437
#> 100 2023-04-07 03:00:00 Blankenberge  51.3106    3.1164          380
#> 101 2023-04-07 04:00:00 Blankenberge  51.3106    3.1164          291
#> 102 2023-04-07 05:00:00 Blankenberge  51.3106    3.1164          191
#> 103 2023-04-07 06:00:00 Blankenberge  51.3106    3.1164           95
#> 104 2023-04-07 07:00:00 Blankenberge  51.3106    3.1164           43
#> 105 2023-04-07 08:00:00 Blankenberge  51.3106    3.1164           99
#> 106 2023-04-07 09:00:00 Blankenberge  51.3106    3.1164          147
#> 107 2023-04-07 10:00:00 Blankenberge  51.3106    3.1164          215
#> 108 2023-04-07 11:00:00 Blankenberge  51.3106    3.1164          371
#> 109 2023-04-07 12:00:00 Blankenberge  51.3106    3.1164          471
#> 110 2023-04-07 13:00:00 Blankenberge  51.3106    3.1164          477
#> 111 2023-04-07 14:00:00 Blankenberge  51.3106    3.1164          460
#> 112 2023-04-07 15:00:00 Blankenberge  51.3106    3.1164          409
#> 113 2023-04-07 16:00:00 Blankenberge  51.3106    3.1164          324
#> 114 2023-04-07 17:00:00 Blankenberge  51.3106    3.1164          223
#> 115 2023-04-07 18:00:00 Blankenberge  51.3106    3.1164          124
#> 116 2023-04-07 19:00:00 Blankenberge  51.3106    3.1164           58
#> 117 2023-04-07 20:00:00 Blankenberge  51.3106    3.1164           93
#> 118 2023-04-07 21:00:00 Blankenberge  51.3106    3.1164          137
#> 119 2023-04-07 22:00:00 Blankenberge  51.3106    3.1164          190
#> 120 2023-04-07 23:00:00 Blankenberge  51.3106    3.1164          300
#> 121 2023-04-08 00:00:00 Blankenberge  51.3106    3.1164          453
#> 122 2023-04-08 01:00:00 Blankenberge  51.3106    3.1164          464
#> 123 2023-04-08 02:00:00 Blankenberge  51.3106    3.1164          454
#> 124 2023-04-08 03:00:00 Blankenberge  51.3106    3.1164          404
#> 125 2023-04-08 04:00:00 Blankenberge  51.3106    3.1164          328
#> 126 2023-04-08 05:00:00 Blankenberge  51.3106    3.1164          226
#> 127 2023-04-08 06:00:00 Blankenberge  51.3106    3.1164          126
#> 128 2023-04-08 07:00:00 Blankenberge  51.3106    3.1164           42
#> 129 2023-04-08 08:00:00 Blankenberge  51.3106    3.1164           40
#> 130 2023-04-08 09:00:00 Blankenberge  51.3106    3.1164           91
#> 131 2023-04-08 10:00:00 Blankenberge  51.3106    3.1164          148
#> 132 2023-04-08 11:00:00 Blankenberge  51.3106    3.1164          247
#> 133 2023-04-08 12:00:00 Blankenberge  51.3106    3.1164          426
#> 134 2023-04-08 13:00:00 Blankenberge  51.3106    3.1164          458
#> 135 2023-04-08 14:00:00 Blankenberge  51.3106    3.1164          455
#> 136 2023-04-08 15:00:00 Blankenberge  51.3106    3.1164          421
#> 137 2023-04-08 16:00:00 Blankenberge  51.3106    3.1164          352
#> 138 2023-04-08 17:00:00 Blankenberge  51.3106    3.1164          258
#> 139 2023-04-08 18:00:00 Blankenberge  51.3106    3.1164          160
#> 140 2023-04-08 19:00:00 Blankenberge  51.3106    3.1164           76
#> 141 2023-04-08 20:00:00 Blankenberge  51.3106    3.1164           51
#> 142 2023-04-08 21:00:00 Blankenberge  51.3106    3.1164          101
#> 143 2023-04-08 22:00:00 Blankenberge  51.3106    3.1164          148
#> 144 2023-04-08 23:00:00 Blankenberge  51.3106    3.1164          223
#> 145 2023-04-09 00:00:00 Blankenberge  51.3106    3.1164          390
#> 146 2023-04-09 01:00:00 Blankenberge  51.3106    3.1164          458
#> 147 2023-04-09 02:00:00 Blankenberge  51.3106    3.1164          458
#> 148 2023-04-09 03:00:00 Blankenberge  51.3106    3.1164          428
#> 149 2023-04-09 04:00:00 Blankenberge  51.3106    3.1164          369
#> 150 2023-04-09 05:00:00 Blankenberge  51.3106    3.1164          279
#> 151 2023-04-09 06:00:00 Blankenberge  51.3106    3.1164          180
#> 152 2023-04-09 07:00:00 Blankenberge  51.3106    3.1164           85
#> 153 2023-04-09 08:00:00 Blankenberge  51.3106    3.1164           18
#> 154 2023-04-09 09:00:00 Blankenberge  51.3106    3.1164           70
#> 155 2023-04-09 10:00:00 Blankenberge  51.3106    3.1164          115
#> 156 2023-04-09 11:00:00 Blankenberge  51.3106    3.1164          185
#> 157 2023-04-09 12:00:00 Blankenberge  51.3106    3.1164          332
#> 158 2023-04-09 13:00:00 Blankenberge  51.3106    3.1164          448
#> 159 2023-04-09 14:00:00 Blankenberge  51.3106    3.1164          459
#> 160 2023-04-09 15:00:00 Blankenberge  51.3106    3.1164          439
#> 161 2023-04-09 16:00:00 Blankenberge  51.3106    3.1164          392
#> 162 2023-04-09 17:00:00 Blankenberge  51.3106    3.1164          312
#> 163 2023-04-09 18:00:00 Blankenberge  51.3106    3.1164          213
#> 164 2023-04-09 19:00:00 Blankenberge  51.3106    3.1164          120
#> 165 2023-04-09 20:00:00 Blankenberge  51.3106    3.1164           47
#> 166 2023-04-09 21:00:00 Blankenberge  51.3106    3.1164           67
#> 167 2023-04-09 22:00:00 Blankenberge  51.3106    3.1164          109
#> 168 2023-04-09 23:00:00 Blankenberge  51.3106    3.1164          169
#> 169 2023-04-10 00:00:00 Blankenberge  51.3106    3.1164          281
#> 170 2023-04-10 01:00:00 Blankenberge  51.3106    3.1164          428
#> 171 2023-04-10 02:00:00 Blankenberge  51.3106    3.1164          444
#> 172 2023-04-10 03:00:00 Blankenberge  51.3106    3.1164          439
#> 173 2023-04-10 04:00:00 Blankenberge  51.3106    3.1164          394
#> 174 2023-04-10 05:00:00 Blankenberge  51.3106    3.1164          322
#> 175 2023-04-10 06:00:00 Blankenberge  51.3106    3.1164          229
#> 176 2023-04-10 07:00:00 Blankenberge  51.3106    3.1164          137
#> 177 2023-04-10 08:00:00 Blankenberge  51.3106    3.1164           54
#> 178 2023-04-10 09:00:00 Blankenberge  51.3106    3.1164           24
#> 179 2023-04-10 10:00:00 Blankenberge  51.3106    3.1164           79
#> 180 2023-04-10 11:00:00 Blankenberge  51.3106    3.1164          129
#> 181 2023-04-10 12:00:00 Blankenberge  51.3106    3.1164          227
#> 182 2023-04-10 13:00:00 Blankenberge  51.3106    3.1164          398
#> 183 2023-04-10 14:00:00 Blankenberge  51.3106    3.1164          437
#> 184 2023-04-10 15:00:00 Blankenberge  51.3106    3.1164          435
#> 185 2023-04-10 16:00:00 Blankenberge  51.3106    3.1164          401
#> 186 2023-04-10 17:00:00 Blankenberge  51.3106    3.1164          344
#> 187 2023-04-10 18:00:00 Blankenberge  51.3106    3.1164          261
#> 188 2023-04-10 19:00:00 Blankenberge  51.3106    3.1164          185
#> 189 2023-04-10 20:00:00 Blankenberge  51.3106    3.1164          110
#> 190 2023-04-10 21:00:00 Blankenberge  51.3106    3.1164           79
#> 191 2023-04-10 22:00:00 Blankenberge  51.3106    3.1164          123
#> 192 2023-04-10 23:00:00 Blankenberge  51.3106    3.1164          167
#> 193 2023-04-11 00:00:00 Blankenberge  51.3106    3.1164          255
#> 194 2023-04-11 01:00:00 Blankenberge  51.3106    3.1164          406
#> 195 2023-04-11 02:00:00 Blankenberge  51.3106    3.1164          464
#> 196 2023-04-11 03:00:00 Blankenberge  51.3106    3.1164          472
#> 197 2023-04-11 04:00:00 Blankenberge  51.3106    3.1164          450
#> 198 2023-04-11 05:00:00 Blankenberge  51.3106    3.1164          402
#> 199 2023-04-11 06:00:00 Blankenberge  51.3106    3.1164          326
#> 200 2023-04-11 07:00:00 Blankenberge  51.3106    3.1164          239
#> 201 2023-04-11 08:00:00 Blankenberge  51.3106    3.1164          156
#> 202 2023-04-11 09:00:00 Blankenberge  51.3106    3.1164           93
#> 203 2023-04-11 10:00:00 Blankenberge  51.3106    3.1164          116
#> 204 2023-04-11 11:00:00 Blankenberge  51.3106    3.1164          159
#> 205 2023-04-11 12:00:00 Blankenberge  51.3106    3.1164          215
#> 206 2023-04-11 13:00:00 Blankenberge  51.3106    3.1164          323
#> 207 2023-04-11 14:00:00 Blankenberge  51.3106    3.1164          441
#> 208 2023-04-11 15:00:00 Blankenberge  51.3106    3.1164          451
#> 209 2023-04-11 16:00:00 Blankenberge  51.3106    3.1164          440
#> 210 2023-04-11 17:00:00 Blankenberge  51.3106    3.1164          398
#> 211 2023-04-11 18:00:00 Blankenberge  51.3106    3.1164          334
#> 212 2023-04-11 19:00:00 Blankenberge  51.3106    3.1164          240
#> 213 2023-04-11 20:00:00 Blankenberge  51.3106    3.1164          158
#> 214 2023-04-11 21:00:00 Blankenberge  51.3106    3.1164           82
#> 215 2023-04-11 22:00:00 Blankenberge  51.3106    3.1164           62
#> 216 2023-04-11 23:00:00 Blankenberge  51.3106    3.1164           99
#> 217 2023-04-12 00:00:00 Blankenberge  51.3106    3.1164          166
#> 218 2023-04-12 01:00:00 Blankenberge  51.3106    3.1164          285
#> 219 2023-04-12 02:00:00 Blankenberge  51.3106    3.1164          397
#> 220 2023-04-12 03:00:00 Blankenberge  51.3106    3.1164          415
#> 221 2023-04-12 04:00:00 Blankenberge  51.3106    3.1164          412
#> 222 2023-04-12 05:00:00 Blankenberge  51.3106    3.1164          385
#> 223 2023-04-12 06:00:00 Blankenberge  51.3106    3.1164          329
#> 224 2023-04-12 07:00:00 Blankenberge  51.3106    3.1164          259
#> 225 2023-04-12 08:00:00 Blankenberge  51.3106    3.1164          188
#> 226 2023-04-12 09:00:00 Blankenberge  51.3106    3.1164          112
#> 227 2023-04-12 10:00:00 Blankenberge  51.3106    3.1164           72
#> 228 2023-04-12 11:00:00 Blankenberge  51.3106    3.1164           95
#> 229 2023-04-12 12:00:00 Blankenberge  51.3106    3.1164          139
#> 230 2023-04-12 13:00:00 Blankenberge  51.3106    3.1164          230
#> 231 2023-04-12 14:00:00 Blankenberge  51.3106    3.1164          362
#> 232 2023-04-12 15:00:00 Blankenberge  51.3106    3.1164          410
#> 233 2023-04-12 16:00:00 Blankenberge  51.3106    3.1164          416
#> 234 2023-04-12 17:00:00 Blankenberge  51.3106    3.1164          398
#> 235 2023-04-12 18:00:00 Blankenberge  51.3106    3.1164          351
#> 236 2023-04-12 19:00:00 Blankenberge  51.3106    3.1164          298
#> 237 2023-04-12 20:00:00 Blankenberge  51.3106    3.1164          237
#> 238 2023-04-12 21:00:00 Blankenberge  51.3106    3.1164          163
#> 239 2023-04-12 22:00:00 Blankenberge  51.3106    3.1164          133
#> 240 2023-04-12 23:00:00 Blankenberge  51.3106    3.1164          135
#> 241 2023-04-13 00:00:00 Blankenberge  51.3106    3.1164          196
#> 242 2023-04-13 01:00:00 Blankenberge  51.3106    3.1164          270
#> 243 2023-04-13 02:00:00 Blankenberge  51.3106    3.1164          357
#> 244 2023-04-13 03:00:00 Blankenberge  51.3106    3.1164          416
#> 245 2023-04-13 04:00:00 Blankenberge  51.3106    3.1164          437
#> 246 2023-04-13 05:00:00 Blankenberge  51.3106    3.1164          436
#> 247 2023-04-13 06:00:00 Blankenberge  51.3106    3.1164          414
#> 248 2023-04-13 07:00:00 Blankenberge  51.3106    3.1164          363
#> 249 2023-04-13 08:00:00 Blankenberge  51.3106    3.1164          299
#> 250 2023-04-13 09:00:00 Blankenberge  51.3106    3.1164          224
#> 251 2023-04-13 10:00:00 Blankenberge  51.3106    3.1164          155
#> 252 2023-04-13 11:00:00 Blankenberge  51.3106    3.1164          115
#> 253 2023-04-13 12:00:00 Blankenberge  51.3106    3.1164          138
#> 254 2023-04-13 13:00:00 Blankenberge  51.3106    3.1164          182
#> 255 2023-04-13 14:00:00 Blankenberge  51.3106    3.1164          246
#> 256 2023-04-13 15:00:00 Blankenberge  51.3106    3.1164          329
#> 257 2023-04-13 16:00:00 Blankenberge  51.3106    3.1164          377
#> 258 2023-04-13 17:00:00 Blankenberge  51.3106    3.1164          382
#> 259 2023-04-13 18:00:00 Blankenberge  51.3106    3.1164          377
#> 260 2023-04-13 19:00:00 Blankenberge  51.3106    3.1164          340
#> 261 2023-04-13 20:00:00 Blankenberge  51.3106    3.1164          284
#> 262 2023-04-13 21:00:00 Blankenberge  51.3106    3.1164          218
#> 263 2023-04-13 22:00:00 Blankenberge  51.3106    3.1164          156
#> 264 2023-04-13 23:00:00 Blankenberge  51.3106    3.1164          105
#> 265 2023-04-14 00:00:00 Blankenberge  51.3106    3.1164           84
#> 266 2023-04-14 01:00:00 Blankenberge  51.3106    3.1164          129
#> 267 2023-04-14 02:00:00 Blankenberge  51.3106    3.1164          186
#> 268 2023-04-14 03:00:00 Blankenberge  51.3106    3.1164          270
#> 269 2023-04-14 04:00:00 Blankenberge  51.3106    3.1164          347
#> 270 2023-04-14 05:00:00 Blankenberge  51.3106    3.1164          383
#> 271 2023-04-14 06:00:00 Blankenberge  51.3106    3.1164          382
#> 272 2023-04-14 07:00:00 Blankenberge  51.3106    3.1164          367
#> 273 2023-04-14 08:00:00 Blankenberge  51.3106    3.1164          327
#> 274 2023-04-14 09:00:00 Blankenberge  51.3106    3.1164          271
#> 275 2023-04-14 10:00:00 Blankenberge  51.3106    3.1164          209
#> 276 2023-04-14 11:00:00 Blankenberge  51.3106    3.1164          147
#> 277 2023-04-14 12:00:00 Blankenberge  51.3106    3.1164          104
#> 278 2023-04-14 13:00:00 Blankenberge  51.3106    3.1164           99
#> 279 2023-04-14 14:00:00 Blankenberge  51.3106    3.1164          141
#> 280 2023-04-14 15:00:00 Blankenberge  51.3106    3.1164          198
#> 281 2023-04-14 16:00:00 Blankenberge  51.3106    3.1164          276
#> 282 2023-04-14 17:00:00 Blankenberge  51.3106    3.1164          343
#> 283 2023-04-14 18:00:00 Blankenberge  51.3106    3.1164          368
#> 284 2023-04-14 19:00:00 Blankenberge  51.3106    3.1164          368
#> 285 2023-04-14 20:00:00 Blankenberge  51.3106    3.1164          349
#> 286 2023-04-14 21:00:00 Blankenberge  51.3106    3.1164          312
#> 287 2023-04-14 22:00:00 Blankenberge  51.3106    3.1164          260
#> 288 2023-04-14 23:00:00 Blankenberge  51.3106    3.1164          204
#> 289 2023-04-15 00:00:00 Blankenberge  51.3106    3.1164          147
#> 290 2023-04-15 01:00:00 Blankenberge  51.3106    3.1164          107
#> 291 2023-04-15 02:00:00 Blankenberge  51.3106    3.1164          123
#> 292 2023-04-15 03:00:00 Blankenberge  51.3106    3.1164          173
#> 293 2023-04-15 04:00:00 Blankenberge  51.3106    3.1164          238
#> 294 2023-04-15 05:00:00 Blankenberge  51.3106    3.1164          321
#> 295 2023-04-15 06:00:00 Blankenberge  51.3106    3.1164          386
#> 296 2023-04-15 07:00:00 Blankenberge  51.3106    3.1164          410
#> 297 2023-04-15 08:00:00 Blankenberge  51.3106    3.1164          407
#> 298 2023-04-15 09:00:00 Blankenberge  51.3106    3.1164          387
#> 299 2023-04-15 10:00:00 Blankenberge  51.3106    3.1164          341
#> 300 2023-04-15 11:00:00 Blankenberge  51.3106    3.1164          279
#> 301 2023-04-15 12:00:00 Blankenberge  51.3106    3.1164          210
#> 302 2023-04-15 13:00:00 Blankenberge  51.3106    3.1164          142
#> 303 2023-04-15 14:00:00 Blankenberge  51.3106    3.1164          103
#> 304 2023-04-15 15:00:00 Blankenberge  51.3106    3.1164          117
#> 305 2023-04-15 16:00:00 Blankenberge  51.3106    3.1164          163
#> 306 2023-04-15 17:00:00 Blankenberge  51.3106    3.1164          218
#> 307 2023-04-15 18:00:00 Blankenberge  51.3106    3.1164          291
#> 308 2023-04-15 19:00:00 Blankenberge  51.3106    3.1164          357
#> 309 2023-04-15 20:00:00 Blankenberge  51.3106    3.1164          370
#> 310 2023-04-15 21:00:00 Blankenberge  51.3106    3.1164          366
#> 311 2023-04-15 22:00:00 Blankenberge  51.3106    3.1164          338
#> 312 2023-04-15 23:00:00 Blankenberge  51.3106    3.1164          292
#> 313 2023-04-16 00:00:00 Blankenberge  51.3106    3.1164          229
#> 314 2023-04-16 01:00:00 Blankenberge  51.3106    3.1164          161
#> 315 2023-04-16 02:00:00 Blankenberge  51.3106    3.1164           99
#> 316 2023-04-16 03:00:00 Blankenberge  51.3106    3.1164           65
#> 317 2023-04-16 04:00:00 Blankenberge  51.3106    3.1164          100
#> 318 2023-04-16 05:00:00 Blankenberge  51.3106    3.1164          157
#> 319 2023-04-16 06:00:00 Blankenberge  51.3106    3.1164          238
#> 320 2023-04-16 07:00:00 Blankenberge  51.3106    3.1164          346
#> 321 2023-04-16 08:00:00 Blankenberge  51.3106    3.1164          402
#> 322 2023-04-16 09:00:00 Blankenberge  51.3106    3.1164          408
#> 323 2023-04-16 10:00:00 Blankenberge  51.3106    3.1164          400
#> 324 2023-04-16 11:00:00 Blankenberge  51.3106    3.1164          360
#> 325 2023-04-16 12:00:00 Blankenberge  51.3106    3.1164          298
#> 326 2023-04-16 13:00:00 Blankenberge  51.3106    3.1164          221
#> 327 2023-04-16 14:00:00 Blankenberge  51.3106    3.1164          141
#> 328 2023-04-16 15:00:00 Blankenberge  51.3106    3.1164           83
#> 329 2023-04-16 16:00:00 Blankenberge  51.3106    3.1164           82
#> 330 2023-04-16 17:00:00 Blankenberge  51.3106    3.1164          120
#> 331 2023-04-16 18:00:00 Blankenberge  51.3106    3.1164          171
#> 332 2023-04-16 19:00:00 Blankenberge  51.3106    3.1164          258
#> 333 2023-04-16 20:00:00 Blankenberge  51.3106    3.1164          365
#> 334 2023-04-16 21:00:00 Blankenberge  51.3106    3.1164          394
#> 335 2023-04-16 22:00:00 Blankenberge  51.3106    3.1164          393
#> 336 2023-04-16 23:00:00 Blankenberge  51.3106    3.1164          365
#> 337 2023-04-17 00:00:00 Blankenberge  51.3106    3.1164          315
#> 338 2023-04-17 01:00:00 Blankenberge  51.3106    3.1164          243
#> 339 2023-04-17 02:00:00 Blankenberge  51.3106    3.1164          167
#> 340 2023-04-17 03:00:00 Blankenberge  51.3106    3.1164           88
#> 341 2023-04-17 04:00:00 Blankenberge  51.3106    3.1164           40
#> 342 2023-04-17 05:00:00 Blankenberge  51.3106    3.1164           80
#> 343 2023-04-17 06:00:00 Blankenberge  51.3106    3.1164          136
#> 344 2023-04-17 07:00:00 Blankenberge  51.3106    3.1164          217
#> 345 2023-04-17 08:00:00 Blankenberge  51.3106    3.1164          354
#> 346 2023-04-17 09:00:00 Blankenberge  51.3106    3.1164          431
#> 347 2023-04-17 10:00:00 Blankenberge  51.3106    3.1164          442
#> 348 2023-04-17 11:00:00 Blankenberge  51.3106    3.1164          432
#> 349 2023-04-17 12:00:00 Blankenberge  51.3106    3.1164          388
#> 350 2023-04-17 13:00:00 Blankenberge  51.3106    3.1164          314
#> 351 2023-04-17 14:00:00 Blankenberge  51.3106    3.1164          228
#> 352 2023-04-17 15:00:00 Blankenberge  51.3106    3.1164          138
#> 353 2023-04-17 16:00:00 Blankenberge  51.3106    3.1164           70
#> 354 2023-04-17 17:00:00 Blankenberge  51.3106    3.1164           80
#> 355 2023-04-17 18:00:00 Blankenberge  51.3106    3.1164          119
#> 356 2023-04-17 19:00:00 Blankenberge  51.3106    3.1164          170
#> 357 2023-04-17 20:00:00 Blankenberge  51.3106    3.1164          251
#> 358 2023-04-17 21:00:00 Blankenberge  51.3106    3.1164          385
#> 359 2023-04-17 22:00:00 Blankenberge  51.3106    3.1164          415
#> 360 2023-04-17 23:00:00 Blankenberge  51.3106    3.1164          411
#> 361 2023-04-18 00:00:00 Blankenberge  51.3106    3.1164          371
#> 362 2023-04-18 01:00:00 Blankenberge  51.3106    3.1164          316
#> 363 2023-04-18 02:00:00 Blankenberge  51.3106    3.1164          231
#> 364 2023-04-18 03:00:00 Blankenberge  51.3106    3.1164          139
#> 365 2023-04-18 04:00:00 Blankenberge  51.3106    3.1164           58
#> 366 2023-04-18 05:00:00 Blankenberge  51.3106    3.1164            9
#> 367 2023-04-18 06:00:00 Blankenberge  51.3106    3.1164           64
#> 368 2023-04-18 07:00:00 Blankenberge  51.3106    3.1164          116
#> 369 2023-04-18 08:00:00 Blankenberge  51.3106    3.1164          200
#> 370 2023-04-18 09:00:00 Blankenberge  51.3106    3.1164          360
#> 371 2023-04-18 10:00:00 Blankenberge  51.3106    3.1164          439
#> 372 2023-04-18 11:00:00 Blankenberge  51.3106    3.1164          447
#> 373 2023-04-18 12:00:00 Blankenberge  51.3106    3.1164          431
#> 374 2023-04-18 13:00:00 Blankenberge  51.3106    3.1164          381
#> 375 2023-04-18 14:00:00 Blankenberge  51.3106    3.1164          298
#> 376 2023-04-18 15:00:00 Blankenberge  51.3106    3.1164          199
#> 377 2023-04-18 16:00:00 Blankenberge  51.3106    3.1164           99
#> 378 2023-04-18 17:00:00 Blankenberge  51.3106    3.1164           40
#> 379 2023-04-18 18:00:00 Blankenberge  51.3106    3.1164           86
#> 380 2023-04-18 19:00:00 Blankenberge  51.3106    3.1164          128
#> 381 2023-04-18 20:00:00 Blankenberge  51.3106    3.1164          178
#> 382 2023-04-18 21:00:00 Blankenberge  51.3106    3.1164          279
#> 383 2023-04-18 22:00:00 Blankenberge  51.3106    3.1164          422
#> 384 2023-04-18 23:00:00 Blankenberge  51.3106    3.1164          433
#> 385 2023-04-19 00:00:00 Blankenberge  51.3106    3.1164          425
#> 386 2023-04-19 01:00:00 Blankenberge  51.3106    3.1164          376
#> 387 2023-04-19 02:00:00 Blankenberge  51.3106    3.1164          306
#> 388 2023-04-19 03:00:00 Blankenberge  51.3106    3.1164          206
#> 389 2023-04-19 04:00:00 Blankenberge  51.3106    3.1164          104
#> 390 2023-04-19 05:00:00 Blankenberge  51.3106    3.1164           26
#> 391 2023-04-19 06:00:00 Blankenberge  51.3106    3.1164            7
#> 392 2023-04-19 07:00:00 Blankenberge  51.3106    3.1164           72
#> 393 2023-04-19 08:00:00 Blankenberge  51.3106    3.1164          126
#> 394 2023-04-19 09:00:00 Blankenberge  51.3106    3.1164          226
#> 395 2023-04-19 10:00:00 Blankenberge  51.3106    3.1164          414
#> 396 2023-04-19 11:00:00 Blankenberge  51.3106    3.1164          461
#> 397 2023-04-19 12:00:00 Blankenberge  51.3106    3.1164          462
#> 398 2023-04-19 13:00:00 Blankenberge  51.3106    3.1164          434
#> 399 2023-04-19 14:00:00 Blankenberge  51.3106    3.1164          368
#> 400 2023-04-19 15:00:00 Blankenberge  51.3106    3.1164          274
#> 401 2023-04-19 16:00:00 Blankenberge  51.3106    3.1164          173
#> 402 2023-04-19 17:00:00 Blankenberge  51.3106    3.1164           80
#> 403 2023-04-19 18:00:00 Blankenberge  51.3106    3.1164           49
#> 404 2023-04-19 19:00:00 Blankenberge  51.3106    3.1164          105
#> 405 2023-04-19 20:00:00 Blankenberge  51.3106    3.1164          151
#> 406 2023-04-19 21:00:00 Blankenberge  51.3106    3.1164          205
#> 407 2023-04-19 22:00:00 Blankenberge  51.3106    3.1164          337
#> 408 2023-04-19 23:00:00 Blankenberge  51.3106    3.1164          451
#> 409 2023-04-20 00:00:00 Blankenberge  51.3106    3.1164          452
#> 410 2023-04-20 01:00:00 Blankenberge  51.3106    3.1164          431
#> 411 2023-04-20 02:00:00 Blankenberge  51.3106    3.1164          374
#> 412 2023-04-20 03:00:00 Blankenberge  51.3106    3.1164          285
#> 413 2023-04-20 04:00:00 Blankenberge  51.3106    3.1164          180
#> 414 2023-04-20 05:00:00 Blankenberge  51.3106    3.1164           81
#> 415 2023-04-20 06:00:00 Blankenberge  51.3106    3.1164            9
#> 416 2023-04-20 07:00:00 Blankenberge  51.3106    3.1164           46
#> 417 2023-04-20 08:00:00 Blankenberge  51.3106    3.1164           97
#> 418 2023-04-20 09:00:00 Blankenberge  51.3106    3.1164          157
#> 419 2023-04-20 10:00:00 Blankenberge  51.3106    3.1164          271
#> 420 2023-04-20 11:00:00 Blankenberge  51.3106    3.1164          455
#> 421 2023-04-20 12:00:00 Blankenberge  51.3106    3.1164          470
#> 422 2023-04-20 13:00:00 Blankenberge  51.3106    3.1164          467
#> 423 2023-04-20 14:00:00 Blankenberge  51.3106    3.1164          422
#> 424 2023-04-20 15:00:00 Blankenberge  51.3106    3.1164          346
#> 425 2023-04-20 16:00:00 Blankenberge  51.3106    3.1164          248
#> 426 2023-04-20 17:00:00 Blankenberge  51.3106    3.1164          139
#> 427 2023-04-20 18:00:00 Blankenberge  51.3106    3.1164           51
#> 428 2023-04-20 19:00:00 Blankenberge  51.3106    3.1164           74
#> 429 2023-04-20 20:00:00 Blankenberge  51.3106    3.1164          122
#> 430 2023-04-20 21:00:00 Blankenberge  51.3106    3.1164          174
#> 431 2023-04-20 22:00:00 Blankenberge  51.3106    3.1164          251
#> 432 2023-04-20 23:00:00 Blankenberge  51.3106    3.1164          431
#> 433 2023-04-21 00:00:00 Blankenberge  51.3106    3.1164          473
#> 434 2023-04-21 01:00:00 Blankenberge  51.3106    3.1164          471
#> 435 2023-04-21 02:00:00 Blankenberge  51.3106    3.1164          432
#> 436 2023-04-21 03:00:00 Blankenberge  51.3106    3.1164          368
#> 437 2023-04-21 04:00:00 Blankenberge  51.3106    3.1164          265
#> 438 2023-04-21 05:00:00 Blankenberge  51.3106    3.1164          159
#> 439 2023-04-21 06:00:00 Blankenberge  51.3106    3.1164           62
#> 440 2023-04-21 07:00:00 Blankenberge  51.3106    3.1164            2
#> 441 2023-04-21 08:00:00 Blankenberge  51.3106    3.1164           63
#> 442 2023-04-21 09:00:00 Blankenberge  51.3106    3.1164          115
#> 443 2023-04-21 10:00:00 Blankenberge  51.3106    3.1164          191
#> 444 2023-04-21 11:00:00 Blankenberge  51.3106    3.1164          368
#> 445 2023-04-21 12:00:00 Blankenberge  51.3106    3.1164          472
#> 446 2023-04-21 13:00:00 Blankenberge  51.3106    3.1164          475
#> 447 2023-04-21 14:00:00 Blankenberge  51.3106    3.1164          455
#> 448 2023-04-21 15:00:00 Blankenberge  51.3106    3.1164          403
#> 449 2023-04-21 16:00:00 Blankenberge  51.3106    3.1164          313
#> 450 2023-04-21 17:00:00 Blankenberge  51.3106    3.1164          212
#> 451 2023-04-21 18:00:00 Blankenberge  51.3106    3.1164          114
#> 452 2023-04-21 19:00:00 Blankenberge  51.3106    3.1164           38
#> 453 2023-04-21 20:00:00 Blankenberge  51.3106    3.1164           93
#> 454 2023-04-21 21:00:00 Blankenberge  51.3106    3.1164          144
#> 455 2023-04-21 22:00:00 Blankenberge  51.3106    3.1164          203
#> 456 2023-04-21 23:00:00 Blankenberge  51.3106    3.1164          329
#> 457 2023-04-22 00:00:00 Blankenberge  51.3106    3.1164          478
#> 458 2023-04-22 01:00:00 Blankenberge  51.3106    3.1164          483
#> 459 2023-04-22 02:00:00 Blankenberge  51.3106    3.1164          472
#> 460 2023-04-22 03:00:00 Blankenberge  51.3106    3.1164          420
#> 461 2023-04-22 04:00:00 Blankenberge  51.3106    3.1164          336
#> 462 2023-04-22 05:00:00 Blankenberge  51.3106    3.1164          231
#> 463 2023-04-22 06:00:00 Blankenberge  51.3106    3.1164          130
#> 464 2023-04-22 07:00:00 Blankenberge  51.3106    3.1164           43
#> 465 2023-04-22 08:00:00 Blankenberge  51.3106    3.1164           45
#> 466 2023-04-22 09:00:00 Blankenberge  51.3106    3.1164           94
#> 467 2023-04-22 10:00:00 Blankenberge  51.3106    3.1164          156
#> 468 2023-04-22 11:00:00 Blankenberge  51.3106    3.1164          257
#> 469 2023-04-22 12:00:00 Blankenberge  51.3106    3.1164          446
#> 470 2023-04-22 13:00:00 Blankenberge  51.3106    3.1164          475
#> 471 2023-04-22 14:00:00 Blankenberge  51.3106    3.1164          473
#> 472 2023-04-22 15:00:00 Blankenberge  51.3106    3.1164          437
#> 473 2023-04-22 16:00:00 Blankenberge  51.3106    3.1164          369
#> 474 2023-04-22 17:00:00 Blankenberge  51.3106    3.1164          273
#> 475 2023-04-22 18:00:00 Blankenberge  51.3106    3.1164          176
#> 476 2023-04-22 19:00:00 Blankenberge  51.3106    3.1164           87
#> 477 2023-04-22 20:00:00 Blankenberge  51.3106    3.1164           49
#> 478 2023-04-22 21:00:00 Blankenberge  51.3106    3.1164          102
#> 479 2023-04-22 22:00:00 Blankenberge  51.3106    3.1164          158
#> 480 2023-04-22 23:00:00 Blankenberge  51.3106    3.1164          241
#> 481 2023-04-23 00:00:00 Blankenberge  51.3106    3.1164          420
#> 482 2023-04-23 01:00:00 Blankenberge  51.3106    3.1164          477
#> 483 2023-04-23 02:00:00 Blankenberge  51.3106    3.1164          479
#> 484 2023-04-23 03:00:00 Blankenberge  51.3106    3.1164          450
#> 485 2023-04-23 04:00:00 Blankenberge  51.3106    3.1164          391
#> 486 2023-04-23 05:00:00 Blankenberge  51.3106    3.1164          301
#> 487 2023-04-23 06:00:00 Blankenberge  51.3106    3.1164          204
#> 488 2023-04-23 07:00:00 Blankenberge  51.3106    3.1164          109
#> 489 2023-04-23 08:00:00 Blankenberge  51.3106    3.1164           43
#> 490 2023-04-23 09:00:00 Blankenberge  51.3106    3.1164           74
#> 491 2023-04-23 10:00:00 Blankenberge  51.3106    3.1164          119
#> 492 2023-04-23 11:00:00 Blankenberge  51.3106    3.1164          189
#> 493 2023-04-23 12:00:00 Blankenberge  51.3106    3.1164          336
#> 494 2023-04-23 13:00:00 Blankenberge  51.3106    3.1164          453
#> 495 2023-04-23 14:00:00 Blankenberge  51.3106    3.1164          460
#> 496 2023-04-23 15:00:00 Blankenberge  51.3106    3.1164          449
#> 497 2023-04-23 16:00:00 Blankenberge  51.3106    3.1164          403
#> 498 2023-04-23 17:00:00 Blankenberge  51.3106    3.1164          331
#> 499 2023-04-23 18:00:00 Blankenberge  51.3106    3.1164          244
#> 500 2023-04-23 19:00:00 Blankenberge  51.3106    3.1164          155
#> 501 2023-04-23 20:00:00 Blankenberge  51.3106    3.1164           80
#> 502 2023-04-23 21:00:00 Blankenberge  51.3106    3.1164           99
#> 503 2023-04-23 22:00:00 Blankenberge  51.3106    3.1164          156
#> 504 2023-04-23 23:00:00 Blankenberge  51.3106    3.1164          212
#> 505 2023-04-24 00:00:00 Blankenberge  51.3106    3.1164          345
#> 506 2023-04-24 01:00:00 Blankenberge  51.3106    3.1164          473
#> 507 2023-04-24 02:00:00 Blankenberge  51.3106    3.1164          492
#> 508 2023-04-24 03:00:00 Blankenberge  51.3106    3.1164          488
#> 509 2023-04-24 04:00:00 Blankenberge  51.3106    3.1164          450
#> 510 2023-04-24 05:00:00 Blankenberge  51.3106    3.1164          382
#> 511 2023-04-24 06:00:00 Blankenberge  51.3106    3.1164          292
#> 512 2023-04-24 07:00:00 Blankenberge  51.3106    3.1164          203
#> 513 2023-04-24 08:00:00 Blankenberge  51.3106    3.1164          123
#> 514 2023-04-24 09:00:00 Blankenberge  51.3106    3.1164          104
#> 515 2023-04-24 10:00:00 Blankenberge  51.3106    3.1164          149
#> 516 2023-04-24 11:00:00 Blankenberge  51.3106    3.1164          198
#> 517 2023-04-24 12:00:00 Blankenberge  51.3106    3.1164          272
#> 518 2023-04-24 13:00:00 Blankenberge  51.3106    3.1164          418
#> 519 2023-04-24 14:00:00 Blankenberge  51.3106    3.1164          469
#> 520 2023-04-24 15:00:00 Blankenberge  51.3106    3.1164          470
#> 521 2023-04-24 16:00:00 Blankenberge  51.3106    3.1164          446
#> 522 2023-04-24 17:00:00 Blankenberge  51.3106    3.1164          392
#> 523 2023-04-24 18:00:00 Blankenberge  51.3106    3.1164          309
#> 524 2023-04-24 19:00:00 Blankenberge  51.3106    3.1164          220
#> 525 2023-04-24 20:00:00 Blankenberge  51.3106    3.1164          140
#> 526 2023-04-24 21:00:00 Blankenberge  51.3106    3.1164           94
#> 527 2023-04-24 22:00:00 Blankenberge  51.3106    3.1164          137
#> 528 2023-04-24 23:00:00 Blankenberge  51.3106    3.1164          188
#> 529 2023-04-25 00:00:00 Blankenberge  51.3106    3.1164          259
#> 530 2023-04-25 01:00:00 Blankenberge  51.3106    3.1164          385
#> 531 2023-04-25 02:00:00 Blankenberge  51.3106    3.1164          466
#> 532 2023-04-25 03:00:00 Blankenberge  51.3106    3.1164          466
#> 533 2023-04-25 04:00:00 Blankenberge  51.3106    3.1164          454
#> 534 2023-04-25 05:00:00 Blankenberge  51.3106    3.1164          406
#> 535 2023-04-25 06:00:00 Blankenberge  51.3106    3.1164          327
#> 536 2023-04-25 07:00:00 Blankenberge  51.3106    3.1164          235
#> 537 2023-04-25 08:00:00 Blankenberge  51.3106    3.1164          150
#> 538 2023-04-25 09:00:00 Blankenberge  51.3106    3.1164           97
#> 539 2023-04-25 10:00:00 Blankenberge  51.3106    3.1164          125
#> 540 2023-04-25 11:00:00 Blankenberge  51.3106    3.1164          161
#> 541 2023-04-25 12:00:00 Blankenberge  51.3106    3.1164          210
#> 542 2023-04-25 13:00:00 Blankenberge  51.3106    3.1164          298
#> 543 2023-04-25 14:00:00 Blankenberge  51.3106    3.1164          407
#> 544 2023-04-25 15:00:00 Blankenberge  51.3106    3.1164          427
#> 545 2023-04-25 16:00:00 Blankenberge  51.3106    3.1164          420
#> 546 2023-04-25 17:00:00 Blankenberge  51.3106    3.1164          382
#> 547 2023-04-25 18:00:00 Blankenberge  51.3106    3.1164          320
#> 548 2023-04-25 19:00:00 Blankenberge  51.3106    3.1164          235
#> 549 2023-04-25 20:00:00 Blankenberge  51.3106    3.1164          154
#> 550 2023-04-25 21:00:00 Blankenberge  51.3106    3.1164           87
#> 551 2023-04-25 22:00:00 Blankenberge  51.3106    3.1164           67
#> 552 2023-04-25 23:00:00 Blankenberge  51.3106    3.1164          109
#> 553 2023-04-26 00:00:00 Blankenberge  51.3106    3.1164          168
#> 554 2023-04-26 01:00:00 Blankenberge  51.3106    3.1164          264
#> 555 2023-04-26 02:00:00 Blankenberge  51.3106    3.1164          379
#> 556 2023-04-26 03:00:00 Blankenberge  51.3106    3.1164          416
#> 557 2023-04-26 04:00:00 Blankenberge  51.3106    3.1164          415
#> 558 2023-04-26 05:00:00 Blankenberge  51.3106    3.1164          395
#> 559 2023-04-26 06:00:00 Blankenberge  51.3106    3.1164          343
#> 560 2023-04-26 07:00:00 Blankenberge  51.3106    3.1164          268
#> 561 2023-04-26 08:00:00 Blankenberge  51.3106    3.1164          193
#> 562 2023-04-26 09:00:00 Blankenberge  51.3106    3.1164          124
#> 563 2023-04-26 10:00:00 Blankenberge  51.3106    3.1164           91
#> 564 2023-04-26 11:00:00 Blankenberge  51.3106    3.1164          122
#> 565 2023-04-26 12:00:00 Blankenberge  51.3106    3.1164          167
#> 566 2023-04-26 13:00:00 Blankenberge  51.3106    3.1164          230
#> 567 2023-04-26 14:00:00 Blankenberge  51.3106    3.1164          327
#> 568 2023-04-26 15:00:00 Blankenberge  51.3106    3.1164          386
#> 569 2023-04-26 16:00:00 Blankenberge  51.3106    3.1164          392
#> 570 2023-04-26 17:00:00 Blankenberge  51.3106    3.1164          379
#> 571 2023-04-26 18:00:00 Blankenberge  51.3106    3.1164          338
#> 572 2023-04-26 19:00:00 Blankenberge  51.3106    3.1164          276
#> 573 2023-04-26 20:00:00 Blankenberge  51.3106    3.1164          204
#> 574 2023-04-26 21:00:00 Blankenberge  51.3106    3.1164          134
#> 575 2023-04-26 22:00:00 Blankenberge  51.3106    3.1164           84
#> 576 2023-04-26 23:00:00 Blankenberge  51.3106    3.1164           82
#> 577 2023-04-27 00:00:00 Blankenberge  51.3106    3.1164          125
#> 578 2023-04-27 01:00:00 Blankenberge  51.3106    3.1164          198
#> 579 2023-04-27 02:00:00 Blankenberge  51.3106    3.1164          300
#> 580 2023-04-27 03:00:00 Blankenberge  51.3106    3.1164          371
#> 581 2023-04-27 04:00:00 Blankenberge  51.3106    3.1164          390
#> 582 2023-04-27 05:00:00 Blankenberge  51.3106    3.1164          388
#> 583 2023-04-27 06:00:00 Blankenberge  51.3106    3.1164          366
#> 584 2023-04-27 07:00:00 Blankenberge  51.3106    3.1164          311
#> 585 2023-04-27 08:00:00 Blankenberge  51.3106    3.1164          249
#> 586 2023-04-27 09:00:00 Blankenberge  51.3106    3.1164          183
#> 587 2023-04-27 10:00:00 Blankenberge  51.3106    3.1164          130
#> 588 2023-04-27 11:00:00 Blankenberge  51.3106    3.1164          109
#> 589 2023-04-27 12:00:00 Blankenberge  51.3106    3.1164          137
#> 590 2023-04-27 13:00:00 Blankenberge  51.3106    3.1164          184
#> 591 2023-04-27 14:00:00 Blankenberge  51.3106    3.1164          244
#> 592 2023-04-27 15:00:00 Blankenberge  51.3106    3.1164          318
#> 593 2023-04-27 16:00:00 Blankenberge  51.3106    3.1164          353
#> 594 2023-04-27 17:00:00 Blankenberge  51.3106    3.1164          356
#> 595 2023-04-27 18:00:00 Blankenberge  51.3106    3.1164          341
#> 596 2023-04-27 19:00:00 Blankenberge  51.3106    3.1164          307
#> 597 2023-04-27 20:00:00 Blankenberge  51.3106    3.1164          253
#> 598 2023-04-27 21:00:00 Blankenberge  51.3106    3.1164          195
#> 599 2023-04-27 22:00:00 Blankenberge  51.3106    3.1164          140
#> 600 2023-04-27 23:00:00 Blankenberge  51.3106    3.1164           98
#> 601 2023-04-28 00:00:00 Blankenberge  51.3106    3.1164           90
#> 602 2023-04-28 01:00:00 Blankenberge  51.3106    3.1164          134
#> 603 2023-04-28 02:00:00 Blankenberge  51.3106    3.1164          218
#> 604 2023-04-28 03:00:00 Blankenberge  51.3106    3.1164          297
#> 605 2023-04-28 04:00:00 Blankenberge  51.3106    3.1164          349
#> 606 2023-04-28 05:00:00 Blankenberge  51.3106    3.1164          365
#> 607 2023-04-28 06:00:00 Blankenberge  51.3106    3.1164          364
#> 608 2023-04-28 07:00:00 Blankenberge  51.3106    3.1164          347
#> 609 2023-04-28 08:00:00 Blankenberge  51.3106    3.1164          310
#> 610 2023-04-28 09:00:00 Blankenberge  51.3106    3.1164          267
#> 611 2023-04-28 10:00:00 Blankenberge  51.3106    3.1164          217
#> 612 2023-04-28 11:00:00 Blankenberge  51.3106    3.1164          173
#> 613 2023-04-28 12:00:00 Blankenberge  51.3106    3.1164          144
#> 614 2023-04-28 13:00:00 Blankenberge  51.3106    3.1164          168
#> 615 2023-04-28 14:00:00 Blankenberge  51.3106    3.1164          205
#> 616 2023-04-28 15:00:00 Blankenberge  51.3106    3.1164          263
#> 617 2023-04-28 16:00:00 Blankenberge  51.3106    3.1164          317
#> 618 2023-04-28 17:00:00 Blankenberge  51.3106    3.1164          345
#> 619 2023-04-28 18:00:00 Blankenberge  51.3106    3.1164          349
#> 620 2023-04-28 19:00:00 Blankenberge  51.3106    3.1164          345
#> 621 2023-04-28 20:00:00 Blankenberge  51.3106    3.1164          321
#> 622 2023-04-28 21:00:00 Blankenberge  51.3106    3.1164          287
#> 623 2023-04-28 22:00:00 Blankenberge  51.3106    3.1164          240
#> 624 2023-04-28 23:00:00 Blankenberge  51.3106    3.1164          188
#> 625 2023-04-29 00:00:00 Blankenberge  51.3106    3.1164          146
#> 626 2023-04-29 01:00:00 Blankenberge  51.3106    3.1164          126
#> 627 2023-04-29 02:00:00 Blankenberge  51.3106    3.1164          155
#> 628 2023-04-29 03:00:00 Blankenberge  51.3106    3.1164          212
#> 629 2023-04-29 04:00:00 Blankenberge  51.3106    3.1164          278
#> 630 2023-04-29 05:00:00 Blankenberge  51.3106    3.1164          331
#> 631 2023-04-29 06:00:00 Blankenberge  51.3106    3.1164          363
#> 632 2023-04-29 07:00:00 Blankenberge  51.3106    3.1164          367
#> 633 2023-04-29 08:00:00 Blankenberge  51.3106    3.1164          363
#> 634 2023-04-29 09:00:00 Blankenberge  51.3106    3.1164          340
#> 635 2023-04-29 10:00:00 Blankenberge  51.3106    3.1164          296
#> 636 2023-04-29 11:00:00 Blankenberge  51.3106    3.1164          241
#> 637 2023-04-29 12:00:00 Blankenberge  51.3106    3.1164          187
#> 638 2023-04-29 13:00:00 Blankenberge  51.3106    3.1164          150
#> 639 2023-04-29 14:00:00 Blankenberge  51.3106    3.1164          142
#> 640 2023-04-29 15:00:00 Blankenberge  51.3106    3.1164          168
#> 641 2023-04-29 16:00:00 Blankenberge  51.3106    3.1164          206
#> 642 2023-04-29 17:00:00 Blankenberge  51.3106    3.1164          253
#> 643 2023-04-29 18:00:00 Blankenberge  51.3106    3.1164          300
#> 644 2023-04-29 19:00:00 Blankenberge  51.3106    3.1164          324
#> 645 2023-04-29 20:00:00 Blankenberge  51.3106    3.1164          326
#> 646 2023-04-29 21:00:00 Blankenberge  51.3106    3.1164          313
#> 647 2023-04-29 22:00:00 Blankenberge  51.3106    3.1164          283
#> 648 2023-04-29 23:00:00 Blankenberge  51.3106    3.1164          239
#> 649 2023-04-30 00:00:00 Blankenberge  51.3106    3.1164          185
#> 650 2023-04-30 01:00:00 Blankenberge  51.3106    3.1164          135
#> 651 2023-04-30 02:00:00 Blankenberge  51.3106    3.1164           97
#> 652 2023-04-30 03:00:00 Blankenberge  51.3106    3.1164          105
#> 653 2023-04-30 04:00:00 Blankenberge  51.3106    3.1164          149
#> 654 2023-04-30 05:00:00 Blankenberge  51.3106    3.1164          211
#> 655 2023-04-30 06:00:00 Blankenberge  51.3106    3.1164          284
#> 656 2023-04-30 07:00:00 Blankenberge  51.3106    3.1164          342
#> 657 2023-04-30 08:00:00 Blankenberge  51.3106    3.1164          360
#> 658 2023-04-30 09:00:00 Blankenberge  51.3106    3.1164          360
#> 659 2023-04-30 10:00:00 Blankenberge  51.3106    3.1164          342
#> 660 2023-04-30 11:00:00 Blankenberge  51.3106    3.1164          302
#> 661 2023-04-30 12:00:00 Blankenberge  51.3106    3.1164          248
#> 662 2023-04-30 13:00:00 Blankenberge  51.3106    3.1164          185
#> 663 2023-04-30 14:00:00 Blankenberge  51.3106    3.1164          132
#> 664 2023-04-30 15:00:00 Blankenberge  51.3106    3.1164          101
#> 665 2023-04-30 16:00:00 Blankenberge  51.3106    3.1164          123
#> 666 2023-04-30 17:00:00 Blankenberge  51.3106    3.1164          167
#> 667 2023-04-30 18:00:00 Blankenberge  51.3106    3.1164          228
#> 668 2023-04-30 19:00:00 Blankenberge  51.3106    3.1164          303
#> 669 2023-04-30 20:00:00 Blankenberge  51.3106    3.1164          349
#> 670 2023-04-30 21:00:00 Blankenberge  51.3106    3.1164          353
#> 671 2023-04-30 22:00:00 Blankenberge  51.3106    3.1164          346
#> 672 2023-04-30 23:00:00 Blankenberge  51.3106    3.1164          313
#> 673 2023-05-01 00:00:00 Blankenberge  51.3106    3.1164          264
#> 674 2023-05-01 01:00:00 Blankenberge  51.3106    3.1164          202
#> 675 2023-05-01 02:00:00 Blankenberge  51.3106    3.1164          140
#> 676 2023-05-01 03:00:00 Blankenberge  51.3106    3.1164           93
#> 677 2023-05-01 04:00:00 Blankenberge  51.3106    3.1164           97
#> 678 2023-05-01 05:00:00 Blankenberge  51.3106    3.1164          144
#> 679 2023-05-01 06:00:00 Blankenberge  51.3106    3.1164          210
#> 680 2023-05-01 07:00:00 Blankenberge  51.3106    3.1164          305
#> 681 2023-05-01 08:00:00 Blankenberge  51.3106    3.1164          375
#> 682 2023-05-01 09:00:00 Blankenberge  51.3106    3.1164          398
#> 683 2023-05-01 10:00:00 Blankenberge  51.3106    3.1164          397
#> 684 2023-05-01 11:00:00 Blankenberge  51.3106    3.1164          378
#> 685 2023-05-01 12:00:00 Blankenberge  51.3106    3.1164          331
#> 686 2023-05-01 13:00:00 Blankenberge  51.3106    3.1164          267
#> 687 2023-05-01 14:00:00 Blankenberge  51.3106    3.1164          198
#> 688 2023-05-01 15:00:00 Blankenberge  51.3106    3.1164          135
#> 689 2023-05-01 16:00:00 Blankenberge  51.3106    3.1164          111
#> 690 2023-05-01 17:00:00 Blankenberge  51.3106    3.1164          145
#> 691 2023-05-01 18:00:00 Blankenberge  51.3106    3.1164          191
#> 692 2023-05-01 19:00:00 Blankenberge  51.3106    3.1164          258
#> 693 2023-05-01 20:00:00 Blankenberge  51.3106    3.1164          355
#> 694 2023-05-01 21:00:00 Blankenberge  51.3106    3.1164          402
#> 695 2023-05-01 22:00:00 Blankenberge  51.3106    3.1164          404
#> 696 2023-05-01 23:00:00 Blankenberge  51.3106    3.1164          388
#> 697 2023-05-02 00:00:00 Blankenberge  51.3106    3.1164          348
#> 698 2023-05-02 01:00:00 Blankenberge  51.3106    3.1164          283
#> 699 2023-05-02 02:00:00 Blankenberge  51.3106    3.1164          208
#> 700 2023-05-02 03:00:00 Blankenberge  51.3106    3.1164          136
#> 701 2023-05-02 04:00:00 Blankenberge  51.3106    3.1164           84
#> 702 2023-05-02 05:00:00 Blankenberge  51.3106    3.1164          106
#> 703 2023-05-02 06:00:00 Blankenberge  51.3106    3.1164          150
#> 704 2023-05-02 07:00:00 Blankenberge  51.3106    3.1164          214
#> 705 2023-05-02 08:00:00 Blankenberge  51.3106    3.1164          320
#> 706 2023-05-02 09:00:00 Blankenberge  51.3106    3.1164          395
#> 707 2023-05-02 10:00:00 Blankenberge  51.3106    3.1164          405
#> 708 2023-05-02 11:00:00 Blankenberge  51.3106    3.1164          399
#> 709 2023-05-02 12:00:00 Blankenberge  51.3106    3.1164          363
#> 710 2023-05-02 13:00:00 Blankenberge  51.3106    3.1164          302
#> 711 2023-05-02 14:00:00 Blankenberge  51.3106    3.1164          221
#> 712 2023-05-02 15:00:00 Blankenberge  51.3106    3.1164          140
#> 713 2023-05-02 16:00:00 Blankenberge  51.3106    3.1164           81
#> 714 2023-05-02 17:00:00 Blankenberge  51.3106    3.1164           80
#> 715 2023-05-02 18:00:00 Blankenberge  51.3106    3.1164          116
#> 716 2023-05-02 19:00:00 Blankenberge  51.3106    3.1164          165
#> 717 2023-05-02 20:00:00 Blankenberge  51.3106    3.1164          248
#> 718 2023-05-02 21:00:00 Blankenberge  51.3106    3.1164          362
#> 719 2023-05-02 22:00:00 Blankenberge  51.3106    3.1164          389
#> 720 2023-05-02 23:00:00 Blankenberge  51.3106    3.1164          388
#> 721 2023-05-03 00:00:00 Blankenberge  51.3106    3.1164          355
#> 722 2023-05-03 01:00:00 Blankenberge  51.3106    3.1164          300
#> 723 2023-05-03 02:00:00 Blankenberge  51.3106    3.1164          221
#> 724 2023-05-03 03:00:00 Blankenberge  51.3106    3.1164          141
#> 725 2023-05-03 04:00:00 Blankenberge  51.3106    3.1164           65
#> 726 2023-05-03 05:00:00 Blankenberge  51.3106    3.1164           34
#> 727 2023-05-03 06:00:00 Blankenberge  51.3106    3.1164           77
#> 728 2023-05-03 07:00:00 Blankenberge  51.3106    3.1164          132
#> 729 2023-05-03 08:00:00 Blankenberge  51.3106    3.1164          183
#> 
#> $par
#> $par$daterange
#> [1] "2023-04-03" "2023-05-04"
#> 
#> $par$type
#> [1] "MVB data"
#> 
#> $par$stations
#> [1] "Blankenberge"
#> 
#> $par$calc
#> [1] "max"
#> 
#> $par$binSize
#> [1] "hour"
#> 
#> $par$parameters
#> [1] "Tide (cmTAW)"
#> 
#>