Retrieve MVB-data from the LifeWatch project

Retrieves the Meetnet Vlaams Banken (MVB) data from the LifeWatch project. Without valid credentials you are only allowed to view the "Tide TAW" parameter for the last 30 days, grouped by day/hour. To get an account, register via the Lifewatch RShiny registration webpage.

getMvbData(
  startdate,
  stopdate,
  parameters,
  stations = NULL,
  by,
  calc,
  usr = NULL,
  pwd = NULL,
  params = FALSE,
  ...
)

Arguments

startdate

Starting date for the query. Without a login, this is restricted to the last month

stopdate

Stopping date for the query. Without a login, this is restricted to the last month

parameters

One (or list) of the $parameter in listMvbStations.

stations

List of stations to be included in the query. See listMvbStations. Use stations="All" or stations=NULL to get all stations.

by

Time grouping for calculation, one of ('day','hour', '10min' ,'none')

calc

Calculation to perform given time grouping, one of ('avg', 'max', 'min', 'none').

usr

Username to connect to database

pwd

Password to connect to database

params

If TRUE, returns a list with the dataset and the query parameters applied in the server side. IF FALSE returns only the data.

...

Reserved for internal use.

Value

Dataframe with the specified MVB data.

Examples

getMvbData(Sys.Date() - 30, Sys.Date() + 1, parameters = 'Tide TAW',
stations = "Nieuwpoort", by = "day", calc = "avg")
#> - Query mode: Post request to OpenCPU server
#> No encoding supplied: defaulting to UTF-8.
#> - Your query:
#>    - daterange: c("2023-10-21", "2023-11-21")
#>    - type: MVB data
#>    - stations: Nieuwpoort
#>    - calc: avg
#>    - binSize: day
#>    - parameters: Tide TAW
#> - Server query:
#>    - daterange: c("2023-10-21", "2023-11-21")
#>    - type: MVB data
#>    - stations: Nieuwpoort
#>    - calc: avg
#>    - binSize: day
#>    - parameters: Tide (cmTAW)
#> Warning: The query applied on the server differ from the parameters you used. 
#>  Hint: You may need an account to fully access the data under moratorium. 
#>  Hint: Request access at: https://rshiny.lifewatch.be/account?p=register
#>          time    station latitude longitude Tide (cmTAW)
#> 1  2023-10-21 Nieuwpoort  51.1506    2.7281     266.2203
#> 2  2023-10-22 Nieuwpoort  51.1506    2.7281     271.0764
#> 3  2023-10-23 Nieuwpoort  51.1506    2.7281     265.5243
#> 4  2023-10-24 Nieuwpoort  51.1506    2.7281     266.2500
#> 5  2023-10-25 Nieuwpoort  51.1506    2.7281     268.1254
#> 6  2023-10-26 Nieuwpoort  51.1506    2.7281     252.9688
#> 7  2023-10-27 Nieuwpoort  51.1506    2.7281     259.9271
#> 8  2023-10-28 Nieuwpoort  51.1506    2.7281     254.1667
#> 9  2023-10-29 Nieuwpoort  51.1506    2.7281     263.8362
#> 10 2023-10-30 Nieuwpoort  51.1506    2.7281     271.6458
#> 11 2023-10-31 Nieuwpoort  51.1506    2.7281     271.5521
#> 12 2023-11-01 Nieuwpoort  51.1506    2.7281     265.9201
#> 13 2023-11-02 Nieuwpoort  51.1506    2.7281     286.1562
#> 14 2023-11-03 Nieuwpoort  51.1506    2.7281     291.0057
#> 15 2023-11-04 Nieuwpoort  51.1506    2.7281     295.0104
#> 16 2023-11-05 Nieuwpoort  51.1506    2.7281     325.9375
#> 17 2023-11-06 Nieuwpoort  51.1506    2.7281     294.6237
#> 18 2023-11-07 Nieuwpoort  51.1506    2.7281     283.7986
#> 19 2023-11-08 Nieuwpoort  51.1506    2.7281     267.9688
#> 20 2023-11-09 Nieuwpoort  51.1506    2.7281     289.7882
#> 21 2023-11-10 Nieuwpoort  51.1506    2.7281     285.2743
#> 22 2023-11-11 Nieuwpoort  51.1506    2.7281     265.8229
#> 23 2023-11-12 Nieuwpoort  51.1506    2.7281     255.8502
#> 24 2023-11-13 Nieuwpoort  51.1506    2.7281     260.1771
#> 25 2023-11-14 Nieuwpoort  51.1506    2.7281     288.0938
#> 26 2023-11-15 Nieuwpoort  51.1506    2.7281     266.0660
#> 27 2023-11-16 Nieuwpoort  51.1506    2.7281     252.3345
#> 28 2023-11-17 Nieuwpoort  51.1506    2.7281     256.6910
#> 29 2023-11-18 Nieuwpoort  51.1506    2.7281     239.7326
#> 30 2023-11-19 Nieuwpoort  51.1506    2.7281     281.6667
#> 31 2023-11-20 Nieuwpoort  51.1506    2.7281     305.7097
getMvbData(Sys.Date() - 30, Sys.Date() + 1, parameters = 'Tide TAW',
stations = "Blankenberge", by = "hour", calc = "max", params = TRUE)
#> - Query mode: Post request to OpenCPU server
#> No encoding supplied: defaulting to UTF-8.
#> - Your query:
#>    - daterange: c("2023-10-21", "2023-11-21")
#>    - type: MVB data
#>    - stations: Blankenberge
#>    - calc: max
#>    - binSize: hour
#>    - parameters: Tide TAW
#> - Server query:
#>    - daterange: c("2023-10-21", "2023-11-21")
#>    - type: MVB data
#>    - stations: Blankenberge
#>    - calc: max
#>    - binSize: hour
#>    - parameters: Tide (cmTAW)
#> Warning: The query applied on the server differ from the parameters you used. 
#>  Hint: You may need an account to fully access the data under moratorium. 
#>  Hint: Request access at: https://rshiny.lifewatch.be/account?p=register
#> $mdf
#>                    time      station latitude longitude Tide (cmTAW)
#> 1   2023-10-21 00:00:00 Blankenberge  51.3106    3.1164          211
#> 2   2023-10-21 01:00:00 Blankenberge  51.3106    3.1164          318
#> 3   2023-10-21 02:00:00 Blankenberge  51.3106    3.1164          412
#> 4   2023-10-21 03:00:00 Blankenberge  51.3106    3.1164          429
#> 5   2023-10-21 04:00:00 Blankenberge  51.3106    3.1164          427
#> 6   2023-10-21 05:00:00 Blankenberge  51.3106    3.1164          399
#> 7   2023-10-21 06:00:00 Blankenberge  51.3106    3.1164          350
#> 8   2023-10-21 07:00:00 Blankenberge  51.3106    3.1164          281
#> 9   2023-10-21 08:00:00 Blankenberge  51.3106    3.1164          204
#> 10  2023-10-21 09:00:00 Blankenberge  51.3106    3.1164          143
#> 11  2023-10-21 10:00:00 Blankenberge  51.3106    3.1164           96
#> 12  2023-10-21 11:00:00 Blankenberge  51.3106    3.1164          116
#> 13  2023-10-21 12:00:00 Blankenberge  51.3106    3.1164          180
#> 14  2023-10-21 13:00:00 Blankenberge  51.3106    3.1164          280
#> 15  2023-10-21 14:00:00 Blankenberge  51.3106    3.1164          373
#> 16  2023-10-21 15:00:00 Blankenberge  51.3106    3.1164          411
#> 17  2023-10-21 16:00:00 Blankenberge  51.3106    3.1164          412
#> 18  2023-10-21 17:00:00 Blankenberge  51.3106    3.1164          403
#> 19  2023-10-21 18:00:00 Blankenberge  51.3106    3.1164          364
#> 20  2023-10-21 19:00:00 Blankenberge  51.3106    3.1164          305
#> 21  2023-10-21 20:00:00 Blankenberge  51.3106    3.1164          243
#> 22  2023-10-21 21:00:00 Blankenberge  51.3106    3.1164          179
#> 23  2023-10-21 22:00:00 Blankenberge  51.3106    3.1164          139
#> 24  2023-10-21 23:00:00 Blankenberge  51.3106    3.1164          131
#> 25  2023-10-22 00:00:00 Blankenberge  51.3106    3.1164          166
#> 26  2023-10-22 01:00:00 Blankenberge  51.3106    3.1164          227
#> 27  2023-10-22 02:00:00 Blankenberge  51.3106    3.1164          311
#> 28  2023-10-22 03:00:00 Blankenberge  51.3106    3.1164          371
#> 29  2023-10-22 04:00:00 Blankenberge  51.3106    3.1164          392
#> 30  2023-10-22 05:00:00 Blankenberge  51.3106    3.1164          390
#> 31  2023-10-22 06:00:00 Blankenberge  51.3106    3.1164          370
#> 32  2023-10-22 07:00:00 Blankenberge  51.3106    3.1164          321
#> 33  2023-10-22 08:00:00 Blankenberge  51.3106    3.1164          261
#> 34  2023-10-22 09:00:00 Blankenberge  51.3106    3.1164          202
#> 35  2023-10-22 10:00:00 Blankenberge  51.3106    3.1164          142
#> 36  2023-10-22 11:00:00 Blankenberge  51.3106    3.1164          104
#> 37  2023-10-22 12:00:00 Blankenberge  51.3106    3.1164          137
#> 38  2023-10-22 13:00:00 Blankenberge  51.3106    3.1164          201
#> 39  2023-10-22 14:00:00 Blankenberge  51.3106    3.1164          287
#> 40  2023-10-22 15:00:00 Blankenberge  51.3106    3.1164          367
#> 41  2023-10-22 16:00:00 Blankenberge  51.3106    3.1164          409
#> 42  2023-10-22 17:00:00 Blankenberge  51.3106    3.1164          417
#> 43  2023-10-22 18:00:00 Blankenberge  51.3106    3.1164          416
#> 44  2023-10-22 19:00:00 Blankenberge  51.3106    3.1164          386
#> 45  2023-10-22 20:00:00 Blankenberge  51.3106    3.1164          338
#> 46  2023-10-22 21:00:00 Blankenberge  51.3106    3.1164          280
#> 47  2023-10-22 22:00:00 Blankenberge  51.3106    3.1164          217
#> 48  2023-10-22 23:00:00 Blankenberge  51.3106    3.1164          169
#> 49  2023-10-23 00:00:00 Blankenberge  51.3106    3.1164          157
#> 50  2023-10-23 01:00:00 Blankenberge  51.3106    3.1164          191
#> 51  2023-10-23 02:00:00 Blankenberge  51.3106    3.1164          232
#> 52  2023-10-23 03:00:00 Blankenberge  51.3106    3.1164          290
#> 53  2023-10-23 04:00:00 Blankenberge  51.3106    3.1164          350
#> 54  2023-10-23 05:00:00 Blankenberge  51.3106    3.1164          381
#> 55  2023-10-23 06:00:00 Blankenberge  51.3106    3.1164          383
#> 56  2023-10-23 07:00:00 Blankenberge  51.3106    3.1164          368
#> 57  2023-10-23 08:00:00 Blankenberge  51.3106    3.1164          332
#> 58  2023-10-23 09:00:00 Blankenberge  51.3106    3.1164          278
#> 59  2023-10-23 10:00:00 Blankenberge  51.3106    3.1164          219
#> 60  2023-10-23 11:00:00 Blankenberge  51.3106    3.1164          162
#> 61  2023-10-23 12:00:00 Blankenberge  51.3106    3.1164          121
#> 62  2023-10-23 13:00:00 Blankenberge  51.3106    3.1164          111
#> 63  2023-10-23 14:00:00 Blankenberge  51.3106    3.1164          160
#> 64  2023-10-23 15:00:00 Blankenberge  51.3106    3.1164          217
#> 65  2023-10-23 16:00:00 Blankenberge  51.3106    3.1164          297
#> 66  2023-10-23 17:00:00 Blankenberge  51.3106    3.1164          366
#> 67  2023-10-23 18:00:00 Blankenberge  51.3106    3.1164          399
#> 68  2023-10-23 19:00:00 Blankenberge  51.3106    3.1164          399
#> 69  2023-10-23 20:00:00 Blankenberge  51.3106    3.1164          387
#> 70  2023-10-23 21:00:00 Blankenberge  51.3106    3.1164          350
#> 71  2023-10-23 22:00:00 Blankenberge  51.3106    3.1164          296
#> 72  2023-10-23 23:00:00 Blankenberge  51.3106    3.1164          235
#> 73  2023-10-24 00:00:00 Blankenberge  51.3106    3.1164          179
#> 74  2023-10-24 01:00:00 Blankenberge  51.3106    3.1164          137
#> 75  2023-10-24 02:00:00 Blankenberge  51.3106    3.1164          142
#> 76  2023-10-24 03:00:00 Blankenberge  51.3106    3.1164          186
#> 77  2023-10-24 04:00:00 Blankenberge  51.3106    3.1164          238
#> 78  2023-10-24 05:00:00 Blankenberge  51.3106    3.1164          305
#> 79  2023-10-24 06:00:00 Blankenberge  51.3106    3.1164          367
#> 80  2023-10-24 07:00:00 Blankenberge  51.3106    3.1164          389
#> 81  2023-10-24 08:00:00 Blankenberge  51.3106    3.1164          386
#> 82  2023-10-24 09:00:00 Blankenberge  51.3106    3.1164          364
#> 83  2023-10-24 10:00:00 Blankenberge  51.3106    3.1164          326
#> 84  2023-10-24 11:00:00 Blankenberge  51.3106    3.1164          270
#> 85  2023-10-24 12:00:00 Blankenberge  51.3106    3.1164          206
#> 86  2023-10-24 13:00:00 Blankenberge  51.3106    3.1164          150
#> 87  2023-10-24 14:00:00 Blankenberge  51.3106    3.1164          107
#> 88  2023-10-24 15:00:00 Blankenberge  51.3106    3.1164          134
#> 89  2023-10-24 16:00:00 Blankenberge  51.3106    3.1164          182
#> 90  2023-10-24 17:00:00 Blankenberge  51.3106    3.1164          262
#> 91  2023-10-24 18:00:00 Blankenberge  51.3106    3.1164          357
#> 92  2023-10-24 19:00:00 Blankenberge  51.3106    3.1164          423
#> 93  2023-10-24 20:00:00 Blankenberge  51.3106    3.1164          438
#> 94  2023-10-24 21:00:00 Blankenberge  51.3106    3.1164          434
#> 95  2023-10-24 22:00:00 Blankenberge  51.3106    3.1164          410
#> 96  2023-10-24 23:00:00 Blankenberge  51.3106    3.1164          359
#> 97  2023-10-25 00:00:00 Blankenberge  51.3106    3.1164          291
#> 98  2023-10-25 01:00:00 Blankenberge  51.3106    3.1164          213
#> 99  2023-10-25 02:00:00 Blankenberge  51.3106    3.1164          151
#> 100 2023-10-25 03:00:00 Blankenberge  51.3106    3.1164          136
#> 101 2023-10-25 04:00:00 Blankenberge  51.3106    3.1164          171
#> 102 2023-10-25 05:00:00 Blankenberge  51.3106    3.1164          214
#> 103 2023-10-25 06:00:00 Blankenberge  51.3106    3.1164          281
#> 104 2023-10-25 07:00:00 Blankenberge  51.3106    3.1164          376
#> 105 2023-10-25 08:00:00 Blankenberge  51.3106    3.1164          426
#> 106 2023-10-25 09:00:00 Blankenberge  51.3106    3.1164          428
#> 107 2023-10-25 10:00:00 Blankenberge  51.3106    3.1164          411
#> 108 2023-10-25 11:00:00 Blankenberge  51.3106    3.1164          369
#> 109 2023-10-25 12:00:00 Blankenberge  51.3106    3.1164          304
#> 110 2023-10-25 13:00:00 Blankenberge  51.3106    3.1164          227
#> 111 2023-10-25 14:00:00 Blankenberge  51.3106    3.1164          157
#> 112 2023-10-25 15:00:00 Blankenberge  51.3106    3.1164           97
#> 113 2023-10-25 16:00:00 Blankenberge  51.3106    3.1164          101
#> 114 2023-10-25 17:00:00 Blankenberge  51.3106    3.1164          155
#> 115 2023-10-25 18:00:00 Blankenberge  51.3106    3.1164          222
#> 116 2023-10-25 19:00:00 Blankenberge  51.3106    3.1164          338
#> 117 2023-10-25 20:00:00 Blankenberge  51.3106    3.1164          444
#> 118 2023-10-25 21:00:00 Blankenberge  51.3106    3.1164          468
#> 119 2023-10-25 22:00:00 Blankenberge  51.3106    3.1164          466
#> 120 2023-10-25 23:00:00 Blankenberge  51.3106    3.1164          438
#> 121 2023-10-26 00:00:00 Blankenberge  51.3106    3.1164          378
#> 122 2023-10-26 01:00:00 Blankenberge  51.3106    3.1164          295
#> 123 2023-10-26 02:00:00 Blankenberge  51.3106    3.1164          209
#> 124 2023-10-26 03:00:00 Blankenberge  51.3106    3.1164          133
#> 125 2023-10-26 04:00:00 Blankenberge  51.3106    3.1164          106
#> 126 2023-10-26 05:00:00 Blankenberge  51.3106    3.1164          144
#> 127 2023-10-26 06:00:00 Blankenberge  51.3106    3.1164          193
#> 128 2023-10-26 07:00:00 Blankenberge  51.3106    3.1164          263
#> 129 2023-10-26 08:00:00 Blankenberge  51.3106    3.1164          397
#> 130 2023-10-26 09:00:00 Blankenberge  51.3106    3.1164          448
#> 131 2023-10-26 10:00:00 Blankenberge  51.3106    3.1164          450
#> 132 2023-10-26 11:00:00 Blankenberge  51.3106    3.1164          422
#> 133 2023-10-26 12:00:00 Blankenberge  51.3106    3.1164          371
#> 134 2023-10-26 13:00:00 Blankenberge  51.3106    3.1164          296
#> 135 2023-10-26 14:00:00 Blankenberge  51.3106    3.1164          215
#> 136 2023-10-26 15:00:00 Blankenberge  51.3106    3.1164          131
#> 137 2023-10-26 16:00:00 Blankenberge  51.3106    3.1164           60
#> 138 2023-10-26 17:00:00 Blankenberge  51.3106    3.1164           65
#> 139 2023-10-26 18:00:00 Blankenberge  51.3106    3.1164          121
#> 140 2023-10-26 19:00:00 Blankenberge  51.3106    3.1164          198
#> 141 2023-10-26 20:00:00 Blankenberge  51.3106    3.1164          348
#> 142 2023-10-26 21:00:00 Blankenberge  51.3106    3.1164          460
#> 143 2023-10-26 22:00:00 Blankenberge  51.3106    3.1164          476
#> 144 2023-10-26 23:00:00 Blankenberge  51.3106    3.1164          471
#> 145 2023-10-27 00:00:00 Blankenberge  51.3106    3.1164          437
#> 146 2023-10-27 01:00:00 Blankenberge  51.3106    3.1164          367
#> 147 2023-10-27 02:00:00 Blankenberge  51.3106    3.1164          280
#> 148 2023-10-27 03:00:00 Blankenberge  51.3106    3.1164          189
#> 149 2023-10-27 04:00:00 Blankenberge  51.3106    3.1164          109
#> 150 2023-10-27 05:00:00 Blankenberge  51.3106    3.1164          112
#> 151 2023-10-27 06:00:00 Blankenberge  51.3106    3.1164          157
#> 152 2023-10-27 07:00:00 Blankenberge  51.3106    3.1164          208
#> 153 2023-10-27 08:00:00 Blankenberge  51.3106    3.1164          302
#> 154 2023-10-27 09:00:00 Blankenberge  51.3106    3.1164          448
#> 155 2023-10-27 10:00:00 Blankenberge  51.3106    3.1164          485
#> 156 2023-10-27 11:00:00 Blankenberge  51.3106    3.1164          482
#> 157 2023-10-27 12:00:00 Blankenberge  51.3106    3.1164          454
#> 158 2023-10-27 13:00:00 Blankenberge  51.3106    3.1164          386
#> 159 2023-10-27 14:00:00 Blankenberge  51.3106    3.1164          298
#> 160 2023-10-27 15:00:00 Blankenberge  51.3106    3.1164          206
#> 161 2023-10-27 16:00:00 Blankenberge  51.3106    3.1164          112
#> 162 2023-10-27 17:00:00 Blankenberge  51.3106    3.1164           42
#> 163 2023-10-27 18:00:00 Blankenberge  51.3106    3.1164           84
#> 164 2023-10-27 19:00:00 Blankenberge  51.3106    3.1164          142
#> 165 2023-10-27 20:00:00 Blankenberge  51.3106    3.1164          231
#> 166 2023-10-27 21:00:00 Blankenberge  51.3106    3.1164          407
#> 167 2023-10-27 22:00:00 Blankenberge  51.3106    3.1164          493
#> 168 2023-10-27 23:00:00 Blankenberge  51.3106    3.1164          499
#> 169 2023-10-28 00:00:00 Blankenberge  51.3106    3.1164          484
#> 170 2023-10-28 01:00:00 Blankenberge  51.3106    3.1164          436
#> 171 2023-10-28 02:00:00 Blankenberge  51.3106    3.1164          357
#> 172 2023-10-28 03:00:00 Blankenberge  51.3106    3.1164          263
#> 173 2023-10-28 04:00:00 Blankenberge  51.3106    3.1164          168
#> 174 2023-10-28 05:00:00 Blankenberge  51.3106    3.1164           87
#> 175 2023-10-28 06:00:00 Blankenberge  51.3106    3.1164          114
#> 176 2023-10-28 07:00:00 Blankenberge  51.3106    3.1164          161
#> 177 2023-10-28 08:00:00 Blankenberge  51.3106    3.1164          219
#> 178 2023-10-28 09:00:00 Blankenberge  51.3106    3.1164          344
#> 179 2023-10-28 10:00:00 Blankenberge  51.3106    3.1164          478
#> 180 2023-10-28 11:00:00 Blankenberge  51.3106    3.1164          497
#> 181 2023-10-28 12:00:00 Blankenberge  51.3106    3.1164          483
#> 182 2023-10-28 13:00:00 Blankenberge  51.3106    3.1164          439
#> 183 2023-10-28 14:00:00 Blankenberge  51.3106    3.1164          362
#> 184 2023-10-28 15:00:00 Blankenberge  51.3106    3.1164          269
#> 185 2023-10-28 16:00:00 Blankenberge  51.3106    3.1164          164
#> 186 2023-10-28 17:00:00 Blankenberge  51.3106    3.1164           70
#> 187 2023-10-28 18:00:00 Blankenberge  51.3106    3.1164           17
#> 188 2023-10-28 19:00:00 Blankenberge  51.3106    3.1164           81
#> 189 2023-10-28 20:00:00 Blankenberge  51.3106    3.1164          139
#> 190 2023-10-28 21:00:00 Blankenberge  51.3106    3.1164          258
#> 191 2023-10-28 22:00:00 Blankenberge  51.3106    3.1164          449
#> 192 2023-10-28 23:00:00 Blankenberge  51.3106    3.1164          496
#> 193 2023-10-29 00:00:00 Blankenberge  51.3106    3.1164          499
#> 194 2023-10-29 01:00:00 Blankenberge  51.3106    3.1164          467
#> 195 2023-10-29 02:00:00 Blankenberge  51.3106    3.1164          414
#> 196 2023-10-29 03:00:00 Blankenberge  51.3106    3.1164          326
#> 197 2023-10-29 04:00:00 Blankenberge  51.3106    3.1164          238
#> 198 2023-10-29 05:00:00 Blankenberge  51.3106    3.1164          149
#> 199 2023-10-29 06:00:00 Blankenberge  51.3106    3.1164           72
#> 200 2023-10-29 07:00:00 Blankenberge  51.3106    3.1164          127
#> 201 2023-10-29 08:00:00 Blankenberge  51.3106    3.1164          176
#> 202 2023-10-29 09:00:00 Blankenberge  51.3106    3.1164          243
#> 203 2023-10-29 10:00:00 Blankenberge  51.3106    3.1164          405
#> 204 2023-10-29 11:00:00 Blankenberge  51.3106    3.1164          494
#> 205 2023-10-29 12:00:00 Blankenberge  51.3106    3.1164          499
#> 206 2023-10-29 13:00:00 Blankenberge  51.3106    3.1164          477
#> 207 2023-10-29 14:00:00 Blankenberge  51.3106    3.1164          420
#> 208 2023-10-29 15:00:00 Blankenberge  51.3106    3.1164          349
#> 209 2023-10-29 16:00:00 Blankenberge  51.3106    3.1164          250
#> 210 2023-10-29 17:00:00 Blankenberge  51.3106    3.1164          158
#> 211 2023-10-29 18:00:00 Blankenberge  51.3106    3.1164           75
#> 212 2023-10-29 19:00:00 Blankenberge  51.3106    3.1164           78
#> 213 2023-10-29 20:00:00 Blankenberge  51.3106    3.1164          130
#> 214 2023-10-29 21:00:00 Blankenberge  51.3106    3.1164          204
#> 215 2023-10-29 22:00:00 Blankenberge  51.3106    3.1164          346
#> 216 2023-10-29 23:00:00 Blankenberge  51.3106    3.1164          495
#> 217 2023-10-30 00:00:00 Blankenberge  51.3106    3.1164          517
#> 218 2023-10-30 01:00:00 Blankenberge  51.3106    3.1164          510
#> 219 2023-10-30 02:00:00 Blankenberge  51.3106    3.1164          469
#> 220 2023-10-30 03:00:00 Blankenberge  51.3106    3.1164          395
#> 221 2023-10-30 04:00:00 Blankenberge  51.3106    3.1164          301
#> 222 2023-10-30 05:00:00 Blankenberge  51.3106    3.1164          205
#> 223 2023-10-30 06:00:00 Blankenberge  51.3106    3.1164          116
#> 224 2023-10-30 07:00:00 Blankenberge  51.3106    3.1164           99
#> 225 2023-10-30 08:00:00 Blankenberge  51.3106    3.1164          152
#> 226 2023-10-30 09:00:00 Blankenberge  51.3106    3.1164          204
#> 227 2023-10-30 10:00:00 Blankenberge  51.3106    3.1164          290
#> 228 2023-10-30 11:00:00 Blankenberge  51.3106    3.1164          467
#> 229 2023-10-30 12:00:00 Blankenberge  51.3106    3.1164          520
#> 230 2023-10-30 13:00:00 Blankenberge  51.3106    3.1164          520
#> 231 2023-10-30 14:00:00 Blankenberge  51.3106    3.1164          483
#> 232 2023-10-30 15:00:00 Blankenberge  51.3106    3.1164          419
#> 233 2023-10-30 16:00:00 Blankenberge  51.3106    3.1164          323
#> 234 2023-10-30 17:00:00 Blankenberge  51.3106    3.1164          222
#> 235 2023-10-30 18:00:00 Blankenberge  51.3106    3.1164          123
#> 236 2023-10-30 19:00:00 Blankenberge  51.3106    3.1164           55
#> 237 2023-10-30 20:00:00 Blankenberge  51.3106    3.1164          108
#> 238 2023-10-30 21:00:00 Blankenberge  51.3106    3.1164          162
#> 239 2023-10-30 22:00:00 Blankenberge  51.3106    3.1164          231
#> 240 2023-10-30 23:00:00 Blankenberge  51.3106    3.1164          390
#> 241 2023-10-31 00:00:00 Blankenberge  51.3106    3.1164          509
#> 242 2023-10-31 01:00:00 Blankenberge  51.3106    3.1164          515
#> 243 2023-10-31 02:00:00 Blankenberge  51.3106    3.1164          500
#> 244 2023-10-31 03:00:00 Blankenberge  51.3106    3.1164          447
#> 245 2023-10-31 04:00:00 Blankenberge  51.3106    3.1164          361
#> 246 2023-10-31 05:00:00 Blankenberge  51.3106    3.1164          259
#> 247 2023-10-31 06:00:00 Blankenberge  51.3106    3.1164          165
#> 248 2023-10-31 07:00:00 Blankenberge  51.3106    3.1164           86
#> 249 2023-10-31 08:00:00 Blankenberge  51.3106    3.1164          110
#> 250 2023-10-31 09:00:00 Blankenberge  51.3106    3.1164          165
#> 251 2023-10-31 10:00:00 Blankenberge  51.3106    3.1164          223
#> 252 2023-10-31 11:00:00 Blankenberge  51.3106    3.1164          346
#> 253 2023-10-31 12:00:00 Blankenberge  51.3106    3.1164          498
#> 254 2023-10-31 13:00:00 Blankenberge  51.3106    3.1164          518
#> 255 2023-10-31 14:00:00 Blankenberge  51.3106    3.1164          513
#> 256 2023-10-31 15:00:00 Blankenberge  51.3106    3.1164          464
#> 257 2023-10-31 16:00:00 Blankenberge  51.3106    3.1164          391
#> 258 2023-10-31 17:00:00 Blankenberge  51.3106    3.1164          294
#> 259 2023-10-31 18:00:00 Blankenberge  51.3106    3.1164          195
#> 260 2023-10-31 19:00:00 Blankenberge  51.3106    3.1164          101
#> 261 2023-10-31 20:00:00 Blankenberge  51.3106    3.1164           65
#> 262 2023-10-31 21:00:00 Blankenberge  51.3106    3.1164          122
#> 263 2023-10-31 22:00:00 Blankenberge  51.3106    3.1164          172
#> 264 2023-10-31 23:00:00 Blankenberge  51.3106    3.1164          256
#> 265 2023-11-01 00:00:00 Blankenberge  51.3106    3.1164          441
#> 266 2023-11-01 01:00:00 Blankenberge  51.3106    3.1164          489
#> 267 2023-11-01 02:00:00 Blankenberge  51.3106    3.1164          489
#> 268 2023-11-01 03:00:00 Blankenberge  51.3106    3.1164          461
#> 269 2023-11-01 04:00:00 Blankenberge  51.3106    3.1164          402
#> 270 2023-11-01 05:00:00 Blankenberge  51.3106    3.1164          311
#> 271 2023-11-01 06:00:00 Blankenberge  51.3106    3.1164          219
#> 272 2023-11-01 07:00:00 Blankenberge  51.3106    3.1164          130
#> 273 2023-11-01 08:00:00 Blankenberge  51.3106    3.1164           66
#> 274 2023-11-01 09:00:00 Blankenberge  51.3106    3.1164          107
#> 275 2023-11-01 10:00:00 Blankenberge  51.3106    3.1164          169
#> 276 2023-11-01 11:00:00 Blankenberge  51.3106    3.1164          257
#> 277 2023-11-01 12:00:00 Blankenberge  51.3106    3.1164          414
#> 278 2023-11-01 13:00:00 Blankenberge  51.3106    3.1164          480
#> 279 2023-11-01 14:00:00 Blankenberge  51.3106    3.1164          485
#> 280 2023-11-01 15:00:00 Blankenberge  51.3106    3.1164          466
#> 281 2023-11-01 16:00:00 Blankenberge  51.3106    3.1164          408
#> 282 2023-11-01 17:00:00 Blankenberge  51.3106    3.1164          337
#> 283 2023-11-01 18:00:00 Blankenberge  51.3106    3.1164          248
#> 284 2023-11-01 19:00:00 Blankenberge  51.3106    3.1164          161
#> 285 2023-11-01 20:00:00 Blankenberge  51.3106    3.1164           83
#> 286 2023-11-01 21:00:00 Blankenberge  51.3106    3.1164           87
#> 287 2023-11-01 22:00:00 Blankenberge  51.3106    3.1164          133
#> 288 2023-11-01 23:00:00 Blankenberge  51.3106    3.1164          201
#> 289 2023-11-02 00:00:00 Blankenberge  51.3106    3.1164          321
#> 290 2023-11-02 01:00:00 Blankenberge  51.3106    3.1164          422
#> 291 2023-11-02 02:00:00 Blankenberge  51.3106    3.1164          438
#> 292 2023-11-02 03:00:00 Blankenberge  51.3106    3.1164          429
#> 293 2023-11-02 04:00:00 Blankenberge  51.3106    3.1164          395
#> 294 2023-11-02 05:00:00 Blankenberge  51.3106    3.1164          324
#> 295 2023-11-02 06:00:00 Blankenberge  51.3106    3.1164          255
#> 296 2023-11-02 07:00:00 Blankenberge  51.3106    3.1164          177
#> 297 2023-11-02 08:00:00 Blankenberge  51.3106    3.1164          123
#> 298 2023-11-02 09:00:00 Blankenberge  51.3106    3.1164          107
#> 299 2023-11-02 10:00:00 Blankenberge  51.3106    3.1164          159
#> 300 2023-11-02 11:00:00 Blankenberge  51.3106    3.1164          245
#> 301 2023-11-02 12:00:00 Blankenberge  51.3106    3.1164          354
#> 302 2023-11-02 13:00:00 Blankenberge  51.3106    3.1164          432
#> 303 2023-11-02 14:00:00 Blankenberge  51.3106    3.1164          456
#> 304 2023-11-02 15:00:00 Blankenberge  51.3106    3.1164          459
#> 305 2023-11-02 16:00:00 Blankenberge  51.3106    3.1164          436
#> 306 2023-11-02 17:00:00 Blankenberge  51.3106    3.1164          332
#> 307 2023-11-02 18:00:00 Blankenberge  51.3106    3.1164          323
#> 308 2023-11-02 19:00:00 Blankenberge  51.3106    3.1164          258
#> 309 2023-11-02 20:00:00 Blankenberge  51.3106    3.1164          197
#> 310 2023-11-02 21:00:00 Blankenberge  51.3106    3.1164          183
#> 311 2023-11-02 22:00:00 Blankenberge  51.3106    3.1164          215
#> 312 2023-11-02 23:00:00 Blankenberge  51.3106    3.1164          271
#> 313 2023-11-03 00:00:00 Blankenberge  51.3106    3.1164          325
#> 314 2023-11-03 06:00:00 Blankenberge  51.3106    3.1164          343
#> 315 2023-11-03 07:00:00 Blankenberge  51.3106    3.1164          285
#> 316 2023-11-03 08:00:00 Blankenberge  51.3106    3.1164          208
#> 317 2023-11-03 12:00:00 Blankenberge  51.3106    3.1164          266
#> 318 2023-11-03 13:00:00 Blankenberge  51.3106    3.1164          379
#> 319 2023-11-03 14:00:00 Blankenberge  51.3106    3.1164          454
#> 320 2023-11-03 15:00:00 Blankenberge  51.3106    3.1164          469
#> 321 2023-11-03 16:00:00 Blankenberge  51.3106    3.1164          461
#> 322 2023-11-03 17:00:00 Blankenberge  51.3106    3.1164          426
#> 323 2023-11-03 18:00:00 Blankenberge  51.3106    3.1164          364
#> 324 2023-11-03 19:00:00 Blankenberge  51.3106    3.1164          291
#> 325 2023-11-03 20:00:00 Blankenberge  51.3106    3.1164          217
#> 326 2023-11-03 21:00:00 Blankenberge  51.3106    3.1164          162
#> 327 2023-11-03 22:00:00 Blankenberge  51.3106    3.1164          144
#> 328 2023-11-03 23:00:00 Blankenberge  51.3106    3.1164          189
#> 329 2023-11-04 00:00:00 Blankenberge  51.3106    3.1164          243
#> 330 2023-11-04 01:00:00 Blankenberge  51.3106    3.1164          321
#> 331 2023-11-04 02:00:00 Blankenberge  51.3106    3.1164          406
#> 332 2023-11-04 03:00:00 Blankenberge  51.3106    3.1164          434
#> 333 2023-11-04 04:00:00 Blankenberge  51.3106    3.1164          434
#> 334 2023-11-04 05:00:00 Blankenberge  51.3106    3.1164          410
#> 335 2023-11-04 06:00:00 Blankenberge  51.3106    3.1164          362
#> 336 2023-11-04 07:00:00 Blankenberge  51.3106    3.1164          294
#> 337 2023-11-04 08:00:00 Blankenberge  51.3106    3.1164          218
#> 338 2023-11-04 09:00:00 Blankenberge  51.3106    3.1164          156
#> 339 2023-11-04 10:00:00 Blankenberge  51.3106    3.1164          114
#> 340 2023-11-04 11:00:00 Blankenberge  51.3106    3.1164          117
#> 341 2023-11-04 12:00:00 Blankenberge  51.3106    3.1164          204
#> 342 2023-11-04 13:00:00 Blankenberge  51.3106    3.1164          326
#> 343 2023-11-04 14:00:00 Blankenberge  51.3106    3.1164          417
#> 344 2023-11-04 15:00:00 Blankenberge  51.3106    3.1164          439
#> 345 2023-11-04 16:00:00 Blankenberge  51.3106    3.1164          453
#> 346 2023-11-04 17:00:00 Blankenberge  51.3106    3.1164          446
#> 347 2023-11-04 18:00:00 Blankenberge  51.3106    3.1164          409
#> 348 2023-11-04 19:00:00 Blankenberge  51.3106    3.1164          352
#> 349 2023-11-04 20:00:00 Blankenberge  51.3106    3.1164          300
#> 350 2023-11-04 21:00:00 Blankenberge  51.3106    3.1164          241
#> 351 2023-11-04 22:00:00 Blankenberge  51.3106    3.1164          195
#> 352 2023-11-04 23:00:00 Blankenberge  51.3106    3.1164          192
#> 353 2023-11-05 00:00:00 Blankenberge  51.3106    3.1164          233
#> 354 2023-11-05 01:00:00 Blankenberge  51.3106    3.1164          300
#> 355 2023-11-05 02:00:00 Blankenberge  51.3106    3.1164          377
#> 356 2023-11-05 03:00:00 Blankenberge  51.3106    3.1164          429
#> 357 2023-11-05 04:00:00 Blankenberge  51.3106    3.1164          442
#> 358 2023-11-05 05:00:00 Blankenberge  51.3106    3.1164          441
#> 359 2023-11-05 06:00:00 Blankenberge  51.3106    3.1164          421
#> 360 2023-11-05 07:00:00 Blankenberge  51.3106    3.1164          389
#> 361 2023-11-05 08:00:00 Blankenberge  51.3106    3.1164          330
#> 362 2023-11-05 09:00:00 Blankenberge  51.3106    3.1164          275
#> 363 2023-11-05 10:00:00 Blankenberge  51.3106    3.1164          226
#> 364 2023-11-05 11:00:00 Blankenberge  51.3106    3.1164          195
#> 365 2023-11-05 12:00:00 Blankenberge  51.3106    3.1164          223
#> 366 2023-11-05 13:00:00 Blankenberge  51.3106    3.1164          280
#> 367 2023-11-05 14:00:00 Blankenberge  51.3106    3.1164          346
#> 368 2023-11-05 15:00:00 Blankenberge  51.3106    3.1164          399
#> 369 2023-11-05 16:00:00 Blankenberge  51.3106    3.1164          431
#> 370 2023-11-05 17:00:00 Blankenberge  51.3106    3.1164          442
#> 371 2023-11-05 18:00:00 Blankenberge  51.3106    3.1164          438
#> 372 2023-11-05 19:00:00 Blankenberge  51.3106    3.1164          418
#> 373 2023-11-05 20:00:00 Blankenberge  51.3106    3.1164          377
#> 374 2023-11-05 21:00:00 Blankenberge  51.3106    3.1164          318
#> 375 2023-11-05 22:00:00 Blankenberge  51.3106    3.1164          260
#> 376 2023-11-05 23:00:00 Blankenberge  51.3106    3.1164          215
#> 377 2023-11-06 00:00:00 Blankenberge  51.3106    3.1164          203
#> 378 2023-11-06 01:00:00 Blankenberge  51.3106    3.1164          230
#> 379 2023-11-06 02:00:00 Blankenberge  51.3106    3.1164          269
#> 380 2023-11-06 03:00:00 Blankenberge  51.3106    3.1164          322
#> 381 2023-11-06 04:00:00 Blankenberge  51.3106    3.1164          364
#> 382 2023-11-06 05:00:00 Blankenberge  51.3106    3.1164          384
#> 383 2023-11-06 06:00:00 Blankenberge  51.3106    3.1164          390
#> 384 2023-11-06 07:00:00 Blankenberge  51.3106    3.1164          375
#> 385 2023-11-06 08:00:00 Blankenberge  51.3106    3.1164          344
#> 386 2023-11-06 09:00:00 Blankenberge  51.3106    3.1164          309
#> 387 2023-11-06 10:00:00 Blankenberge  51.3106    3.1164          255
#> 388 2023-11-06 11:00:00 Blankenberge  51.3106    3.1164          212
#> 389 2023-11-06 12:00:00 Blankenberge  51.3106    3.1164          176
#> 390 2023-11-06 13:00:00 Blankenberge  51.3106    3.1164          172
#> 391 2023-11-06 14:00:00 Blankenberge  51.3106    3.1164          207
#> 392 2023-11-06 15:00:00 Blankenberge  51.3106    3.1164          268
#> 393 2023-11-06 16:00:00 Blankenberge  51.3106    3.1164          326
#> 394 2023-11-06 17:00:00 Blankenberge  51.3106    3.1164          375
#> 395 2023-11-06 18:00:00 Blankenberge  51.3106    3.1164          394
#> 396 2023-11-06 19:00:00 Blankenberge  51.3106    3.1164          397
#> 397 2023-11-06 20:00:00 Blankenberge  51.3106    3.1164          390
#> 398 2023-11-06 21:00:00 Blankenberge  51.3106    3.1164          359
#> 399 2023-11-06 22:00:00 Blankenberge  51.3106    3.1164          319
#> 400 2023-11-06 23:00:00 Blankenberge  51.3106    3.1164          269
#> 401 2023-11-07 00:00:00 Blankenberge  51.3106    3.1164          227
#> 402 2023-11-07 01:00:00 Blankenberge  51.3106    3.1164          200
#> 403 2023-11-07 02:00:00 Blankenberge  51.3106    3.1164          209
#> 404 2023-11-07 03:00:00 Blankenberge  51.3106    3.1164          238
#> 405 2023-11-07 04:00:00 Blankenberge  51.3106    3.1164          283
#> 406 2023-11-07 05:00:00 Blankenberge  51.3106    3.1164          335
#> 407 2023-11-07 06:00:00 Blankenberge  51.3106    3.1164          368
#> 408 2023-11-07 07:00:00 Blankenberge  51.3106    3.1164          377
#> 409 2023-11-07 08:00:00 Blankenberge  51.3106    3.1164          376
#> 410 2023-11-07 09:00:00 Blankenberge  51.3106    3.1164          355
#> 411 2023-11-07 10:00:00 Blankenberge  51.3106    3.1164          315
#> 412 2023-11-07 11:00:00 Blankenberge  51.3106    3.1164          267
#> 413 2023-11-07 12:00:00 Blankenberge  51.3106    3.1164          212
#> 414 2023-11-07 13:00:00 Blankenberge  51.3106    3.1164          168
#> 415 2023-11-07 14:00:00 Blankenberge  51.3106    3.1164          139
#> 416 2023-11-07 15:00:00 Blankenberge  51.3106    3.1164          170
#> 417 2023-11-07 16:00:00 Blankenberge  51.3106    3.1164          225
#> 418 2023-11-07 17:00:00 Blankenberge  51.3106    3.1164          296
#> 419 2023-11-07 18:00:00 Blankenberge  51.3106    3.1164          360
#> 420 2023-11-07 19:00:00 Blankenberge  51.3106    3.1164          402
#> 421 2023-11-07 20:00:00 Blankenberge  51.3106    3.1164          410
#> 422 2023-11-07 21:00:00 Blankenberge  51.3106    3.1164          405
#> 423 2023-11-07 22:00:00 Blankenberge  51.3106    3.1164          383
#> 424 2023-11-07 23:00:00 Blankenberge  51.3106    3.1164          340
#> 425 2023-11-08 00:00:00 Blankenberge  51.3106    3.1164          286
#> 426 2023-11-08 01:00:00 Blankenberge  51.3106    3.1164          228
#> 427 2023-11-08 02:00:00 Blankenberge  51.3106    3.1164          182
#> 428 2023-11-08 03:00:00 Blankenberge  51.3106    3.1164          170
#> 429 2023-11-08 04:00:00 Blankenberge  51.3106    3.1164          201
#> 430 2023-11-08 05:00:00 Blankenberge  51.3106    3.1164          245
#> 431 2023-11-08 06:00:00 Blankenberge  51.3106    3.1164          311
#> 432 2023-11-08 07:00:00 Blankenberge  51.3106    3.1164          375
#> 433 2023-11-08 08:00:00 Blankenberge  51.3106    3.1164          396
#> 434 2023-11-08 09:00:00 Blankenberge  51.3106    3.1164          394
#> 435 2023-11-08 10:00:00 Blankenberge  51.3106    3.1164          371
#> 436 2023-11-08 11:00:00 Blankenberge  51.3106    3.1164          327
#> 437 2023-11-08 12:00:00 Blankenberge  51.3106    3.1164          266
#> 438 2023-11-08 13:00:00 Blankenberge  51.3106    3.1164          209
#> 439 2023-11-08 14:00:00 Blankenberge  51.3106    3.1164          151
#> 440 2023-11-08 15:00:00 Blankenberge  51.3106    3.1164          112
#> 441 2023-11-08 16:00:00 Blankenberge  51.3106    3.1164          120
#> 442 2023-11-08 17:00:00 Blankenberge  51.3106    3.1164          178
#> 443 2023-11-08 18:00:00 Blankenberge  51.3106    3.1164          259
#> 444 2023-11-08 19:00:00 Blankenberge  51.3106    3.1164          345
#> 445 2023-11-08 20:00:00 Blankenberge  51.3106    3.1164          395
#> 446 2023-11-08 21:00:00 Blankenberge  51.3106    3.1164          408
#> 447 2023-11-08 22:00:00 Blankenberge  51.3106    3.1164          407
#> 448 2023-11-08 23:00:00 Blankenberge  51.3106    3.1164          385
#> 449 2023-11-09 00:00:00 Blankenberge  51.3106    3.1164          342
#> 450 2023-11-09 01:00:00 Blankenberge  51.3106    3.1164          283
#> 451 2023-11-09 02:00:00 Blankenberge  51.3106    3.1164          206
#> 452 2023-11-09 03:00:00 Blankenberge  51.3106    3.1164          156
#> 453 2023-11-09 04:00:00 Blankenberge  51.3106    3.1164          171
#> 454 2023-11-09 05:00:00 Blankenberge  51.3106    3.1164          205
#> 455 2023-11-09 06:00:00 Blankenberge  51.3106    3.1164          264
#> 456 2023-11-09 07:00:00 Blankenberge  51.3106    3.1164          354
#> 457 2023-11-09 08:00:00 Blankenberge  51.3106    3.1164          433
#> 458 2023-11-09 09:00:00 Blankenberge  51.3106    3.1164          451
#> 459 2023-11-09 10:00:00 Blankenberge  51.3106    3.1164          445
#> 460 2023-11-09 11:00:00 Blankenberge  51.3106    3.1164          416
#> 461 2023-11-09 12:00:00 Blankenberge  51.3106    3.1164          375
#> 462 2023-11-09 13:00:00 Blankenberge  51.3106    3.1164          294
#> 463 2023-11-09 14:00:00 Blankenberge  51.3106    3.1164          232
#> 464 2023-11-09 15:00:00 Blankenberge  51.3106    3.1164          158
#> 465 2023-11-09 16:00:00 Blankenberge  51.3106    3.1164          127
#> 466 2023-11-09 17:00:00 Blankenberge  51.3106    3.1164          161
#> 467 2023-11-09 18:00:00 Blankenberge  51.3106    3.1164          226
#> 468 2023-11-09 19:00:00 Blankenberge  51.3106    3.1164          315
#> 469 2023-11-09 20:00:00 Blankenberge  51.3106    3.1164          408
#> 470 2023-11-09 21:00:00 Blankenberge  51.3106    3.1164          452
#> 471 2023-11-09 22:00:00 Blankenberge  51.3106    3.1164          450
#> 472 2023-11-09 23:00:00 Blankenberge  51.3106    3.1164          435
#> 473 2023-11-10 00:00:00 Blankenberge  51.3106    3.1164          393
#> 474 2023-11-10 01:00:00 Blankenberge  51.3106    3.1164          330
#> 475 2023-11-10 02:00:00 Blankenberge  51.3106    3.1164          259
#> 476 2023-11-10 03:00:00 Blankenberge  51.3106    3.1164          187
#> 477 2023-11-10 04:00:00 Blankenberge  51.3106    3.1164          133
#> 478 2023-11-10 05:00:00 Blankenberge  51.3106    3.1164          155
#> 479 2023-11-10 06:00:00 Blankenberge  51.3106    3.1164          203
#> 480 2023-11-10 07:00:00 Blankenberge  51.3106    3.1164          265
#> 481 2023-11-10 08:00:00 Blankenberge  51.3106    3.1164          378
#> 482 2023-11-10 09:00:00 Blankenberge  51.3106    3.1164          442
#> 483 2023-11-10 10:00:00 Blankenberge  51.3106    3.1164          452
#> 484 2023-11-10 11:00:00 Blankenberge  51.3106    3.1164          444
#> 485 2023-11-10 12:00:00 Blankenberge  51.3106    3.1164          406
#> 486 2023-11-10 13:00:00 Blankenberge  51.3106    3.1164          344
#> 487 2023-11-10 14:00:00 Blankenberge  51.3106    3.1164          274
#> 488 2023-11-10 15:00:00 Blankenberge  51.3106    3.1164          200
#> 489 2023-11-10 16:00:00 Blankenberge  51.3106    3.1164          134
#> 490 2023-11-10 17:00:00 Blankenberge  51.3106    3.1164          133
#> 491 2023-11-10 18:00:00 Blankenberge  51.3106    3.1164          188
#> 492 2023-11-10 19:00:00 Blankenberge  51.3106    3.1164          251
#> 493 2023-11-10 20:00:00 Blankenberge  51.3106    3.1164          347
#> 494 2023-11-10 21:00:00 Blankenberge  51.3106    3.1164          446
#> 495 2023-11-10 22:00:00 Blankenberge  51.3106    3.1164          474
#> 496 2023-11-10 23:00:00 Blankenberge  51.3106    3.1164          476
#> 497 2023-11-11 00:00:00 Blankenberge  51.3106    3.1164          447
#> 498 2023-11-11 01:00:00 Blankenberge  51.3106    3.1164          386
#> 499 2023-11-11 02:00:00 Blankenberge  51.3106    3.1164          304
#> 500 2023-11-11 03:00:00 Blankenberge  51.3106    3.1164          218
#> 501 2023-11-11 04:00:00 Blankenberge  51.3106    3.1164          143
#> 502 2023-11-11 05:00:00 Blankenberge  51.3106    3.1164          117
#> 503 2023-11-11 06:00:00 Blankenberge  51.3106    3.1164          155
#> 504 2023-11-11 07:00:00 Blankenberge  51.3106    3.1164          202
#> 505 2023-11-11 08:00:00 Blankenberge  51.3106    3.1164          275
#> 506 2023-11-11 09:00:00 Blankenberge  51.3106    3.1164          406
#> 507 2023-11-11 10:00:00 Blankenberge  51.3106    3.1164          458
#> 508 2023-11-11 11:00:00 Blankenberge  51.3106    3.1164          459
#> 509 2023-11-11 12:00:00 Blankenberge  51.3106    3.1164          434
#> 510 2023-11-11 13:00:00 Blankenberge  51.3106    3.1164          383
#> 511 2023-11-11 14:00:00 Blankenberge  51.3106    3.1164          301
#> 512 2023-11-11 15:00:00 Blankenberge  51.3106    3.1164          212
#> 513 2023-11-11 16:00:00 Blankenberge  51.3106    3.1164          130
#> 514 2023-11-11 17:00:00 Blankenberge  51.3106    3.1164           73
#> 515 2023-11-11 18:00:00 Blankenberge  51.3106    3.1164          103
#> 516 2023-11-11 19:00:00 Blankenberge  51.3106    3.1164          155
#> 517 2023-11-11 20:00:00 Blankenberge  51.3106    3.1164          227
#> 518 2023-11-11 21:00:00 Blankenberge  51.3106    3.1164          372
#> 519 2023-11-11 22:00:00 Blankenberge  51.3106    3.1164          457
#> 520 2023-11-11 23:00:00 Blankenberge  51.3106    3.1164          468
#> 521 2023-11-12 00:00:00 Blankenberge  51.3106    3.1164          454
#> 522 2023-11-12 01:00:00 Blankenberge  51.3106    3.1164          407
#> 523 2023-11-12 02:00:00 Blankenberge  51.3106    3.1164          332
#> 524 2023-11-12 03:00:00 Blankenberge  51.3106    3.1164          242
#> 525 2023-11-12 04:00:00 Blankenberge  51.3106    3.1164          155
#> 526 2023-11-12 05:00:00 Blankenberge  51.3106    3.1164           89
#> 527 2023-11-12 06:00:00 Blankenberge  51.3106    3.1164          104
#> 528 2023-11-12 07:00:00 Blankenberge  51.3106    3.1164          149
#> 529 2023-11-12 08:00:00 Blankenberge  51.3106    3.1164          210
#> 530 2023-11-12 09:00:00 Blankenberge  51.3106    3.1164          322
#> 531 2023-11-12 10:00:00 Blankenberge  51.3106    3.1164          450
#> 532 2023-11-12 11:00:00 Blankenberge  51.3106    3.1164          466
#> 533 2023-11-12 12:00:00 Blankenberge  51.3106    3.1164          460
#> 534 2023-11-12 13:00:00 Blankenberge  51.3106    3.1164          416
#> 535 2023-11-12 14:00:00 Blankenberge  51.3106    3.1164          346
#> 536 2023-11-12 15:00:00 Blankenberge  51.3106    3.1164          252
#> 537 2023-11-12 16:00:00 Blankenberge  51.3106    3.1164          161
#> 538 2023-11-12 17:00:00 Blankenberge  51.3106    3.1164           83
#> 539 2023-11-12 18:00:00 Blankenberge  51.3106    3.1164           62
#> 540 2023-11-12 19:00:00 Blankenberge  51.3106    3.1164          109
#> 541 2023-11-12 20:00:00 Blankenberge  51.3106    3.1164          173
#> 542 2023-11-12 21:00:00 Blankenberge  51.3106    3.1164          284
#> 543 2023-11-12 22:00:00 Blankenberge  51.3106    3.1164          439
#> 544 2023-11-12 23:00:00 Blankenberge  51.3106    3.1164          477
#> 545 2023-11-13 00:00:00 Blankenberge  51.3106    3.1164          474
#> 546 2023-11-13 01:00:00 Blankenberge  51.3106    3.1164          444
#> 547 2023-11-13 02:00:00 Blankenberge  51.3106    3.1164          377
#> 548 2023-11-13 03:00:00 Blankenberge  51.3106    3.1164          287
#> 549 2023-11-13 04:00:00 Blankenberge  51.3106    3.1164          198
#> 550 2023-11-13 05:00:00 Blankenberge  51.3106    3.1164          112
#> 551 2023-11-13 06:00:00 Blankenberge  51.3106    3.1164           60
#> 552 2023-11-13 07:00:00 Blankenberge  51.3106    3.1164          102
#> 553 2023-11-13 08:00:00 Blankenberge  51.3106    3.1164          159
#> 554 2023-11-13 09:00:00 Blankenberge  51.3106    3.1164          260
#> 555 2023-11-13 10:00:00 Blankenberge  51.3106    3.1164          426
#> 556 2023-11-13 11:00:00 Blankenberge  51.3106    3.1164          473
#> 557 2023-11-13 12:00:00 Blankenberge  51.3106    3.1164          471
#> 558 2023-11-13 13:00:00 Blankenberge  51.3106    3.1164          441
#> 559 2023-11-13 14:00:00 Blankenberge  51.3106    3.1164          392
#> 560 2023-11-13 15:00:00 Blankenberge  51.3106    3.1164          316
#> 561 2023-11-13 16:00:00 Blankenberge  51.3106    3.1164          231
#> 562 2023-11-13 17:00:00 Blankenberge  51.3106    3.1164          148
#> 563 2023-11-13 18:00:00 Blankenberge  51.3106    3.1164           87
#> 564 2023-11-13 19:00:00 Blankenberge  51.3106    3.1164          110
#> 565 2023-11-13 20:00:00 Blankenberge  51.3106    3.1164          162
#> 566 2023-11-13 21:00:00 Blankenberge  51.3106    3.1164          248
#> 567 2023-11-13 22:00:00 Blankenberge  51.3106    3.1164          394
#> 568 2023-11-13 23:00:00 Blankenberge  51.3106    3.1164          478
#> 569 2023-11-14 00:00:00 Blankenberge  51.3106    3.1164          489
#> 570 2023-11-14 01:00:00 Blankenberge  51.3106    3.1164          482
#> 571 2023-11-14 02:00:00 Blankenberge  51.3106    3.1164          445
#> 572 2023-11-14 03:00:00 Blankenberge  51.3106    3.1164          372
#> 573 2023-11-14 04:00:00 Blankenberge  51.3106    3.1164          294
#> 574 2023-11-14 05:00:00 Blankenberge  51.3106    3.1164          204
#> 575 2023-11-14 06:00:00 Blankenberge  51.3106    3.1164          128
#> 576 2023-11-14 07:00:00 Blankenberge  51.3106    3.1164          135
#> 577 2023-11-14 08:00:00 Blankenberge  51.3106    3.1164          178
#> 578 2023-11-14 09:00:00 Blankenberge  51.3106    3.1164          235
#> 579 2023-11-14 10:00:00 Blankenberge  51.3106    3.1164          347
#> 580 2023-11-14 11:00:00 Blankenberge  51.3106    3.1164          482
#> 581 2023-11-14 12:00:00 Blankenberge  51.3106    3.1164          499
#> 582 2023-11-14 13:00:00 Blankenberge  51.3106    3.1164          493
#> 583 2023-11-14 14:00:00 Blankenberge  51.3106    3.1164          456
#> 584 2023-11-14 15:00:00 Blankenberge  51.3106    3.1164          384
#> 585 2023-11-14 16:00:00 Blankenberge  51.3106    3.1164          300
#> 586 2023-11-14 17:00:00 Blankenberge  51.3106    3.1164          212
#> 587 2023-11-14 18:00:00 Blankenberge  51.3106    3.1164          138
#> 588 2023-11-14 19:00:00 Blankenberge  51.3106    3.1164          105
#> 589 2023-11-14 20:00:00 Blankenberge  51.3106    3.1164          159
#> 590 2023-11-14 21:00:00 Blankenberge  51.3106    3.1164          213
#> 591 2023-11-14 22:00:00 Blankenberge  51.3106    3.1164          304
#> 592 2023-11-14 23:00:00 Blankenberge  51.3106    3.1164          457
#> 593 2023-11-15 00:00:00 Blankenberge  51.3106    3.1164          498
#> 594 2023-11-15 01:00:00 Blankenberge  51.3106    3.1164          497
#> 595 2023-11-15 02:00:00 Blankenberge  51.3106    3.1164          468
#> 596 2023-11-15 03:00:00 Blankenberge  51.3106    3.1164          406
#> 597 2023-11-15 04:00:00 Blankenberge  51.3106    3.1164          315
#> 598 2023-11-15 05:00:00 Blankenberge  51.3106    3.1164          224
#> 599 2023-11-15 06:00:00 Blankenberge  51.3106    3.1164          139
#> 600 2023-11-15 07:00:00 Blankenberge  51.3106    3.1164           79
#> 601 2023-11-15 08:00:00 Blankenberge  51.3106    3.1164          127
#> 602 2023-11-15 09:00:00 Blankenberge  51.3106    3.1164          174
#> 603 2023-11-15 10:00:00 Blankenberge  51.3106    3.1164          244
#> 604 2023-11-15 11:00:00 Blankenberge  51.3106    3.1164          387
#> 605 2023-11-15 12:00:00 Blankenberge  51.3106    3.1164          479
#> 606 2023-11-15 13:00:00 Blankenberge  51.3106    3.1164          482
#> 607 2023-11-15 14:00:00 Blankenberge  51.3106    3.1164          467
#> 608 2023-11-15 15:00:00 Blankenberge  51.3106    3.1164          418
#> 609 2023-11-15 16:00:00 Blankenberge  51.3106    3.1164          339
#> 610 2023-11-15 17:00:00 Blankenberge  51.3106    3.1164          244
#> 611 2023-11-15 18:00:00 Blankenberge  51.3106    3.1164          155
#> 612 2023-11-15 19:00:00 Blankenberge  51.3106    3.1164           81
#> 613 2023-11-15 20:00:00 Blankenberge  51.3106    3.1164          105
#> 614 2023-11-15 21:00:00 Blankenberge  51.3106    3.1164          152
#> 615 2023-11-15 22:00:00 Blankenberge  51.3106    3.1164          209
#> 616 2023-11-15 23:00:00 Blankenberge  51.3106    3.1164          319
#> 617 2023-11-16 00:00:00 Blankenberge  51.3106    3.1164          465
#> 618 2023-11-16 01:00:00 Blankenberge  51.3106    3.1164          482
#> 619 2023-11-16 02:00:00 Blankenberge  51.3106    3.1164          476
#> 620 2023-11-16 03:00:00 Blankenberge  51.3106    3.1164          434
#> 621 2023-11-16 04:00:00 Blankenberge  51.3106    3.1164          361
#> 622 2023-11-16 05:00:00 Blankenberge  51.3106    3.1164          265
#> 623 2023-11-16 06:00:00 Blankenberge  51.3106    3.1164          172
#> 624 2023-11-16 07:00:00 Blankenberge  51.3106    3.1164           86
#> 625 2023-11-16 08:00:00 Blankenberge  51.3106    3.1164           65
#> 626 2023-11-16 09:00:00 Blankenberge  51.3106    3.1164          115
#> 627 2023-11-16 10:00:00 Blankenberge  51.3106    3.1164          169
#> 628 2023-11-16 11:00:00 Blankenberge  51.3106    3.1164          258
#> 629 2023-11-16 12:00:00 Blankenberge  51.3106    3.1164          417
#> 630 2023-11-16 13:00:00 Blankenberge  51.3106    3.1164          463
#> 631 2023-11-16 14:00:00 Blankenberge  51.3106    3.1164          461
#> 632 2023-11-16 15:00:00 Blankenberge  51.3106    3.1164          431
#> 633 2023-11-16 16:00:00 Blankenberge  51.3106    3.1164          372
#> 634 2023-11-16 17:00:00 Blankenberge  51.3106    3.1164          280
#> 635 2023-11-16 18:00:00 Blankenberge  51.3106    3.1164          179
#> 636 2023-11-16 19:00:00 Blankenberge  51.3106    3.1164           98
#> 637 2023-11-16 20:00:00 Blankenberge  51.3106    3.1164           58
#> 638 2023-11-16 21:00:00 Blankenberge  51.3106    3.1164          106
#> 639 2023-11-16 22:00:00 Blankenberge  51.3106    3.1164          157
#> 640 2023-11-16 23:00:00 Blankenberge  51.3106    3.1164          218
#> 641 2023-11-17 00:00:00 Blankenberge  51.3106    3.1164          354
#> 642 2023-11-17 01:00:00 Blankenberge  51.3106    3.1164          452
#> 643 2023-11-17 02:00:00 Blankenberge  51.3106    3.1164          458
#> 644 2023-11-17 03:00:00 Blankenberge  51.3106    3.1164          441
#> 645 2023-11-17 04:00:00 Blankenberge  51.3106    3.1164          389
#> 646 2023-11-17 05:00:00 Blankenberge  51.3106    3.1164          312
#> 647 2023-11-17 06:00:00 Blankenberge  51.3106    3.1164          223
#> 648 2023-11-17 07:00:00 Blankenberge  51.3106    3.1164          139
#> 649 2023-11-17 08:00:00 Blankenberge  51.3106    3.1164           73
#> 650 2023-11-17 09:00:00 Blankenberge  51.3106    3.1164           89
#> 651 2023-11-17 10:00:00 Blankenberge  51.3106    3.1164          136
#> 652 2023-11-17 11:00:00 Blankenberge  51.3106    3.1164          204
#> 653 2023-11-17 12:00:00 Blankenberge  51.3106    3.1164          322
#> 654 2023-11-17 13:00:00 Blankenberge  51.3106    3.1164          450
#> 655 2023-11-17 14:00:00 Blankenberge  51.3106    3.1164          467
#> 656 2023-11-17 15:00:00 Blankenberge  51.3106    3.1164          462
#> 657 2023-11-17 16:00:00 Blankenberge  51.3106    3.1164          423
#> 658 2023-11-17 17:00:00 Blankenberge  51.3106    3.1164          353
#> 659 2023-11-17 18:00:00 Blankenberge  51.3106    3.1164          265
#> 660 2023-11-17 19:00:00 Blankenberge  51.3106    3.1164          173
#> 661 2023-11-17 20:00:00 Blankenberge  51.3106    3.1164           96
#> 662 2023-11-17 21:00:00 Blankenberge  51.3106    3.1164           80
#> 663 2023-11-17 22:00:00 Blankenberge  51.3106    3.1164          120
#> 664 2023-11-17 23:00:00 Blankenberge  51.3106    3.1164          171
#> 665 2023-11-18 00:00:00 Blankenberge  51.3106    3.1164          248
#> 666 2023-11-18 01:00:00 Blankenberge  51.3106    3.1164          390
#> 667 2023-11-18 02:00:00 Blankenberge  51.3106    3.1164          439
#> 668 2023-11-18 03:00:00 Blankenberge  51.3106    3.1164          437
#> 669 2023-11-18 04:00:00 Blankenberge  51.3106    3.1164          405
#> 670 2023-11-18 05:00:00 Blankenberge  51.3106    3.1164          355
#> 671 2023-11-18 06:00:00 Blankenberge  51.3106    3.1164          273
#> 672 2023-11-18 07:00:00 Blankenberge  51.3106    3.1164          188
#> 673 2023-11-18 08:00:00 Blankenberge  51.3106    3.1164          111
#> 674 2023-11-18 09:00:00 Blankenberge  51.3106    3.1164           50
#> 675 2023-11-18 10:00:00 Blankenberge  51.3106    3.1164           55
#> 676 2023-11-18 11:00:00 Blankenberge  51.3106    3.1164          104
#> 677 2023-11-18 12:00:00 Blankenberge  51.3106    3.1164          201
#> 678 2023-11-18 13:00:00 Blankenberge  51.3106    3.1164          345
#> 679 2023-11-18 14:00:00 Blankenberge  51.3106    3.1164          410
#> 680 2023-11-18 15:00:00 Blankenberge  51.3106    3.1164          418
#> 681 2023-11-18 16:00:00 Blankenberge  51.3106    3.1164          404
#> 682 2023-11-18 17:00:00 Blankenberge  51.3106    3.1164          365
#> 683 2023-11-18 18:00:00 Blankenberge  51.3106    3.1164          295
#> 684 2023-11-18 19:00:00 Blankenberge  51.3106    3.1164          224
#> 685 2023-11-18 20:00:00 Blankenberge  51.3106    3.1164          149
#> 686 2023-11-18 21:00:00 Blankenberge  51.3106    3.1164           86
#> 687 2023-11-18 22:00:00 Blankenberge  51.3106    3.1164           87
#> 688 2023-11-18 23:00:00 Blankenberge  51.3106    3.1164          129
#> 689 2023-11-19 00:00:00 Blankenberge  51.3106    3.1164          195
#> 690 2023-11-19 01:00:00 Blankenberge  51.3106    3.1164          297
#> 691 2023-11-19 02:00:00 Blankenberge  51.3106    3.1164          405
#> 692 2023-11-19 03:00:00 Blankenberge  51.3106    3.1164          424
#> 693 2023-11-19 04:00:00 Blankenberge  51.3106    3.1164          420
#> 694 2023-11-19 05:00:00 Blankenberge  51.3106    3.1164          391
#> 695 2023-11-19 06:00:00 Blankenberge  51.3106    3.1164          338
#> 696 2023-11-19 07:00:00 Blankenberge  51.3106    3.1164          276
#> 697 2023-11-19 08:00:00 Blankenberge  51.3106    3.1164          205
#> 698 2023-11-19 09:00:00 Blankenberge  51.3106    3.1164          143
#> 699 2023-11-19 10:00:00 Blankenberge  51.3106    3.1164          103
#> 700 2023-11-19 11:00:00 Blankenberge  51.3106    3.1164          140
#> 701 2023-11-19 12:00:00 Blankenberge  51.3106    3.1164          213
#> 702 2023-11-19 13:00:00 Blankenberge  51.3106    3.1164          303
#> 703 2023-11-19 14:00:00 Blankenberge  51.3106    3.1164          408
#> 704 2023-11-19 15:00:00 Blankenberge  51.3106    3.1164          458
#> 705 2023-11-19 16:00:00 Blankenberge  51.3106    3.1164          461
#> 706 2023-11-19 17:00:00 Blankenberge  51.3106    3.1164          449
#> 707 2023-11-19 18:00:00 Blankenberge  51.3106    3.1164          413
#> 708 2023-11-19 19:00:00 Blankenberge  51.3106    3.1164          353
#> 709 2023-11-19 20:00:00 Blankenberge  51.3106    3.1164          288
#> 710 2023-11-19 21:00:00 Blankenberge  51.3106    3.1164          221
#> 711 2023-11-19 22:00:00 Blankenberge  51.3106    3.1164          162
#> 712 2023-11-19 23:00:00 Blankenberge  51.3106    3.1164          166
#> 713 2023-11-20 00:00:00 Blankenberge  51.3106    3.1164          203
#> 714 2023-11-20 01:00:00 Blankenberge  51.3106    3.1164          258
#> 715 2023-11-20 02:00:00 Blankenberge  51.3106    3.1164          333
#> 716 2023-11-20 03:00:00 Blankenberge  51.3106    3.1164          402
#> 717 2023-11-20 04:00:00 Blankenberge  51.3106    3.1164          423
#> 718 2023-11-20 05:00:00 Blankenberge  51.3106    3.1164          420
#> 719 2023-11-20 06:00:00 Blankenberge  51.3106    3.1164          391
#> 720 2023-11-20 07:00:00 Blankenberge  51.3106    3.1164          340
#> 
#> $par
#> $par$daterange
#> [1] "2023-10-21" "2023-11-21"
#> 
#> $par$type
#> [1] "MVB data"
#> 
#> $par$stations
#> [1] "Blankenberge"
#> 
#> $par$calc
#> [1] "max"
#> 
#> $par$binSize
#> [1] "hour"
#> 
#> $par$parameters
#> [1] "Tide (cmTAW)"
#> 
#>