Retrieve MVB-data from the LifeWatch project

Retrieves the Meetnet Vlaams Banken (MVB) data from the LifeWatch project. Without valid credentials you are only allowed to view the "Tide TAW" parameter for the last 30 days, grouped by day/hour. To get an account, register via the Lifewatch RShiny registration webpage.

getMvbData(
  startdate,
  stopdate,
  parameters,
  stations = NULL,
  by,
  calc,
  usr = NULL,
  pwd = NULL,
  params = FALSE,
  ...
)

Arguments

startdate

Starting date for the query. Without a login, this is restricted to the last month

stopdate

Stopping date for the query. Without a login, this is restricted to the last month

parameters

One (or list) of the $parameter in listMvbStations.

stations

List of stations to be included in the query. See listMvbStations. Use stations="All" or stations=NULL to get all stations.

by

Time grouping for calculation, one of ('day','hour', '10min' ,'none')

calc

Calculation to perform given time grouping, one of ('avg', 'max', 'min', 'none').

usr

Username to connect to database

pwd

Password to connect to database

params

If TRUE, returns a list with the dataset and the query parameters applied in the server side. IF FALSE returns only the data.

...

Reserved for internal use.

Value

Dataframe with the specified MVB data.

Examples

getMvbData(Sys.Date() - 30, Sys.Date() + 1, parameters = 'Tide TAW',
stations = "Nieuwpoort", by = "day", calc = "avg")
#> - Query mode: Post request to OpenCPU server
#> No encoding supplied: defaulting to UTF-8.
#> - Your query:
#>    - daterange: c("2024-07-23", "2024-08-23")
#>    - type: MVB data
#>    - stations: Nieuwpoort
#>    - calc: avg
#>    - binSize: day
#>    - parameters: Tide TAW
#> - Server query:
#>    - daterange: c("2024-07-23", "2024-08-23")
#>    - type: MVB data
#>    - stations: Nieuwpoort
#>    - calc: avg
#>    - binSize: day
#>    - parameters: Tide (cmTAW)
#> Warning: The query applied on the server differ from the parameters you used. 
#>  Hint: You may need an account to fully access the data under moratorium. 
#>  Hint: Request access at: https://rshiny.lifewatch.be/account?p=register
#>          time    station latitude longitude Tide (cmTAW)
#> 1  2024-07-23 Nieuwpoort  51.1506    2.7281     245.0625
#> 2  2024-07-24 Nieuwpoort  51.1506    2.7281     236.2535
#> 3  2024-07-25 Nieuwpoort  51.1506    2.7281     235.1389
#> 4  2024-07-26 Nieuwpoort  51.1506    2.7281     256.4375
#> 5  2024-07-27 Nieuwpoort  51.1506    2.7281     251.1875
#> 6  2024-07-28 Nieuwpoort  51.1506    2.7281     243.3472
#> 7  2024-07-29 Nieuwpoort  51.1506    2.7281     233.6771
#> 8  2024-07-30 Nieuwpoort  51.1506    2.7281     247.1319
#> 9  2024-07-31 Nieuwpoort  51.1506    2.7281     215.8113
#> 10 2024-08-01 Nieuwpoort  51.1506    2.7281     239.2431
#> 11 2024-08-02 Nieuwpoort  51.1506    2.7281     238.2569
#> 12 2024-08-03 Nieuwpoort  51.1506    2.7281     243.7118
#> 13 2024-08-04 Nieuwpoort  51.1506    2.7281     240.8924
#> 14 2024-08-05 Nieuwpoort  51.1506    2.7281     236.7049
#> 15 2024-08-06 Nieuwpoort  51.1506    2.7281     244.1038
#> 16 2024-08-07 Nieuwpoort  51.1506    2.7281     284.1220
#> 17 2024-08-08 Nieuwpoort  51.1506    2.7281     247.7257
#> 18 2024-08-09 Nieuwpoort  51.1506    2.7281     252.8715
#> 19 2024-08-10 Nieuwpoort  51.1506    2.7281     250.2778
#> 20 2024-08-11 Nieuwpoort  51.1506    2.7281     240.3750
#> 21 2024-08-12 Nieuwpoort  51.1506    2.7281     235.5305
#> 22 2024-08-13 Nieuwpoort  51.1506    2.7281     260.7188
#> 23 2024-08-14 Nieuwpoort  51.1506    2.7281     258.2361
#> 24 2024-08-15 Nieuwpoort  51.1506    2.7281     241.1632
#> 25 2024-08-16 Nieuwpoort  51.1506    2.7281     257.4271
#> 26 2024-08-17 Nieuwpoort  51.1506    2.7281     249.5451
#> 27 2024-08-18 Nieuwpoort  51.1506    2.7281     245.7604
#> 28 2024-08-19 Nieuwpoort  51.1506    2.7281     237.2332
#> 29 2024-08-20 Nieuwpoort  51.1506    2.7281     234.8958
#> 30 2024-08-21 Nieuwpoort  51.1506    2.7281     247.9236
#> 31 2024-08-22 Nieuwpoort  51.1506    2.7281     283.8155
getMvbData(Sys.Date() - 30, Sys.Date() + 1, parameters = 'Tide TAW',
stations = "Blankenberge", by = "hour", calc = "max", params = TRUE)
#> - Query mode: Post request to OpenCPU server
#> No encoding supplied: defaulting to UTF-8.
#> - Your query:
#>    - daterange: c("2024-07-23", "2024-08-23")
#>    - type: MVB data
#>    - stations: Blankenberge
#>    - calc: max
#>    - binSize: hour
#>    - parameters: Tide TAW
#> - Server query:
#>    - daterange: c("2024-07-23", "2024-08-23")
#>    - type: MVB data
#>    - stations: Blankenberge
#>    - calc: max
#>    - binSize: hour
#>    - parameters: Tide (cmTAW)
#> Warning: The query applied on the server differ from the parameters you used. 
#>  Hint: You may need an account to fully access the data under moratorium. 
#>  Hint: Request access at: https://rshiny.lifewatch.be/account?p=register
#> $mdf
#>                    time      station latitude longitude Tide (cmTAW)
#> 1   2024-07-23 00:00:00 Blankenberge  51.3106    3.1164          462
#> 2   2024-07-23 01:00:00 Blankenberge  51.3106    3.1164          471
#> 3   2024-07-23 02:00:00 Blankenberge  51.3106    3.1164          459
#> 4   2024-07-23 03:00:00 Blankenberge  51.3106    3.1164          415
#> 5   2024-07-23 04:00:00 Blankenberge  51.3106    3.1164          340
#> 6   2024-07-23 05:00:00 Blankenberge  51.3106    3.1164          249
#> 7   2024-07-23 06:00:00 Blankenberge  51.3106    3.1164          155
#> 8   2024-07-23 07:00:00 Blankenberge  51.3106    3.1164           78
#> 9   2024-07-23 08:00:00 Blankenberge  51.3106    3.1164          101
#> 10  2024-07-23 09:00:00 Blankenberge  51.3106    3.1164          146
#> 11  2024-07-23 10:00:00 Blankenberge  51.3106    3.1164          206
#> 12  2024-07-23 11:00:00 Blankenberge  51.3106    3.1164          298
#> 13  2024-07-23 12:00:00 Blankenberge  51.3106    3.1164          453
#> 14  2024-07-23 13:00:00 Blankenberge  51.3106    3.1164          480
#> 15  2024-07-23 14:00:00 Blankenberge  51.3106    3.1164          476
#> 16  2024-07-23 15:00:00 Blankenberge  51.3106    3.1164          435
#> 17  2024-07-23 16:00:00 Blankenberge  51.3106    3.1164          370
#> 18  2024-07-23 17:00:00 Blankenberge  51.3106    3.1164          278
#> 19  2024-07-23 18:00:00 Blankenberge  51.3106    3.1164          178
#> 20  2024-07-23 19:00:00 Blankenberge  51.3106    3.1164           87
#> 21  2024-07-23 20:00:00 Blankenberge  51.3106    3.1164           40
#> 22  2024-07-23 21:00:00 Blankenberge  51.3106    3.1164           96
#> 23  2024-07-23 22:00:00 Blankenberge  51.3106    3.1164          152
#> 24  2024-07-23 23:00:00 Blankenberge  51.3106    3.1164          238
#> 25  2024-07-24 00:00:00 Blankenberge  51.3106    3.1164          413
#> 26  2024-07-24 01:00:00 Blankenberge  51.3106    3.1164          485
#> 27  2024-07-24 02:00:00 Blankenberge  51.3106    3.1164          487
#> 28  2024-07-24 03:00:00 Blankenberge  51.3106    3.1164          468
#> 29  2024-07-24 04:00:00 Blankenberge  51.3106    3.1164          409
#> 30  2024-07-24 05:00:00 Blankenberge  51.3106    3.1164          320
#> 31  2024-07-24 06:00:00 Blankenberge  51.3106    3.1164          220
#> 32  2024-07-24 07:00:00 Blankenberge  51.3106    3.1164          123
#> 33  2024-07-24 08:00:00 Blankenberge  51.3106    3.1164           56
#> 34  2024-07-24 09:00:00 Blankenberge  51.3106    3.1164          106
#> 35  2024-07-24 10:00:00 Blankenberge  51.3106    3.1164          150
#> 36  2024-07-24 11:00:00 Blankenberge  51.3106    3.1164          204
#> 37  2024-07-24 12:00:00 Blankenberge  51.3106    3.1164          308
#> 38  2024-07-24 13:00:00 Blankenberge  51.3106    3.1164          453
#> 39  2024-07-24 14:00:00 Blankenberge  51.3106    3.1164          463
#> 40  2024-07-24 15:00:00 Blankenberge  51.3106    3.1164          450
#> 41  2024-07-24 16:00:00 Blankenberge  51.3106    3.1164          400
#> 42  2024-07-24 17:00:00 Blankenberge  51.3106    3.1164          325
#> 43  2024-07-24 18:00:00 Blankenberge  51.3106    3.1164          226
#> 44  2024-07-24 19:00:00 Blankenberge  51.3106    3.1164          124
#> 45  2024-07-24 20:00:00 Blankenberge  51.3106    3.1164           38
#> 46  2024-07-24 21:00:00 Blankenberge  51.3106    3.1164           11
#> 47  2024-07-24 22:00:00 Blankenberge  51.3106    3.1164           69
#> 48  2024-07-24 23:00:00 Blankenberge  51.3106    3.1164          131
#> 49  2024-07-25 00:00:00 Blankenberge  51.3106    3.1164          244
#> 50  2024-07-25 01:00:00 Blankenberge  51.3106    3.1164          428
#> 51  2024-07-25 02:00:00 Blankenberge  51.3106    3.1164          467
#> 52  2024-07-25 03:00:00 Blankenberge  51.3106    3.1164          468
#> 53  2024-07-25 04:00:00 Blankenberge  51.3106    3.1164          438
#> 54  2024-07-25 05:00:00 Blankenberge  51.3106    3.1164          371
#> 55  2024-07-25 06:00:00 Blankenberge  51.3106    3.1164          284
#> 56  2024-07-25 07:00:00 Blankenberge  51.3106    3.1164          192
#> 57  2024-07-25 08:00:00 Blankenberge  51.3106    3.1164           95
#> 58  2024-07-25 09:00:00 Blankenberge  51.3106    3.1164           62
#> 59  2024-07-25 10:00:00 Blankenberge  51.3106    3.1164          110
#> 60  2024-07-25 11:00:00 Blankenberge  51.3106    3.1164          160
#> 61  2024-07-25 12:00:00 Blankenberge  51.3106    3.1164          221
#> 62  2024-07-25 13:00:00 Blankenberge  51.3106    3.1164          360
#> 63  2024-07-25 14:00:00 Blankenberge  51.3106    3.1164          460
#> 64  2024-07-25 15:00:00 Blankenberge  51.3106    3.1164          466
#> 65  2024-07-25 16:00:00 Blankenberge  51.3106    3.1164          440
#> 66  2024-07-25 17:00:00 Blankenberge  51.3106    3.1164          384
#> 67  2024-07-25 18:00:00 Blankenberge  51.3106    3.1164          304
#> 68  2024-07-25 19:00:00 Blankenberge  51.3106    3.1164          208
#> 69  2024-07-25 20:00:00 Blankenberge  51.3106    3.1164          106
#> 70  2024-07-25 21:00:00 Blankenberge  51.3106    3.1164           27
#> 71  2024-07-25 22:00:00 Blankenberge  51.3106    3.1164           48
#> 72  2024-07-25 23:00:00 Blankenberge  51.3106    3.1164          106
#> 73  2024-07-26 00:00:00 Blankenberge  51.3106    3.1164          188
#> 74  2024-07-26 01:00:00 Blankenberge  51.3106    3.1164          327
#> 75  2024-07-26 02:00:00 Blankenberge  51.3106    3.1164          466
#> 76  2024-07-26 03:00:00 Blankenberge  51.3106    3.1164          488
#> 77  2024-07-26 04:00:00 Blankenberge  51.3106    3.1164          481
#> 78  2024-07-26 05:00:00 Blankenberge  51.3106    3.1164          448
#> 79  2024-07-26 06:00:00 Blankenberge  51.3106    3.1164          373
#> 80  2024-07-26 07:00:00 Blankenberge  51.3106    3.1164          281
#> 81  2024-07-26 08:00:00 Blankenberge  51.3106    3.1164          191
#> 82  2024-07-26 09:00:00 Blankenberge  51.3106    3.1164          105
#> 83  2024-07-26 10:00:00 Blankenberge  51.3106    3.1164          108
#> 84  2024-07-26 11:00:00 Blankenberge  51.3106    3.1164          158
#> 85  2024-07-26 12:00:00 Blankenberge  51.3106    3.1164          204
#> 86  2024-07-26 13:00:00 Blankenberge  51.3106    3.1164          277
#> 87  2024-07-26 14:00:00 Blankenberge  51.3106    3.1164          417
#> 88  2024-07-26 15:00:00 Blankenberge  51.3106    3.1164          476
#> 89  2024-07-26 16:00:00 Blankenberge  51.3106    3.1164          475
#> 90  2024-07-26 17:00:00 Blankenberge  51.3106    3.1164          435
#> 91  2024-07-26 18:00:00 Blankenberge  51.3106    3.1164          377
#> 92  2024-07-26 19:00:00 Blankenberge  51.3106    3.1164          290
#> 93  2024-07-26 20:00:00 Blankenberge  51.3106    3.1164          189
#> 94  2024-07-26 21:00:00 Blankenberge  51.3106    3.1164           96
#> 95  2024-07-26 22:00:00 Blankenberge  51.3106    3.1164           29
#> 96  2024-07-26 23:00:00 Blankenberge  51.3106    3.1164           59
#> 97  2024-07-27 00:00:00 Blankenberge  51.3106    3.1164          114
#> 98  2024-07-27 01:00:00 Blankenberge  51.3106    3.1164          191
#> 99  2024-07-27 02:00:00 Blankenberge  51.3106    3.1164          344
#> 100 2024-07-27 03:00:00 Blankenberge  51.3106    3.1164          450
#> 101 2024-07-27 04:00:00 Blankenberge  51.3106    3.1164          462
#> 102 2024-07-27 05:00:00 Blankenberge  51.3106    3.1164          452
#> 103 2024-07-27 06:00:00 Blankenberge  51.3106    3.1164          413
#> 104 2024-07-27 07:00:00 Blankenberge  51.3106    3.1164          338
#> 105 2024-07-27 08:00:00 Blankenberge  51.3106    3.1164          249
#> 106 2024-07-27 09:00:00 Blankenberge  51.3106    3.1164          161
#> 107 2024-07-27 10:00:00 Blankenberge  51.3106    3.1164           87
#> 108 2024-07-27 11:00:00 Blankenberge  51.3106    3.1164          113
#> 109 2024-07-27 12:00:00 Blankenberge  51.3106    3.1164          158
#> 110 2024-07-27 13:00:00 Blankenberge  51.3106    3.1164          208
#> 111 2024-07-27 14:00:00 Blankenberge  51.3106    3.1164          296
#> 112 2024-07-27 15:00:00 Blankenberge  51.3106    3.1164          430
#> 113 2024-07-27 16:00:00 Blankenberge  51.3106    3.1164          462
#> 114 2024-07-27 17:00:00 Blankenberge  51.3106    3.1164          458
#> 115 2024-07-27 18:00:00 Blankenberge  51.3106    3.1164          419
#> 116 2024-07-27 19:00:00 Blankenberge  51.3106    3.1164          357
#> 117 2024-07-27 20:00:00 Blankenberge  51.3106    3.1164          269
#> 118 2024-07-27 21:00:00 Blankenberge  51.3106    3.1164          176
#> 119 2024-07-27 22:00:00 Blankenberge  51.3106    3.1164           94
#> 120 2024-07-27 23:00:00 Blankenberge  51.3106    3.1164           36
#> 121 2024-07-28 00:00:00 Blankenberge  51.3106    3.1164           78
#> 122 2024-07-28 01:00:00 Blankenberge  51.3106    3.1164          131
#> 123 2024-07-28 02:00:00 Blankenberge  51.3106    3.1164          215
#> 124 2024-07-28 03:00:00 Blankenberge  51.3106    3.1164          366
#> 125 2024-07-28 04:00:00 Blankenberge  51.3106    3.1164          438
#> 126 2024-07-28 05:00:00 Blankenberge  51.3106    3.1164          443
#> 127 2024-07-28 06:00:00 Blankenberge  51.3106    3.1164          431
#> 128 2024-07-28 07:00:00 Blankenberge  51.3106    3.1164          388
#> 129 2024-07-28 08:00:00 Blankenberge  51.3106    3.1164          315
#> 130 2024-07-28 09:00:00 Blankenberge  51.3106    3.1164          231
#> 131 2024-07-28 10:00:00 Blankenberge  51.3106    3.1164          146
#> 132 2024-07-28 11:00:00 Blankenberge  51.3106    3.1164           88
#> 133 2024-07-28 12:00:00 Blankenberge  51.3106    3.1164          118
#> 134 2024-07-28 13:00:00 Blankenberge  51.3106    3.1164          156
#> 135 2024-07-28 14:00:00 Blankenberge  51.3106    3.1164          211
#> 136 2024-07-28 15:00:00 Blankenberge  51.3106    3.1164          300
#> 137 2024-07-28 16:00:00 Blankenberge  51.3106    3.1164          410
#> 138 2024-07-28 17:00:00 Blankenberge  51.3106    3.1164          427
#> 139 2024-07-28 18:00:00 Blankenberge  51.3106    3.1164          422
#> 140 2024-07-28 19:00:00 Blankenberge  51.3106    3.1164          382
#> 141 2024-07-28 20:00:00 Blankenberge  51.3106    3.1164          320
#> 142 2024-07-28 21:00:00 Blankenberge  51.3106    3.1164          236
#> 143 2024-07-28 22:00:00 Blankenberge  51.3106    3.1164          150
#> 144 2024-07-28 23:00:00 Blankenberge  51.3106    3.1164           76
#> 145 2024-07-29 00:00:00 Blankenberge  51.3106    3.1164           34
#> 146 2024-07-29 01:00:00 Blankenberge  51.3106    3.1164           73
#> 147 2024-07-29 02:00:00 Blankenberge  51.3106    3.1164          126
#> 148 2024-07-29 03:00:00 Blankenberge  51.3106    3.1164          209
#> 149 2024-07-29 04:00:00 Blankenberge  51.3106    3.1164          343
#> 150 2024-07-29 05:00:00 Blankenberge  51.3106    3.1164          403
#> 151 2024-07-29 06:00:00 Blankenberge  51.3106    3.1164          407
#> 152 2024-07-29 07:00:00 Blankenberge  51.3106    3.1164          396
#> 153 2024-07-29 08:00:00 Blankenberge  51.3106    3.1164          357
#> 154 2024-07-29 09:00:00 Blankenberge  51.3106    3.1164          290
#> 155 2024-07-29 10:00:00 Blankenberge  51.3106    3.1164          213
#> 156 2024-07-29 11:00:00 Blankenberge  51.3106    3.1164          139
#> 157 2024-07-29 12:00:00 Blankenberge  51.3106    3.1164           86
#> 158 2024-07-29 13:00:00 Blankenberge  51.3106    3.1164          109
#> 159 2024-07-29 14:00:00 Blankenberge  51.3106    3.1164          149
#> 160 2024-07-29 15:00:00 Blankenberge  51.3106    3.1164          209
#> 161 2024-07-29 16:00:00 Blankenberge  51.3106    3.1164          303
#> 162 2024-07-29 17:00:00 Blankenberge  51.3106    3.1164          390
#> 163 2024-07-29 18:00:00 Blankenberge  51.3106    3.1164          410
#> 164 2024-07-29 19:00:00 Blankenberge  51.3106    3.1164          403
#> 165 2024-07-29 20:00:00 Blankenberge  51.3106    3.1164          372
#> 166 2024-07-29 21:00:00 Blankenberge  51.3106    3.1164          312
#> 167 2024-07-29 22:00:00 Blankenberge  51.3106    3.1164          239
#> 168 2024-07-29 23:00:00 Blankenberge  51.3106    3.1164          165
#> 169 2024-07-30 00:00:00 Blankenberge  51.3106    3.1164           98
#> 170 2024-07-30 01:00:00 Blankenberge  51.3106    3.1164           66
#> 171 2024-07-30 02:00:00 Blankenberge  51.3106    3.1164           97
#> 172 2024-07-30 03:00:00 Blankenberge  51.3106    3.1164          147
#> 173 2024-07-30 04:00:00 Blankenberge  51.3106    3.1164          227
#> 174 2024-07-30 05:00:00 Blankenberge  51.3106    3.1164          337
#> 175 2024-07-30 06:00:00 Blankenberge  51.3106    3.1164          396
#> 176 2024-07-30 07:00:00 Blankenberge  51.3106    3.1164          406
#> 177 2024-07-30 08:00:00 Blankenberge  51.3106    3.1164          400
#> 178 2024-07-30 09:00:00 Blankenberge  51.3106    3.1164          363
#> 179 2024-07-30 10:00:00 Blankenberge  51.3106    3.1164          305
#> 180 2024-07-30 11:00:00 Blankenberge  51.3106    3.1164          239
#> 181 2024-07-30 12:00:00 Blankenberge  51.3106    3.1164          174
#> 182 2024-07-30 13:00:00 Blankenberge  51.3106    3.1164          120
#> 183 2024-07-30 14:00:00 Blankenberge  51.3106    3.1164          127
#> 184 2024-07-30 15:00:00 Blankenberge  51.3106    3.1164          164
#> 185 2024-07-30 16:00:00 Blankenberge  51.3106    3.1164          219
#> 186 2024-07-30 17:00:00 Blankenberge  51.3106    3.1164          303
#> 187 2024-07-30 18:00:00 Blankenberge  51.3106    3.1164          381
#> 188 2024-07-30 19:00:00 Blankenberge  51.3106    3.1164          413
#> 189 2024-07-30 20:00:00 Blankenberge  51.3106    3.1164          410
#> 190 2024-07-30 21:00:00 Blankenberge  51.3106    3.1164          387
#> 191 2024-07-30 22:00:00 Blankenberge  51.3106    3.1164          341
#> 192 2024-07-30 23:00:00 Blankenberge  51.3106    3.1164          275
#> 193 2024-07-31 00:00:00 Blankenberge  51.3106    3.1164          205
#> 194 2024-07-31 01:00:00 Blankenberge  51.3106    3.1164          143
#> 195 2024-07-31 02:00:00 Blankenberge  51.3106    3.1164           97
#> 196 2024-07-31 03:00:00 Blankenberge  51.3106    3.1164          114
#> 197 2024-07-31 04:00:00 Blankenberge  51.3106    3.1164          158
#> 198 2024-07-31 05:00:00 Blankenberge  51.3106    3.1164          217
#> 199 2024-07-31 06:00:00 Blankenberge  51.3106    3.1164          300
#> 200 2024-07-31 07:00:00 Blankenberge  51.3106    3.1164          377
#> 201 2024-07-31 08:00:00 Blankenberge  51.3106    3.1164          399
#> 202 2024-07-31 09:00:00 Blankenberge  51.3106    3.1164          395
#> 203 2024-07-31 10:00:00 Blankenberge  51.3106    3.1164          371
#> 204 2024-07-31 11:00:00 Blankenberge  51.3106    3.1164          320
#> 205 2024-07-31 12:00:00 Blankenberge  51.3106    3.1164          255
#> 206 2024-07-31 13:00:00 Blankenberge  51.3106    3.1164          188
#> 207 2024-07-31 14:00:00 Blankenberge  51.3106    3.1164          126
#> 208 2024-07-31 15:00:00 Blankenberge  51.3106    3.1164           96
#> 209 2024-07-31 16:00:00 Blankenberge  51.3106    3.1164          127
#> 210 2024-07-31 17:00:00 Blankenberge  51.3106    3.1164          178
#> 211 2024-07-31 18:00:00 Blankenberge  51.3106    3.1164          250
#> 212 2024-07-31 19:00:00 Blankenberge  51.3106    3.1164          343
#> 213 2024-07-31 20:00:00 Blankenberge  51.3106    3.1164          395
#> 214 2024-07-31 21:00:00 Blankenberge  51.3106    3.1164          401
#> 215 2024-07-31 22:00:00 Blankenberge  51.3106    3.1164          391
#> 216 2024-07-31 23:00:00 Blankenberge  51.3106    3.1164          352
#> 217 2024-08-01 00:00:00 Blankenberge  51.3106    3.1164          295
#> 218 2024-08-01 01:00:00 Blankenberge  51.3106    3.1164          229
#> 219 2024-08-01 02:00:00 Blankenberge  51.3106    3.1164          160
#> 220 2024-08-01 03:00:00 Blankenberge  51.3106    3.1164          107
#> 221 2024-08-01 04:00:00 Blankenberge  51.3106    3.1164           99
#> 222 2024-08-01 05:00:00 Blankenberge  51.3106    3.1164          140
#> 223 2024-08-01 06:00:00 Blankenberge  51.3106    3.1164          190
#> 224 2024-08-01 07:00:00 Blankenberge  51.3106    3.1164          270
#> 225 2024-08-01 08:00:00 Blankenberge  51.3106    3.1164          363
#> 226 2024-08-01 09:00:00 Blankenberge  51.3106    3.1164          403
#> 227 2024-08-01 10:00:00 Blankenberge  51.3106    3.1164          403
#> 228 2024-08-01 11:00:00 Blankenberge  51.3106    3.1164          386
#> 229 2024-08-01 12:00:00 Blankenberge  51.3106    3.1164          341
#> 230 2024-08-01 13:00:00 Blankenberge  51.3106    3.1164          277
#> 231 2024-08-01 14:00:00 Blankenberge  51.3106    3.1164          207
#> 232 2024-08-01 15:00:00 Blankenberge  51.3106    3.1164          139
#> 233 2024-08-01 16:00:00 Blankenberge  51.3106    3.1164           89
#> 234 2024-08-01 17:00:00 Blankenberge  51.3106    3.1164          115
#> 235 2024-08-01 18:00:00 Blankenberge  51.3106    3.1164          162
#> 236 2024-08-01 19:00:00 Blankenberge  51.3106    3.1164          234
#> 237 2024-08-01 20:00:00 Blankenberge  51.3106    3.1164          343
#> 238 2024-08-01 21:00:00 Blankenberge  51.3106    3.1164          410
#> 239 2024-08-01 22:00:00 Blankenberge  51.3106    3.1164          426
#> 240 2024-08-01 23:00:00 Blankenberge  51.3106    3.1164          416
#> 241 2024-08-02 00:00:00 Blankenberge  51.3106    3.1164          381
#> 242 2024-08-02 01:00:00 Blankenberge  51.3106    3.1164          328
#> 243 2024-08-02 02:00:00 Blankenberge  51.3106    3.1164          252
#> 244 2024-08-02 03:00:00 Blankenberge  51.3106    3.1164          176
#> 245 2024-08-02 04:00:00 Blankenberge  51.3106    3.1164          109
#> 246 2024-08-02 05:00:00 Blankenberge  51.3106    3.1164           95
#> 247 2024-08-02 06:00:00 Blankenberge  51.3106    3.1164          136
#> 248 2024-08-02 07:00:00 Blankenberge  51.3106    3.1164          187
#> 249 2024-08-02 08:00:00 Blankenberge  51.3106    3.1164          264
#> 250 2024-08-02 09:00:00 Blankenberge  51.3106    3.1164          374
#> 251 2024-08-02 10:00:00 Blankenberge  51.3106    3.1164          417
#> 252 2024-08-02 11:00:00 Blankenberge  51.3106    3.1164          418
#> 253 2024-08-02 12:00:00 Blankenberge  51.3106    3.1164          394
#> 254 2024-08-02 13:00:00 Blankenberge  51.3106    3.1164          346
#> 255 2024-08-02 14:00:00 Blankenberge  51.3106    3.1164          279
#> 256 2024-08-02 15:00:00 Blankenberge  51.3106    3.1164          205
#> 257 2024-08-02 16:00:00 Blankenberge  51.3106    3.1164          128
#> 258 2024-08-02 17:00:00 Blankenberge  51.3106    3.1164           73
#> 259 2024-08-02 18:00:00 Blankenberge  51.3106    3.1164          102
#> 260 2024-08-02 19:00:00 Blankenberge  51.3106    3.1164          152
#> 261 2024-08-02 20:00:00 Blankenberge  51.3106    3.1164          231
#> 262 2024-08-02 21:00:00 Blankenberge  51.3106    3.1164          357
#> 263 2024-08-02 22:00:00 Blankenberge  51.3106    3.1164          430
#> 264 2024-08-02 23:00:00 Blankenberge  51.3106    3.1164          439
#> 265 2024-08-03 00:00:00 Blankenberge  51.3106    3.1164          434
#> 266 2024-08-03 01:00:00 Blankenberge  51.3106    3.1164          390
#> 267 2024-08-03 02:00:00 Blankenberge  51.3106    3.1164          328
#> 268 2024-08-03 03:00:00 Blankenberge  51.3106    3.1164          251
#> 269 2024-08-03 04:00:00 Blankenberge  51.3106    3.1164          171
#> 270 2024-08-03 05:00:00 Blankenberge  51.3106    3.1164          102
#> 271 2024-08-03 06:00:00 Blankenberge  51.3106    3.1164          108
#> 272 2024-08-03 07:00:00 Blankenberge  51.3106    3.1164          148
#> 273 2024-08-03 08:00:00 Blankenberge  51.3106    3.1164          200
#> 274 2024-08-03 09:00:00 Blankenberge  51.3106    3.1164          286
#> 275 2024-08-03 10:00:00 Blankenberge  51.3106    3.1164          406
#> 276 2024-08-03 11:00:00 Blankenberge  51.3106    3.1164          441
#> 277 2024-08-03 12:00:00 Blankenberge  51.3106    3.1164          437
#> 278 2024-08-03 13:00:00 Blankenberge  51.3106    3.1164          405
#> 279 2024-08-03 14:00:00 Blankenberge  51.3106    3.1164          353
#> 280 2024-08-03 15:00:00 Blankenberge  51.3106    3.1164          277
#> 281 2024-08-03 16:00:00 Blankenberge  51.3106    3.1164          193
#> 282 2024-08-03 17:00:00 Blankenberge  51.3106    3.1164          116
#> 283 2024-08-03 18:00:00 Blankenberge  51.3106    3.1164           74
#> 284 2024-08-03 19:00:00 Blankenberge  51.3106    3.1164          117
#> 285 2024-08-03 20:00:00 Blankenberge  51.3106    3.1164          171
#> 286 2024-08-03 21:00:00 Blankenberge  51.3106    3.1164          256
#> 287 2024-08-03 22:00:00 Blankenberge  51.3106    3.1164          396
#> 288 2024-08-03 23:00:00 Blankenberge  51.3106    3.1164          458
#> 289 2024-08-04 00:00:00 Blankenberge  51.3106    3.1164          460
#> 290 2024-08-04 01:00:00 Blankenberge  51.3106    3.1164          446
#> 291 2024-08-04 02:00:00 Blankenberge  51.3106    3.1164          397
#> 292 2024-08-04 03:00:00 Blankenberge  51.3106    3.1164          322
#> 293 2024-08-04 04:00:00 Blankenberge  51.3106    3.1164          234
#> 294 2024-08-04 05:00:00 Blankenberge  51.3106    3.1164          150
#> 295 2024-08-04 06:00:00 Blankenberge  51.3106    3.1164           89
#> 296 2024-08-04 07:00:00 Blankenberge  51.3106    3.1164          123
#> 297 2024-08-04 08:00:00 Blankenberge  51.3106    3.1164          166
#> 298 2024-08-04 09:00:00 Blankenberge  51.3106    3.1164          218
#> 299 2024-08-04 10:00:00 Blankenberge  51.3106    3.1164          318
#> 300 2024-08-04 11:00:00 Blankenberge  51.3106    3.1164          429
#> 301 2024-08-04 12:00:00 Blankenberge  51.3106    3.1164          447
#> 302 2024-08-04 13:00:00 Blankenberge  51.3106    3.1164          434
#> 303 2024-08-04 14:00:00 Blankenberge  51.3106    3.1164          394
#> 304 2024-08-04 15:00:00 Blankenberge  51.3106    3.1164          330
#> 305 2024-08-04 16:00:00 Blankenberge  51.3106    3.1164          240
#> 306 2024-08-04 17:00:00 Blankenberge  51.3106    3.1164          150
#> 307 2024-08-04 18:00:00 Blankenberge  51.3106    3.1164           70
#> 308 2024-08-04 19:00:00 Blankenberge  51.3106    3.1164           55
#> 309 2024-08-04 20:00:00 Blankenberge  51.3106    3.1164          104
#> 310 2024-08-04 21:00:00 Blankenberge  51.3106    3.1164          171
#> 311 2024-08-04 22:00:00 Blankenberge  51.3106    3.1164          282
#> 312 2024-08-04 23:00:00 Blankenberge  51.3106    3.1164          426
#> 313 2024-08-05 00:00:00 Blankenberge  51.3106    3.1164          461
#> 314 2024-08-05 01:00:00 Blankenberge  51.3106    3.1164          459
#> 315 2024-08-05 02:00:00 Blankenberge  51.3106    3.1164          426
#> 316 2024-08-05 03:00:00 Blankenberge  51.3106    3.1164          369
#> 317 2024-08-05 04:00:00 Blankenberge  51.3106    3.1164          284
#> 318 2024-08-05 05:00:00 Blankenberge  51.3106    3.1164          195
#> 319 2024-08-05 06:00:00 Blankenberge  51.3106    3.1164          109
#> 320 2024-08-05 07:00:00 Blankenberge  51.3106    3.1164           87
#> 321 2024-08-05 08:00:00 Blankenberge  51.3106    3.1164          125
#> 322 2024-08-05 09:00:00 Blankenberge  51.3106    3.1164          171
#> 323 2024-08-05 10:00:00 Blankenberge  51.3106    3.1164          241
#> 324 2024-08-05 11:00:00 Blankenberge  51.3106    3.1164          378
#> 325 2024-08-05 12:00:00 Blankenberge  51.3106    3.1164          449
#> 326 2024-08-05 13:00:00 Blankenberge  51.3106    3.1164          451
#> 327 2024-08-05 14:00:00 Blankenberge  51.3106    3.1164          428
#> 328 2024-08-05 15:00:00 Blankenberge  51.3106    3.1164          374
#> 329 2024-08-05 16:00:00 Blankenberge  51.3106    3.1164          295
#> 330 2024-08-05 17:00:00 Blankenberge  51.3106    3.1164          203
#> 331 2024-08-05 18:00:00 Blankenberge  51.3106    3.1164          112
#> 332 2024-08-05 19:00:00 Blankenberge  51.3106    3.1164           43
#> 333 2024-08-05 20:00:00 Blankenberge  51.3106    3.1164           68
#> 334 2024-08-05 21:00:00 Blankenberge  51.3106    3.1164          121
#> 335 2024-08-05 22:00:00 Blankenberge  51.3106    3.1164          199
#> 336 2024-08-05 23:00:00 Blankenberge  51.3106    3.1164          351
#> 337 2024-08-06 00:00:00 Blankenberge  51.3106    3.1164          453
#> 338 2024-08-06 01:00:00 Blankenberge  51.3106    3.1164          466
#> 339 2024-08-06 02:00:00 Blankenberge  51.3106    3.1164          455
#> 340 2024-08-06 03:00:00 Blankenberge  51.3106    3.1164          415
#> 341 2024-08-06 04:00:00 Blankenberge  51.3106    3.1164          340
#> 342 2024-08-06 05:00:00 Blankenberge  51.3106    3.1164          250
#> 343 2024-08-06 06:00:00 Blankenberge  51.3106    3.1164          161
#> 344 2024-08-06 14:00:00 Blankenberge  51.3106    3.1164          418
#> 345 2024-08-06 15:00:00 Blankenberge  51.3106    3.1164          416
#> 346 2024-08-06 16:00:00 Blankenberge  51.3106    3.1164          354
#> 347 2024-08-06 17:00:00 Blankenberge  51.3106    3.1164          264
#> 348 2024-08-06 18:00:00 Blankenberge  51.3106    3.1164          172
#> 349 2024-08-06 19:00:00 Blankenberge  51.3106    3.1164           84
#> 350 2024-08-06 20:00:00 Blankenberge  51.3106    3.1164           46
#> 351 2024-08-06 21:00:00 Blankenberge  51.3106    3.1164          103
#> 352 2024-08-06 22:00:00 Blankenberge  51.3106    3.1164          160
#> 353 2024-08-06 23:00:00 Blankenberge  51.3106    3.1164          259
#> 354 2024-08-07 00:00:00 Blankenberge  51.3106    3.1164          425
#> 355 2024-08-07 01:00:00 Blankenberge  51.3106    3.1164          472
#> 356 2024-08-07 02:00:00 Blankenberge  51.3106    3.1164          475
#> 357 2024-08-07 03:00:00 Blankenberge  51.3106    3.1164          449
#> 358 2024-08-07 04:00:00 Blankenberge  51.3106    3.1164          394
#> 359 2024-08-07 05:00:00 Blankenberge  51.3106    3.1164          313
#> 360 2024-08-07 06:00:00 Blankenberge  51.3106    3.1164          217
#> 361 2024-08-07 13:00:00 Blankenberge  51.3106    3.1164          460
#> 362 2024-08-07 14:00:00 Blankenberge  51.3106    3.1164          466
#> 363 2024-08-07 15:00:00 Blankenberge  51.3106    3.1164          444
#> 364 2024-08-07 16:00:00 Blankenberge  51.3106    3.1164          391
#> 365 2024-08-07 17:00:00 Blankenberge  51.3106    3.1164          315
#> 366 2024-08-07 18:00:00 Blankenberge  51.3106    3.1164          221
#> 367 2024-08-07 19:00:00 Blankenberge  51.3106    3.1164          128
#> 368 2024-08-07 20:00:00 Blankenberge  51.3106    3.1164           58
#> 369 2024-08-07 21:00:00 Blankenberge  51.3106    3.1164           67
#> 370 2024-08-07 22:00:00 Blankenberge  51.3106    3.1164          118
#> 371 2024-08-07 23:00:00 Blankenberge  51.3106    3.1164          190
#> 372 2024-08-08 00:00:00 Blankenberge  51.3106    3.1164          328
#> 373 2024-08-08 01:00:00 Blankenberge  51.3106    3.1164          444
#> 374 2024-08-08 02:00:00 Blankenberge  51.3106    3.1164          463
#> 375 2024-08-08 03:00:00 Blankenberge  51.3106    3.1164          454
#> 376 2024-08-08 04:00:00 Blankenberge  51.3106    3.1164          415
#> 377 2024-08-08 05:00:00 Blankenberge  51.3106    3.1164          347
#> 378 2024-08-08 06:00:00 Blankenberge  51.3106    3.1164          258
#> 379 2024-08-08 07:00:00 Blankenberge  51.3106    3.1164          167
#> 380 2024-08-08 08:00:00 Blankenberge  51.3106    3.1164           94
#> 381 2024-08-08 09:00:00 Blankenberge  51.3106    3.1164          115
#> 382 2024-08-08 10:00:00 Blankenberge  51.3106    3.1164          153
#> 383 2024-08-08 11:00:00 Blankenberge  51.3106    3.1164          207
#> 384 2024-08-08 12:00:00 Blankenberge  51.3106    3.1164          292
#> 385 2024-08-08 13:00:00 Blankenberge  51.3106    3.1164          432
#> 386 2024-08-08 14:00:00 Blankenberge  51.3106    3.1164          464
#> 387 2024-08-08 15:00:00 Blankenberge  51.3106    3.1164          458
#> 388 2024-08-08 16:00:00 Blankenberge  51.3106    3.1164          417
#> 389 2024-08-08 17:00:00 Blankenberge  51.3106    3.1164          355
#> 390 2024-08-08 18:00:00 Blankenberge  51.3106    3.1164          267
#> 391 2024-08-08 19:00:00 Blankenberge  51.3106    3.1164          174
#> 392 2024-08-08 20:00:00 Blankenberge  51.3106    3.1164           91
#> 393 2024-08-08 21:00:00 Blankenberge  51.3106    3.1164           37
#> 394 2024-08-08 22:00:00 Blankenberge  51.3106    3.1164           86
#> 395 2024-08-08 23:00:00 Blankenberge  51.3106    3.1164          144
#> 396 2024-08-09 00:00:00 Blankenberge  51.3106    3.1164          237
#> 397 2024-08-09 01:00:00 Blankenberge  51.3106    3.1164          390
#> 398 2024-08-09 02:00:00 Blankenberge  51.3106    3.1164          436
#> 399 2024-08-09 03:00:00 Blankenberge  51.3106    3.1164          437
#> 400 2024-08-09 04:00:00 Blankenberge  51.3106    3.1164          417
#> 401 2024-08-09 05:00:00 Blankenberge  51.3106    3.1164          363
#> 402 2024-08-09 06:00:00 Blankenberge  51.3106    3.1164          286
#> 403 2024-08-09 07:00:00 Blankenberge  51.3106    3.1164          207
#> 404 2024-08-09 08:00:00 Blankenberge  51.3106    3.1164          131
#> 405 2024-08-09 09:00:00 Blankenberge  51.3106    3.1164           95
#> 406 2024-08-09 10:00:00 Blankenberge  51.3106    3.1164          136
#> 407 2024-08-09 11:00:00 Blankenberge  51.3106    3.1164          184
#> 408 2024-08-09 12:00:00 Blankenberge  51.3106    3.1164          252
#> 409 2024-08-09 13:00:00 Blankenberge  51.3106    3.1164          379
#> 410 2024-08-09 14:00:00 Blankenberge  51.3106    3.1164          462
#> 411 2024-08-09 15:00:00 Blankenberge  51.3106    3.1164          467
#> 412 2024-08-09 16:00:00 Blankenberge  51.3106    3.1164          447
#> 413 2024-08-09 17:00:00 Blankenberge  51.3106    3.1164          399
#> 414 2024-08-09 18:00:00 Blankenberge  51.3106    3.1164          327
#> 415 2024-08-09 19:00:00 Blankenberge  51.3106    3.1164          235
#> 416 2024-08-09 20:00:00 Blankenberge  51.3106    3.1164          151
#> 417 2024-08-09 21:00:00 Blankenberge  51.3106    3.1164           81
#> 418 2024-08-09 22:00:00 Blankenberge  51.3106    3.1164           93
#> 419 2024-08-09 23:00:00 Blankenberge  51.3106    3.1164          147
#> 420 2024-08-10 00:00:00 Blankenberge  51.3106    3.1164          212
#> 421 2024-08-10 01:00:00 Blankenberge  51.3106    3.1164          331
#> 422 2024-08-10 02:00:00 Blankenberge  51.3106    3.1164          440
#> 423 2024-08-10 03:00:00 Blankenberge  51.3106    3.1164          455
#> 424 2024-08-10 04:00:00 Blankenberge  51.3106    3.1164          448
#> 425 2024-08-10 05:00:00 Blankenberge  51.3106    3.1164          410
#> 426 2024-08-10 06:00:00 Blankenberge  51.3106    3.1164          339
#> 427 2024-08-10 07:00:00 Blankenberge  51.3106    3.1164          255
#> 428 2024-08-10 08:00:00 Blankenberge  51.3106    3.1164          171
#> 429 2024-08-10 09:00:00 Blankenberge  51.3106    3.1164          101
#> 430 2024-08-10 10:00:00 Blankenberge  51.3106    3.1164          116
#> 431 2024-08-10 11:00:00 Blankenberge  51.3106    3.1164          158
#> 432 2024-08-10 12:00:00 Blankenberge  51.3106    3.1164          201
#> 433 2024-08-10 13:00:00 Blankenberge  51.3106    3.1164          286
#> 434 2024-08-10 14:00:00 Blankenberge  51.3106    3.1164          406
#> 435 2024-08-10 15:00:00 Blankenberge  51.3106    3.1164          437
#> 436 2024-08-10 16:00:00 Blankenberge  51.3106    3.1164          434
#> 437 2024-08-10 17:00:00 Blankenberge  51.3106    3.1164          399
#> 438 2024-08-10 18:00:00 Blankenberge  51.3106    3.1164          345
#> 439 2024-08-10 19:00:00 Blankenberge  51.3106    3.1164          261
#> 440 2024-08-10 20:00:00 Blankenberge  51.3106    3.1164          178
#> 441 2024-08-10 21:00:00 Blankenberge  51.3106    3.1164           99
#> 442 2024-08-10 22:00:00 Blankenberge  51.3106    3.1164           62
#> 443 2024-08-10 23:00:00 Blankenberge  51.3106    3.1164          105
#> 444 2024-08-11 00:00:00 Blankenberge  51.3106    3.1164          158
#> 445 2024-08-11 01:00:00 Blankenberge  51.3106    3.1164          243
#> 446 2024-08-11 02:00:00 Blankenberge  51.3106    3.1164          373
#> 447 2024-08-11 03:00:00 Blankenberge  51.3106    3.1164          430
#> 448 2024-08-11 04:00:00 Blankenberge  51.3106    3.1164          432
#> 449 2024-08-11 05:00:00 Blankenberge  51.3106    3.1164          416
#> 450 2024-08-11 06:00:00 Blankenberge  51.3106    3.1164          368
#> 451 2024-08-11 07:00:00 Blankenberge  51.3106    3.1164          293
#> 452 2024-08-11 08:00:00 Blankenberge  51.3106    3.1164          210
#> 453 2024-08-11 09:00:00 Blankenberge  51.3106    3.1164          133
#> 454 2024-08-11 10:00:00 Blankenberge  51.3106    3.1164           93
#> 455 2024-08-11 11:00:00 Blankenberge  51.3106    3.1164          126
#> 456 2024-08-11 12:00:00 Blankenberge  51.3106    3.1164          167
#> 457 2024-08-11 13:00:00 Blankenberge  51.3106    3.1164          221
#> 458 2024-08-11 14:00:00 Blankenberge  51.3106    3.1164          322
#> 459 2024-08-11 15:00:00 Blankenberge  51.3106    3.1164          405
#> 460 2024-08-11 16:00:00 Blankenberge  51.3106    3.1164          412
#> 461 2024-08-11 17:00:00 Blankenberge  51.3106    3.1164          402
#> 462 2024-08-11 18:00:00 Blankenberge  51.3106    3.1164          357
#> 463 2024-08-11 19:00:00 Blankenberge  51.3106    3.1164          291
#> 464 2024-08-11 20:00:00 Blankenberge  51.3106    3.1164          206
#> 465 2024-08-11 21:00:00 Blankenberge  51.3106    3.1164          128
#> 466 2024-08-11 22:00:00 Blankenberge  51.3106    3.1164           72
#> 467 2024-08-11 23:00:00 Blankenberge  51.3106    3.1164           71
#> 468 2024-08-12 00:00:00 Blankenberge  51.3106    3.1164          109
#> 469 2024-08-12 01:00:00 Blankenberge  51.3106    3.1164          174
#> 470 2024-08-12 02:00:00 Blankenberge  51.3106    3.1164          277
#> 471 2024-08-12 03:00:00 Blankenberge  51.3106    3.1164          377
#> 472 2024-08-12 04:00:00 Blankenberge  51.3106    3.1164          399
#> 473 2024-08-12 05:00:00 Blankenberge  51.3106    3.1164          397
#> 474 2024-08-12 06:00:00 Blankenberge  51.3106    3.1164          370
#> 475 2024-08-12 07:00:00 Blankenberge  51.3106    3.1164          318
#> 476 2024-08-12 08:00:00 Blankenberge  51.3106    3.1164          245
#> 477 2024-08-12 09:00:00 Blankenberge  51.3106    3.1164          174
#> 478 2024-08-12 10:00:00 Blankenberge  51.3106    3.1164          115
#> 479 2024-08-12 11:00:00 Blankenberge  51.3106    3.1164          100
#> 480 2024-08-12 12:00:00 Blankenberge  51.3106    3.1164          129
#> 481 2024-08-12 13:00:00 Blankenberge  51.3106    3.1164          177
#> 482 2024-08-12 14:00:00 Blankenberge  51.3106    3.1164          259
#> 483 2024-08-12 15:00:00 Blankenberge  51.3106    3.1164          359
#> 484 2024-08-12 16:00:00 Blankenberge  51.3106    3.1164          394
#> 485 2024-08-12 17:00:00 Blankenberge  51.3106    3.1164          395
#> 486 2024-08-12 18:00:00 Blankenberge  51.3106    3.1164          370
#> 487 2024-08-12 19:00:00 Blankenberge  51.3106    3.1164          321
#> 488 2024-08-12 20:00:00 Blankenberge  51.3106    3.1164          258
#> 489 2024-08-12 21:00:00 Blankenberge  51.3106    3.1164          193
#> 490 2024-08-12 22:00:00 Blankenberge  51.3106    3.1164          122
#> 491 2024-08-12 23:00:00 Blankenberge  51.3106    3.1164           82
#> 492 2024-08-13 00:00:00 Blankenberge  51.3106    3.1164          108
#> 493 2024-08-13 01:00:00 Blankenberge  51.3106    3.1164          166
#> 494 2024-08-13 02:00:00 Blankenberge  51.3106    3.1164          243
#> 495 2024-08-13 03:00:00 Blankenberge  51.3106    3.1164          347
#> 496 2024-08-13 04:00:00 Blankenberge  51.3106    3.1164          399
#> 497 2024-08-13 05:00:00 Blankenberge  51.3106    3.1164          417
#> 498 2024-08-13 06:00:00 Blankenberge  51.3106    3.1164          407
#> 499 2024-08-13 07:00:00 Blankenberge  51.3106    3.1164          377
#> 500 2024-08-13 08:00:00 Blankenberge  51.3106    3.1164          328
#> 501 2024-08-13 09:00:00 Blankenberge  51.3106    3.1164          263
#> 502 2024-08-13 10:00:00 Blankenberge  51.3106    3.1164          203
#> 503 2024-08-13 11:00:00 Blankenberge  51.3106    3.1164          159
#> 504 2024-08-13 12:00:00 Blankenberge  51.3106    3.1164          165
#> 505 2024-08-13 13:00:00 Blankenberge  51.3106    3.1164          192
#> 506 2024-08-13 14:00:00 Blankenberge  51.3106    3.1164          242
#> 507 2024-08-13 15:00:00 Blankenberge  51.3106    3.1164          313
#> 508 2024-08-13 16:00:00 Blankenberge  51.3106    3.1164          374
#> 509 2024-08-13 17:00:00 Blankenberge  51.3106    3.1164          393
#> 510 2024-08-13 18:00:00 Blankenberge  51.3106    3.1164          387
#> 511 2024-08-13 19:00:00 Blankenberge  51.3106    3.1164          359
#> 512 2024-08-13 20:00:00 Blankenberge  51.3106    3.1164          316
#> 513 2024-08-13 21:00:00 Blankenberge  51.3106    3.1164          255
#> 514 2024-08-13 22:00:00 Blankenberge  51.3106    3.1164          189
#> 515 2024-08-13 23:00:00 Blankenberge  51.3106    3.1164          133
#> 516 2024-08-14 00:00:00 Blankenberge  51.3106    3.1164          101
#> 517 2024-08-14 01:00:00 Blankenberge  51.3106    3.1164          126
#> 518 2024-08-14 02:00:00 Blankenberge  51.3106    3.1164          174
#> 519 2024-08-14 03:00:00 Blankenberge  51.3106    3.1164          243
#> 520 2024-08-14 04:00:00 Blankenberge  51.3106    3.1164          322
#> 521 2024-08-14 05:00:00 Blankenberge  51.3106    3.1164          365
#> 522 2024-08-14 06:00:00 Blankenberge  51.3106    3.1164          381
#> 523 2024-08-14 07:00:00 Blankenberge  51.3106    3.1164          372
#> 524 2024-08-14 08:00:00 Blankenberge  51.3106    3.1164          350
#> 525 2024-08-14 09:00:00 Blankenberge  51.3106    3.1164          309
#> 526 2024-08-14 10:00:00 Blankenberge  51.3106    3.1164          254
#> 527 2024-08-14 11:00:00 Blankenberge  51.3106    3.1164          198
#> 528 2024-08-14 12:00:00 Blankenberge  51.3106    3.1164          173
#> 529 2024-08-14 13:00:00 Blankenberge  51.3106    3.1164          180
#> 530 2024-08-14 14:00:00 Blankenberge  51.3106    3.1164          208
#> 531 2024-08-14 15:00:00 Blankenberge  51.3106    3.1164          254
#> 532 2024-08-14 16:00:00 Blankenberge  51.3106    3.1164          308
#> 533 2024-08-14 17:00:00 Blankenberge  51.3106    3.1164          361
#> 534 2024-08-14 18:00:00 Blankenberge  51.3106    3.1164          377
#> 535 2024-08-14 19:00:00 Blankenberge  51.3106    3.1164          376
#> 536 2024-08-14 20:00:00 Blankenberge  51.3106    3.1164          356
#> 537 2024-08-14 21:00:00 Blankenberge  51.3106    3.1164          323
#> 538 2024-08-14 22:00:00 Blankenberge  51.3106    3.1164          262
#> 539 2024-08-14 23:00:00 Blankenberge  51.3106    3.1164          205
#> 540 2024-08-15 00:00:00 Blankenberge  51.3106    3.1164          154
#> 541 2024-08-15 01:00:00 Blankenberge  51.3106    3.1164          115
#> 542 2024-08-15 02:00:00 Blankenberge  51.3106    3.1164          120
#> 543 2024-08-15 03:00:00 Blankenberge  51.3106    3.1164          152
#> 544 2024-08-15 04:00:00 Blankenberge  51.3106    3.1164          208
#> 545 2024-08-15 05:00:00 Blankenberge  51.3106    3.1164          274
#> 546 2024-08-15 06:00:00 Blankenberge  51.3106    3.1164          326
#> 547 2024-08-15 07:00:00 Blankenberge  51.3106    3.1164          346
#> 548 2024-08-15 08:00:00 Blankenberge  51.3106    3.1164          343
#> 549 2024-08-15 09:00:00 Blankenberge  51.3106    3.1164          326
#> 550 2024-08-15 10:00:00 Blankenberge  51.3106    3.1164          291
#> 551 2024-08-15 11:00:00 Blankenberge  51.3106    3.1164          246
#> 552 2024-08-15 12:00:00 Blankenberge  51.3106    3.1164          203
#> 553 2024-08-15 13:00:00 Blankenberge  51.3106    3.1164          159
#> 554 2024-08-15 14:00:00 Blankenberge  51.3106    3.1164          132
#> 555 2024-08-15 15:00:00 Blankenberge  51.3106    3.1164          159
#> 556 2024-08-15 16:00:00 Blankenberge  51.3106    3.1164          205
#> 557 2024-08-15 17:00:00 Blankenberge  51.3106    3.1164          265
#> 558 2024-08-15 18:00:00 Blankenberge  51.3106    3.1164          331
#> 559 2024-08-15 19:00:00 Blankenberge  51.3106    3.1164          374
#> 560 2024-08-15 20:00:00 Blankenberge  51.3106    3.1164          379
#> 561 2024-08-15 21:00:00 Blankenberge  51.3106    3.1164          375
#> 562 2024-08-15 22:00:00 Blankenberge  51.3106    3.1164          352
#> 563 2024-08-15 23:00:00 Blankenberge  51.3106    3.1164          313
#> 564 2024-08-16 00:00:00 Blankenberge  51.3106    3.1164          272
#> 565 2024-08-16 01:00:00 Blankenberge  51.3106    3.1164          217
#> 566 2024-08-16 02:00:00 Blankenberge  51.3106    3.1164          172
#> 567 2024-08-16 03:00:00 Blankenberge  51.3106    3.1164          158
#> 568 2024-08-16 04:00:00 Blankenberge  51.3106    3.1164          192
#> 569 2024-08-16 05:00:00 Blankenberge  51.3106    3.1164          233
#> 570 2024-08-16 06:00:00 Blankenberge  51.3106    3.1164          288
#> 571 2024-08-16 07:00:00 Blankenberge  51.3106    3.1164          355
#> 572 2024-08-16 08:00:00 Blankenberge  51.3106    3.1164          395
#> 573 2024-08-16 09:00:00 Blankenberge  51.3106    3.1164          401
#> 574 2024-08-16 10:00:00 Blankenberge  51.3106    3.1164          390
#> 575 2024-08-16 11:00:00 Blankenberge  51.3106    3.1164          358
#> 576 2024-08-16 12:00:00 Blankenberge  51.3106    3.1164          307
#> 577 2024-08-16 13:00:00 Blankenberge  51.3106    3.1164          240
#> 578 2024-08-16 14:00:00 Blankenberge  51.3106    3.1164          177
#> 579 2024-08-16 15:00:00 Blankenberge  51.3106    3.1164          124
#> 580 2024-08-16 16:00:00 Blankenberge  51.3106    3.1164          121
#> 581 2024-08-16 17:00:00 Blankenberge  51.3106    3.1164          155
#> 582 2024-08-16 18:00:00 Blankenberge  51.3106    3.1164          211
#> 583 2024-08-16 19:00:00 Blankenberge  51.3106    3.1164          285
#> 584 2024-08-16 20:00:00 Blankenberge  51.3106    3.1164          366
#> 585 2024-08-16 21:00:00 Blankenberge  51.3106    3.1164          398
#> 586 2024-08-16 22:00:00 Blankenberge  51.3106    3.1164          398
#> 587 2024-08-16 23:00:00 Blankenberge  51.3106    3.1164          375
#> 588 2024-08-17 00:00:00 Blankenberge  51.3106    3.1164          328
#> 589 2024-08-17 01:00:00 Blankenberge  51.3106    3.1164          272
#> 590 2024-08-17 02:00:00 Blankenberge  51.3106    3.1164          203
#> 591 2024-08-17 03:00:00 Blankenberge  51.3106    3.1164          139
#> 592 2024-08-17 04:00:00 Blankenberge  51.3106    3.1164          107
#> 593 2024-08-17 05:00:00 Blankenberge  51.3106    3.1164          142
#> 594 2024-08-17 06:00:00 Blankenberge  51.3106    3.1164          183
#> 595 2024-08-17 07:00:00 Blankenberge  51.3106    3.1164          244
#> 596 2024-08-17 08:00:00 Blankenberge  51.3106    3.1164          336
#> 597 2024-08-17 09:00:00 Blankenberge  51.3106    3.1164          401
#> 598 2024-08-17 10:00:00 Blankenberge  51.3106    3.1164          411
#> 599 2024-08-17 11:00:00 Blankenberge  51.3106    3.1164          406
#> 600 2024-08-17 12:00:00 Blankenberge  51.3106    3.1164          368
#> 601 2024-08-17 13:00:00 Blankenberge  51.3106    3.1164          316
#> 602 2024-08-17 14:00:00 Blankenberge  51.3106    3.1164          247
#> 603 2024-08-17 15:00:00 Blankenberge  51.3106    3.1164          175
#> 604 2024-08-17 16:00:00 Blankenberge  51.3106    3.1164          113
#> 605 2024-08-17 17:00:00 Blankenberge  51.3106    3.1164          117
#> 606 2024-08-17 18:00:00 Blankenberge  51.3106    3.1164          157
#> 607 2024-08-17 19:00:00 Blankenberge  51.3106    3.1164          212
#> 608 2024-08-17 20:00:00 Blankenberge  51.3106    3.1164          306
#> 609 2024-08-17 21:00:00 Blankenberge  51.3106    3.1164          411
#> 610 2024-08-17 22:00:00 Blankenberge  51.3106    3.1164          440
#> 611 2024-08-17 23:00:00 Blankenberge  51.3106    3.1164          440
#> 612 2024-08-18 00:00:00 Blankenberge  51.3106    3.1164          410
#> 613 2024-08-18 01:00:00 Blankenberge  51.3106    3.1164          360
#> 614 2024-08-18 02:00:00 Blankenberge  51.3106    3.1164          286
#> 615 2024-08-18 03:00:00 Blankenberge  51.3106    3.1164          203
#> 616 2024-08-18 04:00:00 Blankenberge  51.3106    3.1164          125
#> 617 2024-08-18 05:00:00 Blankenberge  51.3106    3.1164           87
#> 618 2024-08-18 06:00:00 Blankenberge  51.3106    3.1164          129
#> 619 2024-08-18 07:00:00 Blankenberge  51.3106    3.1164          174
#> 620 2024-08-18 08:00:00 Blankenberge  51.3106    3.1164          238
#> 621 2024-08-18 09:00:00 Blankenberge  51.3106    3.1164          355
#> 622 2024-08-18 10:00:00 Blankenberge  51.3106    3.1164          428
#> 623 2024-08-18 11:00:00 Blankenberge  51.3106    3.1164          434
#> 624 2024-08-18 12:00:00 Blankenberge  51.3106    3.1164          420
#> 625 2024-08-18 13:00:00 Blankenberge  51.3106    3.1164          376
#> 626 2024-08-18 14:00:00 Blankenberge  51.3106    3.1164          309
#> 627 2024-08-18 15:00:00 Blankenberge  51.3106    3.1164          229
#> 628 2024-08-18 16:00:00 Blankenberge  51.3106    3.1164          147
#> 629 2024-08-18 17:00:00 Blankenberge  51.3106    3.1164           84
#> 630 2024-08-18 18:00:00 Blankenberge  51.3106    3.1164          105
#> 631 2024-08-18 19:00:00 Blankenberge  51.3106    3.1164          152
#> 632 2024-08-18 20:00:00 Blankenberge  51.3106    3.1164          213
#> 633 2024-08-18 21:00:00 Blankenberge  51.3106    3.1164          329
#> 634 2024-08-18 22:00:00 Blankenberge  51.3106    3.1164          449
#> 635 2024-08-18 23:00:00 Blankenberge  51.3106    3.1164          469
#> 636 2024-08-19 00:00:00 Blankenberge  51.3106    3.1164          465
#> 637 2024-08-19 01:00:00 Blankenberge  51.3106    3.1164          429
#> 638 2024-08-19 02:00:00 Blankenberge  51.3106    3.1164          365
#> 639 2024-08-19 03:00:00 Blankenberge  51.3106    3.1164          278
#> 640 2024-08-19 04:00:00 Blankenberge  51.3106    3.1164          183
#> 641 2024-08-19 05:00:00 Blankenberge  51.3106    3.1164          101
#> 642 2024-08-19 06:00:00 Blankenberge  51.3106    3.1164           70
#> 643 2024-08-19 11:00:00 Blankenberge  51.3106    3.1164          448
#> 644 2024-08-19 12:00:00 Blankenberge  51.3106    3.1164          449
#> 645 2024-08-19 13:00:00 Blankenberge  51.3106    3.1164          419
#> 646 2024-08-19 14:00:00 Blankenberge  51.3106    3.1164          364
#> 647 2024-08-19 15:00:00 Blankenberge  51.3106    3.1164          283
#> 648 2024-08-19 16:00:00 Blankenberge  51.3106    3.1164          192
#> 649 2024-08-19 17:00:00 Blankenberge  51.3106    3.1164           97
#> 650 2024-08-19 18:00:00 Blankenberge  51.3106    3.1164           32
#> 651 2024-08-19 19:00:00 Blankenberge  51.3106    3.1164           64
#> 652 2024-08-19 20:00:00 Blankenberge  51.3106    3.1164          111
#> 653 2024-08-19 21:00:00 Blankenberge  51.3106    3.1164          188
#> 654 2024-08-19 22:00:00 Blankenberge  51.3106    3.1164          349
#> 655 2024-08-19 23:00:00 Blankenberge  51.3106    3.1164          452
#> 656 2024-08-20 00:00:00 Blankenberge  51.3106    3.1164          461
#> 657 2024-08-20 01:00:00 Blankenberge  51.3106    3.1164          449
#> 658 2024-08-20 02:00:00 Blankenberge  51.3106    3.1164          403
#> 659 2024-08-20 03:00:00 Blankenberge  51.3106    3.1164          327
#> 660 2024-08-20 04:00:00 Blankenberge  51.3106    3.1164          235
#> 661 2024-08-20 05:00:00 Blankenberge  51.3106    3.1164          137
#> 662 2024-08-20 06:00:00 Blankenberge  51.3106    3.1164           60
#> 663 2024-08-20 07:00:00 Blankenberge  51.3106    3.1164           78
#> 664 2024-08-20 08:00:00 Blankenberge  51.3106    3.1164          119
#> 665 2024-08-20 09:00:00 Blankenberge  51.3106    3.1164          176
#> 666 2024-08-20 10:00:00 Blankenberge  51.3106    3.1164          277
#> 667 2024-08-20 11:00:00 Blankenberge  51.3106    3.1164          440
#> 668 2024-08-20 12:00:00 Blankenberge  51.3106    3.1164          463
#> 669 2024-08-20 13:00:00 Blankenberge  51.3106    3.1164          460
#> 670 2024-08-20 14:00:00 Blankenberge  51.3106    3.1164          420
#> 671 2024-08-20 15:00:00 Blankenberge  51.3106    3.1164          357
#> 672 2024-08-20 16:00:00 Blankenberge  51.3106    3.1164          264
#> 673 2024-08-20 17:00:00 Blankenberge  51.3106    3.1164          171
#> 674 2024-08-20 18:00:00 Blankenberge  51.3106    3.1164           91
#> 675 2024-08-20 19:00:00 Blankenberge  51.3106    3.1164           48
#> 676 2024-08-20 20:00:00 Blankenberge  51.3106    3.1164          107
#> 677 2024-08-20 21:00:00 Blankenberge  51.3106    3.1164          168
#> 678 2024-08-20 22:00:00 Blankenberge  51.3106    3.1164          270
#> 679 2024-08-20 23:00:00 Blankenberge  51.3106    3.1164          450
#> 680 2024-08-21 00:00:00 Blankenberge  51.3106    3.1164          513
#> 681 2024-08-21 01:00:00 Blankenberge  51.3106    3.1164          517
#> 682 2024-08-21 02:00:00 Blankenberge  51.3106    3.1164          493
#> 683 2024-08-21 03:00:00 Blankenberge  51.3106    3.1164          438
#> 684 2024-08-21 04:00:00 Blankenberge  51.3106    3.1164          346
#> 685 2024-08-21 05:00:00 Blankenberge  51.3106    3.1164          245
#> 686 2024-08-21 06:00:00 Blankenberge  51.3106    3.1164          144
#> 687 2024-08-21 07:00:00 Blankenberge  51.3106    3.1164           76
#> 688 2024-08-21 08:00:00 Blankenberge  51.3106    3.1164          136
#> 689 2024-08-21 09:00:00 Blankenberge  51.3106    3.1164          178
#> 690 2024-08-21 10:00:00 Blankenberge  51.3106    3.1164          225
#> 691 2024-08-21 11:00:00 Blankenberge  51.3106    3.1164          330
#> 692 2024-08-21 12:00:00 Blankenberge  51.3106    3.1164          486
#> 693 2024-08-21 13:00:00 Blankenberge  51.3106    3.1164          494
#> 694 2024-08-21 14:00:00 Blankenberge  51.3106    3.1164          480
#> 695 2024-08-21 15:00:00 Blankenberge  51.3106    3.1164          427
#> 696 2024-08-21 16:00:00 Blankenberge  51.3106    3.1164          350
#> 697 2024-08-21 17:00:00 Blankenberge  51.3106    3.1164          246
#> 698 2024-08-21 18:00:00 Blankenberge  51.3106    3.1164          140
#> 699 2024-08-21 19:00:00 Blankenberge  51.3106    3.1164           48
#> 700 2024-08-21 20:00:00 Blankenberge  51.3106    3.1164           30
#> 701 2024-08-21 21:00:00 Blankenberge  51.3106    3.1164           89
#> 702 2024-08-21 22:00:00 Blankenberge  51.3106    3.1164          145
#> 703 2024-08-21 23:00:00 Blankenberge  51.3106    3.1164          254
#> 704 2024-08-22 00:00:00 Blankenberge  51.3106    3.1164          452
#> 705 2024-08-22 01:00:00 Blankenberge  51.3106    3.1164          481
#> 706 2024-08-22 02:00:00 Blankenberge  51.3106    3.1164          479
#> 707 2024-08-22 03:00:00 Blankenberge  51.3106    3.1164          443
#> 708 2024-08-22 04:00:00 Blankenberge  51.3106    3.1164          372
#> 709 2024-08-22 05:00:00 Blankenberge  51.3106    3.1164          273
#> 710 2024-08-22 06:00:00 Blankenberge  51.3106    3.1164          177
#> 711 2024-08-22 07:00:00 Blankenberge  51.3106    3.1164           77
#> 712 2024-08-22 08:00:00 Blankenberge  51.3106    3.1164           18
#> 
#> $par
#> $par$daterange
#> [1] "2024-07-23" "2024-08-23"
#> 
#> $par$type
#> [1] "MVB data"
#> 
#> $par$stations
#> [1] "Blankenberge"
#> 
#> $par$calc
#> [1] "max"
#> 
#> $par$binSize
#> [1] "hour"
#> 
#> $par$parameters
#> [1] "Tide (cmTAW)"
#> 
#>