Retrieve MVB-data from the LifeWatch project

Retrieves the Meetnet Vlaams Banken (MVB) data from the LifeWatch project. Without valid credentials you are only allowed to view the "Tide TAW" parameter for the last 30 days, grouped by day/hour. To get an account, register via the Lifewatch RShiny registration webpage.

getMvbData(
  startdate,
  stopdate,
  parameters,
  stations = NULL,
  by,
  calc,
  usr = NULL,
  pwd = NULL,
  params = FALSE,
  ...
)

Arguments

startdate

Starting date for the query. Without a login, this is restricted to the last month

stopdate

Stopping date for the query. Without a login, this is restricted to the last month

parameters

One (or list) of the $parameter in listMvbStations.

stations

List of stations to be included in the query. See listMvbStations. Use stations="All" or stations=NULL to get all stations.

by

Time grouping for calculation, one of ('day','hour', '10min' ,'none')

calc

Calculation to perform given time grouping, one of ('avg', 'max', 'min', 'none').

usr

Username to connect to database

pwd

Password to connect to database

params

If TRUE, returns a list with the dataset and the query parameters applied in the server side. IF FALSE returns only the data.

...

Params to be passed to lw_check_lwdataserver().

Value

Dataframe with the specified MVB data.

Examples

getMvbData(Sys.Date() - 30, Sys.Date() + 1, parameters = 'Tide TAW',
stations = "Nieuwpoort", by = "day", calc = "avg")
#> - Query mode: Post request to OpenCPU server
#> No encoding supplied: defaulting to UTF-8.
#> - Query parameters: 
#>    - daterange: c("2021-11-22", "2021-12-23")
#>    - type: MVB data
#>    - stations: Nieuwpoort
#>    - calc: avg
#>    - binSize: day
#>    - parameters: Tide TAW
#>          time    station latitude longitude Tide TAW
#> 1  2021-11-22 Nieuwpoort  51.1506    2.7281 218.6348
#> 2  2021-11-23 Nieuwpoort  51.1506    2.7281 233.0208
#> 3  2021-11-24 Nieuwpoort  51.1506    2.7281 238.7743
#> 4  2021-11-25 Nieuwpoort  51.1506    2.7281 266.0664
#> 5  2021-11-26 Nieuwpoort  51.1506    2.7281 264.7500
#> 6  2021-11-27 Nieuwpoort  51.1506    2.7281 287.6319
#> 7  2021-11-28 Nieuwpoort  51.1506    2.7281 260.2743
#> 8  2021-11-29 Nieuwpoort  51.1506    2.7281 256.0667
#> 9  2021-11-30 Nieuwpoort  51.1506    2.7281 273.7743
#> 10 2021-12-01 Nieuwpoort  51.1506    2.7281 293.0976
#> 11 2021-12-02 Nieuwpoort  51.1506    2.7281 282.2500
#> 12 2021-12-03 Nieuwpoort  51.1506    2.7281 237.4688
#> 13 2021-12-04 Nieuwpoort  51.1506    2.7281 264.6528
#> 14 2021-12-05 Nieuwpoort  51.1506    2.7281 260.9375
#> 15 2021-12-06 Nieuwpoort  51.1506    2.7281 219.2274
#> 16 2021-12-07 Nieuwpoort  51.1506    2.7281 254.5660
#> 17 2021-12-08 Nieuwpoort  51.1506    2.7281 219.1319
#> 18 2021-12-09 Nieuwpoort  51.1506    2.7281 258.0729
#> 19 2021-12-10 Nieuwpoort  51.1506    2.7281 289.8611
#> 20 2021-12-11 Nieuwpoort  51.1506    2.7281 251.2222
#> 21 2021-12-12 Nieuwpoort  51.1506    2.7281 252.8160
#> 22 2021-12-13 Nieuwpoort  51.1506    2.7281 222.1741
#> 23 2021-12-14 Nieuwpoort  51.1506    2.7281 241.1250
#> 24 2021-12-15 Nieuwpoort  51.1506    2.7281 246.2545
#> 25 2021-12-16 Nieuwpoort  51.1506    2.7281 229.6319
#> 26 2021-12-17 Nieuwpoort  51.1506    2.7281 219.9965
#> 27 2021-12-18 Nieuwpoort  51.1506    2.7281 211.7431
#> 28 2021-12-19 Nieuwpoort  51.1506    2.7281 224.3819
#> 29 2021-12-20 Nieuwpoort  51.1506    2.7281 221.9549
#> 30 2021-12-21 Nieuwpoort  51.1506    2.7281 227.7778
#> 31 2021-12-22 Nieuwpoort  51.1506    2.7281 223.7448
getMvbData(Sys.Date() - 30, Sys.Date() + 1, parameters = 'Tide TAW',
stations = "Blankenberge", by = "hour", calc = "max", params = TRUE)
#> - Query mode: Post request to OpenCPU server
#> No encoding supplied: defaulting to UTF-8.
#> - Query parameters: 
#>    - daterange: c("2021-11-22", "2021-12-23")
#>    - type: MVB data
#>    - stations: Blankenberge
#>    - calc: max
#>    - binSize: hour
#>    - parameters: Tide TAW
#> $mdf
#>                    time      station latitude longitude Tide TAW
#> 1   2021-11-22 00:00:00 Blankenberge  51.3106    3.1164      410
#> 2   2021-11-22 01:00:00 Blankenberge  51.3106    3.1164      440
#> 3   2021-11-22 02:00:00 Blankenberge  51.3106    3.1164      432
#> 4   2021-11-22 03:00:00 Blankenberge  51.3106    3.1164      394
#> 5   2021-11-22 04:00:00 Blankenberge  51.3106    3.1164      329
#> 6   2021-11-22 05:00:00 Blankenberge  51.3106    3.1164      232
#> 7   2021-11-22 06:00:00 Blankenberge  51.3106    3.1164      135
#> 8   2021-11-22 07:00:00 Blankenberge  51.3106    3.1164       66
#> 9   2021-11-22 08:00:00 Blankenberge  51.3106    3.1164       60
#> 10  2021-11-22 09:00:00 Blankenberge  51.3106    3.1164      104
#> 11  2021-11-22 10:00:00 Blankenberge  51.3106    3.1164      152
#> 12  2021-11-22 11:00:00 Blankenberge  51.3106    3.1164      226
#> 13  2021-11-22 12:00:00 Blankenberge  51.3106    3.1164      366
#> 14  2021-11-22 13:00:00 Blankenberge  51.3106    3.1164      423
#> 15  2021-11-22 14:00:00 Blankenberge  51.3106    3.1164      425
#> 16  2021-11-22 15:00:00 Blankenberge  51.3106    3.1164      399
#> 17  2021-11-22 16:00:00 Blankenberge  51.3106    3.1164      343
#> 18  2021-11-22 17:00:00 Blankenberge  51.3106    3.1164      256
#> 19  2021-11-22 18:00:00 Blankenberge  51.3106    3.1164      160
#> 20  2021-11-22 19:00:00 Blankenberge  51.3106    3.1164       80
#> 21  2021-11-22 20:00:00 Blankenberge  51.3106    3.1164       70
#> 22  2021-11-22 21:00:00 Blankenberge  51.3106    3.1164      120
#> 23  2021-11-22 22:00:00 Blankenberge  51.3106    3.1164      157
#> 24  2021-11-22 23:00:00 Blankenberge  51.3106    3.1164      222
#> 25  2021-11-23 00:00:00 Blankenberge  51.3106    3.1164      334
#> 26  2021-11-23 01:00:00 Blankenberge  51.3106    3.1164      427
#> 27  2021-11-23 02:00:00 Blankenberge  51.3106    3.1164      431
#> 28  2021-11-23 03:00:00 Blankenberge  51.3106    3.1164      413
#> 29  2021-11-23 04:00:00 Blankenberge  51.3106    3.1164      364
#> 30  2021-11-23 05:00:00 Blankenberge  51.3106    3.1164      290
#> 31  2021-11-23 06:00:00 Blankenberge  51.3106    3.1164      199
#> 32  2021-11-23 07:00:00 Blankenberge  51.3106    3.1164      110
#> 33  2021-11-23 08:00:00 Blankenberge  51.3106    3.1164       60
#> 34  2021-11-23 09:00:00 Blankenberge  51.3106    3.1164       95
#> 35  2021-11-23 10:00:00 Blankenberge  51.3106    3.1164      137
#> 36  2021-11-23 11:00:00 Blankenberge  51.3106    3.1164      200
#> 37  2021-11-23 12:00:00 Blankenberge  51.3106    3.1164      317
#> 38  2021-11-23 13:00:00 Blankenberge  51.3106    3.1164      420
#> 39  2021-11-23 14:00:00 Blankenberge  51.3106    3.1164      433
#> 40  2021-11-23 15:00:00 Blankenberge  51.3106    3.1164      426
#> 41  2021-11-23 16:00:00 Blankenberge  51.3106    3.1164      382
#> 42  2021-11-23 17:00:00 Blankenberge  51.3106    3.1164      313
#> 43  2021-11-23 18:00:00 Blankenberge  51.3106    3.1164      226
#> 44  2021-11-23 19:00:00 Blankenberge  51.3106    3.1164      139
#> 45  2021-11-23 20:00:00 Blankenberge  51.3106    3.1164       79
#> 46  2021-11-23 21:00:00 Blankenberge  51.3106    3.1164       97
#> 47  2021-11-23 22:00:00 Blankenberge  51.3106    3.1164      136
#> 48  2021-11-23 23:00:00 Blankenberge  51.3106    3.1164      187
#> 49  2021-11-24 00:00:00 Blankenberge  51.3106    3.1164      268
#> 50  2021-11-24 01:00:00 Blankenberge  51.3106    3.1164      386
#> 51  2021-11-24 02:00:00 Blankenberge  51.3106    3.1164      416
#> 52  2021-11-24 03:00:00 Blankenberge  51.3106    3.1164      415
#> 53  2021-11-24 04:00:00 Blankenberge  51.3106    3.1164      381
#> 54  2021-11-24 05:00:00 Blankenberge  51.3106    3.1164      324
#> 55  2021-11-24 06:00:00 Blankenberge  51.3106    3.1164      243
#> 56  2021-11-24 07:00:00 Blankenberge  51.3106    3.1164      159
#> 57  2021-11-24 08:00:00 Blankenberge  51.3106    3.1164       89
#> 58  2021-11-24 09:00:00 Blankenberge  51.3106    3.1164       62
#> 59  2021-11-24 10:00:00 Blankenberge  51.3106    3.1164      104
#> 60  2021-11-24 11:00:00 Blankenberge  51.3106    3.1164      156
#> 61  2021-11-24 12:00:00 Blankenberge  51.3106    3.1164      250
#> 62  2021-11-24 13:00:00 Blankenberge  51.3106    3.1164      379
#> 63  2021-11-24 14:00:00 Blankenberge  51.3106    3.1164      424
#> 64  2021-11-24 15:00:00 Blankenberge  51.3106    3.1164      424
#> 65  2021-11-24 16:00:00 Blankenberge  51.3106    3.1164      408
#> 66  2021-11-24 17:00:00 Blankenberge  51.3106    3.1164      361
#> 67  2021-11-24 18:00:00 Blankenberge  51.3106    3.1164      288
#> 68  2021-11-24 19:00:00 Blankenberge  51.3106    3.1164      210
#> 69  2021-11-24 20:00:00 Blankenberge  51.3106    3.1164      137
#> 70  2021-11-24 21:00:00 Blankenberge  51.3106    3.1164      107
#> 71  2021-11-24 22:00:00 Blankenberge  51.3106    3.1164      140
#> 72  2021-11-24 23:00:00 Blankenberge  51.3106    3.1164      178
#> 73  2021-11-25 00:00:00 Blankenberge  51.3106    3.1164      236
#> 74  2021-11-25 01:00:00 Blankenberge  51.3106    3.1164      337
#> 75  2021-11-25 02:00:00 Blankenberge  51.3106    3.1164      413
#> 76  2021-11-25 03:00:00 Blankenberge  51.3106    3.1164      422
#> 77  2021-11-25 04:00:00 Blankenberge  51.3106    3.1164      413
#> 78  2021-11-25 05:00:00 Blankenberge  51.3106    3.1164      372
#> 79  2021-11-25 06:00:00 Blankenberge  51.3106    3.1164      312
#> 80  2021-11-25 07:00:00 Blankenberge  51.3106    3.1164      240
#> 81  2021-11-25 08:00:00 Blankenberge  51.3106    3.1164      172
#> 82  2021-11-25 09:00:00 Blankenberge  51.3106    3.1164      124
#> 83  2021-11-25 10:00:00 Blankenberge  51.3106    3.1164      150
#> 84  2021-11-25 11:00:00 Blankenberge  51.3106    3.1164      180
#> 85  2021-11-25 12:00:00 Blankenberge  51.3106    3.1164      239
#> 86  2021-11-25 13:00:00 Blankenberge  51.3106    3.1164      323
#> 87  2021-11-25 14:00:00 Blankenberge  51.3106    3.1164      417
#> 88  2021-11-25 15:00:00 Blankenberge  51.3106    3.1164      440
#> 89  2021-11-25 16:00:00 Blankenberge  51.3106    3.1164      438
#> 90  2021-11-25 17:00:00 Blankenberge  51.3106    3.1164      412
#> 91  2021-11-25 18:00:00 Blankenberge  51.3106    3.1164      354
#> 92  2021-11-25 19:00:00 Blankenberge  51.3106    3.1164      273
#> 93  2021-11-25 20:00:00 Blankenberge  51.3106    3.1164      203
#> 94  2021-11-25 21:00:00 Blankenberge  51.3106    3.1164      146
#> 95  2021-11-25 22:00:00 Blankenberge  51.3106    3.1164      128
#> 96  2021-11-25 23:00:00 Blankenberge  51.3106    3.1164      155
#> 97  2021-11-26 00:00:00 Blankenberge  51.3106    3.1164      188
#> 98  2021-11-26 01:00:00 Blankenberge  51.3106    3.1164      238
#> 99  2021-11-26 02:00:00 Blankenberge  51.3106    3.1164      328
#> 100 2021-11-26 03:00:00 Blankenberge  51.3106    3.1164      376
#> 101 2021-11-26 04:00:00 Blankenberge  51.3106    3.1164      377
#> 102 2021-11-26 05:00:00 Blankenberge  51.3106    3.1164      360
#> 103 2021-11-26 06:00:00 Blankenberge  51.3106    3.1164      312
#> 104 2021-11-26 07:00:00 Blankenberge  51.3106    3.1164      252
#> 105 2021-11-26 08:00:00 Blankenberge  51.3106    3.1164      185
#> 106 2021-11-26 09:00:00 Blankenberge  51.3106    3.1164      124
#> 107 2021-11-26 10:00:00 Blankenberge  51.3106    3.1164       78
#> 108 2021-11-26 11:00:00 Blankenberge  51.3106    3.1164      103
#> 109 2021-11-26 12:00:00 Blankenberge  51.3106    3.1164      170
#> 110 2021-11-26 13:00:00 Blankenberge  51.3106    3.1164      282
#> 111 2021-11-26 14:00:00 Blankenberge  51.3106    3.1164      382
#> 112 2021-11-26 15:00:00 Blankenberge  51.3106    3.1164      440
#> 113 2021-11-26 16:00:00 Blankenberge  51.3106    3.1164      461
#> 114 2021-11-26 17:00:00 Blankenberge  51.3106    3.1164      461
#> 115 2021-11-26 18:00:00 Blankenberge  51.3106    3.1164      431
#> 116 2021-11-26 19:00:00 Blankenberge  51.3106    3.1164      396
#> 117 2021-11-26 20:00:00 Blankenberge  51.3106    3.1164      318
#> 118 2021-11-26 21:00:00 Blankenberge  51.3106    3.1164      259
#> 119 2021-11-26 22:00:00 Blankenberge  51.3106    3.1164      200
#> 120 2021-11-26 23:00:00 Blankenberge  51.3106    3.1164      188
#> 121 2021-11-27 00:00:00 Blankenberge  51.3106    3.1164      210
#> 122 2021-11-27 01:00:00 Blankenberge  51.3106    3.1164      253
#> 123 2021-11-27 02:00:00 Blankenberge  51.3106    3.1164      322
#> 124 2021-11-27 03:00:00 Blankenberge  51.3106    3.1164      394
#> 125 2021-11-27 04:00:00 Blankenberge  51.3106    3.1164      426
#> 126 2021-11-27 05:00:00 Blankenberge  51.3106    3.1164      432
#> 127 2021-11-27 06:00:00 Blankenberge  51.3106    3.1164      427
#> 128 2021-11-27 07:00:00 Blankenberge  51.3106    3.1164      390
#> 129 2021-11-27 08:00:00 Blankenberge  51.3106    3.1164      326
#> 130 2021-11-27 09:00:00 Blankenberge  51.3106    3.1164      256
#> 131 2021-11-27 10:00:00 Blankenberge  51.3106    3.1164      197
#> 132 2021-11-27 11:00:00 Blankenberge  51.3106    3.1164      149
#> 133 2021-11-27 12:00:00 Blankenberge  51.3106    3.1164      159
#> 134 2021-11-27 13:00:00 Blankenberge  51.3106    3.1164      205
#> 135 2021-11-27 14:00:00 Blankenberge  51.3106    3.1164      254
#> 136 2021-11-27 15:00:00 Blankenberge  51.3106    3.1164      318
#> 137 2021-11-27 16:00:00 Blankenberge  51.3106    3.1164      382
#> 138 2021-11-27 17:00:00 Blankenberge  51.3106    3.1164      411
#> 139 2021-11-27 18:00:00 Blankenberge  51.3106    3.1164      410
#> 140 2021-11-27 19:00:00 Blankenberge  51.3106    3.1164      382
#> 141 2021-11-27 20:00:00 Blankenberge  51.3106    3.1164      344
#> 142 2021-11-27 21:00:00 Blankenberge  51.3106    3.1164      279
#> 143 2021-11-27 22:00:00 Blankenberge  51.3106    3.1164      230
#> 144 2021-11-27 23:00:00 Blankenberge  51.3106    3.1164      202
#> 145 2021-11-28 00:00:00 Blankenberge  51.3106    3.1164      196
#> 146 2021-11-28 01:00:00 Blankenberge  51.3106    3.1164      199
#> 147 2021-11-28 02:00:00 Blankenberge  51.3106    3.1164      228
#> 148 2021-11-28 03:00:00 Blankenberge  51.3106    3.1164      262
#> 149 2021-11-28 04:00:00 Blankenberge  51.3106    3.1164      304
#> 150 2021-11-28 05:00:00 Blankenberge  51.3106    3.1164      361
#> 151 2021-11-28 06:00:00 Blankenberge  51.3106    3.1164      365
#> 152 2021-11-28 07:00:00 Blankenberge  51.3106    3.1164      360
#> 153 2021-11-28 08:00:00 Blankenberge  51.3106    3.1164      332
#> 154 2021-11-28 09:00:00 Blankenberge  51.3106    3.1164      283
#> 155 2021-11-28 10:00:00 Blankenberge  51.3106    3.1164      222
#> 156 2021-11-28 11:00:00 Blankenberge  51.3106    3.1164      170
#> 157 2021-11-28 12:00:00 Blankenberge  51.3106    3.1164      129
#> 158 2021-11-28 13:00:00 Blankenberge  51.3106    3.1164      129
#> 159 2021-11-28 14:00:00 Blankenberge  51.3106    3.1164      167
#> 160 2021-11-28 15:00:00 Blankenberge  51.3106    3.1164      211
#> 161 2021-11-28 16:00:00 Blankenberge  51.3106    3.1164      286
#> 162 2021-11-28 17:00:00 Blankenberge  51.3106    3.1164      359
#> 163 2021-11-28 18:00:00 Blankenberge  51.3106    3.1164      397
#> 164 2021-11-28 19:00:00 Blankenberge  51.3106    3.1164      398
#> 165 2021-11-28 20:00:00 Blankenberge  51.3106    3.1164      382
#> 166 2021-11-28 21:00:00 Blankenberge  51.3106    3.1164      343
#> 167 2021-11-28 22:00:00 Blankenberge  51.3106    3.1164      287
#> 168 2021-11-28 23:00:00 Blankenberge  51.3106    3.1164      225
#> 169 2021-11-29 00:00:00 Blankenberge  51.3106    3.1164      170
#> 170 2021-11-29 01:00:00 Blankenberge  51.3106    3.1164      149
#> 171 2021-11-29 02:00:00 Blankenberge  51.3106    3.1164      181
#> 172 2021-11-29 03:00:00 Blankenberge  51.3106    3.1164      210
#> 173 2021-11-29 04:00:00 Blankenberge  51.3106    3.1164      248
#> 174 2021-11-29 05:00:00 Blankenberge  51.3106    3.1164      322
#> 175 2021-11-29 06:00:00 Blankenberge  51.3106    3.1164      374
#> 176 2021-11-29 07:00:00 Blankenberge  51.3106    3.1164      392
#> 177 2021-11-29 08:00:00 Blankenberge  51.3106    3.1164      380
#> 178 2021-11-29 09:00:00 Blankenberge  51.3106    3.1164      351
#> 179 2021-11-29 10:00:00 Blankenberge  51.3106    3.1164      303
#> 180 2021-11-29 11:00:00 Blankenberge  51.3106    3.1164      245
#> 181 2021-11-29 12:00:00 Blankenberge  51.3106    3.1164      179
#> 182 2021-11-29 13:00:00 Blankenberge  51.3106    3.1164      124
#> 183 2021-11-29 14:00:00 Blankenberge  51.3106    3.1164      105
#> 184 2021-11-29 15:00:00 Blankenberge  51.3106    3.1164      142
#> 185 2021-11-29 16:00:00 Blankenberge  51.3106    3.1164      192
#> 186 2021-11-29 17:00:00 Blankenberge  51.3106    3.1164      276
#> 187 2021-11-29 18:00:00 Blankenberge  51.3106    3.1164      362
#> 188 2021-11-29 19:00:00 Blankenberge  51.3106    3.1164      401
#> 189 2021-11-29 20:00:00 Blankenberge  51.3106    3.1164      402
#> 190 2021-11-29 21:00:00 Blankenberge  51.3106    3.1164      384
#> 191 2021-11-29 22:00:00 Blankenberge  51.3106    3.1164      346
#> 192 2021-11-29 23:00:00 Blankenberge  51.3106    3.1164      284
#> 193 2021-11-30 00:00:00 Blankenberge  51.3106    3.1164      215
#> 194 2021-11-30 01:00:00 Blankenberge  51.3106    3.1164      156
#> 195 2021-11-30 02:00:00 Blankenberge  51.3106    3.1164      124
#> 196 2021-11-30 03:00:00 Blankenberge  51.3106    3.1164      170
#> 197 2021-11-30 04:00:00 Blankenberge  51.3106    3.1164      206
#> 198 2021-11-30 05:00:00 Blankenberge  51.3106    3.1164      274
#> 199 2021-11-30 06:00:00 Blankenberge  51.3106    3.1164      360
#> 200 2021-11-30 07:00:00 Blankenberge  51.3106    3.1164      441
#> 201 2021-11-30 08:00:00 Blankenberge  51.3106    3.1164      455
#> 202 2021-11-30 09:00:00 Blankenberge  51.3106    3.1164      454
#> 203 2021-11-30 10:00:00 Blankenberge  51.3106    3.1164      425
#> 204 2021-11-30 11:00:00 Blankenberge  51.3106    3.1164      371
#> 205 2021-11-30 12:00:00 Blankenberge  51.3106    3.1164      298
#> 206 2021-11-30 13:00:00 Blankenberge  51.3106    3.1164      217
#> 207 2021-11-30 14:00:00 Blankenberge  51.3106    3.1164      152
#> 208 2021-11-30 15:00:00 Blankenberge  51.3106    3.1164      106
#> 209 2021-11-30 16:00:00 Blankenberge  51.3106    3.1164      137
#> 210 2021-11-30 17:00:00 Blankenberge  51.3106    3.1164      194
#> 211 2021-11-30 18:00:00 Blankenberge  51.3106    3.1164      284
#> 212 2021-11-30 19:00:00 Blankenberge  51.3106    3.1164      396
#> 213 2021-11-30 20:00:00 Blankenberge  51.3106    3.1164      431
#> 214 2021-11-30 21:00:00 Blankenberge  51.3106    3.1164      435
#> 215 2021-11-30 22:00:00 Blankenberge  51.3106    3.1164      421
#> 216 2021-11-30 23:00:00 Blankenberge  51.3106    3.1164      373
#> 217 2021-12-01 00:00:00 Blankenberge  51.3106    3.1164      309
#> 218 2021-12-01 01:00:00 Blankenberge  51.3106    3.1164      239
#> 219 2021-12-01 02:00:00 Blankenberge  51.3106    3.1164      168
#> 220 2021-12-01 03:00:00 Blankenberge  51.3106    3.1164      108
#> 221 2021-12-01 04:00:00 Blankenberge  51.3106    3.1164      131
#> 222 2021-12-01 05:00:00 Blankenberge  51.3106    3.1164      178
#> 223 2021-12-01 06:00:00 Blankenberge  51.3106    3.1164      253
#> 224 2021-12-01 07:00:00 Blankenberge  51.3106    3.1164      375
#> 225 2021-12-01 08:00:00 Blankenberge  51.3106    3.1164      454
#> 226 2021-12-01 09:00:00 Blankenberge  51.3106    3.1164      471
#> 227 2021-12-01 10:00:00 Blankenberge  51.3106    3.1164      458
#> 228 2021-12-01 11:00:00 Blankenberge  51.3106    3.1164      434
#> 229 2021-12-01 12:00:00 Blankenberge  51.3106    3.1164      379
#> 230 2021-12-01 13:00:00 Blankenberge  51.3106    3.1164      310
#> 231 2021-12-01 14:00:00 Blankenberge  51.3106    3.1164      231
#> 232 2021-12-01 15:00:00 Blankenberge  51.3106    3.1164      160
#> 233 2021-12-01 16:00:00 Blankenberge  51.3106    3.1164      159
#> 234 2021-12-01 17:00:00 Blankenberge  51.3106    3.1164      201
#> 235 2021-12-01 18:00:00 Blankenberge  51.3106    3.1164      259
#> 236 2021-12-01 19:00:00 Blankenberge  51.3106    3.1164      359
#> 237 2021-12-01 20:00:00 Blankenberge  51.3106    3.1164      470
#> 238 2021-12-01 21:00:00 Blankenberge  51.3106    3.1164      520
#> 239 2021-12-01 22:00:00 Blankenberge  51.3106    3.1164      529
#> 240 2021-12-01 23:00:00 Blankenberge  51.3106    3.1164      505
#> 241 2021-12-02 00:00:00 Blankenberge  51.3106    3.1164      457
#> 242 2021-12-02 01:00:00 Blankenberge  51.3106    3.1164      368
#> 243 2021-12-02 02:00:00 Blankenberge  51.3106    3.1164      268
#> 244 2021-12-02 03:00:00 Blankenberge  51.3106    3.1164      203
#> 245 2021-12-02 04:00:00 Blankenberge  51.3106    3.1164      194
#> 246 2021-12-02 05:00:00 Blankenberge  51.3106    3.1164      214
#> 247 2021-12-02 06:00:00 Blankenberge  51.3106    3.1164      253
#> 248 2021-12-02 07:00:00 Blankenberge  51.3106    3.1164      283
#> 249 2021-12-02 08:00:00 Blankenberge  51.3106    3.1164      363
#> 250 2021-12-02 09:00:00 Blankenberge  51.3106    3.1164      478
#> 251 2021-12-02 10:00:00 Blankenberge  51.3106    3.1164      496
#> 252 2021-12-02 11:00:00 Blankenberge  51.3106    3.1164      488
#> 253 2021-12-02 12:00:00 Blankenberge  51.3106    3.1164      433
#> 254 2021-12-02 13:00:00 Blankenberge  51.3106    3.1164      350
#> 255 2021-12-02 14:00:00 Blankenberge  51.3106    3.1164      269
#> 256 2021-12-02 15:00:00 Blankenberge  51.3106    3.1164      152
#> 257 2021-12-02 16:00:00 Blankenberge  51.3106    3.1164       96
#> 258 2021-12-02 17:00:00 Blankenberge  51.3106    3.1164       97
#> 259 2021-12-02 18:00:00 Blankenberge  51.3106    3.1164      136
#> 260 2021-12-02 19:00:00 Blankenberge  51.3106    3.1164      201
#> 261 2021-12-02 20:00:00 Blankenberge  51.3106    3.1164      284
#> 262 2021-12-02 21:00:00 Blankenberge  51.3106    3.1164      437
#> 263 2021-12-02 22:00:00 Blankenberge  51.3106    3.1164      468
#> 264 2021-12-02 23:00:00 Blankenberge  51.3106    3.1164      465
#> 265 2021-12-03 00:00:00 Blankenberge  51.3106    3.1164      431
#> 266 2021-12-03 01:00:00 Blankenberge  51.3106    3.1164      365
#> 267 2021-12-03 02:00:00 Blankenberge  51.3106    3.1164      272
#> 268 2021-12-03 03:00:00 Blankenberge  51.3106    3.1164      175
#> 269 2021-12-03 04:00:00 Blankenberge  51.3106    3.1164       89
#> 270 2021-12-03 05:00:00 Blankenberge  51.3106    3.1164       52
#> 271 2021-12-03 06:00:00 Blankenberge  51.3106    3.1164      110
#> 272 2021-12-03 07:00:00 Blankenberge  51.3106    3.1164      143
#> 273 2021-12-03 08:00:00 Blankenberge  51.3106    3.1164      214
#> 274 2021-12-03 09:00:00 Blankenberge  51.3106    3.1164      375
#> 275 2021-12-03 10:00:00 Blankenberge  51.3106    3.1164      441
#> 276 2021-12-03 11:00:00 Blankenberge  51.3106    3.1164      442
#> 277 2021-12-03 12:00:00 Blankenberge  51.3106    3.1164      415
#> 278 2021-12-03 13:00:00 Blankenberge  51.3106    3.1164      361
#> 279 2021-12-03 14:00:00 Blankenberge  51.3106    3.1164      285
#> 280 2021-12-03 15:00:00 Blankenberge  51.3106    3.1164      199
#> 281 2021-12-03 16:00:00 Blankenberge  51.3106    3.1164      119
#> 282 2021-12-03 17:00:00 Blankenberge  51.3106    3.1164       50
#> 283 2021-12-03 18:00:00 Blankenberge  51.3106    3.1164       98
#> 284 2021-12-03 19:00:00 Blankenberge  51.3106    3.1164      145
#> 285 2021-12-03 20:00:00 Blankenberge  51.3106    3.1164      211
#> 286 2021-12-03 21:00:00 Blankenberge  51.3106    3.1164      354
#> 287 2021-12-03 22:00:00 Blankenberge  51.3106    3.1164      489
#> 288 2021-12-03 23:00:00 Blankenberge  51.3106    3.1164      502
#> 289 2021-12-04 00:00:00 Blankenberge  51.3106    3.1164      496
#> 290 2021-12-04 01:00:00 Blankenberge  51.3106    3.1164      456
#> 291 2021-12-04 02:00:00 Blankenberge  51.3106    3.1164      377
#> 292 2021-12-04 03:00:00 Blankenberge  51.3106    3.1164      288
#> 293 2021-12-04 04:00:00 Blankenberge  51.3106    3.1164      189
#> 294 2021-12-04 05:00:00 Blankenberge  51.3106    3.1164      100
#> 295 2021-12-04 06:00:00 Blankenberge  51.3106    3.1164       84
#> 296 2021-12-04 07:00:00 Blankenberge  51.3106    3.1164      138
#> 297 2021-12-04 08:00:00 Blankenberge  51.3106    3.1164      192
#> 298 2021-12-04 09:00:00 Blankenberge  51.3106    3.1164      288
#> 299 2021-12-04 10:00:00 Blankenberge  51.3106    3.1164      467
#> 300 2021-12-04 11:00:00 Blankenberge  51.3106    3.1164      504
#> 301 2021-12-04 12:00:00 Blankenberge  51.3106    3.1164      504
#> 302 2021-12-04 13:00:00 Blankenberge  51.3106    3.1164      466
#> 303 2021-12-04 14:00:00 Blankenberge  51.3106    3.1164      404
#> 304 2021-12-04 15:00:00 Blankenberge  51.3106    3.1164      312
#> 305 2021-12-04 16:00:00 Blankenberge  51.3106    3.1164      216
#> 306 2021-12-04 17:00:00 Blankenberge  51.3106    3.1164      128
#> 307 2021-12-04 18:00:00 Blankenberge  51.3106    3.1164       53
#> 308 2021-12-04 19:00:00 Blankenberge  51.3106    3.1164      113
#> 309 2021-12-04 20:00:00 Blankenberge  51.3106    3.1164      168
#> 310 2021-12-04 21:00:00 Blankenberge  51.3106    3.1164      240
#> 311 2021-12-04 22:00:00 Blankenberge  51.3106    3.1164      403
#> 312 2021-12-04 23:00:00 Blankenberge  51.3106    3.1164      506
#> 313 2021-12-05 00:00:00 Blankenberge  51.3106    3.1164      513
#> 314 2021-12-05 01:00:00 Blankenberge  51.3106    3.1164      497
#> 315 2021-12-05 02:00:00 Blankenberge  51.3106    3.1164      447
#> 316 2021-12-05 03:00:00 Blankenberge  51.3106    3.1164      366
#> 317 2021-12-05 04:00:00 Blankenberge  51.3106    3.1164      268
#> 318 2021-12-05 05:00:00 Blankenberge  51.3106    3.1164      173
#> 319 2021-12-05 06:00:00 Blankenberge  51.3106    3.1164       90
#> 320 2021-12-05 07:00:00 Blankenberge  51.3106    3.1164      119
#> 321 2021-12-05 08:00:00 Blankenberge  51.3106    3.1164      166
#> 322 2021-12-05 09:00:00 Blankenberge  51.3106    3.1164      216
#> 323 2021-12-05 10:00:00 Blankenberge  51.3106    3.1164      312
#> 324 2021-12-05 11:00:00 Blankenberge  51.3106    3.1164      483
#> 325 2021-12-05 12:00:00 Blankenberge  51.3106    3.1164      508
#> 326 2021-12-05 13:00:00 Blankenberge  51.3106    3.1164      503
#> 327 2021-12-05 14:00:00 Blankenberge  51.3106    3.1164      461
#> 328 2021-12-05 15:00:00 Blankenberge  51.3106    3.1164      391
#> 329 2021-12-05 16:00:00 Blankenberge  51.3106    3.1164      290
#> 330 2021-12-05 17:00:00 Blankenberge  51.3106    3.1164      186
#> 331 2021-12-05 18:00:00 Blankenberge  51.3106    3.1164       84
#> 332 2021-12-05 19:00:00 Blankenberge  51.3106    3.1164       44
#> 333 2021-12-05 20:00:00 Blankenberge  51.3106    3.1164      101
#> 334 2021-12-05 21:00:00 Blankenberge  51.3106    3.1164      148
#> 335 2021-12-05 22:00:00 Blankenberge  51.3106    3.1164      211
#> 336 2021-12-05 23:00:00 Blankenberge  51.3106    3.1164      363
#> 337 2021-12-06 00:00:00 Blankenberge  51.3106    3.1164      466
#> 338 2021-12-06 01:00:00 Blankenberge  51.3106    3.1164      465
#> 339 2021-12-06 02:00:00 Blankenberge  51.3106    3.1164      443
#> 340 2021-12-06 03:00:00 Blankenberge  51.3106    3.1164      381
#> 341 2021-12-06 04:00:00 Blankenberge  51.3106    3.1164      294
#> 342 2021-12-06 05:00:00 Blankenberge  51.3106    3.1164      190
#> 343 2021-12-06 06:00:00 Blankenberge  51.3106    3.1164       98
#> 344 2021-12-06 07:00:00 Blankenberge  51.3106    3.1164       20
#> 345 2021-12-06 08:00:00 Blankenberge  51.3106    3.1164       51
#> 346 2021-12-06 09:00:00 Blankenberge  51.3106    3.1164      102
#> 347 2021-12-06 10:00:00 Blankenberge  51.3106    3.1164      158
#> 348 2021-12-06 11:00:00 Blankenberge  51.3106    3.1164      290
#> 349 2021-12-06 12:00:00 Blankenberge  51.3106    3.1164      443
#> 350 2021-12-06 13:00:00 Blankenberge  51.3106    3.1164      456
#> 351 2021-12-06 14:00:00 Blankenberge  51.3106    3.1164      445
#> 352 2021-12-06 15:00:00 Blankenberge  51.3106    3.1164      392
#> 353 2021-12-06 16:00:00 Blankenberge  51.3106    3.1164      321
#> 354 2021-12-06 17:00:00 Blankenberge  51.3106    3.1164      227
#> 355 2021-12-06 18:00:00 Blankenberge  51.3106    3.1164      136
#> 356 2021-12-06 19:00:00 Blankenberge  51.3106    3.1164       51
#> 357 2021-12-06 20:00:00 Blankenberge  51.3106    3.1164       47
#> 358 2021-12-06 21:00:00 Blankenberge  51.3106    3.1164      112
#> 359 2021-12-06 22:00:00 Blankenberge  51.3106    3.1164      162
#> 360 2021-12-06 23:00:00 Blankenberge  51.3106    3.1164      267
#> 361 2021-12-07 00:00:00 Blankenberge  51.3106    3.1164      445
#> 362 2021-12-07 01:00:00 Blankenberge  51.3106    3.1164      494
#> 363 2021-12-07 02:00:00 Blankenberge  51.3106    3.1164      496
#> 364 2021-12-07 03:00:00 Blankenberge  51.3106    3.1164      470
#> 365 2021-12-07 04:00:00 Blankenberge  51.3106    3.1164      405
#> 366 2021-12-07 05:00:00 Blankenberge  51.3106    3.1164      315
#> 367 2021-12-07 06:00:00 Blankenberge  51.3106    3.1164      225
#> 368 2021-12-07 07:00:00 Blankenberge  51.3106    3.1164      131
#> 369 2021-12-07 08:00:00 Blankenberge  51.3106    3.1164       67
#> 370 2021-12-07 09:00:00 Blankenberge  51.3106    3.1164      119
#> 371 2021-12-07 10:00:00 Blankenberge  51.3106    3.1164      166
#> 372 2021-12-07 11:00:00 Blankenberge  51.3106    3.1164      233
#> 373 2021-12-07 12:00:00 Blankenberge  51.3106    3.1164      380
#> 374 2021-12-07 13:00:00 Blankenberge  51.3106    3.1164      485
#> 375 2021-12-07 14:00:00 Blankenberge  51.3106    3.1164      491
#> 376 2021-12-07 15:00:00 Blankenberge  51.3106    3.1164      466
#> 377 2021-12-07 16:00:00 Blankenberge  51.3106    3.1164      410
#> 378 2021-12-07 17:00:00 Blankenberge  51.3106    3.1164      323
#> 379 2021-12-07 18:00:00 Blankenberge  51.3106    3.1164      234
#> 380 2021-12-07 19:00:00 Blankenberge  51.3106    3.1164      145
#> 381 2021-12-07 20:00:00 Blankenberge  51.3106    3.1164       68
#> 382 2021-12-07 21:00:00 Blankenberge  51.3106    3.1164       37
#> 383 2021-12-07 22:00:00 Blankenberge  51.3106    3.1164       75
#> 384 2021-12-07 23:00:00 Blankenberge  51.3106    3.1164      141
#> 385 2021-12-08 00:00:00 Blankenberge  51.3106    3.1164      271
#> 386 2021-12-08 01:00:00 Blankenberge  51.3106    3.1164      405
#> 387 2021-12-08 02:00:00 Blankenberge  51.3106    3.1164      426
#> 388 2021-12-08 03:00:00 Blankenberge  51.3106    3.1164      417
#> 389 2021-12-08 04:00:00 Blankenberge  51.3106    3.1164      374
#> 390 2021-12-08 05:00:00 Blankenberge  51.3106    3.1164      314
#> 391 2021-12-08 06:00:00 Blankenberge  51.3106    3.1164      231
#> 392 2021-12-08 07:00:00 Blankenberge  51.3106    3.1164      155
#> 393 2021-12-08 08:00:00 Blankenberge  51.3106    3.1164       71
#> 394 2021-12-08 09:00:00 Blankenberge  51.3106    3.1164        2
#> 395 2021-12-08 10:00:00 Blankenberge  51.3106    3.1164       19
#> 396 2021-12-08 11:00:00 Blankenberge  51.3106    3.1164       86
#> 397 2021-12-08 12:00:00 Blankenberge  51.3106    3.1164      208
#> 398 2021-12-08 13:00:00 Blankenberge  51.3106    3.1164      353
#> 399 2021-12-08 14:00:00 Blankenberge  51.3106    3.1164      408
#> 400 2021-12-08 15:00:00 Blankenberge  51.3106    3.1164      411
#> 401 2021-12-08 16:00:00 Blankenberge  51.3106    3.1164      396
#> 402 2021-12-08 17:00:00 Blankenberge  51.3106    3.1164      349
#> 403 2021-12-08 18:00:00 Blankenberge  51.3106    3.1164      281
#> 404 2021-12-08 19:00:00 Blankenberge  51.3106    3.1164      211
#> 405 2021-12-08 20:00:00 Blankenberge  51.3106    3.1164      134
#> 406 2021-12-08 21:00:00 Blankenberge  51.3106    3.1164       69
#> 407 2021-12-08 22:00:00 Blankenberge  51.3106    3.1164       94
#> 408 2021-12-08 23:00:00 Blankenberge  51.3106    3.1164      140
#> 409 2021-12-09 00:00:00 Blankenberge  51.3106    3.1164      208
#> 410 2021-12-09 01:00:00 Blankenberge  51.3106    3.1164      324
#> 411 2021-12-09 02:00:00 Blankenberge  51.3106    3.1164      433
#> 412 2021-12-09 03:00:00 Blankenberge  51.3106    3.1164      451
#> 413 2021-12-09 04:00:00 Blankenberge  51.3106    3.1164      446
#> 414 2021-12-09 05:00:00 Blankenberge  51.3106    3.1164      412
#> 415 2021-12-09 06:00:00 Blankenberge  51.3106    3.1164      349
#> 416 2021-12-09 07:00:00 Blankenberge  51.3106    3.1164      270
#> 417 2021-12-09 08:00:00 Blankenberge  51.3106    3.1164      190
#> 418 2021-12-09 09:00:00 Blankenberge  51.3106    3.1164      108
#> 419 2021-12-09 10:00:00 Blankenberge  51.3106    3.1164       64
#> 420 2021-12-09 11:00:00 Blankenberge  51.3106    3.1164      107
#> 421 2021-12-09 12:00:00 Blankenberge  51.3106    3.1164      164
#> 422 2021-12-09 13:00:00 Blankenberge  51.3106    3.1164      264
#> 423 2021-12-09 14:00:00 Blankenberge  51.3106    3.1164      401
#> 424 2021-12-09 15:00:00 Blankenberge  51.3106    3.1164      448
#> 425 2021-12-09 16:00:00 Blankenberge  51.3106    3.1164      449
#> 426 2021-12-09 17:00:00 Blankenberge  51.3106    3.1164      429
#> 427 2021-12-09 18:00:00 Blankenberge  51.3106    3.1164      378
#> 428 2021-12-09 19:00:00 Blankenberge  51.3106    3.1164      301
#> 429 2021-12-09 20:00:00 Blankenberge  51.3106    3.1164      221
#> 430 2021-12-09 21:00:00 Blankenberge  51.3106    3.1164      136
#> 431 2021-12-09 22:00:00 Blankenberge  51.3106    3.1164       77
#> 432 2021-12-09 23:00:00 Blankenberge  51.3106    3.1164       93
#> 433 2021-12-10 00:00:00 Blankenberge  51.3106    3.1164      134
#> 434 2021-12-10 01:00:00 Blankenberge  51.3106    3.1164      195
#> 435 2021-12-10 02:00:00 Blankenberge  51.3106    3.1164      310
#> 436 2021-12-10 03:00:00 Blankenberge  51.3106    3.1164      406
#> 437 2021-12-10 04:00:00 Blankenberge  51.3106    3.1164      413
#> 438 2021-12-10 05:00:00 Blankenberge  51.3106    3.1164      405
#> 439 2021-12-10 06:00:00 Blankenberge  51.3106    3.1164      370
#> 440 2021-12-10 07:00:00 Blankenberge  51.3106    3.1164      318
#> 441 2021-12-10 08:00:00 Blankenberge  51.3106    3.1164      245
#> 442 2021-12-10 09:00:00 Blankenberge  51.3106    3.1164      179
#> 443 2021-12-10 10:00:00 Blankenberge  51.3106    3.1164      129
#> 444 2021-12-10 11:00:00 Blankenberge  51.3106    3.1164      133
#> 445 2021-12-10 12:00:00 Blankenberge  51.3106    3.1164      195
#> 446 2021-12-10 13:00:00 Blankenberge  51.3106    3.1164      252
#> 447 2021-12-10 14:00:00 Blankenberge  51.3106    3.1164      345
#> 448 2021-12-10 15:00:00 Blankenberge  51.3106    3.1164      450
#> 449 2021-12-10 16:00:00 Blankenberge  51.3106    3.1164      503
#> 450 2021-12-10 17:00:00 Blankenberge  51.3106    3.1164      506
#> 451 2021-12-10 18:00:00 Blankenberge  51.3106    3.1164      477
#> 452 2021-12-10 19:00:00 Blankenberge  51.3106    3.1164      438
#> 453 2021-12-10 20:00:00 Blankenberge  51.3106    3.1164      371
#> 454 2021-12-10 21:00:00 Blankenberge  51.3106    3.1164      294
#> 455 2021-12-10 22:00:00 Blankenberge  51.3106    3.1164      209
#> 456 2021-12-10 23:00:00 Blankenberge  51.3106    3.1164      166
#> 457 2021-12-11 00:00:00 Blankenberge  51.3106    3.1164      187
#> 458 2021-12-11 01:00:00 Blankenberge  51.3106    3.1164      217
#> 459 2021-12-11 02:00:00 Blankenberge  51.3106    3.1164      257
#> 460 2021-12-11 03:00:00 Blankenberge  51.3106    3.1164      321
#> 461 2021-12-11 04:00:00 Blankenberge  51.3106    3.1164      400
#> 462 2021-12-11 05:00:00 Blankenberge  51.3106    3.1164      411
#> 463 2021-12-11 06:00:00 Blankenberge  51.3106    3.1164      401
#> 464 2021-12-11 07:00:00 Blankenberge  51.3106    3.1164      366
#> 465 2021-12-11 08:00:00 Blankenberge  51.3106    3.1164      308
#> 466 2021-12-11 09:00:00 Blankenberge  51.3106    3.1164      232
#> 467 2021-12-11 10:00:00 Blankenberge  51.3106    3.1164      155
#> 468 2021-12-11 11:00:00 Blankenberge  51.3106    3.1164       94
#> 469 2021-12-11 12:00:00 Blankenberge  51.3106    3.1164       58
#> 470 2021-12-11 13:00:00 Blankenberge  51.3106    3.1164       92
#> 471 2021-12-11 14:00:00 Blankenberge  51.3106    3.1164      150
#> 472 2021-12-11 15:00:00 Blankenberge  51.3106    3.1164      238
#> 473 2021-12-11 16:00:00 Blankenberge  51.3106    3.1164      350
#> 474 2021-12-11 17:00:00 Blankenberge  51.3106    3.1164      399
#> 475 2021-12-11 18:00:00 Blankenberge  51.3106    3.1164      405
#> 476 2021-12-11 19:00:00 Blankenberge  51.3106    3.1164      395
#> 477 2021-12-11 20:00:00 Blankenberge  51.3106    3.1164      352
#> 478 2021-12-11 21:00:00 Blankenberge  51.3106    3.1164      294
#> 479 2021-12-11 22:00:00 Blankenberge  51.3106    3.1164      223
#> 480 2021-12-11 23:00:00 Blankenberge  51.3106    3.1164      154
#> 481 2021-12-12 00:00:00 Blankenberge  51.3106    3.1164      105
#> 482 2021-12-12 01:00:00 Blankenberge  51.3106    3.1164      109
#> 483 2021-12-12 02:00:00 Blankenberge  51.3106    3.1164      146
#> 484 2021-12-12 03:00:00 Blankenberge  51.3106    3.1164      207
#> 485 2021-12-12 04:00:00 Blankenberge  51.3106    3.1164      299
#> 486 2021-12-12 05:00:00 Blankenberge  51.3106    3.1164      377
#> 487 2021-12-12 06:00:00 Blankenberge  51.3106    3.1164      394
#> 488 2021-12-12 07:00:00 Blankenberge  51.3106    3.1164      391
#> 489 2021-12-12 08:00:00 Blankenberge  51.3106    3.1164      361
#> 490 2021-12-12 09:00:00 Blankenberge  51.3106    3.1164      313
#> 491 2021-12-12 10:00:00 Blankenberge  51.3106    3.1164      252
#> 492 2021-12-12 11:00:00 Blankenberge  51.3106    3.1164      189
#> 493 2021-12-12 12:00:00 Blankenberge  51.3106    3.1164      127
#> 494 2021-12-12 13:00:00 Blankenberge  51.3106    3.1164       87
#> 495 2021-12-12 14:00:00 Blankenberge  51.3106    3.1164      117
#> 496 2021-12-12 15:00:00 Blankenberge  51.3106    3.1164      180
#> 497 2021-12-12 16:00:00 Blankenberge  51.3106    3.1164      265
#> 498 2021-12-12 17:00:00 Blankenberge  51.3106    3.1164      363
#> 499 2021-12-12 18:00:00 Blankenberge  51.3106    3.1164      412
#> 500 2021-12-12 19:00:00 Blankenberge  51.3106    3.1164      419
#> 501 2021-12-12 20:00:00 Blankenberge  51.3106    3.1164      412
#> 502 2021-12-12 21:00:00 Blankenberge  51.3106    3.1164      379
#> 503 2021-12-12 22:00:00 Blankenberge  51.3106    3.1164      320
#> 504 2021-12-12 23:00:00 Blankenberge  51.3106    3.1164      253
#> 505 2021-12-13 00:00:00 Blankenberge  51.3106    3.1164      186
#> 506 2021-12-13 01:00:00 Blankenberge  51.3106    3.1164      130
#> 507 2021-12-13 02:00:00 Blankenberge  51.3106    3.1164      122
#> 508 2021-12-13 03:00:00 Blankenberge  51.3106    3.1164      155
#> 509 2021-12-13 04:00:00 Blankenberge  51.3106    3.1164      203
#> 510 2021-12-13 05:00:00 Blankenberge  51.3106    3.1164      272
#> 511 2021-12-13 06:00:00 Blankenberge  51.3106    3.1164      344
#> 512 2021-12-13 07:00:00 Blankenberge  51.3106    3.1164      363
#> 513 2021-12-13 08:00:00 Blankenberge  51.3106    3.1164      362
#> 514 2021-12-13 09:00:00 Blankenberge  51.3106    3.1164      331
#> 515 2021-12-13 10:00:00 Blankenberge  51.3106    3.1164      289
#> 516 2021-12-13 11:00:00 Blankenberge  51.3106    3.1164      229
#> 517 2021-12-13 12:00:00 Blankenberge  51.3106    3.1164      164
#> 518 2021-12-13 13:00:00 Blankenberge  51.3106    3.1164      105
#> 519 2021-12-13 14:00:00 Blankenberge  51.3106    3.1164       68
#> 520 2021-12-13 22:00:00 Blankenberge  51.3106    3.1164      402
#> 521 2021-12-13 23:00:00 Blankenberge  51.3106    3.1164      361
#> 522 2021-12-14 00:00:00 Blankenberge  51.3106    3.1164      296
#> 523 2021-12-14 01:00:00 Blankenberge  51.3106    3.1164      218
#> 524 2021-12-14 02:00:00 Blankenberge  51.3106    3.1164      166
#> 525 2021-12-14 03:00:00 Blankenberge  51.3106    3.1164      158
#> 526 2021-12-14 04:00:00 Blankenberge  51.3106    3.1164      185
#> 527 2021-12-14 05:00:00 Blankenberge  51.3106    3.1164      225
#> 528 2021-12-14 06:00:00 Blankenberge  51.3106    3.1164      282
#> 529 2021-12-14 07:00:00 Blankenberge  51.3106    3.1164      354
#> 530 2021-12-14 08:00:00 Blankenberge  51.3106    3.1164      383
#> 531 2021-12-14 09:00:00 Blankenberge  51.3106    3.1164      381
#> 532 2021-12-14 10:00:00 Blankenberge  51.3106    3.1164      354
#> 533 2021-12-14 11:00:00 Blankenberge  51.3106    3.1164      309
#> 534 2021-12-14 12:00:00 Blankenberge  51.3106    3.1164      249
#> 535 2021-12-14 13:00:00 Blankenberge  51.3106    3.1164      173
#> 536 2021-12-14 14:00:00 Blankenberge  51.3106    3.1164      112
#> 537 2021-12-14 15:00:00 Blankenberge  51.3106    3.1164       73
#> 538 2021-12-14 16:00:00 Blankenberge  51.3106    3.1164       81
#> 539 2021-12-14 17:00:00 Blankenberge  51.3106    3.1164      126
#> 540 2021-12-14 18:00:00 Blankenberge  51.3106    3.1164      200
#> 541 2021-12-14 19:00:00 Blankenberge  51.3106    3.1164      305
#> 542 2021-12-14 20:00:00 Blankenberge  51.3106    3.1164      382
#> 543 2021-12-14 21:00:00 Blankenberge  51.3106    3.1164      398
#> 544 2021-12-14 22:00:00 Blankenberge  51.3106    3.1164      395
#> 545 2021-12-14 23:00:00 Blankenberge  51.3106    3.1164      368
#> 546 2021-12-15 00:00:00 Blankenberge  51.3106    3.1164      314
#> 547 2021-12-15 01:00:00 Blankenberge  51.3106    3.1164      246
#> 548 2021-12-15 02:00:00 Blankenberge  51.3106    3.1164      180
#> 549 2021-12-15 03:00:00 Blankenberge  51.3106    3.1164      127
#> 550 2021-12-15 04:00:00 Blankenberge  51.3106    3.1164      138
#> 551 2021-12-15 05:00:00 Blankenberge  51.3106    3.1164      177
#> 552 2021-12-15 06:00:00 Blankenberge  51.3106    3.1164      221
#> 553 2021-12-15 07:00:00 Blankenberge  51.3106    3.1164      297
#> 554 2021-12-15 08:00:00 Blankenberge  51.3106    3.1164      382
#> 555 2021-12-15 09:00:00 Blankenberge  51.3106    3.1164      412
#> 556 2021-12-15 10:00:00 Blankenberge  51.3106    3.1164      410
#> 557 2021-12-15 11:00:00 Blankenberge  51.3106    3.1164      377
#> 558 2021-12-15 12:00:00 Blankenberge  51.3106    3.1164      330
#> 559 2021-12-15 13:00:00 Blankenberge  51.3106    3.1164      260
#> 560 2021-12-15 14:00:00 Blankenberge  51.3106    3.1164      179
#> 561 2021-12-15 15:00:00 Blankenberge  51.3106    3.1164      114
#> 562 2021-12-15 16:00:00 Blankenberge  51.3106    3.1164       70
#> 563 2021-12-15 17:00:00 Blankenberge  51.3106    3.1164      100
#> 564 2021-12-15 18:00:00 Blankenberge  51.3106    3.1164      152
#> 565 2021-12-15 19:00:00 Blankenberge  51.3106    3.1164      226
#> 566 2021-12-15 20:00:00 Blankenberge  51.3106    3.1164      338
#> 567 2021-12-15 21:00:00 Blankenberge  51.3106    3.1164      417
#> 568 2021-12-15 22:00:00 Blankenberge  51.3106    3.1164      426
#> 569 2021-12-15 23:00:00 Blankenberge  51.3106    3.1164      415
#> 570 2021-12-16 00:00:00 Blankenberge  51.3106    3.1164      379
#> 571 2021-12-16 01:00:00 Blankenberge  51.3106    3.1164      313
#> 572 2021-12-16 02:00:00 Blankenberge  51.3106    3.1164      233
#> 573 2021-12-16 03:00:00 Blankenberge  51.3106    3.1164      155
#> 574 2021-12-16 04:00:00 Blankenberge  51.3106    3.1164      102
#> 575 2021-12-16 05:00:00 Blankenberge  51.3106    3.1164      122
#> 576 2021-12-16 06:00:00 Blankenberge  51.3106    3.1164      162
#> 577 2021-12-16 07:00:00 Blankenberge  51.3106    3.1164      215
#> 578 2021-12-16 08:00:00 Blankenberge  51.3106    3.1164      300
#> 579 2021-12-16 09:00:00 Blankenberge  51.3106    3.1164      398
#> 580 2021-12-16 10:00:00 Blankenberge  51.3106    3.1164      413
#> 581 2021-12-16 11:00:00 Blankenberge  51.3106    3.1164      407
#> 582 2021-12-16 12:00:00 Blankenberge  51.3106    3.1164      371
#> 583 2021-12-16 13:00:00 Blankenberge  51.3106    3.1164      311
#> 584 2021-12-16 14:00:00 Blankenberge  51.3106    3.1164      227
#> 585 2021-12-16 15:00:00 Blankenberge  51.3106    3.1164      143
#> 586 2021-12-16 16:00:00 Blankenberge  51.3106    3.1164       75
#> 587 2021-12-16 17:00:00 Blankenberge  51.3106    3.1164       51
#> 588 2021-12-16 18:00:00 Blankenberge  51.3106    3.1164       91
#> 589 2021-12-16 19:00:00 Blankenberge  51.3106    3.1164      142
#> 590 2021-12-16 20:00:00 Blankenberge  51.3106    3.1164      227
#> 591 2021-12-16 21:00:00 Blankenberge  51.3106    3.1164      353
#> 592 2021-12-16 22:00:00 Blankenberge  51.3106    3.1164      407
#> 593 2021-12-16 23:00:00 Blankenberge  51.3106    3.1164      411
#> 594 2021-12-17 00:00:00 Blankenberge  51.3106    3.1164      393
#> 595 2021-12-17 01:00:00 Blankenberge  51.3106    3.1164      344
#> 596 2021-12-17 02:00:00 Blankenberge  51.3106    3.1164      269
#> 597 2021-12-17 03:00:00 Blankenberge  51.3106    3.1164      181
#> 598 2021-12-17 04:00:00 Blankenberge  51.3106    3.1164      100
#> 599 2021-12-17 05:00:00 Blankenberge  51.3106    3.1164       72
#> 600 2021-12-17 06:00:00 Blankenberge  51.3106    3.1164      107
#> 601 2021-12-17 07:00:00 Blankenberge  51.3106    3.1164      150
#> 602 2021-12-17 08:00:00 Blankenberge  51.3106    3.1164      217
#> 603 2021-12-17 09:00:00 Blankenberge  51.3106    3.1164      334
#> 604 2021-12-17 10:00:00 Blankenberge  51.3106    3.1164      413
#> 605 2021-12-17 11:00:00 Blankenberge  51.3106    3.1164      419
#> 606 2021-12-17 12:00:00 Blankenberge  51.3106    3.1164      410
#> 607 2021-12-17 13:00:00 Blankenberge  51.3106    3.1164      363
#> 608 2021-12-17 14:00:00 Blankenberge  51.3106    3.1164      290
#> 609 2021-12-17 15:00:00 Blankenberge  51.3106    3.1164      206
#> 610 2021-12-17 16:00:00 Blankenberge  51.3106    3.1164      116
#> 611 2021-12-17 17:00:00 Blankenberge  51.3106    3.1164       57
#> 612 2021-12-17 18:00:00 Blankenberge  51.3106    3.1164       70
#> 613 2021-12-17 19:00:00 Blankenberge  51.3106    3.1164      112
#> 614 2021-12-17 20:00:00 Blankenberge  51.3106    3.1164      168
#> 615 2021-12-17 21:00:00 Blankenberge  51.3106    3.1164      260
#> 616 2021-12-17 22:00:00 Blankenberge  51.3106    3.1164      378
#> 617 2021-12-17 23:00:00 Blankenberge  51.3106    3.1164      406
#> 618 2021-12-18 00:00:00 Blankenberge  51.3106    3.1164      405
#> 619 2021-12-18 01:00:00 Blankenberge  51.3106    3.1164      374
#> 620 2021-12-18 02:00:00 Blankenberge  51.3106    3.1164      311
#> 621 2021-12-18 03:00:00 Blankenberge  51.3106    3.1164      223
#> 622 2021-12-18 04:00:00 Blankenberge  51.3106    3.1164      133
#> 623 2021-12-18 05:00:00 Blankenberge  51.3106    3.1164       66
#> 624 2021-12-18 06:00:00 Blankenberge  51.3106    3.1164       59
#> 625 2021-12-18 07:00:00 Blankenberge  51.3106    3.1164      101
#> 626 2021-12-18 08:00:00 Blankenberge  51.3106    3.1164      150
#> 627 2021-12-18 09:00:00 Blankenberge  51.3106    3.1164      226
#> 628 2021-12-18 10:00:00 Blankenberge  51.3106    3.1164      362
#> 629 2021-12-18 11:00:00 Blankenberge  51.3106    3.1164      411
#> 630 2021-12-18 12:00:00 Blankenberge  51.3106    3.1164      413
#> 631 2021-12-18 13:00:00 Blankenberge  51.3106    3.1164      388
#> 632 2021-12-18 14:00:00 Blankenberge  51.3106    3.1164      333
#> 633 2021-12-18 15:00:00 Blankenberge  51.3106    3.1164      253
#> 634 2021-12-18 16:00:00 Blankenberge  51.3106    3.1164      163
#> 635 2021-12-18 17:00:00 Blankenberge  51.3106    3.1164       83
#> 636 2021-12-18 18:00:00 Blankenberge  51.3106    3.1164       45
#> 637 2021-12-18 19:00:00 Blankenberge  51.3106    3.1164       91
#> 638 2021-12-18 20:00:00 Blankenberge  51.3106    3.1164      137
#> 639 2021-12-18 21:00:00 Blankenberge  51.3106    3.1164      203
#> 640 2021-12-18 22:00:00 Blankenberge  51.3106    3.1164      323
#> 641 2021-12-18 23:00:00 Blankenberge  51.3106    3.1164      416
#> 642 2021-12-19 00:00:00 Blankenberge  51.3106    3.1164      425
#> 643 2021-12-19 01:00:00 Blankenberge  51.3106    3.1164      412
#> 644 2021-12-19 02:00:00 Blankenberge  51.3106    3.1164      366
#> 645 2021-12-19 03:00:00 Blankenberge  51.3106    3.1164      292
#> 646 2021-12-19 04:00:00 Blankenberge  51.3106    3.1164      200
#> 647 2021-12-19 05:00:00 Blankenberge  51.3106    3.1164      110
#> 648 2021-12-19 06:00:00 Blankenberge  51.3106    3.1164       56
#> 649 2021-12-19 07:00:00 Blankenberge  51.3106    3.1164       82
#> 650 2021-12-19 08:00:00 Blankenberge  51.3106    3.1164      122
#> 651 2021-12-19 09:00:00 Blankenberge  51.3106    3.1164      187
#> 652 2021-12-19 10:00:00 Blankenberge  51.3106    3.1164      296
#> 653 2021-12-19 11:00:00 Blankenberge  51.3106    3.1164      415
#> 654 2021-12-19 12:00:00 Blankenberge  51.3106    3.1164      432
#> 655 2021-12-19 13:00:00 Blankenberge  51.3106    3.1164      426
#> 656 2021-12-19 14:00:00 Blankenberge  51.3106    3.1164      388
#> 657 2021-12-19 15:00:00 Blankenberge  51.3106    3.1164      322
#> 658 2021-12-19 16:00:00 Blankenberge  51.3106    3.1164      232
#> 659 2021-12-19 17:00:00 Blankenberge  51.3106    3.1164      142
#> 660 2021-12-19 18:00:00 Blankenberge  51.3106    3.1164       73
#> 661 2021-12-19 19:00:00 Blankenberge  51.3106    3.1164       84
#> 662 2021-12-19 20:00:00 Blankenberge  51.3106    3.1164      126
#> 663 2021-12-19 21:00:00 Blankenberge  51.3106    3.1164      174
#> 664 2021-12-19 22:00:00 Blankenberge  51.3106    3.1164      255
#> 665 2021-12-19 23:00:00 Blankenberge  51.3106    3.1164      392
#> 666 2021-12-20 00:00:00 Blankenberge  51.3106    3.1164      433
#> 667 2021-12-20 01:00:00 Blankenberge  51.3106    3.1164      433
#> 668 2021-12-20 02:00:00 Blankenberge  51.3106    3.1164      399
#> 669 2021-12-20 03:00:00 Blankenberge  51.3106    3.1164      342
#> 670 2021-12-20 04:00:00 Blankenberge  51.3106    3.1164      255
#> 671 2021-12-20 05:00:00 Blankenberge  51.3106    3.1164      158
#> 672 2021-12-20 06:00:00 Blankenberge  51.3106    3.1164       78
#> 673 2021-12-20 07:00:00 Blankenberge  51.3106    3.1164       44
#> 674 2021-12-20 08:00:00 Blankenberge  51.3106    3.1164       93
#> 675 2021-12-20 09:00:00 Blankenberge  51.3106    3.1164      141
#> 676 2021-12-20 10:00:00 Blankenberge  51.3106    3.1164      220
#> 677 2021-12-20 11:00:00 Blankenberge  51.3106    3.1164      365
#> 678 2021-12-20 12:00:00 Blankenberge  51.3106    3.1164      429
#> 679 2021-12-20 13:00:00 Blankenberge  51.3106    3.1164      432
#> 680 2021-12-20 14:00:00 Blankenberge  51.3106    3.1164      412
#> 681 2021-12-20 15:00:00 Blankenberge  51.3106    3.1164      360
#> 682 2021-12-20 16:00:00 Blankenberge  51.3106    3.1164      280
#> 683 2021-12-20 17:00:00 Blankenberge  51.3106    3.1164      188
#> 684 2021-12-20 18:00:00 Blankenberge  51.3106    3.1164      102
#> 685 2021-12-20 19:00:00 Blankenberge  51.3106    3.1164       50
#> 686 2021-12-20 20:00:00 Blankenberge  51.3106    3.1164       91
#> 687 2021-12-20 21:00:00 Blankenberge  51.3106    3.1164      136
#> 688 2021-12-20 22:00:00 Blankenberge  51.3106    3.1164      198
#> 689 2021-12-20 23:00:00 Blankenberge  51.3106    3.1164      314
#> 690 2021-12-21 00:00:00 Blankenberge  51.3106    3.1164      421
#> 691 2021-12-21 01:00:00 Blankenberge  51.3106    3.1164      429
#> 692 2021-12-21 02:00:00 Blankenberge  51.3106    3.1164      420
#> 693 2021-12-21 03:00:00 Blankenberge  51.3106    3.1164      374
#> 694 2021-12-21 04:00:00 Blankenberge  51.3106    3.1164      304
#> 695 2021-12-21 05:00:00 Blankenberge  51.3106    3.1164      212
#> 696 2021-12-21 06:00:00 Blankenberge  51.3106    3.1164      121
#> 697 2021-12-21 07:00:00 Blankenberge  51.3106    3.1164       55
#> 698 2021-12-21 08:00:00 Blankenberge  51.3106    3.1164       65
#> 699 2021-12-21 09:00:00 Blankenberge  51.3106    3.1164      110
#> 700 2021-12-21 10:00:00 Blankenberge  51.3106    3.1164      172
#> 701 2021-12-21 11:00:00 Blankenberge  51.3106    3.1164      288
#> 702 2021-12-21 12:00:00 Blankenberge  51.3106    3.1164      417
#> 703 2021-12-21 13:00:00 Blankenberge  51.3106    3.1164      440
#> 704 2021-12-21 14:00:00 Blankenberge  51.3106    3.1164      435
#> 705 2021-12-21 15:00:00 Blankenberge  51.3106    3.1164      399
#> 706 2021-12-21 16:00:00 Blankenberge  51.3106    3.1164      333
#> 707 2021-12-21 17:00:00 Blankenberge  51.3106    3.1164      244
#> 708 2021-12-21 18:00:00 Blankenberge  51.3106    3.1164      157
#> 709 2021-12-21 19:00:00 Blankenberge  51.3106    3.1164       82
#> 710 2021-12-21 20:00:00 Blankenberge  51.3106    3.1164       79
#> 711 2021-12-21 21:00:00 Blankenberge  51.3106    3.1164      118
#> 712 2021-12-21 22:00:00 Blankenberge  51.3106    3.1164      166
#> 713 2021-12-21 23:00:00 Blankenberge  51.3106    3.1164      243
#> 714 2021-12-22 00:00:00 Blankenberge  51.3106    3.1164      381
#> 715 2021-12-22 01:00:00 Blankenberge  51.3106    3.1164      430
#> 716 2021-12-22 02:00:00 Blankenberge  51.3106    3.1164      433
#> 717 2021-12-22 03:00:00 Blankenberge  51.3106    3.1164      403
#> 718 2021-12-22 04:00:00 Blankenberge  51.3106    3.1164      345
#> 719 2021-12-22 05:00:00 Blankenberge  51.3106    3.1164      266
#> 720 2021-12-22 06:00:00 Blankenberge  51.3106    3.1164      176
#> 721 2021-12-22 07:00:00 Blankenberge  51.3106    3.1164       91
#> 722 2021-12-22 08:00:00 Blankenberge  51.3106    3.1164       41
#> 723 2021-12-22 09:00:00 Blankenberge  51.3106    3.1164       82
#> 724 2021-12-22 10:00:00 Blankenberge  51.3106    3.1164      129
#> 725 2021-12-22 11:00:00 Blankenberge  51.3106    3.1164      212
#> 726 2021-12-22 12:00:00 Blankenberge  51.3106    3.1164      223
#> 
#> $par
#> $par$daterange
#> [1] "2021-11-22" "2021-12-23"
#> 
#> $par$type
#> [1] "MVB data"
#> 
#> $par$stations
#> [1] "Blankenberge"
#> 
#> $par$calc
#> [1] "max"
#> 
#> $par$binSize
#> [1] "hour"
#> 
#> $par$parameters
#> [1] "Tide TAW"
#> 
#>