Retrieve MVB-data from the LifeWatch project

Retrieves the Meetnet Vlaams Banken (MVB) data from the LifeWatch project. Without valid credentials you are only allowed to view the "Tide TAW" parameter for the last 30 days, grouped by day/hour. To get an account, register via the Lifewatch RShiny registration webpage.

getMvbData(
  startdate,
  stopdate,
  parameters,
  stations = NULL,
  by,
  calc,
  usr = NULL,
  pwd = NULL,
  params = FALSE,
  ...
)

Arguments

startdate

Starting date for the query. Without a login, this is restricted to the last month

stopdate

Stopping date for the query. Without a login, this is restricted to the last month

parameters

One (or list) of the $parameter in listMvbStations.

stations

List of stations to be included in the query. See listMvbStations. Use stations="All" or stations=NULL to get all stations.

by

Time grouping for calculation, one of ('day','hour', '10min' ,'none')

calc

Calculation to perform given time grouping, one of ('avg', 'max', 'min', 'none').

usr

Username to connect to database

pwd

Password to connect to database

params

If TRUE, returns a list with the dataset and the query parameters applied in the server side. IF FALSE returns only the data.

...

Params to be passed to lw_check_lwdataserver().

Value

Dataframe with the specified MVB data.

Examples

getMvbData(Sys.Date() - 30, Sys.Date() + 1, parameters = 'Tide TAW',
stations = "Nieuwpoort", by = "day", calc = "avg")
#> - Query mode: Post request to OpenCPU server
#> No encoding supplied: defaulting to UTF-8.
#> - Your query:
#>    - daterange: c("2022-12-19", "2023-01-19")
#>    - type: MVB data
#>    - stations: Nieuwpoort
#>    - calc: avg
#>    - binSize: day
#>    - parameters: Tide TAW
#> - Server query:
#>    - daterange: c("2022-12-19", "2023-01-19")
#>    - type: MVB data
#>    - stations: Nieuwpoort
#>    - calc: avg
#>    - binSize: day
#>    - parameters: Tide (cmTAW)
#> Warning: The query applied on the server differ from the parameters you used.
#> Hint: You may need an account to fully access the data under moratorium.
#> Hint: Request access at: https://rshiny.lifewatch.be/account?p=register
#>          time    station latitude longitude Tide (cmTAW)
#> 1  2022-12-19 Nieuwpoort  51.1506    2.7281     236.1319
#> 2  2022-12-20 Nieuwpoort  51.1506    2.7281     273.7073
#> 3  2022-12-21 Nieuwpoort  51.1506    2.7281     256.2569
#> 4  2022-12-22 Nieuwpoort  51.1506    2.7281     268.4653
#> 5  2022-12-23 Nieuwpoort  51.1506    2.7281     266.9896
#> 6  2022-12-24 Nieuwpoort  51.1506    2.7281     245.3368
#> 7  2022-12-25 Nieuwpoort  51.1506    2.7281     244.6042
#> 8  2022-12-26 Nieuwpoort  51.1506    2.7281     277.1944
#> 9  2022-12-27 Nieuwpoort  51.1506    2.7281     256.8403
#> 10 2022-12-28 Nieuwpoort  51.1506    2.7281     249.3937
#> 11 2022-12-29 Nieuwpoort  51.1506    2.7281     271.4688
#> 12 2022-12-30 Nieuwpoort  51.1506    2.7281     247.4132
#> 13 2022-12-31 Nieuwpoort  51.1506    2.7281     280.3264
#> 14 2023-01-01 Nieuwpoort  51.1506    2.7281     265.8328
#> 15 2023-01-02 Nieuwpoort  51.1506    2.7281     265.8090
#> 16 2023-01-03 Nieuwpoort  51.1506    2.7281     237.7569
#> 17 2023-01-04 Nieuwpoort  51.1506    2.7281     266.4634
#> 18 2023-01-05 Nieuwpoort  51.1506    2.7281     258.7708
#> 19 2023-01-06 Nieuwpoort  51.1506    2.7281     248.1354
#> 20 2023-01-07 Nieuwpoort  51.1506    2.7281     240.2986
#> 21 2023-01-08 Nieuwpoort  51.1506    2.7281     263.7569
#> 22 2023-01-09 Nieuwpoort  51.1506    2.7281     296.9271
#> 23 2023-01-10 Nieuwpoort  51.1506    2.7281     241.4757
#> 24 2023-01-11 Nieuwpoort  51.1506    2.7281     269.7431
#> 25 2023-01-12 Nieuwpoort  51.1506    2.7281     276.4653
#> 26 2023-01-13 Nieuwpoort  51.1506    2.7281     283.2431
#> 27 2023-01-14 Nieuwpoort  51.1506    2.7281     252.6899
#> 28 2023-01-15 Nieuwpoort  51.1506    2.7281     294.4271
#> 29 2023-01-16 Nieuwpoort  51.1506    2.7281     280.8715
#> 30 2023-01-17 Nieuwpoort  51.1506    2.7281     269.6354
#> 31 2023-01-18 Nieuwpoort  51.1506    2.7281     262.3254
getMvbData(Sys.Date() - 30, Sys.Date() + 1, parameters = 'Tide TAW',
stations = "Blankenberge", by = "hour", calc = "max", params = TRUE)
#> - Query mode: Post request to OpenCPU server
#> No encoding supplied: defaulting to UTF-8.
#> - Your query:
#>    - daterange: c("2022-12-19", "2023-01-19")
#>    - type: MVB data
#>    - stations: Blankenberge
#>    - calc: max
#>    - binSize: hour
#>    - parameters: Tide TAW
#> - Server query:
#>    - daterange: c("2022-12-19", "2023-01-19")
#>    - type: MVB data
#>    - stations: Blankenberge
#>    - calc: max
#>    - binSize: hour
#>    - parameters: Tide (cmTAW)
#> Warning: The query applied on the server differ from the parameters you used.
#> Hint: You may need an account to fully access the data under moratorium.
#> Hint: Request access at: https://rshiny.lifewatch.be/account?p=register
#> $mdf
#>                    time      station latitude longitude Tide (cmTAW)
#> 1   2022-12-19 00:00:00 Blankenberge  51.3106    3.1164          198
#> 2   2022-12-19 01:00:00 Blankenberge  51.3106    3.1164          137
#> 3   2022-12-19 02:00:00 Blankenberge  51.3106    3.1164           94
#> 4   2022-12-19 03:00:00 Blankenberge  51.3106    3.1164          102
#> 5   2022-12-19 04:00:00 Blankenberge  51.3106    3.1164          145
#> 6   2022-12-19 05:00:00 Blankenberge  51.3106    3.1164          230
#> 7   2022-12-19 06:00:00 Blankenberge  51.3106    3.1164          312
#> 8   2022-12-19 07:00:00 Blankenberge  51.3106    3.1164          350
#> 9   2022-12-19 08:00:00 Blankenberge  51.3106    3.1164          357
#> 10  2022-12-19 09:00:00 Blankenberge  51.3106    3.1164          346
#> 11  2022-12-19 10:00:00 Blankenberge  51.3106    3.1164          318
#> 12  2022-12-19 11:00:00 Blankenberge  51.3106    3.1164          266
#> 13  2022-12-19 12:00:00 Blankenberge  51.3106    3.1164          213
#> 14  2022-12-19 13:00:00 Blankenberge  51.3106    3.1164          152
#> 15  2022-12-19 14:00:00 Blankenberge  51.3106    3.1164           96
#> 16  2022-12-19 15:00:00 Blankenberge  51.3106    3.1164           83
#> 17  2022-12-19 16:00:00 Blankenberge  51.3106    3.1164          133
#> 18  2022-12-19 17:00:00 Blankenberge  51.3106    3.1164          212
#> 19  2022-12-19 18:00:00 Blankenberge  51.3106    3.1164          320
#> 20  2022-12-19 19:00:00 Blankenberge  51.3106    3.1164          392
#> 21  2022-12-19 20:00:00 Blankenberge  51.3106    3.1164          414
#> 22  2022-12-19 21:00:00 Blankenberge  51.3106    3.1164          415
#> 23  2022-12-19 22:00:00 Blankenberge  51.3106    3.1164          398
#> 24  2022-12-19 23:00:00 Blankenberge  51.3106    3.1164          356
#> 25  2022-12-20 00:00:00 Blankenberge  51.3106    3.1164          298
#> 26  2022-12-20 01:00:00 Blankenberge  51.3106    3.1164          228
#> 27  2022-12-20 02:00:00 Blankenberge  51.3106    3.1164          161
#> 28  2022-12-20 03:00:00 Blankenberge  51.3106    3.1164          121
#> 29  2022-12-20 04:00:00 Blankenberge  51.3106    3.1164          149
#> 30  2022-12-20 05:00:00 Blankenberge  51.3106    3.1164          189
#> 31  2022-12-20 06:00:00 Blankenberge  51.3106    3.1164          256
#> 32  2022-12-20 07:00:00 Blankenberge  51.3106    3.1164          355
#> 33  2022-12-20 08:00:00 Blankenberge  51.3106    3.1164          420
#> 34  2022-12-20 09:00:00 Blankenberge  51.3106    3.1164          426
#> 35  2022-12-20 10:00:00 Blankenberge  51.3106    3.1164          415
#> 36  2022-12-20 11:00:00 Blankenberge  51.3106    3.1164          379
#> 37  2022-12-20 12:00:00 Blankenberge  51.3106    3.1164          318
#> 38  2022-12-20 13:00:00 Blankenberge  51.3106    3.1164          250
#> 39  2022-12-20 14:00:00 Blankenberge  51.3106    3.1164          177
#> 40  2022-12-20 15:00:00 Blankenberge  51.3106    3.1164          111
#> 41  2022-12-20 16:00:00 Blankenberge  51.3106    3.1164           99
#> 42  2022-12-20 20:00:00 Blankenberge  51.3106    3.1164          415
#> 43  2022-12-20 21:00:00 Blankenberge  51.3106    3.1164          454
#> 44  2022-12-20 22:00:00 Blankenberge  51.3106    3.1164          453
#> 45  2022-12-20 23:00:00 Blankenberge  51.3106    3.1164          430
#> 46  2022-12-21 00:00:00 Blankenberge  51.3106    3.1164          378
#> 47  2022-12-21 01:00:00 Blankenberge  51.3106    3.1164          305
#> 48  2022-12-21 02:00:00 Blankenberge  51.3106    3.1164          225
#> 49  2022-12-21 03:00:00 Blankenberge  51.3106    3.1164          149
#> 50  2022-12-21 04:00:00 Blankenberge  51.3106    3.1164          109
#> 51  2022-12-21 05:00:00 Blankenberge  51.3106    3.1164          148
#> 52  2022-12-21 06:00:00 Blankenberge  51.3106    3.1164          186
#> 53  2022-12-21 07:00:00 Blankenberge  51.3106    3.1164          250
#> 54  2022-12-21 08:00:00 Blankenberge  51.3106    3.1164          366
#> 55  2022-12-21 09:00:00 Blankenberge  51.3106    3.1164          438
#> 56  2022-12-21 10:00:00 Blankenberge  51.3106    3.1164          443
#> 57  2022-12-21 11:00:00 Blankenberge  51.3106    3.1164          430
#> 58  2022-12-21 12:00:00 Blankenberge  51.3106    3.1164          380
#> 59  2022-12-21 13:00:00 Blankenberge  51.3106    3.1164          313
#> 60  2022-12-21 14:00:00 Blankenberge  51.3106    3.1164          232
#> 61  2022-12-21 15:00:00 Blankenberge  51.3106    3.1164          152
#> 62  2022-12-21 16:00:00 Blankenberge  51.3106    3.1164           86
#> 63  2022-12-21 17:00:00 Blankenberge  51.3106    3.1164           90
#> 64  2022-12-21 18:00:00 Blankenberge  51.3106    3.1164          141
#> 65  2022-12-21 19:00:00 Blankenberge  51.3106    3.1164          203
#> 66  2022-12-21 20:00:00 Blankenberge  51.3106    3.1164          313
#> 67  2022-12-21 21:00:00 Blankenberge  51.3106    3.1164          438
#> 68  2022-12-21 22:00:00 Blankenberge  51.3106    3.1164          464
#> 69  2022-12-21 23:00:00 Blankenberge  51.3106    3.1164          463
#> 70  2022-12-22 00:00:00 Blankenberge  51.3106    3.1164          431
#> 71  2022-12-22 01:00:00 Blankenberge  51.3106    3.1164          375
#> 72  2022-12-22 02:00:00 Blankenberge  51.3106    3.1164          296
#> 73  2022-12-22 03:00:00 Blankenberge  51.3106    3.1164          212
#> 74  2022-12-22 04:00:00 Blankenberge  51.3106    3.1164          129
#> 75  2022-12-22 05:00:00 Blankenberge  51.3106    3.1164          107
#> 76  2022-12-22 06:00:00 Blankenberge  51.3106    3.1164          156
#> 77  2022-12-22 07:00:00 Blankenberge  51.3106    3.1164          199
#> 78  2022-12-22 08:00:00 Blankenberge  51.3106    3.1164          276
#> 79  2022-12-22 09:00:00 Blankenberge  51.3106    3.1164          418
#> 80  2022-12-22 10:00:00 Blankenberge  51.3106    3.1164          478
#> 81  2022-12-22 11:00:00 Blankenberge  51.3106    3.1164          480
#> 82  2022-12-22 12:00:00 Blankenberge  51.3106    3.1164          456
#> 83  2022-12-22 13:00:00 Blankenberge  51.3106    3.1164          405
#> 84  2022-12-22 14:00:00 Blankenberge  51.3106    3.1164          328
#> 85  2022-12-22 15:00:00 Blankenberge  51.3106    3.1164          241
#> 86  2022-12-22 16:00:00 Blankenberge  51.3106    3.1164          157
#> 87  2022-12-22 17:00:00 Blankenberge  51.3106    3.1164           89
#> 88  2022-12-22 18:00:00 Blankenberge  51.3106    3.1164          119
#> 89  2022-12-22 19:00:00 Blankenberge  51.3106    3.1164          165
#> 90  2022-12-22 20:00:00 Blankenberge  51.3106    3.1164          227
#> 91  2022-12-22 21:00:00 Blankenberge  51.3106    3.1164          354
#> 92  2022-12-22 22:00:00 Blankenberge  51.3106    3.1164          479
#> 93  2022-12-22 23:00:00 Blankenberge  51.3106    3.1164          491
#> 94  2022-12-23 00:00:00 Blankenberge  51.3106    3.1164          483
#> 95  2022-12-23 01:00:00 Blankenberge  51.3106    3.1164          444
#> 96  2022-12-23 02:00:00 Blankenberge  51.3106    3.1164          374
#> 97  2022-12-23 03:00:00 Blankenberge  51.3106    3.1164          283
#> 98  2022-12-23 04:00:00 Blankenberge  51.3106    3.1164          190
#> 99  2022-12-23 05:00:00 Blankenberge  51.3106    3.1164          109
#> 100 2022-12-23 06:00:00 Blankenberge  51.3106    3.1164           93
#> 101 2022-12-23 07:00:00 Blankenberge  51.3106    3.1164          139
#> 102 2022-12-23 08:00:00 Blankenberge  51.3106    3.1164          190
#> 103 2022-12-23 09:00:00 Blankenberge  51.3106    3.1164          277
#> 104 2022-12-23 10:00:00 Blankenberge  51.3106    3.1164          443
#> 105 2022-12-23 11:00:00 Blankenberge  51.3106    3.1164          487
#> 106 2022-12-23 12:00:00 Blankenberge  51.3106    3.1164          493
#> 107 2022-12-23 13:00:00 Blankenberge  51.3106    3.1164          472
#> 108 2022-12-23 14:00:00 Blankenberge  51.3106    3.1164          408
#> 109 2022-12-23 15:00:00 Blankenberge  51.3106    3.1164          328
#> 110 2022-12-23 16:00:00 Blankenberge  51.3106    3.1164          233
#> 111 2022-12-23 17:00:00 Blankenberge  51.3106    3.1164          149
#> 112 2022-12-23 18:00:00 Blankenberge  51.3106    3.1164           84
#> 113 2022-12-23 19:00:00 Blankenberge  51.3106    3.1164          129
#> 114 2022-12-23 20:00:00 Blankenberge  51.3106    3.1164          173
#> 115 2022-12-23 21:00:00 Blankenberge  51.3106    3.1164          237
#> 116 2022-12-23 22:00:00 Blankenberge  51.3106    3.1164          380
#> 117 2022-12-23 23:00:00 Blankenberge  51.3106    3.1164          478
#> 118 2022-12-24 00:00:00 Blankenberge  51.3106    3.1164          485
#> 119 2022-12-24 01:00:00 Blankenberge  51.3106    3.1164          468
#> 120 2022-12-24 02:00:00 Blankenberge  51.3106    3.1164          422
#> 121 2022-12-24 03:00:00 Blankenberge  51.3106    3.1164          344
#> 122 2022-12-24 04:00:00 Blankenberge  51.3106    3.1164          255
#> 123 2022-12-24 05:00:00 Blankenberge  51.3106    3.1164          165
#> 124 2022-12-24 06:00:00 Blankenberge  51.3106    3.1164           84
#> 125 2022-12-24 07:00:00 Blankenberge  51.3106    3.1164           79
#> 126 2022-12-24 08:00:00 Blankenberge  51.3106    3.1164          128
#> 127 2022-12-24 09:00:00 Blankenberge  51.3106    3.1164          187
#> 128 2022-12-24 10:00:00 Blankenberge  51.3106    3.1164          297
#> 129 2022-12-24 11:00:00 Blankenberge  51.3106    3.1164          459
#> 130 2022-12-24 12:00:00 Blankenberge  51.3106    3.1164          482
#> 131 2022-12-24 13:00:00 Blankenberge  51.3106    3.1164          480
#> 132 2022-12-24 14:00:00 Blankenberge  51.3106    3.1164          441
#> 133 2022-12-24 15:00:00 Blankenberge  51.3106    3.1164          374
#> 134 2022-12-24 16:00:00 Blankenberge  51.3106    3.1164          281
#> 135 2022-12-24 17:00:00 Blankenberge  51.3106    3.1164          184
#> 136 2022-12-24 18:00:00 Blankenberge  51.3106    3.1164           92
#> 137 2022-12-24 19:00:00 Blankenberge  51.3106    3.1164           41
#> 138 2022-12-24 20:00:00 Blankenberge  51.3106    3.1164           91
#> 139 2022-12-24 21:00:00 Blankenberge  51.3106    3.1164          142
#> 140 2022-12-24 22:00:00 Blankenberge  51.3106    3.1164          216
#> 141 2022-12-24 23:00:00 Blankenberge  51.3106    3.1164          383
#> 142 2022-12-25 00:00:00 Blankenberge  51.3106    3.1164          461
#> 143 2022-12-25 01:00:00 Blankenberge  51.3106    3.1164          464
#> 144 2022-12-25 02:00:00 Blankenberge  51.3106    3.1164          438
#> 145 2022-12-25 03:00:00 Blankenberge  51.3106    3.1164          386
#> 146 2022-12-25 04:00:00 Blankenberge  51.3106    3.1164          304
#> 147 2022-12-25 05:00:00 Blankenberge  51.3106    3.1164          209
#> 148 2022-12-25 06:00:00 Blankenberge  51.3106    3.1164          114
#> 149 2022-12-25 07:00:00 Blankenberge  51.3106    3.1164           44
#> 150 2022-12-25 08:00:00 Blankenberge  51.3106    3.1164           65
#> 151 2022-12-25 09:00:00 Blankenberge  51.3106    3.1164          113
#> 152 2022-12-25 10:00:00 Blankenberge  51.3106    3.1164          184
#> 153 2022-12-25 11:00:00 Blankenberge  51.3106    3.1164          325
#> 154 2022-12-25 12:00:00 Blankenberge  51.3106    3.1164          471
#> 155 2022-12-25 13:00:00 Blankenberge  51.3106    3.1164          485
#> 156 2022-12-25 14:00:00 Blankenberge  51.3106    3.1164          480
#> 157 2022-12-25 15:00:00 Blankenberge  51.3106    3.1164          437
#> 158 2022-12-25 16:00:00 Blankenberge  51.3106    3.1164          365
#> 159 2022-12-25 17:00:00 Blankenberge  51.3106    3.1164          272
#> 160 2022-12-25 18:00:00 Blankenberge  51.3106    3.1164          177
#> 161 2022-12-25 19:00:00 Blankenberge  51.3106    3.1164           91
#> 162 2022-12-25 20:00:00 Blankenberge  51.3106    3.1164           86
#> 163 2022-12-25 21:00:00 Blankenberge  51.3106    3.1164          135
#> 164 2022-12-25 22:00:00 Blankenberge  51.3106    3.1164          190
#> 165 2022-12-25 23:00:00 Blankenberge  51.3106    3.1164          266
#> 166 2022-12-26 00:00:00 Blankenberge  51.3106    3.1164          436
#> 167 2022-12-26 01:00:00 Blankenberge  51.3106    3.1164          494
#> 168 2022-12-26 02:00:00 Blankenberge  51.3106    3.1164          492
#> 169 2022-12-26 03:00:00 Blankenberge  51.3106    3.1164          456
#> 170 2022-12-26 04:00:00 Blankenberge  51.3106    3.1164          399
#> 171 2022-12-26 05:00:00 Blankenberge  51.3106    3.1164          312
#> 172 2022-12-26 06:00:00 Blankenberge  51.3106    3.1164          222
#> 173 2022-12-26 07:00:00 Blankenberge  51.3106    3.1164          133
#> 174 2022-12-26 08:00:00 Blankenberge  51.3106    3.1164           79
#> 175 2022-12-26 09:00:00 Blankenberge  51.3106    3.1164          133
#> 176 2022-12-26 10:00:00 Blankenberge  51.3106    3.1164          178
#> 177 2022-12-26 11:00:00 Blankenberge  51.3106    3.1164          246
#> 178 2022-12-26 12:00:00 Blankenberge  51.3106    3.1164          397
#> 179 2022-12-26 13:00:00 Blankenberge  51.3106    3.1164          509
#> 180 2022-12-26 14:00:00 Blankenberge  51.3106    3.1164          514
#> 181 2022-12-26 15:00:00 Blankenberge  51.3106    3.1164          498
#> 182 2022-12-26 16:00:00 Blankenberge  51.3106    3.1164          450
#> 183 2022-12-26 17:00:00 Blankenberge  51.3106    3.1164          372
#> 184 2022-12-26 18:00:00 Blankenberge  51.3106    3.1164          275
#> 185 2022-12-26 19:00:00 Blankenberge  51.3106    3.1164          181
#> 186 2022-12-26 20:00:00 Blankenberge  51.3106    3.1164          105
#> 187 2022-12-26 21:00:00 Blankenberge  51.3106    3.1164          112
#> 188 2022-12-26 22:00:00 Blankenberge  51.3106    3.1164          157
#> 189 2022-12-26 23:00:00 Blankenberge  51.3106    3.1164          205
#> 190 2022-12-27 00:00:00 Blankenberge  51.3106    3.1164          283
#> 191 2022-12-27 01:00:00 Blankenberge  51.3106    3.1164          444
#> 192 2022-12-27 02:00:00 Blankenberge  51.3106    3.1164          485
#> 193 2022-12-27 03:00:00 Blankenberge  51.3106    3.1164          484
#> 194 2022-12-27 04:00:00 Blankenberge  51.3106    3.1164          451
#> 195 2022-12-27 05:00:00 Blankenberge  51.3106    3.1164          390
#> 196 2022-12-27 06:00:00 Blankenberge  51.3106    3.1164          303
#> 197 2022-12-27 07:00:00 Blankenberge  51.3106    3.1164          208
#> 198 2022-12-27 08:00:00 Blankenberge  51.3106    3.1164          122
#> 199 2022-12-27 09:00:00 Blankenberge  51.3106    3.1164           65
#> 200 2022-12-27 10:00:00 Blankenberge  51.3106    3.1164          110
#> 201 2022-12-27 11:00:00 Blankenberge  51.3106    3.1164          160
#> 202 2022-12-27 12:00:00 Blankenberge  51.3106    3.1164          227
#> 203 2022-12-27 13:00:00 Blankenberge  51.3106    3.1164          383
#> 204 2022-12-27 14:00:00 Blankenberge  51.3106    3.1164          468
#> 205 2022-12-27 15:00:00 Blankenberge  51.3106    3.1164          470
#> 206 2022-12-27 16:00:00 Blankenberge  51.3106    3.1164          449
#> 207 2022-12-27 17:00:00 Blankenberge  51.3106    3.1164          397
#> 208 2022-12-27 18:00:00 Blankenberge  51.3106    3.1164          317
#> 209 2022-12-27 19:00:00 Blankenberge  51.3106    3.1164          221
#> 210 2022-12-27 20:00:00 Blankenberge  51.3106    3.1164          134
#> 211 2022-12-27 21:00:00 Blankenberge  51.3106    3.1164           64
#> 212 2022-12-27 22:00:00 Blankenberge  51.3106    3.1164           61
#> 213 2022-12-27 23:00:00 Blankenberge  51.3106    3.1164           99
#> 214 2022-12-28 00:00:00 Blankenberge  51.3106    3.1164          158
#> 215 2022-12-28 01:00:00 Blankenberge  51.3106    3.1164          267
#> 216 2022-12-28 02:00:00 Blankenberge  51.3106    3.1164          410
#> 217 2022-12-28 03:00:00 Blankenberge  51.3106    3.1164          422
#> 218 2022-12-28 04:00:00 Blankenberge  51.3106    3.1164          418
#> 219 2022-12-28 05:00:00 Blankenberge  51.3106    3.1164          380
#> 220 2022-12-28 06:00:00 Blankenberge  51.3106    3.1164          321
#> 221 2022-12-28 07:00:00 Blankenberge  51.3106    3.1164          246
#> 222 2022-12-28 08:00:00 Blankenberge  51.3106    3.1164          167
#> 223 2022-12-28 09:00:00 Blankenberge  51.3106    3.1164           90
#> 224 2022-12-28 10:00:00 Blankenberge  51.3106    3.1164           37
#> 225 2022-12-28 11:00:00 Blankenberge  51.3106    3.1164           84
#> 226 2022-12-28 12:00:00 Blankenberge  51.3106    3.1164          155
#> 227 2022-12-28 13:00:00 Blankenberge  51.3106    3.1164          266
#> 228 2022-12-28 14:00:00 Blankenberge  51.3106    3.1164          412
#> 229 2022-12-28 15:00:00 Blankenberge  51.3106    3.1164          458
#> 230 2022-12-28 16:00:00 Blankenberge  51.3106    3.1164          461
#> 231 2022-12-28 17:00:00 Blankenberge  51.3106    3.1164          449
#> 232 2022-12-28 18:00:00 Blankenberge  51.3106    3.1164          400
#> 233 2022-12-28 19:00:00 Blankenberge  51.3106    3.1164          325
#> 234 2022-12-28 20:00:00 Blankenberge  51.3106    3.1164          252
#> 235 2022-12-28 21:00:00 Blankenberge  51.3106    3.1164          166
#> 236 2022-12-28 22:00:00 Blankenberge  51.3106    3.1164          100
#> 237 2022-12-28 23:00:00 Blankenberge  51.3106    3.1164          121
#> 238 2022-12-29 00:00:00 Blankenberge  51.3106    3.1164          175
#> 239 2022-12-29 01:00:00 Blankenberge  51.3106    3.1164          234
#> 240 2022-12-29 02:00:00 Blankenberge  51.3106    3.1164          349
#> 241 2022-12-29 03:00:00 Blankenberge  51.3106    3.1164          431
#> 242 2022-12-29 04:00:00 Blankenberge  51.3106    3.1164          449
#> 243 2022-12-29 05:00:00 Blankenberge  51.3106    3.1164          435
#> 244 2022-12-29 06:00:00 Blankenberge  51.3106    3.1164          396
#> 245 2022-12-29 07:00:00 Blankenberge  51.3106    3.1164          323
#> 246 2022-12-29 08:00:00 Blankenberge  51.3106    3.1164          253
#> 247 2022-12-29 09:00:00 Blankenberge  51.3106    3.1164          192
#> 248 2022-12-29 10:00:00 Blankenberge  51.3106    3.1164          117
#> 249 2022-12-29 11:00:00 Blankenberge  51.3106    3.1164           66
#> 250 2022-12-29 12:00:00 Blankenberge  51.3106    3.1164          109
#> 251 2022-12-29 13:00:00 Blankenberge  51.3106    3.1164          200
#> 252 2022-12-29 14:00:00 Blankenberge  51.3106    3.1164          297
#> 253 2022-12-29 15:00:00 Blankenberge  51.3106    3.1164          412
#> 254 2022-12-29 16:00:00 Blankenberge  51.3106    3.1164          459
#> 255 2022-12-29 17:00:00 Blankenberge  51.3106    3.1164          462
#> 256 2022-12-29 18:00:00 Blankenberge  51.3106    3.1164          444
#> 257 2022-12-29 19:00:00 Blankenberge  51.3106    3.1164          400
#> 258 2022-12-29 20:00:00 Blankenberge  51.3106    3.1164          330
#> 259 2022-12-29 21:00:00 Blankenberge  51.3106    3.1164          245
#> 260 2022-12-29 22:00:00 Blankenberge  51.3106    3.1164          175
#> 261 2022-12-29 23:00:00 Blankenberge  51.3106    3.1164          129
#> 262 2022-12-30 00:00:00 Blankenberge  51.3106    3.1164          161
#> 263 2022-12-30 01:00:00 Blankenberge  51.3106    3.1164          207
#> 264 2022-12-30 02:00:00 Blankenberge  51.3106    3.1164          259
#> 265 2022-12-30 03:00:00 Blankenberge  51.3106    3.1164          353
#> 266 2022-12-30 04:00:00 Blankenberge  51.3106    3.1164          428
#> 267 2022-12-30 05:00:00 Blankenberge  51.3106    3.1164          437
#> 268 2022-12-30 06:00:00 Blankenberge  51.3106    3.1164          426
#> 269 2022-12-30 07:00:00 Blankenberge  51.3106    3.1164          381
#> 270 2022-12-30 08:00:00 Blankenberge  51.3106    3.1164          314
#> 271 2022-12-30 09:00:00 Blankenberge  51.3106    3.1164          228
#> 272 2022-12-30 10:00:00 Blankenberge  51.3106    3.1164          150
#> 273 2022-12-30 11:00:00 Blankenberge  51.3106    3.1164           85
#> 274 2022-12-30 12:00:00 Blankenberge  51.3106    3.1164           29
#> 275 2022-12-30 13:00:00 Blankenberge  51.3106    3.1164           62
#> 276 2022-12-30 14:00:00 Blankenberge  51.3106    3.1164          153
#> 277 2022-12-30 15:00:00 Blankenberge  51.3106    3.1164          276
#> 278 2022-12-30 16:00:00 Blankenberge  51.3106    3.1164          356
#> 279 2022-12-30 17:00:00 Blankenberge  51.3106    3.1164          386
#> 280 2022-12-30 18:00:00 Blankenberge  51.3106    3.1164          389
#> 281 2022-12-30 19:00:00 Blankenberge  51.3106    3.1164          372
#> 282 2022-12-30 20:00:00 Blankenberge  51.3106    3.1164          331
#> 283 2022-12-30 21:00:00 Blankenberge  51.3106    3.1164          278
#> 284 2022-12-30 22:00:00 Blankenberge  51.3106    3.1164          217
#> 285 2022-12-30 23:00:00 Blankenberge  51.3106    3.1164          151
#> 286 2022-12-31 00:00:00 Blankenberge  51.3106    3.1164           99
#> 287 2022-12-31 01:00:00 Blankenberge  51.3106    3.1164          123
#> 288 2022-12-31 02:00:00 Blankenberge  51.3106    3.1164          174
#> 289 2022-12-31 03:00:00 Blankenberge  51.3106    3.1164          244
#> 290 2022-12-31 04:00:00 Blankenberge  51.3106    3.1164          346
#> 291 2022-12-31 05:00:00 Blankenberge  51.3106    3.1164          424
#> 292 2022-12-31 06:00:00 Blankenberge  51.3106    3.1164          436
#> 293 2022-12-31 07:00:00 Blankenberge  51.3106    3.1164          428
#> 294 2022-12-31 08:00:00 Blankenberge  51.3106    3.1164          397
#> 295 2022-12-31 09:00:00 Blankenberge  51.3106    3.1164          346
#> 296 2022-12-31 10:00:00 Blankenberge  51.3106    3.1164          274
#> 297 2022-12-31 11:00:00 Blankenberge  51.3106    3.1164          210
#> 298 2022-12-31 12:00:00 Blankenberge  51.3106    3.1164          147
#> 299 2022-12-31 13:00:00 Blankenberge  51.3106    3.1164          124
#> 300 2022-12-31 14:00:00 Blankenberge  51.3106    3.1164          167
#> 301 2022-12-31 15:00:00 Blankenberge  51.3106    3.1164          223
#> 302 2022-12-31 16:00:00 Blankenberge  51.3106    3.1164          304
#> 303 2022-12-31 17:00:00 Blankenberge  51.3106    3.1164          400
#> 304 2022-12-31 18:00:00 Blankenberge  51.3106    3.1164          442
#> 305 2022-12-31 19:00:00 Blankenberge  51.3106    3.1164          444
#> 306 2022-12-31 20:00:00 Blankenberge  51.3106    3.1164          428
#> 307 2022-12-31 21:00:00 Blankenberge  51.3106    3.1164          381
#> 308 2022-12-31 22:00:00 Blankenberge  51.3106    3.1164          326
#> 309 2022-12-31 23:00:00 Blankenberge  51.3106    3.1164          260
#> 310 2023-01-01 00:00:00 Blankenberge  51.3106    3.1164          205
#> 311 2023-01-01 01:00:00 Blankenberge  51.3106    3.1164          148
#> 312 2023-01-01 02:00:00 Blankenberge  51.3106    3.1164          183
#> 313 2023-01-01 03:00:00 Blankenberge  51.3106    3.1164          223
#> 314 2023-01-01 04:00:00 Blankenberge  51.3106    3.1164          280
#> 315 2023-01-01 05:00:00 Blankenberge  51.3106    3.1164          366
#> 316 2023-01-01 06:00:00 Blankenberge  51.3106    3.1164          417
#> 317 2023-01-01 07:00:00 Blankenberge  51.3106    3.1164          424
#> 318 2023-01-01 08:00:00 Blankenberge  51.3106    3.1164          413
#> 319 2023-01-01 09:00:00 Blankenberge  51.3106    3.1164          380
#> 320 2023-01-01 10:00:00 Blankenberge  51.3106    3.1164          321
#> 321 2023-01-01 11:00:00 Blankenberge  51.3106    3.1164          255
#> 322 2023-01-01 12:00:00 Blankenberge  51.3106    3.1164          186
#> 323 2023-01-01 13:00:00 Blankenberge  51.3106    3.1164          124
#> 324 2023-01-01 14:00:00 Blankenberge  51.3106    3.1164           82
#> 325 2023-01-01 15:00:00 Blankenberge  51.3106    3.1164           96
#> 326 2023-01-01 16:00:00 Blankenberge  51.3106    3.1164          161
#> 327 2023-01-01 17:00:00 Blankenberge  51.3106    3.1164          255
#> 328 2023-01-01 18:00:00 Blankenberge  51.3106    3.1164          353
#> 329 2023-01-01 19:00:00 Blankenberge  51.3106    3.1164          403
#> 330 2023-01-01 20:00:00 Blankenberge  51.3106    3.1164          410
#> 331 2023-01-01 21:00:00 Blankenberge  51.3106    3.1164          404
#> 332 2023-01-01 22:00:00 Blankenberge  51.3106    3.1164          367
#> 333 2023-01-01 23:00:00 Blankenberge  51.3106    3.1164          313
#> 334 2023-01-02 00:00:00 Blankenberge  51.3106    3.1164          251
#> 335 2023-01-02 01:00:00 Blankenberge  51.3106    3.1164          185
#> 336 2023-01-02 02:00:00 Blankenberge  51.3106    3.1164          136
#> 337 2023-01-02 03:00:00 Blankenberge  51.3106    3.1164          149
#> 338 2023-01-02 04:00:00 Blankenberge  51.3106    3.1164          196
#> 339 2023-01-02 05:00:00 Blankenberge  51.3106    3.1164          253
#> 340 2023-01-02 06:00:00 Blankenberge  51.3106    3.1164          335
#> 341 2023-01-02 07:00:00 Blankenberge  51.3106    3.1164          404
#> 342 2023-01-02 08:00:00 Blankenberge  51.3106    3.1164          419
#> 343 2023-01-02 09:00:00 Blankenberge  51.3106    3.1164          416
#> 344 2023-01-02 10:00:00 Blankenberge  51.3106    3.1164          389
#> 345 2023-01-02 11:00:00 Blankenberge  51.3106    3.1164          336
#> 346 2023-01-02 12:00:00 Blankenberge  51.3106    3.1164          273
#> 347 2023-01-02 13:00:00 Blankenberge  51.3106    3.1164          206
#> 348 2023-01-02 14:00:00 Blankenberge  51.3106    3.1164          148
#> 349 2023-01-02 15:00:00 Blankenberge  51.3106    3.1164          107
#> 350 2023-01-02 16:00:00 Blankenberge  51.3106    3.1164          130
#> 351 2023-01-02 17:00:00 Blankenberge  51.3106    3.1164          181
#> 352 2023-01-02 18:00:00 Blankenberge  51.3106    3.1164          257
#> 353 2023-01-02 19:00:00 Blankenberge  51.3106    3.1164          352
#> 354 2023-01-02 20:00:00 Blankenberge  51.3106    3.1164          413
#> 355 2023-01-02 21:00:00 Blankenberge  51.3106    3.1164          422
#> 356 2023-01-02 22:00:00 Blankenberge  51.3106    3.1164          416
#> 357 2023-01-02 23:00:00 Blankenberge  51.3106    3.1164          384
#> 358 2023-01-03 00:00:00 Blankenberge  51.3106    3.1164          326
#> 359 2023-01-03 01:00:00 Blankenberge  51.3106    3.1164          255
#> 360 2023-01-03 02:00:00 Blankenberge  51.3106    3.1164          179
#> 361 2023-01-03 03:00:00 Blankenberge  51.3106    3.1164          127
#> 362 2023-01-03 04:00:00 Blankenberge  51.3106    3.1164          121
#> 363 2023-01-03 05:00:00 Blankenberge  51.3106    3.1164          163
#> 364 2023-01-03 06:00:00 Blankenberge  51.3106    3.1164          218
#> 365 2023-01-03 07:00:00 Blankenberge  51.3106    3.1164          306
#> 366 2023-01-03 08:00:00 Blankenberge  51.3106    3.1164          389
#> 367 2023-01-03 09:00:00 Blankenberge  51.3106    3.1164          410
#> 368 2023-01-03 10:00:00 Blankenberge  51.3106    3.1164          408
#> 369 2023-01-03 11:00:00 Blankenberge  51.3106    3.1164          375
#> 370 2023-01-03 12:00:00 Blankenberge  51.3106    3.1164          325
#> 371 2023-01-03 13:00:00 Blankenberge  51.3106    3.1164          259
#> 372 2023-01-03 14:00:00 Blankenberge  51.3106    3.1164          188
#> 373 2023-01-03 15:00:00 Blankenberge  51.3106    3.1164          125
#> 374 2023-01-03 16:00:00 Blankenberge  51.3106    3.1164           75
#> 375 2023-01-03 17:00:00 Blankenberge  51.3106    3.1164           73
#> 376 2023-01-03 18:00:00 Blankenberge  51.3106    3.1164          126
#> 377 2023-01-03 19:00:00 Blankenberge  51.3106    3.1164          214
#> 378 2023-01-03 20:00:00 Blankenberge  51.3106    3.1164          325
#> 379 2023-01-03 21:00:00 Blankenberge  51.3106    3.1164          369
#> 380 2023-01-03 22:00:00 Blankenberge  51.3106    3.1164          371
#> 381 2023-01-03 23:00:00 Blankenberge  51.3106    3.1164          364
#> 382 2023-01-04 00:00:00 Blankenberge  51.3106    3.1164          330
#> 383 2023-01-04 01:00:00 Blankenberge  51.3106    3.1164          278
#> 384 2023-01-04 02:00:00 Blankenberge  51.3106    3.1164          219
#> 385 2023-01-04 03:00:00 Blankenberge  51.3106    3.1164          155
#> 386 2023-01-04 04:00:00 Blankenberge  51.3106    3.1164          110
#> 387 2023-01-04 05:00:00 Blankenberge  51.3106    3.1164          115
#> 388 2023-01-04 06:00:00 Blankenberge  51.3106    3.1164          172
#> 389 2023-01-04 07:00:00 Blankenberge  51.3106    3.1164          251
#> 390 2023-01-04 08:00:00 Blankenberge  51.3106    3.1164          353
#> 391 2023-01-04 09:00:00 Blankenberge  51.3106    3.1164          426
#> 392 2023-01-04 10:00:00 Blankenberge  51.3106    3.1164          440
#> 393 2023-01-04 11:00:00 Blankenberge  51.3106    3.1164          438
#> 394 2023-01-04 12:00:00 Blankenberge  51.3106    3.1164          415
#> 395 2023-01-04 13:00:00 Blankenberge  51.3106    3.1164          362
#> 396 2023-01-04 14:00:00 Blankenberge  51.3106    3.1164          299
#> 397 2023-01-04 15:00:00 Blankenberge  51.3106    3.1164          226
#> 398 2023-01-04 16:00:00 Blankenberge  51.3106    3.1164          165
#> 399 2023-01-04 17:00:00 Blankenberge  51.3106    3.1164          134
#> 400 2023-01-04 18:00:00 Blankenberge  51.3106    3.1164          170
#> 401 2023-01-04 19:00:00 Blankenberge  51.3106    3.1164          224
#> 402 2023-01-04 20:00:00 Blankenberge  51.3106    3.1164          302
#> 403 2023-01-04 21:00:00 Blankenberge  51.3106    3.1164          410
#> 404 2023-01-04 22:00:00 Blankenberge  51.3106    3.1164          461
#> 405 2023-01-04 23:00:00 Blankenberge  51.3106    3.1164          462
#> 406 2023-01-05 00:00:00 Blankenberge  51.3106    3.1164          449
#> 407 2023-01-05 01:00:00 Blankenberge  51.3106    3.1164          402
#> 408 2023-01-05 02:00:00 Blankenberge  51.3106    3.1164          337
#> 409 2023-01-05 03:00:00 Blankenberge  51.3106    3.1164          258
#> 410 2023-01-05 04:00:00 Blankenberge  51.3106    3.1164          178
#> 411 2023-01-05 05:00:00 Blankenberge  51.3106    3.1164          124
#> 412 2023-01-05 06:00:00 Blankenberge  51.3106    3.1164          146
#> 413 2023-01-05 07:00:00 Blankenberge  51.3106    3.1164          190
#> 414 2023-01-05 08:00:00 Blankenberge  51.3106    3.1164          252
#> 415 2023-01-05 09:00:00 Blankenberge  51.3106    3.1164          353
#> 416 2023-01-05 10:00:00 Blankenberge  51.3106    3.1164          441
#> 417 2023-01-05 11:00:00 Blankenberge  51.3106    3.1164          453
#> 418 2023-01-05 12:00:00 Blankenberge  51.3106    3.1164          443
#> 419 2023-01-05 13:00:00 Blankenberge  51.3106    3.1164          404
#> 420 2023-01-05 14:00:00 Blankenberge  51.3106    3.1164          336
#> 421 2023-01-05 15:00:00 Blankenberge  51.3106    3.1164          253
#> 422 2023-01-05 16:00:00 Blankenberge  51.3106    3.1164          172
#> 423 2023-01-05 17:00:00 Blankenberge  51.3106    3.1164          104
#> 424 2023-01-05 18:00:00 Blankenberge  51.3106    3.1164           86
#> 425 2023-01-05 19:00:00 Blankenberge  51.3106    3.1164          125
#> 426 2023-01-05 20:00:00 Blankenberge  51.3106    3.1164          182
#> 427 2023-01-05 21:00:00 Blankenberge  51.3106    3.1164          268
#> 428 2023-01-05 22:00:00 Blankenberge  51.3106    3.1164          390
#> 429 2023-01-05 23:00:00 Blankenberge  51.3106    3.1164          419
#> 430 2023-01-06 00:00:00 Blankenberge  51.3106    3.1164          418
#> 431 2023-01-06 01:00:00 Blankenberge  51.3106    3.1164          382
#> 432 2023-01-06 02:00:00 Blankenberge  51.3106    3.1164          324
#> 433 2023-01-06 03:00:00 Blankenberge  51.3106    3.1164          255
#> 434 2023-01-06 04:00:00 Blankenberge  51.3106    3.1164          176
#> 435 2023-01-06 05:00:00 Blankenberge  51.3106    3.1164          107
#> 436 2023-01-06 06:00:00 Blankenberge  51.3106    3.1164           65
#> 437 2023-01-06 07:00:00 Blankenberge  51.3106    3.1164          103
#> 438 2023-01-06 08:00:00 Blankenberge  51.3106    3.1164          157
#> 439 2023-01-06 09:00:00 Blankenberge  51.3106    3.1164          248
#> 440 2023-01-06 10:00:00 Blankenberge  51.3106    3.1164          388
#> 441 2023-01-06 11:00:00 Blankenberge  51.3106    3.1164          437
#> 442 2023-01-06 12:00:00 Blankenberge  51.3106    3.1164          437
#> 443 2023-01-06 13:00:00 Blankenberge  51.3106    3.1164          422
#> 444 2023-01-06 14:00:00 Blankenberge  51.3106    3.1164          371
#> 445 2023-01-06 15:00:00 Blankenberge  51.3106    3.1164          305
#> 446 2023-01-06 16:00:00 Blankenberge  51.3106    3.1164          230
#> 447 2023-01-06 17:00:00 Blankenberge  51.3106    3.1164          151
#> 448 2023-01-06 18:00:00 Blankenberge  51.3106    3.1164          108
#> 449 2023-01-06 19:00:00 Blankenberge  51.3106    3.1164          147
#> 450 2023-01-06 20:00:00 Blankenberge  51.3106    3.1164          192
#> 451 2023-01-06 21:00:00 Blankenberge  51.3106    3.1164          260
#> 452 2023-01-06 22:00:00 Blankenberge  51.3106    3.1164          372
#> 453 2023-01-06 23:00:00 Blankenberge  51.3106    3.1164          467
#> 454 2023-01-07 00:00:00 Blankenberge  51.3106    3.1164          476
#> 455 2023-01-07 01:00:00 Blankenberge  51.3106    3.1164          464
#> 456 2023-01-07 02:00:00 Blankenberge  51.3106    3.1164          424
#> 457 2023-01-07 03:00:00 Blankenberge  51.3106    3.1164          351
#> 458 2023-01-07 04:00:00 Blankenberge  51.3106    3.1164          267
#> 459 2023-01-07 05:00:00 Blankenberge  51.3106    3.1164          179
#> 460 2023-01-07 06:00:00 Blankenberge  51.3106    3.1164          105
#> 461 2023-01-07 07:00:00 Blankenberge  51.3106    3.1164           75
#> 462 2023-01-07 08:00:00 Blankenberge  51.3106    3.1164          109
#> 463 2023-01-07 09:00:00 Blankenberge  51.3106    3.1164          166
#> 464 2023-01-07 10:00:00 Blankenberge  51.3106    3.1164          283
#> 465 2023-01-07 11:00:00 Blankenberge  51.3106    3.1164          397
#> 466 2023-01-07 12:00:00 Blankenberge  51.3106    3.1164          414
#> 467 2023-01-07 13:00:00 Blankenberge  51.3106    3.1164          406
#> 468 2023-01-07 14:00:00 Blankenberge  51.3106    3.1164          373
#> 469 2023-01-07 15:00:00 Blankenberge  51.3106    3.1164          312
#> 470 2023-01-07 16:00:00 Blankenberge  51.3106    3.1164          240
#> 471 2023-01-07 17:00:00 Blankenberge  51.3106    3.1164          164
#> 472 2023-01-07 18:00:00 Blankenberge  51.3106    3.1164           94
#> 473 2023-01-07 19:00:00 Blankenberge  51.3106    3.1164           54
#> 474 2023-01-07 20:00:00 Blankenberge  51.3106    3.1164          115
#> 475 2023-01-07 21:00:00 Blankenberge  51.3106    3.1164          177
#> 476 2023-01-07 22:00:00 Blankenberge  51.3106    3.1164          283
#> 477 2023-01-07 23:00:00 Blankenberge  51.3106    3.1164          410
#> 478 2023-01-08 00:00:00 Blankenberge  51.3106    3.1164          451
#> 479 2023-01-08 01:00:00 Blankenberge  51.3106    3.1164          451
#> 480 2023-01-08 02:00:00 Blankenberge  51.3106    3.1164          430
#> 481 2023-01-08 03:00:00 Blankenberge  51.3106    3.1164          381
#> 482 2023-01-08 04:00:00 Blankenberge  51.3106    3.1164          308
#> 483 2023-01-08 05:00:00 Blankenberge  51.3106    3.1164          227
#> 484 2023-01-08 06:00:00 Blankenberge  51.3106    3.1164          144
#> 485 2023-01-08 07:00:00 Blankenberge  51.3106    3.1164           81
#> 486 2023-01-08 08:00:00 Blankenberge  51.3106    3.1164           98
#> 487 2023-01-08 09:00:00 Blankenberge  51.3106    3.1164          156
#> 488 2023-01-08 10:00:00 Blankenberge  51.3106    3.1164          248
#> 489 2023-01-08 11:00:00 Blankenberge  51.3106    3.1164          390
#> 490 2023-01-08 12:00:00 Blankenberge  51.3106    3.1164          450
#> 491 2023-01-08 13:00:00 Blankenberge  51.3106    3.1164          456
#> 492 2023-01-08 14:00:00 Blankenberge  51.3106    3.1164          443
#> 493 2023-01-08 15:00:00 Blankenberge  51.3106    3.1164          400
#> 494 2023-01-08 16:00:00 Blankenberge  51.3106    3.1164          330
#> 495 2023-01-08 17:00:00 Blankenberge  51.3106    3.1164          255
#> 496 2023-01-08 18:00:00 Blankenberge  51.3106    3.1164          176
#> 497 2023-01-08 19:00:00 Blankenberge  51.3106    3.1164          120
#> 498 2023-01-08 20:00:00 Blankenberge  51.3106    3.1164          150
#> 499 2023-01-08 21:00:00 Blankenberge  51.3106    3.1164          185
#> 500 2023-01-08 22:00:00 Blankenberge  51.3106    3.1164          240
#> 501 2023-01-08 23:00:00 Blankenberge  51.3106    3.1164          353
#> 502 2023-01-09 00:00:00 Blankenberge  51.3106    3.1164          471
#> 503 2023-01-09 01:00:00 Blankenberge  51.3106    3.1164          487
#> 504 2023-01-09 02:00:00 Blankenberge  51.3106    3.1164          480
#> 505 2023-01-09 03:00:00 Blankenberge  51.3106    3.1164          443
#> 506 2023-01-09 04:00:00 Blankenberge  51.3106    3.1164          394
#> 507 2023-01-09 05:00:00 Blankenberge  51.3106    3.1164          316
#> 508 2023-01-09 06:00:00 Blankenberge  51.3106    3.1164          231
#> 509 2023-01-09 07:00:00 Blankenberge  51.3106    3.1164          159
#> 510 2023-01-09 08:00:00 Blankenberge  51.3106    3.1164          135
#> 511 2023-01-09 09:00:00 Blankenberge  51.3106    3.1164          184
#> 512 2023-01-09 10:00:00 Blankenberge  51.3106    3.1164          246
#> 513 2023-01-09 11:00:00 Blankenberge  51.3106    3.1164          343
#> 514 2023-01-09 12:00:00 Blankenberge  51.3106    3.1164          469
#> 515 2023-01-09 13:00:00 Blankenberge  51.3106    3.1164          499
#> 516 2023-01-09 14:00:00 Blankenberge  51.3106    3.1164          498
#> 517 2023-01-09 15:00:00 Blankenberge  51.3106    3.1164          475
#> 518 2023-01-09 16:00:00 Blankenberge  51.3106    3.1164          419
#> 519 2023-01-09 17:00:00 Blankenberge  51.3106    3.1164          339
#> 520 2023-01-09 18:00:00 Blankenberge  51.3106    3.1164          247
#> 521 2023-01-09 19:00:00 Blankenberge  51.3106    3.1164          171
#> 522 2023-01-09 20:00:00 Blankenberge  51.3106    3.1164          126
#> 523 2023-01-09 21:00:00 Blankenberge  51.3106    3.1164          167
#> 524 2023-01-09 22:00:00 Blankenberge  51.3106    3.1164          210
#> 525 2023-01-09 23:00:00 Blankenberge  51.3106    3.1164          271
#> 526 2023-01-10 00:00:00 Blankenberge  51.3106    3.1164          383
#> 527 2023-01-10 01:00:00 Blankenberge  51.3106    3.1164          455
#> 528 2023-01-10 02:00:00 Blankenberge  51.3106    3.1164          458
#> 529 2023-01-10 03:00:00 Blankenberge  51.3106    3.1164          438
#> 530 2023-01-10 04:00:00 Blankenberge  51.3106    3.1164          388
#> 531 2023-01-10 05:00:00 Blankenberge  51.3106    3.1164          309
#> 532 2023-01-10 06:00:00 Blankenberge  51.3106    3.1164          220
#> 533 2023-01-10 07:00:00 Blankenberge  51.3106    3.1164          135
#> 534 2023-01-10 08:00:00 Blankenberge  51.3106    3.1164           69
#> 535 2023-01-10 09:00:00 Blankenberge  51.3106    3.1164           77
#> 536 2023-01-10 10:00:00 Blankenberge  51.3106    3.1164          125
#> 537 2023-01-10 11:00:00 Blankenberge  51.3106    3.1164          195
#> 538 2023-01-10 12:00:00 Blankenberge  51.3106    3.1164          330
#> 539 2023-01-10 13:00:00 Blankenberge  51.3106    3.1164          423
#> 540 2023-01-10 14:00:00 Blankenberge  51.3106    3.1164          429
#> 541 2023-01-10 15:00:00 Blankenberge  51.3106    3.1164          416
#> 542 2023-01-10 16:00:00 Blankenberge  51.3106    3.1164          375
#> 543 2023-01-10 17:00:00 Blankenberge  51.3106    3.1164          313
#> 544 2023-01-10 18:00:00 Blankenberge  51.3106    3.1164          232
#> 545 2023-01-10 19:00:00 Blankenberge  51.3106    3.1164          160
#> 546 2023-01-10 20:00:00 Blankenberge  51.3106    3.1164           93
#> 547 2023-01-10 21:00:00 Blankenberge  51.3106    3.1164           75
#> 548 2023-01-10 22:00:00 Blankenberge  51.3106    3.1164          114
#> 549 2023-01-10 23:00:00 Blankenberge  51.3106    3.1164          178
#> 550 2023-01-11 00:00:00 Blankenberge  51.3106    3.1164          308
#> 551 2023-01-11 01:00:00 Blankenberge  51.3106    3.1164          407
#> 552 2023-01-11 02:00:00 Blankenberge  51.3106    3.1164          425
#> 553 2023-01-11 03:00:00 Blankenberge  51.3106    3.1164          425
#> 554 2023-01-11 04:00:00 Blankenberge  51.3106    3.1164          397
#> 555 2023-01-11 05:00:00 Blankenberge  51.3106    3.1164          337
#> 556 2023-01-11 06:00:00 Blankenberge  51.3106    3.1164          266
#> 557 2023-01-11 07:00:00 Blankenberge  51.3106    3.1164          192
#> 558 2023-01-11 08:00:00 Blankenberge  51.3106    3.1164          115
#> 559 2023-01-11 09:00:00 Blankenberge  51.3106    3.1164           74
#> 560 2023-01-11 10:00:00 Blankenberge  51.3106    3.1164          118
#> 561 2023-01-11 11:00:00 Blankenberge  51.3106    3.1164          196
#> 562 2023-01-11 12:00:00 Blankenberge  51.3106    3.1164          303
#> 563 2023-01-11 13:00:00 Blankenberge  51.3106    3.1164          428
#> 564 2023-01-11 14:00:00 Blankenberge  51.3106    3.1164          468
#> 565 2023-01-11 15:00:00 Blankenberge  51.3106    3.1164          466
#> 566 2023-01-11 16:00:00 Blankenberge  51.3106    3.1164          449
#> 567 2023-01-11 17:00:00 Blankenberge  51.3106    3.1164          404
#> 568 2023-01-11 18:00:00 Blankenberge  51.3106    3.1164          326
#> 569 2023-01-11 19:00:00 Blankenberge  51.3106    3.1164          262
#> 570 2023-01-11 20:00:00 Blankenberge  51.3106    3.1164          185
#> 571 2023-01-11 21:00:00 Blankenberge  51.3106    3.1164          144
#> 572 2023-01-11 22:00:00 Blankenberge  51.3106    3.1164          190
#> 573 2023-01-11 23:00:00 Blankenberge  51.3106    3.1164          216
#> 574 2023-01-12 00:00:00 Blankenberge  51.3106    3.1164          291
#> 575 2023-01-12 01:00:00 Blankenberge  51.3106    3.1164          391
#> 576 2023-01-12 02:00:00 Blankenberge  51.3106    3.1164          449
#> 577 2023-01-12 03:00:00 Blankenberge  51.3106    3.1164          457
#> 578 2023-01-12 04:00:00 Blankenberge  51.3106    3.1164          442
#> 579 2023-01-12 05:00:00 Blankenberge  51.3106    3.1164          395
#> 580 2023-01-12 06:00:00 Blankenberge  51.3106    3.1164          325
#> 581 2023-01-12 07:00:00 Blankenberge  51.3106    3.1164          257
#> 582 2023-01-12 08:00:00 Blankenberge  51.3106    3.1164          183
#> 583 2023-01-12 09:00:00 Blankenberge  51.3106    3.1164          123
#> 584 2023-01-12 10:00:00 Blankenberge  51.3106    3.1164          113
#> 585 2023-01-12 11:00:00 Blankenberge  51.3106    3.1164          166
#> 586 2023-01-12 12:00:00 Blankenberge  51.3106    3.1164          256
#> 587 2023-01-12 13:00:00 Blankenberge  51.3106    3.1164          363
#> 588 2023-01-12 14:00:00 Blankenberge  51.3106    3.1164          437
#> 589 2023-01-12 15:00:00 Blankenberge  51.3106    3.1164          454
#> 590 2023-01-12 16:00:00 Blankenberge  51.3106    3.1164          448
#> 591 2023-01-12 17:00:00 Blankenberge  51.3106    3.1164          419
#> 592 2023-01-12 18:00:00 Blankenberge  51.3106    3.1164          350
#> 593 2023-01-12 19:00:00 Blankenberge  51.3106    3.1164          280
#> 594 2023-01-12 20:00:00 Blankenberge  51.3106    3.1164          210
#> 595 2023-01-12 21:00:00 Blankenberge  51.3106    3.1164          146
#> 596 2023-01-12 22:00:00 Blankenberge  51.3106    3.1164          135
#> 597 2023-01-12 23:00:00 Blankenberge  51.3106    3.1164          174
#> 598 2023-01-13 00:00:00 Blankenberge  51.3106    3.1164          219
#> 599 2023-01-13 01:00:00 Blankenberge  51.3106    3.1164          304
#> 600 2023-01-13 02:00:00 Blankenberge  51.3106    3.1164          392
#> 601 2023-01-13 03:00:00 Blankenberge  51.3106    3.1164          436
#> 602 2023-01-13 04:00:00 Blankenberge  51.3106    3.1164          428
#> 603 2023-01-13 05:00:00 Blankenberge  51.3106    3.1164          405
#> 604 2023-01-13 06:00:00 Blankenberge  51.3106    3.1164          349
#> 605 2023-01-13 07:00:00 Blankenberge  51.3106    3.1164          285
#> 606 2023-01-13 08:00:00 Blankenberge  51.3106    3.1164          217
#> 607 2023-01-13 09:00:00 Blankenberge  51.3106    3.1164          149
#> 608 2023-01-13 10:00:00 Blankenberge  51.3106    3.1164          114
#> 609 2023-01-13 11:00:00 Blankenberge  51.3106    3.1164          156
#> 610 2023-01-13 12:00:00 Blankenberge  51.3106    3.1164          217
#> 611 2023-01-13 13:00:00 Blankenberge  51.3106    3.1164          321
#> 612 2023-01-13 14:00:00 Blankenberge  51.3106    3.1164          420
#> 613 2023-01-13 15:00:00 Blankenberge  51.3106    3.1164          467
#> 614 2023-01-13 16:00:00 Blankenberge  51.3106    3.1164          475
#> 615 2023-01-13 17:00:00 Blankenberge  51.3106    3.1164          467
#> 616 2023-01-13 18:00:00 Blankenberge  51.3106    3.1164          425
#> 617 2023-01-13 19:00:00 Blankenberge  51.3106    3.1164          358
#> 618 2023-01-13 20:00:00 Blankenberge  51.3106    3.1164          284
#> 619 2023-01-13 21:00:00 Blankenberge  51.3106    3.1164          207
#> 620 2023-01-13 22:00:00 Blankenberge  51.3106    3.1164          156
#> 621 2023-01-13 23:00:00 Blankenberge  51.3106    3.1164          164
#> 622 2023-01-14 00:00:00 Blankenberge  51.3106    3.1164          200
#> 623 2023-01-14 01:00:00 Blankenberge  51.3106    3.1164          242
#> 624 2023-01-14 02:00:00 Blankenberge  51.3106    3.1164          323
#> 625 2023-01-14 03:00:00 Blankenberge  51.3106    3.1164          402
#> 626 2023-01-14 04:00:00 Blankenberge  51.3106    3.1164          413
#> 627 2023-01-14 05:00:00 Blankenberge  51.3106    3.1164          407
#> 628 2023-01-14 06:00:00 Blankenberge  51.3106    3.1164          365
#> 629 2023-01-14 07:00:00 Blankenberge  51.3106    3.1164          300
#> 630 2023-01-14 08:00:00 Blankenberge  51.3106    3.1164          231
#> 631 2023-01-14 09:00:00 Blankenberge  51.3106    3.1164          161
#> 632 2023-01-14 10:00:00 Blankenberge  51.3106    3.1164           94
#> 633 2023-01-14 11:00:00 Blankenberge  51.3106    3.1164           67
#> 634 2023-01-14 12:00:00 Blankenberge  51.3106    3.1164           98
#> 635 2023-01-14 13:00:00 Blankenberge  51.3106    3.1164          187
#> 636 2023-01-14 14:00:00 Blankenberge  51.3106    3.1164          291
#> 637 2023-01-14 15:00:00 Blankenberge  51.3106    3.1164          370
#> 638 2023-01-14 16:00:00 Blankenberge  51.3106    3.1164          400
#> 639 2023-01-14 17:00:00 Blankenberge  51.3106    3.1164          403
#> 640 2023-01-14 18:00:00 Blankenberge  51.3106    3.1164          396
#> 641 2023-01-14 19:00:00 Blankenberge  51.3106    3.1164          357
#> 642 2023-01-14 20:00:00 Blankenberge  51.3106    3.1164          301
#> 643 2023-01-14 21:00:00 Blankenberge  51.3106    3.1164          234
#> 644 2023-01-14 22:00:00 Blankenberge  51.3106    3.1164          171
#> 645 2023-01-14 23:00:00 Blankenberge  51.3106    3.1164          130
#> 646 2023-01-15 00:00:00 Blankenberge  51.3106    3.1164          160
#> 647 2023-01-15 01:00:00 Blankenberge  51.3106    3.1164          199
#> 648 2023-01-15 02:00:00 Blankenberge  51.3106    3.1164          281
#> 649 2023-01-15 03:00:00 Blankenberge  51.3106    3.1164          375
#> 650 2023-01-15 04:00:00 Blankenberge  51.3106    3.1164          423
#> 651 2023-01-15 05:00:00 Blankenberge  51.3106    3.1164          428
#> 652 2023-01-15 06:00:00 Blankenberge  51.3106    3.1164          419
#> 653 2023-01-15 07:00:00 Blankenberge  51.3106    3.1164          384
#> 654 2023-01-15 08:00:00 Blankenberge  51.3106    3.1164          332
#> 655 2023-01-15 09:00:00 Blankenberge  51.3106    3.1164          268
#> 656 2023-01-15 10:00:00 Blankenberge  51.3106    3.1164          201
#> 657 2023-01-15 11:00:00 Blankenberge  51.3106    3.1164          161
#> 658 2023-01-15 12:00:00 Blankenberge  51.3106    3.1164          171
#> 659 2023-01-15 13:00:00 Blankenberge  51.3106    3.1164          217
#> 660 2023-01-15 14:00:00 Blankenberge  51.3106    3.1164          288
#> 661 2023-01-15 15:00:00 Blankenberge  51.3106    3.1164          371
#> 662 2023-01-15 16:00:00 Blankenberge  51.3106    3.1164          441
#> 663 2023-01-15 17:00:00 Blankenberge  51.3106    3.1164          464
#> 664 2023-01-15 18:00:00 Blankenberge  51.3106    3.1164          466
#> 665 2023-01-15 19:00:00 Blankenberge  51.3106    3.1164          432
#> 666 2023-01-15 20:00:00 Blankenberge  51.3106    3.1164          387
#> 667 2023-01-15 21:00:00 Blankenberge  51.3106    3.1164          320
#> 668 2023-01-15 22:00:00 Blankenberge  51.3106    3.1164          248
#> 669 2023-01-15 23:00:00 Blankenberge  51.3106    3.1164          185
#> 670 2023-01-16 00:00:00 Blankenberge  51.3106    3.1164          144
#> 671 2023-01-16 01:00:00 Blankenberge  51.3106    3.1164          166
#> 672 2023-01-16 02:00:00 Blankenberge  51.3106    3.1164          199
#> 673 2023-01-16 03:00:00 Blankenberge  51.3106    3.1164          260
#> 674 2023-01-16 04:00:00 Blankenberge  51.3106    3.1164          330
#> 675 2023-01-16 05:00:00 Blankenberge  51.3106    3.1164          396
#> 676 2023-01-16 06:00:00 Blankenberge  51.3106    3.1164          401
#> 677 2023-01-16 07:00:00 Blankenberge  51.3106    3.1164          385
#> 678 2023-01-16 08:00:00 Blankenberge  51.3106    3.1164          357
#> 679 2023-01-16 09:00:00 Blankenberge  51.3106    3.1164          300
#> 680 2023-01-16 10:00:00 Blankenberge  51.3106    3.1164          230
#> 681 2023-01-16 11:00:00 Blankenberge  51.3106    3.1164          169
#> 682 2023-01-16 12:00:00 Blankenberge  51.3106    3.1164          115
#> 683 2023-01-16 13:00:00 Blankenberge  51.3106    3.1164          151
#> 684 2023-01-16 14:00:00 Blankenberge  51.3106    3.1164          207
#> 685 2023-01-16 15:00:00 Blankenberge  51.3106    3.1164          261
#> 686 2023-01-16 16:00:00 Blankenberge  51.3106    3.1164          338
#> 687 2023-01-16 17:00:00 Blankenberge  51.3106    3.1164          420
#> 688 2023-01-16 18:00:00 Blankenberge  51.3106    3.1164          459
#> 689 2023-01-16 19:00:00 Blankenberge  51.3106    3.1164          457
#> 690 2023-01-16 20:00:00 Blankenberge  51.3106    3.1164          439
#> 691 2023-01-16 21:00:00 Blankenberge  51.3106    3.1164          397
#> 692 2023-01-16 22:00:00 Blankenberge  51.3106    3.1164          329
#> 693 2023-01-16 23:00:00 Blankenberge  51.3106    3.1164          252
#> 694 2023-01-17 00:00:00 Blankenberge  51.3106    3.1164          191
#> 695 2023-01-17 01:00:00 Blankenberge  51.3106    3.1164          154
#> 696 2023-01-17 02:00:00 Blankenberge  51.3106    3.1164          174
#> 697 2023-01-17 03:00:00 Blankenberge  51.3106    3.1164          206
#> 698 2023-01-17 04:00:00 Blankenberge  51.3106    3.1164          260
#> 699 2023-01-17 05:00:00 Blankenberge  51.3106    3.1164          334
#> 700 2023-01-17 06:00:00 Blankenberge  51.3106    3.1164          393
#> 701 2023-01-17 07:00:00 Blankenberge  51.3106    3.1164          398
#> 702 2023-01-17 08:00:00 Blankenberge  51.3106    3.1164          393
#> 703 2023-01-17 09:00:00 Blankenberge  51.3106    3.1164          358
#> 704 2023-01-17 10:00:00 Blankenberge  51.3106    3.1164          309
#> 705 2023-01-17 11:00:00 Blankenberge  51.3106    3.1164          246
#> 706 2023-01-17 12:00:00 Blankenberge  51.3106    3.1164          186
#> 707 2023-01-17 13:00:00 Blankenberge  51.3106    3.1164          135
#> 708 2023-01-17 14:00:00 Blankenberge  51.3106    3.1164          122
#> 709 2023-01-17 15:00:00 Blankenberge  51.3106    3.1164          162
#> 710 2023-01-17 16:00:00 Blankenberge  51.3106    3.1164          212
#> 711 2023-01-17 17:00:00 Blankenberge  51.3106    3.1164          290
#> 712 2023-01-17 18:00:00 Blankenberge  51.3106    3.1164          373
#> 713 2023-01-17 19:00:00 Blankenberge  51.3106    3.1164          414
#> 714 2023-01-17 20:00:00 Blankenberge  51.3106    3.1164          417
#> 715 2023-01-17 21:00:00 Blankenberge  51.3106    3.1164          405
#> 716 2023-01-17 22:00:00 Blankenberge  51.3106    3.1164          367
#> 717 2023-01-17 23:00:00 Blankenberge  51.3106    3.1164          307
#> 718 2023-01-18 00:00:00 Blankenberge  51.3106    3.1164          241
#> 719 2023-01-18 01:00:00 Blankenberge  51.3106    3.1164          181
#> 720 2023-01-18 02:00:00 Blankenberge  51.3106    3.1164          142
#> 721 2023-01-18 03:00:00 Blankenberge  51.3106    3.1164          155
#> 722 2023-01-18 04:00:00 Blankenberge  51.3106    3.1164          190
#> 723 2023-01-18 05:00:00 Blankenberge  51.3106    3.1164          242
#> 724 2023-01-18 06:00:00 Blankenberge  51.3106    3.1164          320
#> 725 2023-01-18 07:00:00 Blankenberge  51.3106    3.1164          388
#> 726 2023-01-18 08:00:00 Blankenberge  51.3106    3.1164          403
#> 727 2023-01-18 09:00:00 Blankenberge  51.3106    3.1164          400
#> 728 2023-01-18 10:00:00 Blankenberge  51.3106    3.1164          371
#> 
#> $par
#> $par$daterange
#> [1] "2022-12-19" "2023-01-19"
#> 
#> $par$type
#> [1] "MVB data"
#> 
#> $par$stations
#> [1] "Blankenberge"
#> 
#> $par$calc
#> [1] "max"
#> 
#> $par$binSize
#> [1] "hour"
#> 
#> $par$parameters
#> [1] "Tide (cmTAW)"
#> 
#>