Retrieve MVB-data from the LifeWatch project

Retrieves the Meetnet Vlaams Banken (MVB) data from the LifeWatch project. Without valid credentials you are only allowed to view the "Tide TAW" parameter for the last 30 days, grouped by day/hour. To get an account, register via the Lifewatch RShiny registration webpage.

getMvbData(
  startdate,
  stopdate,
  parameters,
  stations = NULL,
  by,
  calc,
  usr = NULL,
  pwd = NULL,
  params = FALSE
)

Arguments

startdate
Starting date for the query. Without a login, this is restricted to the last month

stopdate
Stopping date for the query. Without a login, this is restricted to the last month

parameters
One (or list) of the $parameter in listMvbStations.

stations
List of stations to be included in the query. See listMvbStations. Use stations="All" or stations=NULL to get all stations.

by
Time grouping for calculation, one of ('day','hour', '10min' ,'none')

calc
Calculation to perform given time grouping, one of ('avg', 'max', 'min', 'none').

usr
Username to connect to database

pwd
Password to connect to database

params
If TRUE, returns a list with the dataset and the query parameters applied in the server side. IF FALSE returns only the data.

Value

Dataframe with the specified MVB data.

Examples

getMvbData(Sys.Date() - 30, Sys.Date() + 1, parameters = 'Tide TAW',
stations = "Nieuwpoort", by = "day", calc = "avg")

#> No encoding supplied: defaulting to UTF-8.

#> Query parameters

#> ---------------------------------------------

#> daterange : c("2021-08-23", "2021-09-23")

#> type : MVB data

#> stations : Nieuwpoort

#> calc : avg

#> binSize : day

#> parameters : Tide TAW

#> ---------------------------------------------

#> Data dimension: (31 x 5)

#>          TIME    STATION LATITUDE LONGITUDE TIDE TAW
#> 1  2021-08-23 Nieuwpoort  51.1506    2.7281 232.2439
#> 2  2021-08-24 Nieuwpoort  51.1506    2.7281 217.9132
#> 3  2021-08-25 Nieuwpoort  51.1506    2.7281 230.3090
#> 4  2021-08-26 Nieuwpoort  51.1506    2.7281 254.7049
#> 5  2021-08-27 Nieuwpoort  51.1506    2.7281 239.3542
#> 6  2021-08-28 Nieuwpoort  51.1506    2.7281 239.4514
#> 7  2021-08-29 Nieuwpoort  51.1506    2.7281 244.1562
#> 8  2021-08-30 Nieuwpoort  51.1506    2.7281 234.7743
#> 9  2021-08-31 Nieuwpoort  51.1506    2.7281 232.8160
#> 10 2021-09-01 Nieuwpoort  51.1506    2.7281 223.4306
#> 11 2021-09-02 Nieuwpoort  51.1506    2.7281 219.0451
#> 12 2021-09-03 Nieuwpoort  51.1506    2.7281 223.8368
#> 13 2021-09-04 Nieuwpoort  51.1506    2.7281 226.2257
#> 14 2021-09-05 Nieuwpoort  51.1506    2.7281 224.4132
#> 15 2021-09-06 Nieuwpoort  51.1506    2.7281 224.4704
#> 16 2021-09-07 Nieuwpoort  51.1506    2.7281 228.6667
#> 17 2021-09-08 Nieuwpoort  51.1506    2.7281 233.4861
#> 18 2021-09-09 Nieuwpoort  51.1506    2.7281 241.4757
#> 19 2021-09-10 Nieuwpoort  51.1506    2.7281 244.2491
#> 20 2021-09-11 Nieuwpoort  51.1506    2.7281 247.3472
#> 21 2021-09-12 Nieuwpoort  51.1506    2.7281 254.2951
#> 22 2021-09-13 Nieuwpoort  51.1506    2.7281 244.3507
#> 23 2021-09-14 Nieuwpoort  51.1506    2.7281 247.7535
#> 24 2021-09-15 Nieuwpoort  51.1506    2.7281 256.1756
#> 25 2021-09-16 Nieuwpoort  51.1506    2.7281 248.3160
#> 26 2021-09-17 Nieuwpoort  51.1506    2.7281 236.0868
#> 27 2021-09-18 Nieuwpoort  51.1506    2.7281 230.9201
#> 28 2021-09-19 Nieuwpoort  51.1506    2.7281 229.5868
#> 29 2021-09-20 Nieuwpoort  51.1506    2.7281 230.5660
#> 30 2021-09-21 Nieuwpoort  51.1506    2.7281 229.4618
#> 31 2021-09-22 Nieuwpoort  51.1506    2.7281 289.1809

getMvbData(Sys.Date() - 30, Sys.Date() + 1, parameters = 'Tide TAW',
stations = "Blankenberge", by = "hour", calc = "max", params = TRUE)

#> No encoding supplied: defaulting to UTF-8.

#> Query parameters

#> ---------------------------------------------

#> daterange : c("2021-08-23", "2021-09-23")

#> type : MVB data

#> stations : Blankenberge

#> calc : max

#> binSize : hour

#> parameters : Tide TAW

#> ---------------------------------------------

#> Data dimension: (736 x 5)

#> $mdf
#>                    TIME      STATION LATITUDE LONGITUDE TIDE TAW
#> 1   2021-08-23 00:00:00 Blankenberge  51.3106    3.1164      483
#> 2   2021-08-23 01:00:00 Blankenberge  51.3106    3.1164      480
#> 3   2021-08-23 02:00:00 Blankenberge  51.3106    3.1164      445
#> 4   2021-08-23 03:00:00 Blankenberge  51.3106    3.1164      376
#> 5   2021-08-23 04:00:00 Blankenberge  51.3106    3.1164      282
#> 6   2021-08-23 05:00:00 Blankenberge  51.3106    3.1164      178
#> 7   2021-08-23 06:00:00 Blankenberge  51.3106    3.1164       90
#> 8   2021-08-23 07:00:00 Blankenberge  51.3106    3.1164       81
#> 9   2021-08-23 08:00:00 Blankenberge  51.3106    3.1164      131
#> 10  2021-08-23 09:00:00 Blankenberge  51.3106    3.1164      173
#> 11  2021-08-23 10:00:00 Blankenberge  51.3106    3.1164      235
#> 12  2021-08-23 11:00:00 Blankenberge  51.3106    3.1164      382
#> 13  2021-08-23 12:00:00 Blankenberge  51.3106    3.1164      460
#> 14  2021-08-23 13:00:00 Blankenberge  51.3106    3.1164      458
#> 15  2021-08-23 14:00:00 Blankenberge  51.3106    3.1164      426
#> 16  2021-08-23 15:00:00 Blankenberge  51.3106    3.1164      371
#> 17  2021-08-23 16:00:00 Blankenberge  51.3106    3.1164      284
#> 18  2021-08-23 17:00:00 Blankenberge  51.3106    3.1164      180
#> 19  2021-08-23 18:00:00 Blankenberge  51.3106    3.1164       82
#> 20  2021-08-23 19:00:00 Blankenberge  51.3106    3.1164       14
#> 21  2021-08-23 20:00:00 Blankenberge  51.3106    3.1164       59
#> 22  2021-08-23 21:00:00 Blankenberge  51.3106    3.1164      109
#> 23  2021-08-23 22:00:00 Blankenberge  51.3106    3.1164      175
#> 24  2021-08-23 23:00:00 Blankenberge  51.3106    3.1164      324
#> 25  2021-08-24 00:00:00 Blankenberge  51.3106    3.1164      457
#> 26  2021-08-24 01:00:00 Blankenberge  51.3106    3.1164      464
#> 27  2021-08-24 02:00:00 Blankenberge  51.3106    3.1164      456
#> 28  2021-08-24 03:00:00 Blankenberge  51.3106    3.1164      410
#> 29  2021-08-24 04:00:00 Blankenberge  51.3106    3.1164      330
#> 30  2021-08-24 05:00:00 Blankenberge  51.3106    3.1164      227
#> 31  2021-08-24 06:00:00 Blankenberge  51.3106    3.1164      123
#> 32  2021-08-24 07:00:00 Blankenberge  51.3106    3.1164       46
#> 33  2021-08-24 08:00:00 Blankenberge  51.3106    3.1164       83
#> 34  2021-08-24 09:00:00 Blankenberge  51.3106    3.1164      129
#> 35  2021-08-24 10:00:00 Blankenberge  51.3106    3.1164      178
#> 36  2021-08-24 11:00:00 Blankenberge  51.3106    3.1164      260
#> 37  2021-08-24 12:00:00 Blankenberge  51.3106    3.1164      425
#> 38  2021-08-24 13:00:00 Blankenberge  51.3106    3.1164      452
#> 39  2021-08-24 14:00:00 Blankenberge  51.3106    3.1164      446
#> 40  2021-08-24 15:00:00 Blankenberge  51.3106    3.1164      401
#> 41  2021-08-24 16:00:00 Blankenberge  51.3106    3.1164      334
#> 42  2021-08-24 17:00:00 Blankenberge  51.3106    3.1164      235
#> 43  2021-08-24 18:00:00 Blankenberge  51.3106    3.1164      129
#> 44  2021-08-24 19:00:00 Blankenberge  51.3106    3.1164       39
#> 45  2021-08-24 20:00:00 Blankenberge  51.3106    3.1164        4
#> 46  2021-08-24 21:00:00 Blankenberge  51.3106    3.1164       65
#> 47  2021-08-24 22:00:00 Blankenberge  51.3106    3.1164      114
#> 48  2021-08-24 23:00:00 Blankenberge  51.3106    3.1164      200
#> 49  2021-08-25 00:00:00 Blankenberge  51.3106    3.1164      390
#> 50  2021-08-25 01:00:00 Blankenberge  51.3106    3.1164      455
#> 51  2021-08-25 02:00:00 Blankenberge  51.3106    3.1164      460
#> 52  2021-08-25 03:00:00 Blankenberge  51.3106    3.1164      434
#> 53  2021-08-25 04:00:00 Blankenberge  51.3106    3.1164      378
#> 54  2021-08-25 05:00:00 Blankenberge  51.3106    3.1164      288
#> 55  2021-08-25 06:00:00 Blankenberge  51.3106    3.1164      187
#> 56  2021-08-25 07:00:00 Blankenberge  51.3106    3.1164       88
#> 57  2021-08-25 08:00:00 Blankenberge  51.3106    3.1164       56
#> 58  2021-08-25 09:00:00 Blankenberge  51.3106    3.1164      112
#> 59  2021-08-25 10:00:00 Blankenberge  51.3106    3.1164      151
#> 60  2021-08-25 11:00:00 Blankenberge  51.3106    3.1164      209
#> 61  2021-08-25 12:00:00 Blankenberge  51.3106    3.1164      326
#> 62  2021-08-25 13:00:00 Blankenberge  51.3106    3.1164      456
#> 63  2021-08-25 14:00:00 Blankenberge  51.3106    3.1164      463
#> 64  2021-08-25 15:00:00 Blankenberge  51.3106    3.1164      444
#> 65  2021-08-25 16:00:00 Blankenberge  51.3106    3.1164      391
#> 66  2021-08-25 17:00:00 Blankenberge  51.3106    3.1164      312
#> 67  2021-08-25 18:00:00 Blankenberge  51.3106    3.1164      207
#> 68  2021-08-25 19:00:00 Blankenberge  51.3106    3.1164      110
#> 69  2021-08-25 20:00:00 Blankenberge  51.3106    3.1164       38
#> 70  2021-08-25 21:00:00 Blankenberge  51.3106    3.1164       68
#> 71  2021-08-25 22:00:00 Blankenberge  51.3106    3.1164      114
#> 72  2021-08-25 23:00:00 Blankenberge  51.3106    3.1164      178
#> 73  2021-08-26 00:00:00 Blankenberge  51.3106    3.1164      300
#> 74  2021-08-26 01:00:00 Blankenberge  51.3106    3.1164      461
#> 75  2021-08-26 02:00:00 Blankenberge  51.3106    3.1164      484
#> 76  2021-08-26 03:00:00 Blankenberge  51.3106    3.1164      483
#> 77  2021-08-26 04:00:00 Blankenberge  51.3106    3.1164      445
#> 78  2021-08-26 05:00:00 Blankenberge  51.3106    3.1164      373
#> 79  2021-08-26 06:00:00 Blankenberge  51.3106    3.1164      271
#> 80  2021-08-26 07:00:00 Blankenberge  51.3106    3.1164      176
#> 81  2021-08-26 08:00:00 Blankenberge  51.3106    3.1164       96
#> 82  2021-08-26 09:00:00 Blankenberge  51.3106    3.1164      127
#> 83  2021-08-26 10:00:00 Blankenberge  51.3106    3.1164      170
#> 84  2021-08-26 11:00:00 Blankenberge  51.3106    3.1164      206
#> 85  2021-08-26 12:00:00 Blankenberge  51.3106    3.1164      283
#> 86  2021-08-26 13:00:00 Blankenberge  51.3106    3.1164      432
#> 87  2021-08-26 14:00:00 Blankenberge  51.3106    3.1164      478
#> 88  2021-08-26 15:00:00 Blankenberge  51.3106    3.1164      476
#> 89  2021-08-26 16:00:00 Blankenberge  51.3106    3.1164      440
#> 90  2021-08-26 17:00:00 Blankenberge  51.3106    3.1164      374
#> 91  2021-08-26 18:00:00 Blankenberge  51.3106    3.1164      277
#> 92  2021-08-26 19:00:00 Blankenberge  51.3106    3.1164      174
#> 93  2021-08-26 20:00:00 Blankenberge  51.3106    3.1164       84
#> 94  2021-08-26 21:00:00 Blankenberge  51.3106    3.1164       28
#> 95  2021-08-26 22:00:00 Blankenberge  51.3106    3.1164       79
#> 96  2021-08-26 23:00:00 Blankenberge  51.3106    3.1164      131
#> 97  2021-08-27 00:00:00 Blankenberge  51.3106    3.1164      206
#> 98  2021-08-27 01:00:00 Blankenberge  51.3106    3.1164      365
#> 99  2021-08-27 02:00:00 Blankenberge  51.3106    3.1164      452
#> 100 2021-08-27 03:00:00 Blankenberge  51.3106    3.1164      454
#> 101 2021-08-27 04:00:00 Blankenberge  51.3106    3.1164      438
#> 102 2021-08-27 05:00:00 Blankenberge  51.3106    3.1164      387
#> 103 2021-08-27 06:00:00 Blankenberge  51.3106    3.1164      302
#> 104 2021-08-27 07:00:00 Blankenberge  51.3106    3.1164      206
#> 105 2021-08-27 08:00:00 Blankenberge  51.3106    3.1164      113
#> 106 2021-08-27 09:00:00 Blankenberge  51.3106    3.1164       66
#> 107 2021-08-27 10:00:00 Blankenberge  51.3106    3.1164      114
#> 108 2021-08-27 11:00:00 Blankenberge  51.3106    3.1164      158
#> 109 2021-08-27 12:00:00 Blankenberge  51.3106    3.1164      214
#> 110 2021-08-27 13:00:00 Blankenberge  51.3106    3.1164      324
#> 111 2021-08-27 14:00:00 Blankenberge  51.3106    3.1164      446
#> 112 2021-08-27 15:00:00 Blankenberge  51.3106    3.1164      454
#> 113 2021-08-27 16:00:00 Blankenberge  51.3106    3.1164      442
#> 114 2021-08-27 17:00:00 Blankenberge  51.3106    3.1164      395
#> 115 2021-08-27 18:00:00 Blankenberge  51.3106    3.1164      322
#> 116 2021-08-27 19:00:00 Blankenberge  51.3106    3.1164      224
#> 117 2021-08-27 20:00:00 Blankenberge  51.3106    3.1164      130
#> 118 2021-08-27 21:00:00 Blankenberge  51.3106    3.1164       64
#> 119 2021-08-27 22:00:00 Blankenberge  51.3106    3.1164       65
#> 120 2021-08-27 23:00:00 Blankenberge  51.3106    3.1164      109
#> 121 2021-08-28 00:00:00 Blankenberge  51.3106    3.1164      168
#> 122 2021-08-28 01:00:00 Blankenberge  51.3106    3.1164      274
#> 123 2021-08-28 02:00:00 Blankenberge  51.3106    3.1164      416
#> 124 2021-08-28 03:00:00 Blankenberge  51.3106    3.1164      442
#> 125 2021-08-28 04:00:00 Blankenberge  51.3106    3.1164      440
#> 126 2021-08-28 05:00:00 Blankenberge  51.3106    3.1164      410
#> 127 2021-08-28 06:00:00 Blankenberge  51.3106    3.1164      347
#> 128 2021-08-28 07:00:00 Blankenberge  51.3106    3.1164      260
#> 129 2021-08-28 08:00:00 Blankenberge  51.3106    3.1164      165
#> 130 2021-08-28 09:00:00 Blankenberge  51.3106    3.1164       95
#> 131 2021-08-28 10:00:00 Blankenberge  51.3106    3.1164       95
#> 132 2021-08-28 11:00:00 Blankenberge  51.3106    3.1164      139
#> 133 2021-08-28 12:00:00 Blankenberge  51.3106    3.1164      187
#> 134 2021-08-28 13:00:00 Blankenberge  51.3106    3.1164      258
#> 135 2021-08-28 14:00:00 Blankenberge  51.3106    3.1164      384
#> 136 2021-08-28 15:00:00 Blankenberge  51.3106    3.1164      434
#> 137 2021-08-28 16:00:00 Blankenberge  51.3106    3.1164      435
#> 138 2021-08-28 17:00:00 Blankenberge  51.3106    3.1164      406
#> 139 2021-08-28 18:00:00 Blankenberge  51.3106    3.1164      351
#> 140 2021-08-28 19:00:00 Blankenberge  51.3106    3.1164      268
#> 141 2021-08-28 20:00:00 Blankenberge  51.3106    3.1164      178
#> 142 2021-08-28 21:00:00 Blankenberge  51.3106    3.1164       97
#> 143 2021-08-28 22:00:00 Blankenberge  51.3106    3.1164       51
#> 144 2021-08-28 23:00:00 Blankenberge  51.3106    3.1164       84
#> 145 2021-08-29 00:00:00 Blankenberge  51.3106    3.1164      133
#> 146 2021-08-29 01:00:00 Blankenberge  51.3106    3.1164      206
#> 147 2021-08-29 02:00:00 Blankenberge  51.3106    3.1164      338
#> 148 2021-08-29 03:00:00 Blankenberge  51.3106    3.1164      413
#> 149 2021-08-29 04:00:00 Blankenberge  51.3106    3.1164      421
#> 150 2021-08-29 05:00:00 Blankenberge  51.3106    3.1164      413
#> 151 2021-08-29 06:00:00 Blankenberge  51.3106    3.1164      372
#> 152 2021-08-29 07:00:00 Blankenberge  51.3106    3.1164      302
#> 153 2021-08-29 08:00:00 Blankenberge  51.3106    3.1164      221
#> 154 2021-08-29 09:00:00 Blankenberge  51.3106    3.1164      143
#> 155 2021-08-29 10:00:00 Blankenberge  51.3106    3.1164       95
#> 156 2021-08-29 11:00:00 Blankenberge  51.3106    3.1164      130
#> 157 2021-08-29 12:00:00 Blankenberge  51.3106    3.1164      166
#> 158 2021-08-29 13:00:00 Blankenberge  51.3106    3.1164      223
#> 159 2021-08-29 14:00:00 Blankenberge  51.3106    3.1164      318
#> 160 2021-08-29 15:00:00 Blankenberge  51.3106    3.1164      409
#> 161 2021-08-29 16:00:00 Blankenberge  51.3106    3.1164      423
#> 162 2021-08-29 17:00:00 Blankenberge  51.3106    3.1164      419
#> 163 2021-08-29 18:00:00 Blankenberge  51.3106    3.1164      381
#> 164 2021-08-29 19:00:00 Blankenberge  51.3106    3.1164      318
#> 165 2021-08-29 20:00:00 Blankenberge  51.3106    3.1164      234
#> 166 2021-08-29 21:00:00 Blankenberge  51.3106    3.1164      160
#> 167 2021-08-29 22:00:00 Blankenberge  51.3106    3.1164       96
#> 168 2021-08-29 23:00:00 Blankenberge  51.3106    3.1164       84
#> 169 2021-08-30 00:00:00 Blankenberge  51.3106    3.1164      121
#> 170 2021-08-30 01:00:00 Blankenberge  51.3106    3.1164      174
#> 171 2021-08-30 02:00:00 Blankenberge  51.3106    3.1164      252
#> 172 2021-08-30 03:00:00 Blankenberge  51.3106    3.1164      356
#> 173 2021-08-30 04:00:00 Blankenberge  51.3106    3.1164      390
#> 174 2021-08-30 05:00:00 Blankenberge  51.3106    3.1164      391
#> 175 2021-08-30 06:00:00 Blankenberge  51.3106    3.1164      372
#> 176 2021-08-30 07:00:00 Blankenberge  51.3106    3.1164      324
#> 177 2021-08-30 08:00:00 Blankenberge  51.3106    3.1164      257
#> 178 2021-08-30 09:00:00 Blankenberge  51.3106    3.1164      182
#> 179 2021-08-30 10:00:00 Blankenberge  51.3106    3.1164      120
#> 180 2021-08-30 11:00:00 Blankenberge  51.3106    3.1164       98
#> 181 2021-08-30 12:00:00 Blankenberge  51.3106    3.1164      134
#> 182 2021-08-30 13:00:00 Blankenberge  51.3106    3.1164      173
#> 183 2021-08-30 14:00:00 Blankenberge  51.3106    3.1164      238
#> 184 2021-08-30 15:00:00 Blankenberge  51.3106    3.1164      328
#> 185 2021-08-30 16:00:00 Blankenberge  51.3106    3.1164      380
#> 186 2021-08-30 17:00:00 Blankenberge  51.3106    3.1164      384
#> 187 2021-08-30 18:00:00 Blankenberge  51.3106    3.1164      373
#> 188 2021-08-30 19:00:00 Blankenberge  51.3106    3.1164      334
#> 189 2021-08-30 20:00:00 Blankenberge  51.3106    3.1164      276
#> 190 2021-08-30 21:00:00 Blankenberge  51.3106    3.1164      205
#> 191 2021-08-30 22:00:00 Blankenberge  51.3106    3.1164      141
#> 192 2021-08-30 23:00:00 Blankenberge  51.3106    3.1164      101
#> 193 2021-08-31 00:00:00 Blankenberge  51.3106    3.1164      111
#> 194 2021-08-31 01:00:00 Blankenberge  51.3106    3.1164      145
#> 195 2021-08-31 02:00:00 Blankenberge  51.3106    3.1164      199
#> 196 2021-08-31 03:00:00 Blankenberge  51.3106    3.1164      270
#> 197 2021-08-31 04:00:00 Blankenberge  51.3106    3.1164      340
#> 198 2021-08-31 05:00:00 Blankenberge  51.3106    3.1164      361
#> 199 2021-08-31 06:00:00 Blankenberge  51.3106    3.1164      361
#> 200 2021-08-31 07:00:00 Blankenberge  51.3106    3.1164      340
#> 201 2021-08-31 08:00:00 Blankenberge  51.3106    3.1164      297
#> 202 2021-08-31 09:00:00 Blankenberge  51.3106    3.1164      238
#> 203 2021-08-31 10:00:00 Blankenberge  51.3106    3.1164      177
#> 204 2021-08-31 11:00:00 Blankenberge  51.3106    3.1164      134
#> 205 2021-08-31 12:00:00 Blankenberge  51.3106    3.1164      115
#> 206 2021-08-31 13:00:00 Blankenberge  51.3106    3.1164      143
#> 207 2021-08-31 14:00:00 Blankenberge  51.3106    3.1164      187
#> 208 2021-08-31 15:00:00 Blankenberge  51.3106    3.1164      249
#> 209 2021-08-31 16:00:00 Blankenberge  51.3106    3.1164      316
#> 210 2021-08-31 17:00:00 Blankenberge  51.3106    3.1164      350
#> 211 2021-08-31 18:00:00 Blankenberge  51.3106    3.1164      353
#> 212 2021-08-31 19:00:00 Blankenberge  51.3106    3.1164      346
#> 213 2021-08-31 20:00:00 Blankenberge  51.3106    3.1164      313
#> 214 2021-08-31 21:00:00 Blankenberge  51.3106    3.1164      262
#> 215 2021-08-31 22:00:00 Blankenberge  51.3106    3.1164      205
#> 216 2021-08-31 23:00:00 Blankenberge  51.3106    3.1164      153
#> 217 2021-09-01 00:00:00 Blankenberge  51.3106    3.1164      121
#> 218 2021-09-01 01:00:00 Blankenberge  51.3106    3.1164      121
#> 219 2021-09-01 02:00:00 Blankenberge  51.3106    3.1164      154
#> 220 2021-09-01 03:00:00 Blankenberge  51.3106    3.1164      195
#> 221 2021-09-01 04:00:00 Blankenberge  51.3106    3.1164      252
#> 222 2021-09-01 05:00:00 Blankenberge  51.3106    3.1164      304
#> 223 2021-09-01 06:00:00 Blankenberge  51.3106    3.1164      326
#> 224 2021-09-01 07:00:00 Blankenberge  51.3106    3.1164      328
#> 225 2021-09-01 08:00:00 Blankenberge  51.3106    3.1164      319
#> 226 2021-09-01 09:00:00 Blankenberge  51.3106    3.1164      286
#> 227 2021-09-01 10:00:00 Blankenberge  51.3106    3.1164      240
#> 228 2021-09-01 11:00:00 Blankenberge  51.3106    3.1164      189
#> 229 2021-09-01 12:00:00 Blankenberge  51.3106    3.1164      140
#> 230 2021-09-01 13:00:00 Blankenberge  51.3106    3.1164      116
#> 231 2021-09-01 14:00:00 Blankenberge  51.3106    3.1164      131
#> 232 2021-09-01 15:00:00 Blankenberge  51.3106    3.1164      166
#> 233 2021-09-01 16:00:00 Blankenberge  51.3106    3.1164      213
#> 234 2021-09-01 17:00:00 Blankenberge  51.3106    3.1164      273
#> 235 2021-09-01 18:00:00 Blankenberge  51.3106    3.1164      316
#> 236 2021-09-01 19:00:00 Blankenberge  51.3106    3.1164      331
#> 237 2021-09-01 20:00:00 Blankenberge  51.3106    3.1164      330
#> 238 2021-09-01 21:00:00 Blankenberge  51.3106    3.1164      309
#> 239 2021-09-01 22:00:00 Blankenberge  51.3106    3.1164      273
#> 240 2021-09-01 23:00:00 Blankenberge  51.3106    3.1164      222
#> 241 2021-09-02 00:00:00 Blankenberge  51.3106    3.1164      174
#> 242 2021-09-02 01:00:00 Blankenberge  51.3106    3.1164      133
#> 243 2021-09-02 02:00:00 Blankenberge  51.3106    3.1164      111
#> 244 2021-09-02 03:00:00 Blankenberge  51.3106    3.1164      135
#> 245 2021-09-02 04:00:00 Blankenberge  51.3106    3.1164      171
#> 246 2021-09-02 05:00:00 Blankenberge  51.3106    3.1164      215
#> 247 2021-09-02 06:00:00 Blankenberge  51.3106    3.1164      270
#> 248 2021-09-02 07:00:00 Blankenberge  51.3106    3.1164      313
#> 249 2021-09-02 08:00:00 Blankenberge  51.3106    3.1164      326
#> 250 2021-09-02 09:00:00 Blankenberge  51.3106    3.1164      326
#> 251 2021-09-02 10:00:00 Blankenberge  51.3106    3.1164      309
#> 252 2021-09-02 11:00:00 Blankenberge  51.3106    3.1164      269
#> 253 2021-09-02 12:00:00 Blankenberge  51.3106    3.1164      219
#> 254 2021-09-02 13:00:00 Blankenberge  51.3106    3.1164      165
#> 255 2021-09-02 14:00:00 Blankenberge  51.3106    3.1164      123
#> 256 2021-09-02 15:00:00 Blankenberge  51.3106    3.1164      103
#> 257 2021-09-02 16:00:00 Blankenberge  51.3106    3.1164      133
#> 258 2021-09-02 17:00:00 Blankenberge  51.3106    3.1164      174
#> 259 2021-09-02 18:00:00 Blankenberge  51.3106    3.1164      237
#> 260 2021-09-02 19:00:00 Blankenberge  51.3106    3.1164      301
#> 261 2021-09-02 20:00:00 Blankenberge  51.3106    3.1164      340
#> 262 2021-09-02 21:00:00 Blankenberge  51.3106    3.1164      345
#> 263 2021-09-02 22:00:00 Blankenberge  51.3106    3.1164      338
#> 264 2021-09-02 23:00:00 Blankenberge  51.3106    3.1164      309
#> 265 2021-09-03 00:00:00 Blankenberge  51.3106    3.1164      260
#> 266 2021-09-03 01:00:00 Blankenberge  51.3106    3.1164      202
#> 267 2021-09-03 02:00:00 Blankenberge  51.3106    3.1164      145
#> 268 2021-09-03 03:00:00 Blankenberge  51.3106    3.1164      105
#> 269 2021-09-03 04:00:00 Blankenberge  51.3106    3.1164      113
#> 270 2021-09-03 05:00:00 Blankenberge  51.3106    3.1164      149
#> 271 2021-09-03 06:00:00 Blankenberge  51.3106    3.1164      197
#> 272 2021-09-03 07:00:00 Blankenberge  51.3106    3.1164      258
#> 273 2021-09-03 08:00:00 Blankenberge  51.3106    3.1164      330
#> 274 2021-09-03 09:00:00 Blankenberge  51.3106    3.1164      359
#> 275 2021-09-03 10:00:00 Blankenberge  51.3106    3.1164      359
#> 276 2021-09-03 11:00:00 Blankenberge  51.3106    3.1164      342
#> 277 2021-09-03 12:00:00 Blankenberge  51.3106    3.1164      300
#> 278 2021-09-03 13:00:00 Blankenberge  51.3106    3.1164      244
#> 279 2021-09-03 14:00:00 Blankenberge  51.3106    3.1164      179
#> 280 2021-09-03 15:00:00 Blankenberge  51.3106    3.1164      120
#> 281 2021-09-03 16:00:00 Blankenberge  51.3106    3.1164       85
#> 282 2021-09-03 17:00:00 Blankenberge  51.3106    3.1164      113
#> 283 2021-09-03 18:00:00 Blankenberge  51.3106    3.1164      162
#> 284 2021-09-03 19:00:00 Blankenberge  51.3106    3.1164      233
#> 285 2021-09-03 20:00:00 Blankenberge  51.3106    3.1164      325
#> 286 2021-09-03 21:00:00 Blankenberge  51.3106    3.1164      380
#> 287 2021-09-03 22:00:00 Blankenberge  51.3106    3.1164      388
#> 288 2021-09-03 23:00:00 Blankenberge  51.3106    3.1164      383
#> 289 2021-09-04 00:00:00 Blankenberge  51.3106    3.1164      348
#> 290 2021-09-04 01:00:00 Blankenberge  51.3106    3.1164      289
#> 291 2021-09-04 02:00:00 Blankenberge  51.3106    3.1164      217
#> 292 2021-09-04 03:00:00 Blankenberge  51.3106    3.1164      149
#> 293 2021-09-04 04:00:00 Blankenberge  51.3106    3.1164       96
#> 294 2021-09-04 05:00:00 Blankenberge  51.3106    3.1164      106
#> 295 2021-09-04 06:00:00 Blankenberge  51.3106    3.1164      142
#> 296 2021-09-04 07:00:00 Blankenberge  51.3106    3.1164      194
#> 297 2021-09-04 08:00:00 Blankenberge  51.3106    3.1164      275
#> 298 2021-09-04 09:00:00 Blankenberge  51.3106    3.1164      367
#> 299 2021-09-04 10:00:00 Blankenberge  51.3106    3.1164      391
#> 300 2021-09-04 11:00:00 Blankenberge  51.3106    3.1164      388
#> 301 2021-09-04 12:00:00 Blankenberge  51.3106    3.1164      361
#> 302 2021-09-04 13:00:00 Blankenberge  51.3106    3.1164      310
#> 303 2021-09-04 14:00:00 Blankenberge  51.3106    3.1164      236
#> 304 2021-09-04 15:00:00 Blankenberge  51.3106    3.1164      166
#> 305 2021-09-04 16:00:00 Blankenberge  51.3106    3.1164       95
#> 306 2021-09-04 17:00:00 Blankenberge  51.3106    3.1164       73
#> 307 2021-09-04 18:00:00 Blankenberge  51.3106    3.1164      109
#> 308 2021-09-04 19:00:00 Blankenberge  51.3106    3.1164      163
#> 309 2021-09-04 20:00:00 Blankenberge  51.3106    3.1164      246
#> 310 2021-09-04 21:00:00 Blankenberge  51.3106    3.1164      363
#> 311 2021-09-04 22:00:00 Blankenberge  51.3106    3.1164      412
#> 312 2021-09-04 23:00:00 Blankenberge  51.3106    3.1164      414
#> 313 2021-09-05 00:00:00 Blankenberge  51.3106    3.1164      400
#> 314 2021-09-05 01:00:00 Blankenberge  51.3106    3.1164      354
#> 315 2021-09-05 02:00:00 Blankenberge  51.3106    3.1164      281
#> 316 2021-09-05 03:00:00 Blankenberge  51.3106    3.1164      203
#> 317 2021-09-05 04:00:00 Blankenberge  51.3106    3.1164      124
#> 318 2021-09-05 05:00:00 Blankenberge  51.3106    3.1164       73
#> 319 2021-09-05 06:00:00 Blankenberge  51.3106    3.1164      101
#> 320 2021-09-05 07:00:00 Blankenberge  51.3106    3.1164      141
#> 321 2021-09-05 08:00:00 Blankenberge  51.3106    3.1164      201
#> 322 2021-09-05 09:00:00 Blankenberge  51.3106    3.1164      311
#> 323 2021-09-05 10:00:00 Blankenberge  51.3106    3.1164      405
#> 324 2021-09-05 11:00:00 Blankenberge  51.3106    3.1164      413
#> 325 2021-09-05 12:00:00 Blankenberge  51.3106    3.1164      405
#> 326 2021-09-05 13:00:00 Blankenberge  51.3106    3.1164      365
#> 327 2021-09-05 14:00:00 Blankenberge  51.3106    3.1164      302
#> 328 2021-09-05 15:00:00 Blankenberge  51.3106    3.1164      219
#> 329 2021-09-05 16:00:00 Blankenberge  51.3106    3.1164      134
#> 330 2021-09-05 17:00:00 Blankenberge  51.3106    3.1164       65
#> 331 2021-09-05 18:00:00 Blankenberge  51.3106    3.1164       61
#> 332 2021-09-05 19:00:00 Blankenberge  51.3106    3.1164      105
#> 333 2021-09-05 20:00:00 Blankenberge  51.3106    3.1164      171
#> 334 2021-09-05 21:00:00 Blankenberge  51.3106    3.1164      283
#> 335 2021-09-05 22:00:00 Blankenberge  51.3106    3.1164      411
#> 336 2021-09-05 23:00:00 Blankenberge  51.3106    3.1164      439
#> 337 2021-09-06 00:00:00 Blankenberge  51.3106    3.1164      436
#> 338 2021-09-06 01:00:00 Blankenberge  51.3106    3.1164      408
#> 339 2021-09-06 02:00:00 Blankenberge  51.3106    3.1164      346
#> 340 2021-09-06 03:00:00 Blankenberge  51.3106    3.1164      262
#> 341 2021-09-06 04:00:00 Blankenberge  51.3106    3.1164      173
#> 342 2021-09-06 05:00:00 Blankenberge  51.3106    3.1164       92
#> 343 2021-09-06 06:00:00 Blankenberge  51.3106    3.1164       76
#> 344 2021-09-06 07:00:00 Blankenberge  51.3106    3.1164      111
#> 345 2021-09-06 08:00:00 Blankenberge  51.3106    3.1164      162
#> 346 2021-09-06 09:00:00 Blankenberge  51.3106    3.1164      231
#> 347 2021-09-06 10:00:00 Blankenberge  51.3106    3.1164      372
#> 348 2021-09-06 11:00:00 Blankenberge  51.3106    3.1164      434
#> 349 2021-09-06 12:00:00 Blankenberge  51.3106    3.1164      434
#> 350 2021-09-06 13:00:00 Blankenberge  51.3106    3.1164      410
#> 351 2021-09-06 14:00:00 Blankenberge  51.3106    3.1164      361
#> 352 2021-09-06 15:00:00 Blankenberge  51.3106    3.1164      281
#> 353 2021-09-06 16:00:00 Blankenberge  51.3106    3.1164      190
#> 354 2021-09-06 17:00:00 Blankenberge  51.3106    3.1164       95
#> 355 2021-09-06 18:00:00 Blankenberge  51.3106    3.1164       32
#> 356 2021-09-06 19:00:00 Blankenberge  51.3106    3.1164       73
#> 357 2021-09-06 20:00:00 Blankenberge  51.3106    3.1164      120
#> 358 2021-09-06 21:00:00 Blankenberge  51.3106    3.1164      194
#> 359 2021-09-06 22:00:00 Blankenberge  51.3106    3.1164      346
#> 360 2021-09-06 23:00:00 Blankenberge  51.3106    3.1164      446
#> 361 2021-09-07 00:00:00 Blankenberge  51.3106    3.1164      457
#> 362 2021-09-07 01:00:00 Blankenberge  51.3106    3.1164      446
#> 363 2021-09-07 02:00:00 Blankenberge  51.3106    3.1164      401
#> 364 2021-09-07 03:00:00 Blankenberge  51.3106    3.1164      326
#> 365 2021-09-07 04:00:00 Blankenberge  51.3106    3.1164      231
#> 366 2021-09-07 05:00:00 Blankenberge  51.3106    3.1164      135
#> 367 2021-09-07 06:00:00 Blankenberge  51.3106    3.1164       62
#> 368 2021-09-07 07:00:00 Blankenberge  51.3106    3.1164       85
#> 369 2021-09-07 08:00:00 Blankenberge  51.3106    3.1164      129
#> 370 2021-09-07 09:00:00 Blankenberge  51.3106    3.1164      180
#> 371 2021-09-07 10:00:00 Blankenberge  51.3106    3.1164      278
#> 372 2021-09-07 11:00:00 Blankenberge  51.3106    3.1164      437
#> 373 2021-09-07 12:00:00 Blankenberge  51.3106    3.1164      458
#> 374 2021-09-07 13:00:00 Blankenberge  51.3106    3.1164      454
#> 375 2021-09-07 14:00:00 Blankenberge  51.3106    3.1164      412
#> 376 2021-09-07 15:00:00 Blankenberge  51.3106    3.1164      345
#> 377 2021-09-07 16:00:00 Blankenberge  51.3106    3.1164      251
#> 378 2021-09-07 17:00:00 Blankenberge  51.3106    3.1164      150
#> 379 2021-09-07 18:00:00 Blankenberge  51.3106    3.1164       61
#> 380 2021-09-07 19:00:00 Blankenberge  51.3106    3.1164       36
#> 381 2021-09-07 20:00:00 Blankenberge  51.3106    3.1164       92
#> 382 2021-09-07 21:00:00 Blankenberge  51.3106    3.1164      146
#> 383 2021-09-07 22:00:00 Blankenberge  51.3106    3.1164      241
#> 384 2021-09-07 23:00:00 Blankenberge  51.3106    3.1164      425
#> 385 2021-09-08 00:00:00 Blankenberge  51.3106    3.1164      474
#> 386 2021-09-08 01:00:00 Blankenberge  51.3106    3.1164      476
#> 387 2021-09-08 02:00:00 Blankenberge  51.3106    3.1164      447
#> 388 2021-09-08 03:00:00 Blankenberge  51.3106    3.1164      389
#> 389 2021-09-08 04:00:00 Blankenberge  51.3106    3.1164      296
#> 390 2021-09-08 05:00:00 Blankenberge  51.3106    3.1164      196
#> 391 2021-09-08 06:00:00 Blankenberge  51.3106    3.1164       99
#> 392 2021-09-08 07:00:00 Blankenberge  51.3106    3.1164       55
#> 393 2021-09-08 08:00:00 Blankenberge  51.3106    3.1164      109
#> 394 2021-09-08 09:00:00 Blankenberge  51.3106    3.1164      155
#> 395 2021-09-08 10:00:00 Blankenberge  51.3106    3.1164      214
#> 396 2021-09-08 11:00:00 Blankenberge  51.3106    3.1164      352
#> 397 2021-09-08 12:00:00 Blankenberge  51.3106    3.1164      470
#> 398 2021-09-08 13:00:00 Blankenberge  51.3106    3.1164      472
#> 399 2021-09-08 14:00:00 Blankenberge  51.3106    3.1164      454
#> 400 2021-09-08 15:00:00 Blankenberge  51.3106    3.1164      402
#> 401 2021-09-08 16:00:00 Blankenberge  51.3106    3.1164      317
#> 402 2021-09-08 17:00:00 Blankenberge  51.3106    3.1164      215
#> 403 2021-09-08 18:00:00 Blankenberge  51.3106    3.1164      108
#> 404 2021-09-08 19:00:00 Blankenberge  51.3106    3.1164       29
#> 405 2021-09-08 20:00:00 Blankenberge  51.3106    3.1164       55
#> 406 2021-09-08 21:00:00 Blankenberge  51.3106    3.1164      107
#> 407 2021-09-08 22:00:00 Blankenberge  51.3106    3.1164      175
#> 408 2021-09-08 23:00:00 Blankenberge  51.3106    3.1164      314
#> 409 2021-09-09 00:00:00 Blankenberge  51.3106    3.1164      471
#> 410 2021-09-09 01:00:00 Blankenberge  51.3106    3.1164      490
#> 411 2021-09-09 02:00:00 Blankenberge  51.3106    3.1164      486
#> 412 2021-09-09 03:00:00 Blankenberge  51.3106    3.1164      439
#> 413 2021-09-09 04:00:00 Blankenberge  51.3106    3.1164      365
#> 414 2021-09-09 05:00:00 Blankenberge  51.3106    3.1164      264
#> 415 2021-09-09 06:00:00 Blankenberge  51.3106    3.1164      164
#> 416 2021-09-09 07:00:00 Blankenberge  51.3106    3.1164       77
#> 417 2021-09-09 08:00:00 Blankenberge  51.3106    3.1164       82
#> 418 2021-09-09 09:00:00 Blankenberge  51.3106    3.1164      132
#> 419 2021-09-09 10:00:00 Blankenberge  51.3106    3.1164      182
#> 420 2021-09-09 11:00:00 Blankenberge  51.3106    3.1164      262
#> 421 2021-09-09 12:00:00 Blankenberge  51.3106    3.1164      446
#> 422 2021-09-09 13:00:00 Blankenberge  51.3106    3.1164      486
#> 423 2021-09-09 14:00:00 Blankenberge  51.3106    3.1164      483
#> 424 2021-09-09 15:00:00 Blankenberge  51.3106    3.1164      446
#> 425 2021-09-09 16:00:00 Blankenberge  51.3106    3.1164      379
#> 426 2021-09-09 17:00:00 Blankenberge  51.3106    3.1164      284
#> 427 2021-09-09 18:00:00 Blankenberge  51.3106    3.1164      177
#> 428 2021-09-09 19:00:00 Blankenberge  51.3106    3.1164       78
#> 429 2021-09-09 20:00:00 Blankenberge  51.3106    3.1164        9
#> 430 2021-09-09 21:00:00 Blankenberge  51.3106    3.1164       72
#> 431 2021-09-09 22:00:00 Blankenberge  51.3106    3.1164      125
#> 432 2021-09-09 23:00:00 Blankenberge  51.3106    3.1164      209
#> 433 2021-09-10 00:00:00 Blankenberge  51.3106    3.1164      403
#> 434 2021-09-10 01:00:00 Blankenberge  51.3106    3.1164      482
#> 435 2021-09-10 02:00:00 Blankenberge  51.3106    3.1164      487
#> 436 2021-09-10 03:00:00 Blankenberge  51.3106    3.1164      466
#> 437 2021-09-10 04:00:00 Blankenberge  51.3106    3.1164      411
#> 438 2021-09-10 05:00:00 Blankenberge  51.3106    3.1164      321
#> 439 2021-09-10 06:00:00 Blankenberge  51.3106    3.1164      220
#> 440 2021-09-10 07:00:00 Blankenberge  51.3106    3.1164      121
#> 441 2021-09-10 08:00:00 Blankenberge  51.3106    3.1164       47
#> 442 2021-09-10 09:00:00 Blankenberge  51.3106    3.1164       94
#> 443 2021-09-10 10:00:00 Blankenberge  51.3106    3.1164      143
#> 444 2021-09-10 11:00:00 Blankenberge  51.3106    3.1164      201
#> 445 2021-09-10 12:00:00 Blankenberge  51.3106    3.1164      321
#> 446 2021-09-10 13:00:00 Blankenberge  51.3106    3.1164      475
#> 447 2021-09-10 14:00:00 Blankenberge  51.3106    3.1164      483
#> 448 2021-09-10 15:00:00 Blankenberge  51.3106    3.1164      472
#> 449 2021-09-10 16:00:00 Blankenberge  51.3106    3.1164      420
#> 450 2021-09-10 17:00:00 Blankenberge  51.3106    3.1164      343
#> 451 2021-09-10 18:00:00 Blankenberge  51.3106    3.1164      245
#> 452 2021-09-10 19:00:00 Blankenberge  51.3106    3.1164      143
#> 453 2021-09-10 20:00:00 Blankenberge  51.3106    3.1164       53
#> 454 2021-09-10 21:00:00 Blankenberge  51.3106    3.1164       36
#> 455 2021-09-10 22:00:00 Blankenberge  51.3106    3.1164       94
#> 456 2021-09-10 23:00:00 Blankenberge  51.3106    3.1164      159
#> 457 2021-09-11 00:00:00 Blankenberge  51.3106    3.1164      270
#> 458 2021-09-11 01:00:00 Blankenberge  51.3106    3.1164      453
#> 459 2021-09-11 02:00:00 Blankenberge  51.3106    3.1164      481
#> 460 2021-09-11 03:00:00 Blankenberge  51.3106    3.1164      480
#> 461 2021-09-11 04:00:00 Blankenberge  51.3106    3.1164      449
#> 462 2021-09-11 05:00:00 Blankenberge  51.3106    3.1164      383
#> 463 2021-09-11 06:00:00 Blankenberge  51.3106    3.1164      290
#> 464 2021-09-11 07:00:00 Blankenberge  51.3106    3.1164      195
#> 465 2021-09-11 08:00:00 Blankenberge  51.3106    3.1164      101
#> 466 2021-09-11 09:00:00 Blankenberge  51.3106    3.1164       69
#> 467 2021-09-11 10:00:00 Blankenberge  51.3106    3.1164      122
#> 468 2021-09-11 11:00:00 Blankenberge  51.3106    3.1164      170
#> 469 2021-09-11 12:00:00 Blankenberge  51.3106    3.1164      236
#> 470 2021-09-11 13:00:00 Blankenberge  51.3106    3.1164      393
#> 471 2021-09-11 14:00:00 Blankenberge  51.3106    3.1164      476
#> 472 2021-09-11 15:00:00 Blankenberge  51.3106    3.1164      476
#> 473 2021-09-11 16:00:00 Blankenberge  51.3106    3.1164      450
#> 474 2021-09-11 17:00:00 Blankenberge  51.3106    3.1164      396
#> 475 2021-09-11 18:00:00 Blankenberge  51.3106    3.1164      310
#> 476 2021-09-11 19:00:00 Blankenberge  51.3106    3.1164      211
#> 477 2021-09-11 20:00:00 Blankenberge  51.3106    3.1164      113
#> 478 2021-09-11 21:00:00 Blankenberge  51.3106    3.1164       37
#> 479 2021-09-11 22:00:00 Blankenberge  51.3106    3.1164       59
#> 480 2021-09-11 23:00:00 Blankenberge  51.3106    3.1164      109
#> 481 2021-09-12 00:00:00 Blankenberge  51.3106    3.1164      179
#> 482 2021-09-12 01:00:00 Blankenberge  51.3106    3.1164      318
#> 483 2021-09-12 02:00:00 Blankenberge  51.3106    3.1164      453
#> 484 2021-09-12 03:00:00 Blankenberge  51.3106    3.1164      467
#> 485 2021-09-12 04:00:00 Blankenberge  51.3106    3.1164      458
#> 486 2021-09-12 05:00:00 Blankenberge  51.3106    3.1164      420
#> 487 2021-09-12 06:00:00 Blankenberge  51.3106    3.1164      349
#> 488 2021-09-12 07:00:00 Blankenberge  51.3106    3.1164      256
#> 489 2021-09-12 08:00:00 Blankenberge  51.3106    3.1164      164
#> 490 2021-09-12 09:00:00 Blankenberge  51.3106    3.1164       89
#> 491 2021-09-12 10:00:00 Blankenberge  51.3106    3.1164      102
#> 492 2021-09-12 11:00:00 Blankenberge  51.3106    3.1164      145
#> 493 2021-09-12 12:00:00 Blankenberge  51.3106    3.1164      199
#> 494 2021-09-12 13:00:00 Blankenberge  51.3106    3.1164      284
#> 495 2021-09-12 14:00:00 Blankenberge  51.3106    3.1164      432
#> 496 2021-09-12 15:00:00 Blankenberge  51.3106    3.1164      474
#> 497 2021-09-12 16:00:00 Blankenberge  51.3106    3.1164      473
#> 498 2021-09-12 17:00:00 Blankenberge  51.3106    3.1164      439
#> 499 2021-09-12 18:00:00 Blankenberge  51.3106    3.1164      381
#> 500 2021-09-12 19:00:00 Blankenberge  51.3106    3.1164      291
#> 501 2021-09-12 20:00:00 Blankenberge  51.3106    3.1164      198
#> 502 2021-09-12 21:00:00 Blankenberge  51.3106    3.1164      105
#> 503 2021-09-12 22:00:00 Blankenberge  51.3106    3.1164       47
#> 504 2021-09-12 23:00:00 Blankenberge  51.3106    3.1164       95
#> 505 2021-09-13 00:00:00 Blankenberge  51.3106    3.1164      144
#> 506 2021-09-13 01:00:00 Blankenberge  51.3106    3.1164      215
#> 507 2021-09-13 02:00:00 Blankenberge  51.3106    3.1164      360
#> 508 2021-09-13 03:00:00 Blankenberge  51.3106    3.1164      446
#> 509 2021-09-13 04:00:00 Blankenberge  51.3106    3.1164      452
#> 510 2021-09-13 05:00:00 Blankenberge  51.3106    3.1164      436
#> 511 2021-09-13 06:00:00 Blankenberge  51.3106    3.1164      391
#> 512 2021-09-13 07:00:00 Blankenberge  51.3106    3.1164      314
#> 513 2021-09-13 08:00:00 Blankenberge  51.3106    3.1164      225
#> 514 2021-09-13 09:00:00 Blankenberge  51.3106    3.1164      141
#> 515 2021-09-13 10:00:00 Blankenberge  51.3106    3.1164       80
#> 516 2021-09-13 11:00:00 Blankenberge  51.3106    3.1164      102
#> 517 2021-09-13 12:00:00 Blankenberge  51.3106    3.1164      146
#> 518 2021-09-13 13:00:00 Blankenberge  51.3106    3.1164      198
#> 519 2021-09-13 14:00:00 Blankenberge  51.3106    3.1164      295
#> 520 2021-09-13 15:00:00 Blankenberge  51.3106    3.1164      414
#> 521 2021-09-13 16:00:00 Blankenberge  51.3106    3.1164      438
#> 522 2021-09-13 17:00:00 Blankenberge  51.3106    3.1164      430
#> 523 2021-09-13 18:00:00 Blankenberge  51.3106    3.1164      394
#> 524 2021-09-13 19:00:00 Blankenberge  51.3106    3.1164      334
#> 525 2021-09-13 20:00:00 Blankenberge  51.3106    3.1164      251
#> 526 2021-09-13 21:00:00 Blankenberge  51.3106    3.1164      158
#> 527 2021-09-13 22:00:00 Blankenberge  51.3106    3.1164       84
#> 528 2021-09-13 23:00:00 Blankenberge  51.3106    3.1164       46
#> 529 2021-09-14 00:00:00 Blankenberge  51.3106    3.1164       90
#> 530 2021-09-14 01:00:00 Blankenberge  51.3106    3.1164      141
#> 531 2021-09-14 02:00:00 Blankenberge  51.3106    3.1164      222
#> 532 2021-09-14 03:00:00 Blankenberge  51.3106    3.1164      351
#> 533 2021-09-14 04:00:00 Blankenberge  51.3106    3.1164      413
#> 534 2021-09-14 05:00:00 Blankenberge  51.3106    3.1164      418
#> 535 2021-09-14 06:00:00 Blankenberge  51.3106    3.1164      403
#> 536 2021-09-14 07:00:00 Blankenberge  51.3106    3.1164      358
#> 537 2021-09-14 08:00:00 Blankenberge  51.3106    3.1164      294
#> 538 2021-09-14 09:00:00 Blankenberge  51.3106    3.1164      217
#> 539 2021-09-14 10:00:00 Blankenberge  51.3106    3.1164      142
#> 540 2021-09-14 11:00:00 Blankenberge  51.3106    3.1164       91
#> 541 2021-09-14 12:00:00 Blankenberge  51.3106    3.1164      114
#> 542 2021-09-14 13:00:00 Blankenberge  51.3106    3.1164      156
#> 543 2021-09-14 14:00:00 Blankenberge  51.3106    3.1164      220
#> 544 2021-09-14 15:00:00 Blankenberge  51.3106    3.1164      317
#> 545 2021-09-14 16:00:00 Blankenberge  51.3106    3.1164      405
#> 546 2021-09-14 17:00:00 Blankenberge  51.3106    3.1164      422
#> 547 2021-09-14 18:00:00 Blankenberge  51.3106    3.1164      418
#> 548 2021-09-14 19:00:00 Blankenberge  51.3106    3.1164      388
#> 549 2021-09-14 20:00:00 Blankenberge  51.3106    3.1164      334
#> 550 2021-09-14 21:00:00 Blankenberge  51.3106    3.1164      263
#> 551 2021-09-14 22:00:00 Blankenberge  51.3106    3.1164      193
#> 552 2021-09-14 23:00:00 Blankenberge  51.3106    3.1164      127
#> 553 2021-09-15 00:00:00 Blankenberge  51.3106    3.1164       86
#> 554 2021-09-15 01:00:00 Blankenberge  51.3106    3.1164      117
#> 555 2021-09-15 02:00:00 Blankenberge  51.3106    3.1164      171
#> 556 2021-09-15 03:00:00 Blankenberge  51.3106    3.1164      238
#> 557 2021-09-15 04:00:00 Blankenberge  51.3106    3.1164      338
#> 558 2021-09-15 05:00:00 Blankenberge  51.3106    3.1164      398
#> 559 2021-09-15 06:00:00 Blankenberge  51.3106    3.1164      404
#> 560 2021-09-15 07:00:00 Blankenberge  51.3106    3.1164      394
#> 561 2021-09-15 08:00:00 Blankenberge  51.3106    3.1164      362
#> 562 2021-09-15 09:00:00 Blankenberge  51.3106    3.1164      308
#> 563 2021-09-15 10:00:00 Blankenberge  51.3106    3.1164      241
#> 564 2021-09-15 11:00:00 Blankenberge  51.3106    3.1164      178
#> 565 2021-09-15 12:00:00 Blankenberge  51.3106    3.1164      127
#> 566 2021-09-15 13:00:00 Blankenberge  51.3106    3.1164      122
#> 567 2021-09-15 14:00:00 Blankenberge  51.3106    3.1164      160
#> 568 2021-09-15 15:00:00 Blankenberge  51.3106    3.1164      214
#> 569 2021-09-15 16:00:00 Blankenberge  51.3106    3.1164      289
#> 570 2021-09-15 17:00:00 Blankenberge  51.3106    3.1164      368
#> 571 2021-09-15 18:00:00 Blankenberge  51.3106    3.1164      406
#> 572 2021-09-15 19:00:00 Blankenberge  51.3106    3.1164      405
#> 573 2021-09-15 20:00:00 Blankenberge  51.3106    3.1164      390
#> 574 2021-09-15 21:00:00 Blankenberge  51.3106    3.1164      350
#> 575 2021-09-15 22:00:00 Blankenberge  51.3106    3.1164      290
#> 576 2021-09-15 23:00:00 Blankenberge  51.3106    3.1164      227
#> 577 2021-09-16 00:00:00 Blankenberge  51.3106    3.1164      166
#> 578 2021-09-16 01:00:00 Blankenberge  51.3106    3.1164      116
#> 579 2021-09-16 02:00:00 Blankenberge  51.3106    3.1164      122
#> 580 2021-09-16 03:00:00 Blankenberge  51.3106    3.1164      161
#> 581 2021-09-16 04:00:00 Blankenberge  51.3106    3.1164      211
#> 582 2021-09-16 05:00:00 Blankenberge  51.3106    3.1164      283
#> 583 2021-09-16 06:00:00 Blankenberge  51.3106    3.1164      357
#> 584 2021-09-16 07:00:00 Blankenberge  51.3106    3.1164      384
#> 585 2021-09-16 08:00:00 Blankenberge  51.3106    3.1164      384
#> 586 2021-09-16 09:00:00 Blankenberge  51.3106    3.1164      366
#> 587 2021-09-16 10:00:00 Blankenberge  51.3106    3.1164      322
#> 588 2021-09-16 11:00:00 Blankenberge  51.3106    3.1164      269
#> 589 2021-09-16 12:00:00 Blankenberge  51.3106    3.1164      209
#> 590 2021-09-16 13:00:00 Blankenberge  51.3106    3.1164      151
#> 591 2021-09-16 14:00:00 Blankenberge  51.3106    3.1164      105
#> 592 2021-09-16 15:00:00 Blankenberge  51.3106    3.1164      125
#> 593 2021-09-16 16:00:00 Blankenberge  51.3106    3.1164      171
#> 594 2021-09-16 17:00:00 Blankenberge  51.3106    3.1164      230
#> 595 2021-09-16 18:00:00 Blankenberge  51.3106    3.1164      322
#> 596 2021-09-16 19:00:00 Blankenberge  51.3106    3.1164      392
#> 597 2021-09-16 20:00:00 Blankenberge  51.3106    3.1164      411
#> 598 2021-09-16 21:00:00 Blankenberge  51.3106    3.1164      408
#> 599 2021-09-16 22:00:00 Blankenberge  51.3106    3.1164      384
#> 600 2021-09-16 23:00:00 Blankenberge  51.3106    3.1164      333
#> 601 2021-09-17 00:00:00 Blankenberge  51.3106    3.1164      270
#> 602 2021-09-17 01:00:00 Blankenberge  51.3106    3.1164      203
#> 603 2021-09-17 02:00:00 Blankenberge  51.3106    3.1164      135
#> 604 2021-09-17 03:00:00 Blankenberge  51.3106    3.1164       98
#> 605 2021-09-17 04:00:00 Blankenberge  51.3106    3.1164      133
#> 606 2021-09-17 05:00:00 Blankenberge  51.3106    3.1164      173
#> 607 2021-09-17 06:00:00 Blankenberge  51.3106    3.1164      231
#> 608 2021-09-17 07:00:00 Blankenberge  51.3106    3.1164      319
#> 609 2021-09-17 08:00:00 Blankenberge  51.3106    3.1164      382
#> 610 2021-09-17 09:00:00 Blankenberge  51.3106    3.1164      388
#> 611 2021-09-17 10:00:00 Blankenberge  51.3106    3.1164      383
#> 612 2021-09-17 11:00:00 Blankenberge  51.3106    3.1164      347
#> 613 2021-09-17 12:00:00 Blankenberge  51.3106    3.1164      295
#> 614 2021-09-17 13:00:00 Blankenberge  51.3106    3.1164      229
#> 615 2021-09-17 14:00:00 Blankenberge  51.3106    3.1164      158
#> 616 2021-09-17 15:00:00 Blankenberge  51.3106    3.1164       98
#> 617 2021-09-17 16:00:00 Blankenberge  51.3106    3.1164       79
#> 618 2021-09-17 17:00:00 Blankenberge  51.3106    3.1164      124
#> 619 2021-09-17 18:00:00 Blankenberge  51.3106    3.1164      183
#> 620 2021-09-17 19:00:00 Blankenberge  51.3106    3.1164      275
#> 621 2021-09-17 20:00:00 Blankenberge  51.3106    3.1164      387
#> 622 2021-09-17 21:00:00 Blankenberge  51.3106    3.1164      427
#> 623 2021-09-17 22:00:00 Blankenberge  51.3106    3.1164      427
#> 624 2021-09-17 23:00:00 Blankenberge  51.3106    3.1164      411
#> 625 2021-09-18 00:00:00 Blankenberge  51.3106    3.1164      364
#> 626 2021-09-18 01:00:00 Blankenberge  51.3106    3.1164      294
#> 627 2021-09-18 02:00:00 Blankenberge  51.3106    3.1164      215
#> 628 2021-09-18 03:00:00 Blankenberge  51.3106    3.1164      136
#> 629 2021-09-18 04:00:00 Blankenberge  51.3106    3.1164       83
#> 630 2021-09-18 05:00:00 Blankenberge  51.3106    3.1164      112
#> 631 2021-09-18 06:00:00 Blankenberge  51.3106    3.1164      152
#> 632 2021-09-18 07:00:00 Blankenberge  51.3106    3.1164      204
#> 633 2021-09-18 08:00:00 Blankenberge  51.3106    3.1164      305
#> 634 2021-09-18 09:00:00 Blankenberge  51.3106    3.1164      401
#> 635 2021-09-18 10:00:00 Blankenberge  51.3106    3.1164      414
#> 636 2021-09-18 11:00:00 Blankenberge  51.3106    3.1164      408
#> 637 2021-09-18 12:00:00 Blankenberge  51.3106    3.1164      371
#> 638 2021-09-18 13:00:00 Blankenberge  51.3106    3.1164      316
#> 639 2021-09-18 14:00:00 Blankenberge  51.3106    3.1164      240
#> 640 2021-09-18 15:00:00 Blankenberge  51.3106    3.1164      156
#> 641 2021-09-18 16:00:00 Blankenberge  51.3106    3.1164       83
#> 642 2021-09-18 17:00:00 Blankenberge  51.3106    3.1164       60
#> 643 2021-09-18 18:00:00 Blankenberge  51.3106    3.1164      105
#> 644 2021-09-18 19:00:00 Blankenberge  51.3106    3.1164      164
#> 645 2021-09-18 20:00:00 Blankenberge  51.3106    3.1164      258
#> 646 2021-09-18 21:00:00 Blankenberge  51.3106    3.1164      400
#> 647 2021-09-18 22:00:00 Blankenberge  51.3106    3.1164      444
#> 648 2021-09-18 23:00:00 Blankenberge  51.3106    3.1164      446
#> 649 2021-09-19 00:00:00 Blankenberge  51.3106    3.1164      424
#> 650 2021-09-19 01:00:00 Blankenberge  51.3106    3.1164      373
#> 651 2021-09-19 02:00:00 Blankenberge  51.3106    3.1164      295
#> 652 2021-09-19 03:00:00 Blankenberge  51.3106    3.1164      208
#> 653 2021-09-19 04:00:00 Blankenberge  51.3106    3.1164      122
#> 654 2021-09-19 05:00:00 Blankenberge  51.3106    3.1164       72
#> 655 2021-09-19 06:00:00 Blankenberge  51.3106    3.1164      111
#> 656 2021-09-19 07:00:00 Blankenberge  51.3106    3.1164      150
#> 657 2021-09-19 08:00:00 Blankenberge  51.3106    3.1164      209
#> 658 2021-09-19 09:00:00 Blankenberge  51.3106    3.1164      332
#> 659 2021-09-19 10:00:00 Blankenberge  51.3106    3.1164      427
#> 660 2021-09-19 11:00:00 Blankenberge  51.3106    3.1164      432
#> 661 2021-09-19 12:00:00 Blankenberge  51.3106    3.1164      419
#> 662 2021-09-19 13:00:00 Blankenberge  51.3106    3.1164      375
#> 663 2021-09-19 14:00:00 Blankenberge  51.3106    3.1164      308
#> 664 2021-09-19 15:00:00 Blankenberge  51.3106    3.1164      221
#> 665 2021-09-19 16:00:00 Blankenberge  51.3106    3.1164      131
#> 666 2021-09-19 17:00:00 Blankenberge  51.3106    3.1164       60
#> 667 2021-09-19 18:00:00 Blankenberge  51.3106    3.1164       55
#> 668 2021-09-19 19:00:00 Blankenberge  51.3106    3.1164      106
#> 669 2021-09-19 20:00:00 Blankenberge  51.3106    3.1164      170
#> 670 2021-09-19 21:00:00 Blankenberge  51.3106    3.1164      291
#> 671 2021-09-19 22:00:00 Blankenberge  51.3106    3.1164      436
#> 672 2021-09-19 23:00:00 Blankenberge  51.3106    3.1164      463
#> 673 2021-09-20 00:00:00 Blankenberge  51.3106    3.1164      462
#> 674 2021-09-20 01:00:00 Blankenberge  51.3106    3.1164      430
#> 675 2021-09-20 02:00:00 Blankenberge  51.3106    3.1164      368
#> 676 2021-09-20 03:00:00 Blankenberge  51.3106    3.1164      280
#> 677 2021-09-20 04:00:00 Blankenberge  51.3106    3.1164      185
#> 678 2021-09-20 05:00:00 Blankenberge  51.3106    3.1164       98
#> 679 2021-09-20 06:00:00 Blankenberge  51.3106    3.1164       83
#> 680 2021-09-20 07:00:00 Blankenberge  51.3106    3.1164      126
#> 681 2021-09-20 08:00:00 Blankenberge  51.3106    3.1164      174
#> 682 2021-09-20 09:00:00 Blankenberge  51.3106    3.1164      237
#> 683 2021-09-20 10:00:00 Blankenberge  51.3106    3.1164      382
#> 684 2021-09-20 11:00:00 Blankenberge  51.3106    3.1164      449
#> 685 2021-09-20 12:00:00 Blankenberge  51.3106    3.1164      450
#> 686 2021-09-20 13:00:00 Blankenberge  51.3106    3.1164      421
#> 687 2021-09-20 14:00:00 Blankenberge  51.3106    3.1164      368
#> 688 2021-09-20 15:00:00 Blankenberge  51.3106    3.1164      287
#> 689 2021-09-20 16:00:00 Blankenberge  51.3106    3.1164      190
#> 690 2021-09-20 17:00:00 Blankenberge  51.3106    3.1164       97
#> 691 2021-09-20 18:00:00 Blankenberge  51.3106    3.1164       30
#> 692 2021-09-20 19:00:00 Blankenberge  51.3106    3.1164       60
#> 693 2021-09-20 20:00:00 Blankenberge  51.3106    3.1164      112
#> 694 2021-09-20 21:00:00 Blankenberge  51.3106    3.1164      181
#> 695 2021-09-20 22:00:00 Blankenberge  51.3106    3.1164      337
#> 696 2021-09-20 23:00:00 Blankenberge  51.3106    3.1164      458
#> 697 2021-09-21 00:00:00 Blankenberge  51.3106    3.1164      469
#> 698 2021-09-21 01:00:00 Blankenberge  51.3106    3.1164      459
#> 699 2021-09-21 02:00:00 Blankenberge  51.3106    3.1164      415
#> 700 2021-09-21 03:00:00 Blankenberge  51.3106    3.1164      340
#> 701 2021-09-21 04:00:00 Blankenberge  51.3106    3.1164      244
#> 702 2021-09-21 05:00:00 Blankenberge  51.3106    3.1164      144
#> 703 2021-09-21 06:00:00 Blankenberge  51.3106    3.1164       69
#> 704 2021-09-21 07:00:00 Blankenberge  51.3106    3.1164      107
#> 705 2021-09-21 08:00:00 Blankenberge  51.3106    3.1164      152
#> 706 2021-09-21 09:00:00 Blankenberge  51.3106    3.1164      200
#> 707 2021-09-21 10:00:00 Blankenberge  51.3106    3.1164      289
#> 708 2021-09-21 11:00:00 Blankenberge  51.3106    3.1164      443
#> 709 2021-09-21 12:00:00 Blankenberge  51.3106    3.1164      468
#> 710 2021-09-21 13:00:00 Blankenberge  51.3106    3.1164      462
#> 711 2021-09-21 14:00:00 Blankenberge  51.3106    3.1164      417
#> 712 2021-09-21 15:00:00 Blankenberge  51.3106    3.1164      350
#> 713 2021-09-21 16:00:00 Blankenberge  51.3106    3.1164      253
#> 714 2021-09-21 17:00:00 Blankenberge  51.3106    3.1164      145
#> 715 2021-09-21 18:00:00 Blankenberge  51.3106    3.1164       57
#> 716 2021-09-21 19:00:00 Blankenberge  51.3106    3.1164       10
#> 717 2021-09-21 20:00:00 Blankenberge  51.3106    3.1164       73
#> 718 2021-09-21 21:00:00 Blankenberge  51.3106    3.1164      126
#> 719 2021-09-21 22:00:00 Blankenberge  51.3106    3.1164      215
#> 720 2021-09-21 23:00:00 Blankenberge  51.3106    3.1164      405
#> 721 2021-09-22 00:00:00 Blankenberge  51.3106    3.1164      464
#> 722 2021-09-22 01:00:00 Blankenberge  51.3106    3.1164      465
#> 723 2021-09-22 02:00:00 Blankenberge  51.3106    3.1164      440
#> 724 2021-09-22 03:00:00 Blankenberge  51.3106    3.1164      380
#> 725 2021-09-22 04:00:00 Blankenberge  51.3106    3.1164      290
#> 726 2021-09-22 05:00:00 Blankenberge  51.3106    3.1164      193
#> 727 2021-09-22 06:00:00 Blankenberge  51.3106    3.1164       97
#> 728 2021-09-22 07:00:00 Blankenberge  51.3106    3.1164       72
#> 729 2021-09-22 08:00:00 Blankenberge  51.3106    3.1164      124
#> 730 2021-09-22 09:00:00 Blankenberge  51.3106    3.1164      168
#> 731 2021-09-22 10:00:00 Blankenberge  51.3106    3.1164      229
#> 732 2021-09-22 11:00:00 Blankenberge  51.3106    3.1164      363
#> 733 2021-09-22 12:00:00 Blankenberge  51.3106    3.1164      472
#> 734 2021-09-22 13:00:00 Blankenberge  51.3106    3.1164      474
#> 735 2021-09-22 14:00:00 Blankenberge  51.3106    3.1164      454
#> 736 2021-09-22 15:00:00 Blankenberge  51.3106    3.1164      399
#> 
#> $par
#> $par$daterange
#> [1] "2021-08-23" "2021-09-23"
#> 
#> $par$type
#> [1] "MVB data"
#> 
#> $par$stations
#> [1] "Blankenberge"
#> 
#> $par$calc
#> [1] "max"
#> 
#> $par$binSize
#> [1] "hour"
#> 
#> $par$parameters
#> [1] "Tide TAW"
#> 
#>