Retrieve an overview of the ETN data.

Retrieves the name of network/animals? To get an account, register via the Lifewatch RShiny registration webpage.

listEtnProjects(usr = NULL, pwd = NULL, ...)

Arguments

usr

Username to connect to ETN database

pwd

Password to connect to ETN database

...

Reserved for internal use.

Value

Dataframe with name and type of networks.

Examples

listEtnProjects()
#> - Query mode: Post request to OpenCPU server
#> No encoding supplied: defaulting to UTF-8.
#>                                                                                 name
#> 1                                                                  2011_Bovenschelde
#> 2                                                                         2011_Loire
#> 3                                                                         2013_Foyle
#> 4                                                                          2013_Maas
#> 5                                                                          2014_Nene
#> 6                                                           2015_PhD_Gutmann_Roberts
#> 7                                                               2016_Diaccia_Botrona
#> 8                                                                        2017_Fremur
#> 9                                                                    2020_PhD_Winter
#> 10                                                                     Aberdeen 2017
#> 11                                       Acousic receiver array for the MOVE project
#> 12                     Acoustic array in the Ria de Vigo (Galicia, Spain) since 2024
#> 13   Acoustic data between 2026 and 2028 for rheophilic fish in weir-regulated Meuse
#> 14                                      Acoustic Ecology of Pelagic Fish Communities
#> 15                             Acoustic Monitoring of Bluefin Tuna in the Aegean Sea
#> 16              Acoustic receiver array in the Gudena River, Denmark, from 2004-2006
#> 17                          Acoustic receiver array in the River Hevring (2023-2028)
#> 18              Acoustic receiver array in the Zuidelijke Randmeren, the Netherlands
#> 19    Acoustic telemetry array active since 2019 to track species in Bristol Channel
#> 20    Acoustic telemetry array data for elasmobranch in Northern Ireland (2024-2030)
#> 21   Acoustic telemetry array in the river Oude IJssel (Germany and the Netherlands)
#> 22                      Acoustic telemetry array on the east cost of Corsica, France
#> 23    Acoustic telemetry data for Atlantic cod at an oil and gas platform in Denmark
#> 24         Acoustic telemetry data for blue crab in the S'Ena Arrubia lagoon (Italy)
#> 25          Acoustic telemetry data for coastal fishes in Marintique Island (France)
#> 26    Acoustic telemetry data for European eel (Anguilla anguilla) in the River Test
#> 27       Acoustic telemetry data for Lionfish distribution in Gökova Bay, Aegean Sea
#> 28    Acoustic telemetry data from the Clew Bay and Achill region in west of Ireland
#> 29  Acoustic telemetry data of marine species monitored 6 French offshore wind farms
#> 30                                                   Acoustic telemetry in lake Osby
#> 31           Acoustic telemetry network in the Ecowende wind farm in the Netherlands
#> 32              Acoustic telemetry network in the harbours of IJmuiden and Rotterdam
#> 33     Acoustic telemetry network in the non-tidal River Scheldt and its tributaries
#> 34   Acoustic telemetry network in the North Sea and the Eastern Scheldt (2025-2028)
#> 35   Acoustic telemetry network in the western River Rhine and hinterlands 2025–2026
#> 36      Acoustic telemetry on Atlantic cod in the Horns Rev 2 offshore windmill farm
#> 37   Acoustic telemetry receiver array in Galway Bay and Dunkellin River (2015-2019)
#> 38                         Acoustic telemetry receiver array in Murcia Region, Spain
#> 39            Acoustic telemetry receiver array in the harinxma channel in Harlingen
#> 40               Acoustic telemetry receiver array in the Nederrijn-Lek, Netherlands
#> 41                 Acoustic telemetry receiver array in the North Sea from 2017-2018
#> 42                  Acoustic telemetry receiver array in the Orbetello lagoon, Italy
#> 43    Acoustic telemetry receiver array in the River Warnow (Germany) from 2011-2012
#> 44         Acoustic telemetry receiver array in the Waal river between 2025 and 2027
#> 45               Acoustic telemetry study in the Meuse river basin at Lith and Linne
#> 46                                                                      Albertkanaal
#> 47                                                              Albertkanaal_VPS_Ham
#> 48                                                          Albertkanaal_VPS_Hasselt
#> 49           An acoustic receiver network operating at 69 kHz in and around Blekinge
#> 50                                                                        ARAISOLA02
#> 51                                                                        ARAISOLA03
#> 52                                                                               Ars
#> 53                                                                          Artevigo
#> 54                                                 Azorean acoustic receiver network
#> 55         Balearic strategy for climate change monitoring and prediction in the Med
#> 56                                                         Balearic Tracking Network
#> 57                   Balearic Tracking Network - Spanish Deep Water Receiver Network
#> 58                                                       Baltic Sea Tracking Network
#> 59                                                            Basking_shark_Hebrides
#> 60                                                                            BECORV
#> 61    Behaviour of Cancer pagurus in proximity to a fish farm in Storfjorden, Norway
#> 62                           Belwind temporary acoustic receiver network 2020 - 2021
#> 63                                           Bergen's Ocean Acoustic Tracking System
#> 64  Blue crab telemetry monitoring in three lagoons, channels and coast of Occitanie
#> 65                                                                       Boddenhecht
#> 66                                                                              BPNS
#> 67                                                           Brasem Ijsselmeergebied
#> 68                                               Brown trout in the Falkland Islands
#> 69                                                                    BTN_LOST_ARRAY
#> 70                                                      Canaries Angel Shark Project
#> 71                                                          Catalan Tracking Network
#> 72                                                   CMRC European Sea Bass Tracking
#> 73                                                            CODEVCO fish detectors
#> 74                          Connectivity of coastal fish in the French Mediterranean
#> 75                                                                        Conon 2017
#> 76                                                                       cpodnetwork
#> 77                                                                     DAbecasis PhD
#> 78                                                                     DAK 2018-2019
#> 79                                                          Dalalven Smolt Migration
#> 80                                                        Danish_southeastern_strait
#> 81                                            Danish Straits Acoustic Infrastructure
#> 82                                                                  Danube_Sturgeons
#> 83                                                                Delfzijl 2014-2015
#> 84                                                                             Demer
#> 85    Deployment of acoustic receivers in Carlingford Lough for the NorTrack project
#> 86                             Deployment of acoustic receivers in inner Lough Foyle
#> 87                             Deployment of acoustic receivers in outer Lough Foyle
#> 88             Deployment of acoustic receivers in rivers within the Foyle Catchment
#> 89      Deployment of an acoustic array in the North Channel for the STRAITS project
#> 90                                                                      Deveron 2016
#> 91                                                                             Dijle
#> 92                                                                         Dijle_VPS
#> 93                                                                        DTU-Skjern
#> 94    Dusky grouper´s resettlement - Acoustic telemetry in Maro-Cerro Gordo´s cliffs
#> 95    Duurzame economische dragers in Vlaamse kustvisserijgemeenschappen post-Brexit
#> 96                                                  Eel migration from source to sea
#> 97                                                                      Eemskanaal I
#> 98                                                                     Eemskanaal II
#> 99                                                                    Eemskanaal III
#> 100                                                                             EMMN
#> 101                                                                             ESGL
#> 102                                                      ETN_network_project_group_1
#> 103                                                      ETN_network_project_group_2
#> 104                                                      ETN_network_project_group_3
#> 105                                                      ETN_network_project_group_4
#> 106                                                      ETN_network_project_group_5
#> 107                                                      ETN_network_project_group_6
#> 108                                                      ETN_network_project_group_7
#> 109                                                      ETN_network_project_group_8
#> 110                                                      ETN_network_project_group_9
#> 111                                                          Euskal Tracking Network
#> 112   Evaluation of the effectiveness of the Balearic Islands Marine Reserve Network
#> 113                                                                FarmTrack_Spain22
#> 114     Fine scale acoustic telemetry array at Hinkley Point C water intakes in 2025
#> 115     Fine scale acoustic telemetry array at Hinkley Point C water intakes in 2026
#> 116  Fine-scale multispecies acoustic telemetry positioning system data at Helgoland
#> 117                                           Fisheries Industry Science Partnership
#> 118               Fisheries Innovation for sustainable SHared INTerchannEL resources
#> 119                                          Fish migration from source to Waddensea
#> 120                                                             FISHOWF_Array_France
#> 121                                                              Friesland 2011-2012
#> 122                                    Galician (NW Spain) acoustic Tracking Network
#> 123                                                        GIBRALTAR_STRAITS_CURTAIN
#> 124                                                                   Gudena_network
#> 125                              Haringvlietdam - Rhine-Meuse Delta Acoustic Network
#> 126    High-frequency acoustic telemetry array in the northwestern Mediterranean sea
#> 127  HR2 tech and 3D tracking of European eel in landlocked drinking water reservoir
#> 128 ICOD: array to monitor interaction between cod spawning grounds and salmon farms
#> 129                                                                  IG II Fish Pass
#> 130                                                                        Iller_VPS
#> 131                                        Immature bass acoustic stock surveillance
#> 132                                                                          IMR_OWF
#> 133                                                                    Inforbiomares
#> 134                                                       Infragravity waves project
#> 135                                                                  INTERREG_CANAPE
#> 136         Intra-lagoon ecological connectivity: migration surveys of European eels
#> 137                                                     Jersey Coastal Fish Tracking
#> 138                                                              JSATS-PalmaBay-2019
#> 139                                                   Kalmar Baltic Tracking Network
#> 140                                                           Kattegat Islands Array
#> 141                                                      Kiel Bight Acoustic Network
#> 142                                                                   Kornwerderzand
#> 143                                                                Lake_Anundsjo_VPS
#> 144                       Land-use change and human activity on anadromous salmonids
#> 145                                                                      Lauwersmeer
#> 146                                                                    Leopoldkanaal
#> 147        Le Tréport harbour receiver array in the Bresle estuary, northwest France
#> 148                                                                        Lifewatch
#> 149                                                   Lifewatch: additional stations
#> 150                                             Lifewatch: Porpoise Performance test
#> 151                                                                         Limfjord
#> 152                                             Ljusne Bay Salmon Spawning Migration
#> 153 Long-term array MOnitoring BEhaviour in Illas Atlánticas National Park, NW Spain
#> 154                                                                       LSTSTJ_MPA
#> 155                                                                          MacFish
#> 156       Mapping critical habitats of threatened shark species in the Algarve coast
#> 157                                                          Marble vs rainbow trout
#> 158                                                            MariagerFjord_network
#> 159        Marine Ecological connectivity and conseRvation in the MOZambique Channel
#> 160                                                                       MBA_Massmo
#> 161                                                                      MBA_Wavehub
#> 162                                                                     MBA_Whitsand
#> 163                                                                           MEPNSW
#> 164                                                          Middle_Danube_sturgeons
#> 165   Migration of pikeperch within the Lake Peipus complex (Estonia) from 2021-2023
#> 166                                                                      MIGRATOEBRE
#> 167                                                            MMERMAID_Array_France
#> 168                                                                      Mobula Imar
#> 169                             Monitoring Gibraltar Strait Using Acoustic Telemetry
#> 170                            Monitoring Gibraltral Strait Using Acoustic Telemetry
#> 171                                                                 Moray Firth 2016
#> 172      Movement and habitat connectivity of Atlantic cod in a coastal fjord system
#> 173     Movement, Range Size and Habitat Utilization of magafish in Turkey 2019-2021
#> 174   Movements of cod, plaice, flounder and turbot in the western German Baltic Sea
#> 175   MPAs' effect on Balearic chondrichthyes' biodiversity using acoustic telemetry
#> 176                                                                         mrc_vliz
#> 177                                                                         NARVAEEL
#> 178                                                          Northern_Norfolk_Broads
#> 179                                                                 North_sea_wrecks
#> 180                                                                    NTNU-Gaulosen
#> 181                                                       NTNU River Nidelva Project
#> 182   Observing the effect of group dynamics on individual behavior in basking shark
#> 183                                                Open Protocols scientific testing
#> 184                                                                       Orsted Cod
#> 185                                                                  OTN-Hemnfjorden
#> 186                                                          OTN Loan - Beiarfjorden
#> 187                                                             OTN-Skjerstadfjorden
#> 188                                                     OTN Strait of Gibraltar Line
#> 189                                                                 OTN-Tosenfjorden
#> 190                                                           OTN VR2W Loan - Norway
#> 191                                                                    paintedcomber
#> 192                                         PAM Harbour porpoise Borssele, 2019-2024
#> 193                                        Passive Acoustic Monitoring (PAM) STRAITS
#> 194                                                                             PC4C
#> 195             Permanent acoustic receiver array in the River Grote Nete since 2019
#> 196                                                                PhD Barbara Koeck
#> 197                                                                PhD Jeremy Pastor
#> 198                                                                        PhD_Nolan
#> 199                                                                         PhysFish
#> 200                       Pilot permanent deployment acoustic array Gibraltar Strait
#> 201                                   Porbeagle Tracking Network in Brittany, France
#> 202                          Predators and prey around renewable energy developments
#> 203                                                                        PTN-MARSW
#> 204                                                                  PTN/PROTECT2012
#> 205                                                                  PTN/PROTECT2013
#> 206                                                           PTN-Silver eel-Mondego
#> 207                                                          Purewind fish detectors
#> 208                                                                     RECCRU_array
#> 209              Receiver network in Killary Harbour and River Erriff from 2014-2024
#> 210  Redes de Sensores submarinos autonomos para la medida de indicadores biologicos
#> 211                                                                         Reelease
#> 212                 Research on the impact of management actions on beaver behaviour
#> 213                                                                           ResMed
#> 214   Restoration of Baltic sturgeon stock in Oder river & in Mecklenburg-Vorpommern
#> 215                                                River_lamprey_River Trent_2020/21
#> 216                                             River_Lamprey_Yorkshire Ouse_2018-20
#> 217                                             River Usk receiver network 2021-2024
#> 218                                                                rt2020_zeeschelde
#> 219                                                             Ruim_baan_voor_vis_2
#> 220                                                                      Running_eel
#> 221                                                                       Saeftinghe
#> 222                                               Salmon_migration_eastern_Greenland
#> 223                                                                          SAMARCH
#> 224                                                                       SARTELARAM
#> 225                                                                         SARTELTG
#> 226                                                                    SARTELZINGARO
#> 227                                                                           SCHUFI
#> 228                                                                         Scytrack
#> 229                                                                       SeaMonitor
#> 230           SeaMonitor - Costs and Pathways of Marine Migration in Atlantic salmon
#> 231                                                        SeaMonitor - GMIT Project
#> 232                                                        sea trout_2018-9_gullmarn
#> 233 SEAWave: small cetacean presence at the EMEC Orkney Billia Croo wave energy site
#> 234           SIARC network of scientific moorings in Pen Llŷn a’r Sarnau SAC, Wales
#> 235                                                                 Siganid_East_Med
#> 236                                                               Siganid_Gulf_Aqaba
#> 237              Silver Eel downstream migration in the Walloon Meuse - 2021 to 2024
#> 238       Silver European Eel escapement success and migration patterns in River Ems
#> 239                                                         Skagerrak-NorthSea-array
#> 240                                                 Skagerrak_temeletry_array_Norway
#> 241                                                                        Skye 2017
#> 242                                                             SMOLTRACK-I-Engeland
#> 243                                                            SMOLTRACK-II-Engeland
#> 244         Spatial Ecology of spurdog (Squalus acanthias) in a Scottish loch system
#> 245 Spawning migration of pike in MatsaluBay and RiverKasari (Baltic Sea, 2018-2020)
#> 246  Spawning migration of pike in Teorehe-Sauemeri floodplain (BalticSea 2017-2019)
#> 247                                                                        SPAWNSEIS
#> 248                                                                 SPIDER_GNB_array
#> 249                                                                S_rivulatus_Aqaba
#> 250                                                S_rivulatus_Eastern_Mediterranean
#> 251                                                                      STR_MOV_ANG
#> 252                                                     Sturgeon_reintroduction_Ebre
#> 253                                                  Sud_le_Impuls_voor_Diadrome_Vis
#> 254                                                    Sud_le_Innovatieve_monitoring
#> 255                                          survival_movement_elasmobranch_Adriatic
#> 256                                                                  SVNL-FISH-WATCH
#> 257                                                            Swedish eel migration
#> 258                                                                     SWIMWAY_2021
#> 259                                                           Swimway Rijnland Array
#> 260                                                                   Swimway Vechte
#> 261                                                 Swiss Acoustic Telemetry Network
#> 262   Tagging and tracking butterfly ray with acoustic telemetry in the Balearic Sea
#> 263                                  Temporary acoustic network in the Sines harbour
#> 264    Temporary acoustic network in the SW Alentejo and Vicentine Coast Marine Park
#> 265                                  Temporary acoustic receiver network in Bråviken
#> 266                     Temporary receiver array in Lithuanian rivers from 2014-2015
#> 267                                                                             TeRi
#> 268                                                                       Test VR2AR
#> 269                                                     The Bergen Telemetry Network
#> 270         The National Programme for Monitoring the Aquatic Environment and Nature
#> 271                                                                     Thorntonbank
#> 272                                                                  Trent¨_Tracking
#> 273                                                   Turkish Straits Detection Data
#> 274                         Turkish Telemetry Array in the eastern Mediterranean sea
#> 275       Understanding the movement of dolphinfish in the Mediterranean FAD fishery
#> 276                      Understand movement and life history of trout: Isle of Skye
#> 277    Underwater acoustic data from the Belgian ARMS-MBON hard-substrate monitoring
#> 278                                                             Unlocking the Severn
#> 279                  Unravelling the migrations of the invasive blue crab in Algarve
#> 280                                                                 VFM_BalticProper
#> 281                                                                 VFM_GulfOfBotnia
#> 282                                                                    VFM_Hjalmaren
#> 283                                                                      VFM_Malaren
#> 284                                                                    VFM_RiverGota
#> 285                                                                       VFM_Siljan
#> 286                                                                      VFM_Vattern
#> 287                                                              VFM_WestCoastCostal
#> 288                                                           Vissen_voor_verbinding
#> 289                                       West Coast Tracking Project Receiver Array
#> 290                                                                  Westerschelde 1
#> 291                                                                  Westerschelde 2
#> 292                                                                  Westerschelde 3
#> 293                                                       Winde Tjeukemeer 2023-2026
#> 294                                      Yellow eel movement behaviour in the Azores
#> 295                                                                       Zeeschelde
#> 296                                                                 2010_phd_reubens
#> 297                                                            2010_phd_reubens_sync
#> 298                                                                       2011_Loire
#> 299                                                                  2011 Rivierprik
#> 300                                                               2012 Leopoldkanaal
#> 301                                                                2013 Albertkanaal
#> 302                                                                       2014 Demer
#> 303                                                                       2014_Frome
#> 304                                                                        2014_Nene
#> 305                                                        2015_Albertkanaal_VPS_Ham
#> 306                                                                       2015 Dijle
#> 307                                                                     2015_Homarus
#> 308                                                            2015 PhD Verhelst_cod
#> 309                                                            2015 PhD Verhelst_eel
#> 310                                                        2016_Albertkanaal_VPS_Ham
#> 311                                                               2016 PhD Vergeynst
#> 312                                                        2017_Albertkanaal_VPS_Ham
#> 313                                                                   2019_Grotenete
#> 314                                                                   2021_Dijle_VPS
#> 315                                                         2021_Sturgeons_Danube_BG
#> 316                                                    2022_Albertkanaal_VPS_Hasselt
#> 317   Acoustic telemetry data for Atlantic cod at an oil and gas platform in Denmark
#> 318        Acoustic telemetry data for blue crab in the S'Ena Arrubia lagoon (Italy)
#> 319         Acoustic telemetry data for coastal fishes in Marintique Island (France)
#> 320    Acoustic telemetry data for Dicentrarchus labrax in the Gulf of Lion (France)
#> 321 Acoustic telemetry data for elasmobranch species in Northern Ireland (2024-2030)
#> 322         Acoustic telemetry data for European seabass tagged in 2026 in the BPNS.
#> 323    Acoustic telemetry data for garfish (Belone belone), Western Baltic (Denmark)
#> 324  Acoustic telemetry data for porbeagle sharks (Lamna nasus) tagged across Europe
#> 325          Acoustic telemetry data for Spicara smaris on the east coast of Corsica
#> 326 Acoustic telemetry data of Atlantic salmon and sea trout in Galway and Dunkellin
#> 327              Acoustic telemetry data of Carp in lake Markermeer, the Netherlands
#> 328    Acoustic telemetry for Atlantic mackerel in the German Bight of the North Sea
#> 329 Acoustic Telemetry on blue sharks (Prionace glauca) in the Basque country, Spain
#> 330                                                                           AFICHE
#> 331                                                                       ARAISOLA02
#> 332                                                                       ARAISOLA03
#> 333                                                                              Ars
#> 334                                                                         Artevigo
#> 335        Balearic strategy for climate change monitoring and prediction in the Med
#> 336                                                        Balearic Tracking Network
#> 337                  Balearic Tracking Network - Spanish Deep Water Receiver Network
#> 338                                                                         CodNoise
#> 339      Comparative study of marble and rainbow trout behaviour in Kanomljica River
#> 340 Conventional and electronic telemetry data for starry smooth-hound in NEAtlantic
#> 341                                                                    DAbecasis PhD
#> 342                                                                    DAK 2018-2019
#> 343                                                                 Danube_Sturgeons
#> 344                                                         Delfzijl Drie Delfzijlen
#> 345   Dusky grouper´s resettlement - Acoustic telemetry in Maro-Cerro Gordo´s cliffs
#> 346                                                                             EMMN
#> 347                                                                             ESGL
#> 348                                                       ETN_animal_project_group_1
#> 349                                                       ETN_animal_project_group_2
#> 350                                                       ETN_animal_project_group_3
#> 351                                                       ETN_animal_project_group_4
#> 352                                                       ETN_animal_project_group_5
#> 353                                                       ETN_animal_project_group_6
#> 354                                                       ETN_animal_project_group_7
#> 355                                                       ETN_animal_project_group_8
#> 356                                                       ETN_animal_project_group_9
#> 357   Evaluation of the effectiveness of the Balearic Islands Marine Reserve Network
#> 358  High-frequency data for rocky reef fishes in the northwestern Mediterranean sea
#> 359                                                                        Hunze_eel
#> 360                                                              JSATS-PalmaBay-2019
#> 361                                                                         Limfjord
#> 362                                                          Marble vs rainbow trout
#> 363                                                                       MBA_Massmo
#> 364                                                                      MBA_Wavehub
#> 365                                                                     MBA_Whitsand
#> 366                                                          Middle_Danube_sturgeons
#> 367          Migration of common bream in the Zuidelijke Randmeren, the Netherlands.
#> 368                                                                      Mobula Imar
#> 369      Movement and habitat connectivity of Atlantic cod in a coastal fjord system
#> 370   MPAs' effect on Balearic chondrichthyes' biodiversity using acoustic telemetry
#> 371  Multispecies acoustic telemetry data from fine-scale positioning system MarGate
#> 372                                                         Noordzeekanaal 2017-2018
#> 373                                                   North Aegean Anguilla Tracking
#> 374                                                                    NTNU-Gaulosen
#> 375                                                       NTNU River Nidelva Project
#> 376                                                Open Protocols scientific testing
#> 377                                                                       Orsted Cod
#> 378                                                                  OTN-Hemnfjorden
#> 379                                                                 OTN-Tosenfjorden
#> 380                                                           OTN VR2W Loan - Norway
#> 381                                                                PhD Barbara Koeck
#> 382                                                                     PhD_Goossens
#> 383                                                                PhD Jeremy Pastor
#> 384                                                                         PhysFish
#> 385                                                           PTN-Silver eel-Mondego
#> 386                                                                         Reelease
#> 387                                                                   Rijke_Noordzee
#> 388                                                                       SARTELARAM
#> 389                                                                         SARTELTG
#> 390                                                                    SARTELZINGARO
#> 391                                                                         Scytrack
#> 392                                                        sea trout_2018-9_gullmarn
#> 393                                                                 Siganid_East_Med
#> 394                                                               Siganid_Gulf_Aqaba
#> 395                                Silver eel migration in the Gudena River, Denmark
#> 396                                     Silver eel migration in the Warnow catchment
#> 397          Silver eel migration patterns in lowland rivers and lagoon in Lithuania
#> 398       Silver European Eel escapement success and migration patterns in River Ems
#> 399         Spatial Ecology of spurdog (Squalus acanthias) in a Scottish loch system
#> 400                                                S_rivulatus_Eastern_Mediterranean
#> 401                                                     Sturgeon_reintroduction_Ebre
#> 402   Tagging and tracking butterfly ray with acoustic telemetry in the Balearic Sea
#> 403                                                                             TeRi
#> 404                      Tracking anadromous brown trout at sea by data storage tags
#> 405  Tracking movements of 15 Dusky groupers released in the harbor of Sines in 2021
#> 406                          Tracking twaite shads at sea by using data storage tags
#> 407                 Trap and transport of silver eels from a reservoir to an estuary
#> 408       Understanding the movement of dolphinfish in the Mediterranean FAD fishery
#> 409                                                                    VPS Duurswold
#> 410                                                                  VPS Garmerwolde
#> 411                                      Yellow eel movement behaviour in the Azores
#>        type
#> 1   network
#> 2   network
#> 3   network
#> 4   network
#> 5   network
#> 6   network
#> 7   network
#> 8   network
#> 9   network
#> 10  network
#> 11  network
#> 12  network
#> 13  network
#> 14  network
#> 15  network
#> 16  network
#> 17  network
#> 18  network
#> 19  network
#> 20  network
#> 21  network
#> 22  network
#> 23  network
#> 24  network
#> 25  network
#> 26  network
#> 27  network
#> 28  network
#> 29  network
#> 30  network
#> 31  network
#> 32  network
#> 33  network
#> 34  network
#> 35  network
#> 36  network
#> 37  network
#> 38  network
#> 39  network
#> 40  network
#> 41  network
#> 42  network
#> 43  network
#> 44  network
#> 45  network
#> 46  network
#> 47  network
#> 48  network
#> 49  network
#> 50  network
#> 51  network
#> 52  network
#> 53  network
#> 54  network
#> 55  network
#> 56  network
#> 57  network
#> 58  network
#> 59  network
#> 60  network
#> 61  network
#> 62  network
#> 63  network
#> 64  network
#> 65  network
#> 66  network
#> 67  network
#> 68  network
#> 69  network
#> 70  network
#> 71  network
#> 72  network
#> 73  network
#> 74  network
#> 75  network
#> 76  network
#> 77  network
#> 78  network
#> 79  network
#> 80  network
#> 81  network
#> 82  network
#> 83  network
#> 84  network
#> 85  network
#> 86  network
#> 87  network
#> 88  network
#> 89  network
#> 90  network
#> 91  network
#> 92  network
#> 93  network
#> 94  network
#> 95  network
#> 96  network
#> 97  network
#> 98  network
#> 99  network
#> 100 network
#> 101 network
#> 102 network
#> 103 network
#> 104 network
#> 105 network
#> 106 network
#> 107 network
#> 108 network
#> 109 network
#> 110 network
#> 111 network
#> 112 network
#> 113 network
#> 114 network
#> 115 network
#> 116 network
#> 117 network
#> 118 network
#> 119 network
#> 120 network
#> 121 network
#> 122 network
#> 123 network
#> 124 network
#> 125 network
#> 126 network
#> 127 network
#> 128 network
#> 129 network
#> 130 network
#> 131 network
#> 132 network
#> 133 network
#> 134 network
#> 135 network
#> 136 network
#> 137 network
#> 138 network
#> 139 network
#> 140 network
#> 141 network
#> 142 network
#> 143 network
#> 144 network
#> 145 network
#> 146 network
#> 147 network
#> 148 network
#> 149 network
#> 150 network
#> 151 network
#> 152 network
#> 153 network
#> 154 network
#> 155 network
#> 156 network
#> 157 network
#> 158 network
#> 159 network
#> 160 network
#> 161 network
#> 162 network
#> 163 network
#> 164 network
#> 165 network
#> 166 network
#> 167 network
#> 168 network
#> 169 network
#> 170 network
#> 171 network
#> 172 network
#> 173 network
#> 174 network
#> 175 network
#> 176 network
#> 177 network
#> 178 network
#> 179 network
#> 180 network
#> 181 network
#> 182 network
#> 183 network
#> 184 network
#> 185 network
#> 186 network
#> 187 network
#> 188 network
#> 189 network
#> 190 network
#> 191 network
#> 192 network
#> 193 network
#> 194 network
#> 195 network
#> 196 network
#> 197 network
#> 198 network
#> 199 network
#> 200 network
#> 201 network
#> 202 network
#> 203 network
#> 204 network
#> 205 network
#> 206 network
#> 207 network
#> 208 network
#> 209 network
#> 210 network
#> 211 network
#> 212 network
#> 213 network
#> 214 network
#> 215 network
#> 216 network
#> 217 network
#> 218 network
#> 219 network
#> 220 network
#> 221 network
#> 222 network
#> 223 network
#> 224 network
#> 225 network
#> 226 network
#> 227 network
#> 228 network
#> 229 network
#> 230 network
#> 231 network
#> 232 network
#> 233 network
#> 234 network
#> 235 network
#> 236 network
#> 237 network
#> 238 network
#> 239 network
#> 240 network
#> 241 network
#> 242 network
#> 243 network
#> 244 network
#> 245 network
#> 246 network
#> 247 network
#> 248 network
#> 249 network
#> 250 network
#> 251 network
#> 252 network
#> 253 network
#> 254 network
#> 255 network
#> 256 network
#> 257 network
#> 258 network
#> 259 network
#> 260 network
#> 261 network
#> 262 network
#> 263 network
#> 264 network
#> 265 network
#> 266 network
#> 267 network
#> 268 network
#> 269 network
#> 270 network
#> 271 network
#> 272 network
#> 273 network
#> 274 network
#> 275 network
#> 276 network
#> 277 network
#> 278 network
#> 279 network
#> 280 network
#> 281 network
#> 282 network
#> 283 network
#> 284 network
#> 285 network
#> 286 network
#> 287 network
#> 288 network
#> 289 network
#> 290 network
#> 291 network
#> 292 network
#> 293 network
#> 294 network
#> 295 network
#> 296  animal
#> 297  animal
#> 298  animal
#> 299  animal
#> 300  animal
#> 301  animal
#> 302  animal
#> 303  animal
#> 304  animal
#> 305  animal
#> 306  animal
#> 307  animal
#> 308  animal
#> 309  animal
#> 310  animal
#> 311  animal
#> 312  animal
#> 313  animal
#> 314  animal
#> 315  animal
#> 316  animal
#> 317  animal
#> 318  animal
#> 319  animal
#> 320  animal
#> 321  animal
#> 322  animal
#> 323  animal
#> 324  animal
#> 325  animal
#> 326  animal
#> 327  animal
#> 328  animal
#> 329  animal
#> 330  animal
#> 331  animal
#> 332  animal
#> 333  animal
#> 334  animal
#> 335  animal
#> 336  animal
#> 337  animal
#> 338  animal
#> 339  animal
#> 340  animal
#> 341  animal
#> 342  animal
#> 343  animal
#> 344  animal
#> 345  animal
#> 346  animal
#> 347  animal
#> 348  animal
#> 349  animal
#> 350  animal
#> 351  animal
#> 352  animal
#> 353  animal
#> 354  animal
#> 355  animal
#> 356  animal
#> 357  animal
#> 358  animal
#> 359  animal
#> 360  animal
#> 361  animal
#> 362  animal
#> 363  animal
#> 364  animal
#> 365  animal
#> 366  animal
#> 367  animal
#> 368  animal
#> 369  animal
#> 370  animal
#> 371  animal
#> 372  animal
#> 373  animal
#> 374  animal
#> 375  animal
#> 376  animal
#> 377  animal
#> 378  animal
#> 379  animal
#> 380  animal
#> 381  animal
#> 382  animal
#> 383  animal
#> 384  animal
#> 385  animal
#> 386  animal
#> 387  animal
#> 388  animal
#> 389  animal
#> 390  animal
#> 391  animal
#> 392  animal
#> 393  animal
#> 394  animal
#> 395  animal
#> 396  animal
#> 397  animal
#> 398  animal
#> 399  animal
#> 400  animal
#> 401  animal
#> 402  animal
#> 403  animal
#> 404  animal
#> 405  animal
#> 406  animal
#> 407  animal
#> 408  animal
#> 409  animal
#> 410  animal
#> 411  animal