Retrieve an overview of the ETN data.

Retrieves the name of network/animals? To get an account, register via the Lifewatch RShiny registration webpage.

listEtnProjects(usr = NULL, pwd = NULL, ...)

Arguments

usr

Username to connect to ETN database

pwd

Password to connect to ETN database

...

Reserved for internal use.

Value

Dataframe with name and type of networks.

Examples

listEtnProjects()
#> - Query mode: Post request to OpenCPU server
#> No encoding supplied: defaulting to UTF-8.
#>                                                                                 name
#> 1                                                                        2004_Gudena
#> 2                                                                  2011_Bovenschelde
#> 3                                                                         2011_Loire
#> 4                                                                        2011_Warnow
#> 5                                                                         2013_Foyle
#> 6                                                                          2013_Maas
#> 7                                                                          2014_Nene
#> 8                                                           2015_PhD_Gutmann_Roberts
#> 9                                                               2016_Diaccia_Botrona
#> 10                                                                       2017_Fremur
#> 11                                                                    2019_Grotenete
#> 12                                                                   2020_PhD_Winter
#> 13                                                                     Aberdeen 2017
#> 14                                      Acoustic Ecology of Pelagic Fish Communities
#> 15    Acoustic telemetry data for Atlantic cod at an oil and gas platform in Denmark
#> 16    Acoustic telemetry data from the Clew Bay and Achill region in west of Ireland
#> 17     Acoustic telemetry network in the non-tidal River Scheldt and its tributaries
#> 18      Acoustic telemetry on Atlantic cod in the Horns Rev 2 offshore windmill farm
#> 19               Acoustic telemetry study in the Meuse river basin at Lith and Linne
#> 20                                                                      Albertkanaal
#> 21                                                              Albertkanaal_VPS_Ham
#> 22                                                          Albertkanaal_VPS_Hasselt
#> 23                                                                        ARAISOLA02
#> 24                                                                        ARAISOLA03
#> 25                                                                               Ars
#> 26                                                                          Artevigo
#> 27                                                 Azorean acoustic receiver network
#> 28                                                         Balearic Tracking Network
#> 29                   Balearic Tracking Network - Spanish Deep Water Receiver Network
#> 30                                                            Basking_shark_Hebrides
#> 31                                                                            BECORV
#> 32    Behaviour of Cancer pagurus in proximity to a fish farm in Storfjorden, Norway
#> 33                           Belwind temporary acoustic receiver network 2020 - 2021
#> 34  Blue crab telemetry monitoring in three lagoons, channels and coast of Occitanie
#> 35                                                                       Boddenhecht
#> 36                                                                              BPNS
#> 37                                                           Brasem Ijsselmeergebied
#> 38                                               Brown trout in the Falkland Islands
#> 39                                                      Canaries Angel Shark Project
#> 40                                                          Catalan Tracking Network
#> 41                                                            CODEVCO fish detectors
#> 42                          Connectivity of coastal fish in the French Mediterranean
#> 43                                                                        Conon 2017
#> 44                                                                       cpodnetwork
#> 45                                                                     DAbecasis PhD
#> 46                                                                     DAK 2018-2019
#> 47                                                          Dalalven Smolt Migration
#> 48                                                        Danish_southeastern_strait
#> 49                                            Danish Straits Acoustic Infrastructure
#> 50                                                                  Danube_Sturgeons
#> 51                                                                Delfzijl 2014-2015
#> 52                                                                             Demer
#> 53                                                                      Deveron 2016
#> 54                                                                             Dijle
#> 55                                                                         Dijle_VPS
#> 56                                                                        DTU-Skjern
#> 57    Dusky grouper´s resettlement - Acoustic telemetry in Maro-Cerro Gordo´s cliffs
#> 58    Duurzame economische dragers in Vlaamse kustvisserijgemeenschappen post-Brexit
#> 59                                                  Eel migration from source to sea
#> 60                                                                      Eemskanaal I
#> 61                                                                     Eemskanaal II
#> 62                                                                    Eemskanaal III
#> 63                                                                              EMMN
#> 64                                                                              ESGL
#> 65                                                       ETN_network_project_group_1
#> 66                                                       ETN_network_project_group_2
#> 67                                                       ETN_network_project_group_3
#> 68    Evaluation of the effectiveness of the Balearic Islands Marine Reserve Network
#> 69                                                                 FarmTrack_Spain22
#> 70                                            Fisheries Industry Science Partnership
#> 71                Fisheries Innovation for sustainable SHared INTerchannEL resources
#> 72                                           Fish migration from source to Waddensea
#> 73                                                              FISHOWF_Array_France
#> 74                                                               Friesland 2011-2012
#> 75                                     Galician (NW Spain) acoustic Tracking Network
#> 76                                                         GIBRALTAR_STRAITS_CURTAIN
#> 77                                                                    Gudena_network
#> 78   HR2 tech and 3D tracking of European eel in landlocked drinking water reservoir
#> 79                                                                   IG II Fish Pass
#> 80                                                                         Iller_VPS
#> 81                                         Immature bass acoustic stock surveillance
#> 82                                                                     Inforbiomares
#> 83                                                        Infragravity waves project
#> 84                                                                   INTERREG_CANAPE
#> 85          Intra-lagoon ecological connectivity: migration surveys of European eels
#> 86                                                      Jersey Coastal Fish Tracking
#> 87                                                               JSATS-PalmaBay-2019
#> 88                                                    Kalmar Baltic Tracking Network
#> 89                                                                    Kornwerderzand
#> 90                                                                 Lake_Anundsjo_VPS
#> 91                        Land-use change and human activity on anadromous salmonids
#> 92                                              Large_Migrants_Haringvliet_2023-2025
#> 93                                                                       Lauwersmeer
#> 94                                                                     Leopoldkanaal
#> 95                                                                         Lifewatch
#> 96                                                                          Limfjord
#> 97                                              Ljusne Bay Salmon Spawning Migration
#> 98  Long-term array MOnitoring BEhaviour in Illas Atlánticas National Park, NW Spain
#> 99                                                                           MacFish
#> 100                                                          Marble vs rainbow trout
#> 101                                                            MariagerFjord_network
#> 102                                                                       MBA_Massmo
#> 103                                                                      MBA_Wavehub
#> 104                                                                     MBA_Whitsand
#> 105                                                                           MEPNSW
#> 106                                                          Middle_Danube_sturgeons
#> 107   Migration of pikeperch within the Lake Peipus complex (Estonia) from 2021-2023
#> 108                                                                      MIGRATOEBRE
#> 109                                                            MMERMAID_Array_France
#> 110                                                                      Mobula Imar
#> 111                            Monitoring Gibraltral Strait Using Acoustic Telemetry
#> 112                                                                 Moray Firth 2016
#> 113      Movement and habitat connectivity of Atlantic cod in a coastal fjord system
#> 114     Movement, Range Size and Habitat Utilization of magafish in Turkey 2019-2021
#> 115   Movements of cod, plaice, flounder and turbot in the western German Baltic Sea
#> 116                                                                         mrc_vliz
#> 117                                                                         NARVAEEL
#> 118                                                         Noordzeekanaal 2017-2018
#> 119                                                          Northern_Norfolk_Broads
#> 120                                                                 North_sea_wrecks
#> 121                                                                    NTNU-Gaulosen
#> 122                                                       NTNU River Nidelva Project
#> 123   Observing the effect of group dynamics on individual behavior in basking shark
#> 124                                                                       Orsted Cod
#> 125                                                                  OTN-Hemnfjorden
#> 126                                                          OTN Loan - Beiarfjorden
#> 127                                                             OTN-Skjerstadfjorden
#> 128                                                     OTN Strait of Gibraltar Line
#> 129                                                                 OTN-Tosenfjorden
#> 130                                                           OTN VR2W Loan - Norway
#> 131                                                                    paintedcomber
#> 132                                         PAM Harbour porpoise Borssele, 2019-2024
#> 133                                        Passive Acoustic Monitoring (PAM) STRAITS
#> 134                                                                             PC4C
#> 135                                                                PhD Barbara Koeck
#> 136                                                                PhD Jeremy Pastor
#> 137                                                                        PhD_Nolan
#> 138                                                                         PhysFish
#> 139                       Pilot permanent deployment acoustic array Gibraltar Strait
#> 140                          Predators and prey around renewable energy developments
#> 141                                                                  PTN/PROTECT2012
#> 142                                                                  PTN/PROTECT2013
#> 143                                                           PTN-Silver eel-Mondego
#> 144                                                          Purewind fish detectors
#> 145  Redes de Sensores submarinos autonomos para la medida de indicadores biologicos
#> 146                                                                         Reelease
#> 147                 Research on the impact of management actions on beaver behaviour
#> 148                                                                           ResMed
#> 149                                                River_lamprey_River Trent_2020/21
#> 150                                             River_Lamprey_Yorkshire Ouse_2018-20
#> 151                                             River Usk receiver network 2021-2024
#> 152                                                                rt2020_zeeschelde
#> 153                                                             Ruim_baan_voor_vis_2
#> 154                                                                      Running_eel
#> 155                                                                       Saeftinghe
#> 156                                                                          SAMARCH
#> 157                                                                       SARTELARAM
#> 158                                                                         SARTELTG
#> 159                                                                    SARTELZINGARO
#> 160                                                                         Scytrack
#> 161                                                                       SeaMonitor
#> 162           SeaMonitor - Costs and Pathways of Marine Migration in Atlantic salmon
#> 163                                                        SeaMonitor - GMIT Project
#> 164                                                        sea trout_2018-9_gullmarn
#> 165 SEAWave: small cetacean presence at the EMEC Orkney Billia Croo wave energy site
#> 166                                                                   Semp Lithuania
#> 167                                                                 Siganid_East_Med
#> 168                                                               Siganid_Gulf_Aqaba
#> 169              Silver Eel downstream migration in the Walloon Meuse - 2021 to 2024
#> 170                                                 Skagerrak_temeletry_array_Norway
#> 171                                                                        Skye 2017
#> 172                                                             SMOLTRACK-I-Engeland
#> 173                                                            SMOLTRACK-II-Engeland
#> 174         Spatial Ecology of spurdog (Squalus acanthias) in a Scottish loch system
#> 175 Spawning migration of pike in MatsaluBay and RiverKasari (Baltic Sea, 2018-2020)
#> 176  Spawning migration of pike in Teorehe-Sauemeri floodplain (BalticSea 2017-2019)
#> 177                                                                        SPAWNSEIS
#> 178                                                                S_rivulatus_Aqaba
#> 179                                                S_rivulatus_Eastern_Mediterranean
#> 180                                                         STRAITS_Portugal_network
#> 181                                                                      STR_MOV_ANG
#> 182                                                  Sud_le_Impuls_voor_Diadrome_Vis
#> 183                                                    Sud_le_Innovatieve_monitoring
#> 184                                          survival_movement_elasmobranch_Adriatic
#> 185                                                                  SVNL-FISH-WATCH
#> 186 Swansea Uni acoustic receiver array: Bristol Channel, Swansea Bay and River Tawe
#> 187                                                            Swedish eel migration
#> 188                                                                     SWIMWAY_2021
#> 189                                                                   Swimway Vechte
#> 190                                                 Swiss Acoustic Telemetry Network
#> 191                                  Temporary acoustic network in the Sines harbour
#> 192    Temporary acoustic network in the SW Alentejo and Vicentine Coast Marine Park
#> 193                                                                             TeRi
#> 194                                                                       Test VR2AR
#> 195                                                     The Bergen Telemetry Network
#> 196                                                                     Thorntonbank
#> 197                                                                  Trent¨_Tracking
#> 198                                                   Turkish Straits Detection Data
#> 199       Understanding the movement of dolphinfish in the Mediterranean FAD fishery
#> 200                      Understand movement and life history of trout: Isle of Skye
#> 201    Underwater acoustic data from the Belgian ARMS-MBON hard-substrate monitoring
#> 202                                                             Unlocking the Severn
#> 203                  Unravelling the migrations of the invasive blue crab in Algarve
#> 204                                                                 VFM_BalticProper
#> 205                                                                 VFM_GulfOfBotnia
#> 206                                                                    VFM_Hjalmaren
#> 207                                                                      VFM_Malaren
#> 208                                                                    VFM_RiverGota
#> 209                                                                       VFM_Siljan
#> 210                                                                      VFM_Vattern
#> 211                                                              VFM_WestCoastCostal
#> 212                                                           Vissen_voor_verbinding
#> 213                                                                  Westerschelde 1
#> 214                                                                  Westerschelde 2
#> 215                                                                  Westerschelde 3
#> 216                                      Yellow eel movement behaviour in the Azores
#> 217                                                                       Zeeschelde
#> 218                                                                 2010_phd_reubens
#> 219                                                            2010_phd_reubens_sync
#> 220                                                                  2011 Rivierprik
#> 221                                                               2012 Leopoldkanaal
#> 222                                                                2013 Albertkanaal
#> 223                                                                       2014 Demer
#> 224                                                        2015_Albertkanaal_VPS_Ham
#> 225                                                                       2015 Dijle
#> 226                                                                     2015_Homarus
#> 227                                                            2015 PhD Verhelst_cod
#> 228                                                            2015 PhD Verhelst_eel
#> 229                                                        2016_Albertkanaal_VPS_Ham
#> 230                                                               2016 PhD Vergeynst
#> 231                                                        2017_Albertkanaal_VPS_Ham
#> 232                                                         2021_Sturgeons_Danube_BG
#> 233                                                    2022_Albertkanaal_VPS_Hasselt
#> 234   Acoustic telemetry data for Atlantic cod at an oil and gas platform in Denmark
#> 235                                                                       ARAISOLA02
#> 236                                                                       ARAISOLA03
#> 237                                                                              Ars
#> 238                                                                         Artevigo
#> 239                                                        Balearic Tracking Network
#> 240                  Balearic Tracking Network - Spanish Deep Water Receiver Network
#> 241 Blue crab telemetry monitoring in three lagoons, channels and coast of Occitanie
#> 242      Comparative study of marble and rainbow trout behaviour in Kanomljica River
#> 243                                                                    DAbecasis PhD
#> 244                                                                 Danube_Sturgeons
#> 245                                                         Delfzijl Drie Delfzijlen
#> 246   Dusky grouper´s resettlement - Acoustic telemetry in Maro-Cerro Gordo´s cliffs
#> 247                                                                             EMMN
#> 248                                                                             ESGL
#> 249                                                       ETN_animal_project_group_1
#> 250                                                       ETN_animal_project_group_2
#> 251                                                       ETN_animal_project_group_3
#> 252   Evaluation of the effectiveness of the Balearic Islands Marine Reserve Network
#> 253                                                                        Hunze_eel
#> 254                                                              JSATS-PalmaBay-2019
#> 255                                                                         Limfjord
#> 256                                                          Marble vs rainbow trout
#> 257                                                                       MBA_Massmo
#> 258                                                                      MBA_Wavehub
#> 259                                                                     MBA_Whitsand
#> 260                                                          Middle_Danube_sturgeons
#> 261                                                                      Mobula Imar
#> 262                            Monitoring Gibraltral Strait Using Acoustic Telemetry
#> 263      Movement and habitat connectivity of Atlantic cod in a coastal fjord system
#> 264                                                   North Aegean Anguilla Tracking
#> 265                                                                    NTNU-Gaulosen
#> 266                                                       NTNU River Nidelva Project
#> 267                                                                       Orsted Cod
#> 268                                                                  OTN-Hemnfjorden
#> 269                                                                 OTN-Tosenfjorden
#> 270                                                           OTN VR2W Loan - Norway
#> 271                                                                PhD Barbara Koeck
#> 272                                                                     PhD_Goossens
#> 273                                                                PhD Jeremy Pastor
#> 274                                                                     PhD_Marrocco
#> 275                                                                         PhysFish
#> 276                                                                       SARTELARAM
#> 277                                                                         SARTELTG
#> 278                                                                    SARTELZINGARO
#> 279                                                                         Scytrack
#> 280                                                        sea trout_2018-9_gullmarn
#> 281                                                                   Semp Lithuania
#> 282                                                                 Siganid_East_Med
#> 283                                                               Siganid_Gulf_Aqaba
#> 284              Silver Eel downstream migration in the Walloon Meuse - 2021 to 2024
#> 285         Spatial Ecology of spurdog (Squalus acanthias) in a Scottish loch system
#> 286                                                S_rivulatus_Eastern_Mediterranean
#> 287                                                          STRAITS_Portugal_animal
#> 288                                                                             TeRi
#> 289       Understanding the movement of dolphinfish in the Mediterranean FAD fishery
#> 290                                                                    VPS Duurswold
#> 291                                                                  VPS Garmerwolde
#>        type
#> 1   network
#> 2   network
#> 3   network
#> 4   network
#> 5   network
#> 6   network
#> 7   network
#> 8   network
#> 9   network
#> 10  network
#> 11  network
#> 12  network
#> 13  network
#> 14  network
#> 15  network
#> 16  network
#> 17  network
#> 18  network
#> 19  network
#> 20  network
#> 21  network
#> 22  network
#> 23  network
#> 24  network
#> 25  network
#> 26  network
#> 27  network
#> 28  network
#> 29  network
#> 30  network
#> 31  network
#> 32  network
#> 33  network
#> 34  network
#> 35  network
#> 36  network
#> 37  network
#> 38  network
#> 39  network
#> 40  network
#> 41  network
#> 42  network
#> 43  network
#> 44  network
#> 45  network
#> 46  network
#> 47  network
#> 48  network
#> 49  network
#> 50  network
#> 51  network
#> 52  network
#> 53  network
#> 54  network
#> 55  network
#> 56  network
#> 57  network
#> 58  network
#> 59  network
#> 60  network
#> 61  network
#> 62  network
#> 63  network
#> 64  network
#> 65  network
#> 66  network
#> 67  network
#> 68  network
#> 69  network
#> 70  network
#> 71  network
#> 72  network
#> 73  network
#> 74  network
#> 75  network
#> 76  network
#> 77  network
#> 78  network
#> 79  network
#> 80  network
#> 81  network
#> 82  network
#> 83  network
#> 84  network
#> 85  network
#> 86  network
#> 87  network
#> 88  network
#> 89  network
#> 90  network
#> 91  network
#> 92  network
#> 93  network
#> 94  network
#> 95  network
#> 96  network
#> 97  network
#> 98  network
#> 99  network
#> 100 network
#> 101 network
#> 102 network
#> 103 network
#> 104 network
#> 105 network
#> 106 network
#> 107 network
#> 108 network
#> 109 network
#> 110 network
#> 111 network
#> 112 network
#> 113 network
#> 114 network
#> 115 network
#> 116 network
#> 117 network
#> 118 network
#> 119 network
#> 120 network
#> 121 network
#> 122 network
#> 123 network
#> 124 network
#> 125 network
#> 126 network
#> 127 network
#> 128 network
#> 129 network
#> 130 network
#> 131 network
#> 132 network
#> 133 network
#> 134 network
#> 135 network
#> 136 network
#> 137 network
#> 138 network
#> 139 network
#> 140 network
#> 141 network
#> 142 network
#> 143 network
#> 144 network
#> 145 network
#> 146 network
#> 147 network
#> 148 network
#> 149 network
#> 150 network
#> 151 network
#> 152 network
#> 153 network
#> 154 network
#> 155 network
#> 156 network
#> 157 network
#> 158 network
#> 159 network
#> 160 network
#> 161 network
#> 162 network
#> 163 network
#> 164 network
#> 165 network
#> 166 network
#> 167 network
#> 168 network
#> 169 network
#> 170 network
#> 171 network
#> 172 network
#> 173 network
#> 174 network
#> 175 network
#> 176 network
#> 177 network
#> 178 network
#> 179 network
#> 180 network
#> 181 network
#> 182 network
#> 183 network
#> 184 network
#> 185 network
#> 186 network
#> 187 network
#> 188 network
#> 189 network
#> 190 network
#> 191 network
#> 192 network
#> 193 network
#> 194 network
#> 195 network
#> 196 network
#> 197 network
#> 198 network
#> 199 network
#> 200 network
#> 201 network
#> 202 network
#> 203 network
#> 204 network
#> 205 network
#> 206 network
#> 207 network
#> 208 network
#> 209 network
#> 210 network
#> 211 network
#> 212 network
#> 213 network
#> 214 network
#> 215 network
#> 216 network
#> 217 network
#> 218  animal
#> 219  animal
#> 220  animal
#> 221  animal
#> 222  animal
#> 223  animal
#> 224  animal
#> 225  animal
#> 226  animal
#> 227  animal
#> 228  animal
#> 229  animal
#> 230  animal
#> 231  animal
#> 232  animal
#> 233  animal
#> 234  animal
#> 235  animal
#> 236  animal
#> 237  animal
#> 238  animal
#> 239  animal
#> 240  animal
#> 241  animal
#> 242  animal
#> 243  animal
#> 244  animal
#> 245  animal
#> 246  animal
#> 247  animal
#> 248  animal
#> 249  animal
#> 250  animal
#> 251  animal
#> 252  animal
#> 253  animal
#> 254  animal
#> 255  animal
#> 256  animal
#> 257  animal
#> 258  animal
#> 259  animal
#> 260  animal
#> 261  animal
#> 262  animal
#> 263  animal
#> 264  animal
#> 265  animal
#> 266  animal
#> 267  animal
#> 268  animal
#> 269  animal
#> 270  animal
#> 271  animal
#> 272  animal
#> 273  animal
#> 274  animal
#> 275  animal
#> 276  animal
#> 277  animal
#> 278  animal
#> 279  animal
#> 280  animal
#> 281  animal
#> 282  animal
#> 283  animal
#> 284  animal
#> 285  animal
#> 286  animal
#> 287  animal
#> 288  animal
#> 289  animal
#> 290  animal
#> 291  animal