Retrieve an overview of the ETN data.

Retrieves the name of network/animals? To get an account, register via the Lifewatch RShiny registration webpage.

listEtnProjects(usr = NULL, pwd = NULL, ...)

Arguments

usr

Username to connect to ETN database

pwd

Password to connect to ETN database

...

Reserved for internal use.

Value

Dataframe with name and type of networks.

Examples

listEtnProjects()
#> - Query mode: Post request to OpenCPU server
#> No encoding supplied: defaulting to UTF-8.
#>                                                                                 name
#> 1                                                                        2004_Gudena
#> 2                                                                         2011_Loire
#> 3                                                                        2011_Warnow
#> 4                                                                         2013_Foyle
#> 5                                                                          2013_Maas
#> 6                                                                          2014_Nene
#> 7                                                           2015_PhD_Gutmann_Roberts
#> 8                                                               2016_Diaccia_Botrona
#> 9                                                                        2017_Fremur
#> 10                                                                    2019_Grotenete
#> 11                                                                   2020_PhD_Winter
#> 12                                                                     Aberdeen 2017
#> 13                                      Acoustic Ecology of Pelagic Fish Communities
#> 14    Acoustic telemetry data for Atlantic cod at an oil and gas platform in Denmark
#> 15      Acoustic telemetry on Atlantic cod in the Horns Rev 2 offshore windmill farm
#> 16                                                                      Albertkanaal
#> 17                                                              Albertkanaal_VPS_Ham
#> 18                                                          Albertkanaal_VPS_Hasselt
#> 19                                                                        ARAISOLA02
#> 20                                                                        ARAISOLA03
#> 21                                                                               Ars
#> 22                                                                          Artevigo
#> 23                                                 Azorean acoustic receiver network
#> 24                                                         Balearic Tracking Network
#> 25                   Balearic Tracking Network - Spanish Deep Water Receiver Network
#> 26                                                            Basking_shark_Hebrides
#> 27                                                                            BECORV
#> 28                           Belwind temporary acoustic receiver network 2020 - 2021
#> 29                                                                       Boddenhecht
#> 30                                                                      Bovenschelde
#> 31                                                                              BPNS
#> 32                                                           Brasem Ijsselmeergebied
#> 33                                               Brown trout in the Falkland Islands
#> 34                                                      Canaries Angel Shark Project
#> 35                                                          Catalan Tracking Network
#> 36                                                            CODEVCO fish detectors
#> 37                          Connectivity of coastal fish in the French Mediterranean
#> 38                                                                        Conon 2017
#> 39                                                                       cpodnetwork
#> 40                                                                     DAbecasis PhD
#> 41                                                                     DAK 2018-2019
#> 42                                                          Dalalven Smolt Migration
#> 43                                                        Danish_southeastern_strait
#> 44                                            Danish Straits Acoustic Infrastructure
#> 45                                                                  Danube_Sturgeons
#> 46                                                                Delfzijl 2014-2015
#> 47                                                                             Demer
#> 48                                                                      Deveron 2016
#> 49                                                                             Dijle
#> 50                                                                         Dijle_VPS
#> 51                                                                        DTU-Skjern
#> 52    Dusky grouper´s resettlement - Acoustic telemetry in Maro-Cerro Gordo´s cliffs
#> 53    Duurzame economische dragers in Vlaamse kustvisserijgemeenschappen post-Brexit
#> 54                                                  Eel migration from source to sea
#> 55                                                                      Eemskanaal I
#> 56                                                                     Eemskanaal II
#> 57                                                                    Eemskanaal III
#> 58                                                                              EMMN
#> 59                                                                              ESGL
#> 60                                                               ETN_project_group_1
#> 61                                                               ETN_project_group_4
#> 62                                                               ETN_Workshop_group2
#> 63    Evaluation of the effectiveness of the Balearic Islands Marine Reserve Network
#> 64                                                                 FarmTrack_Spain22
#> 65                                            Fisheries Industry Science Partnership
#> 66                Fisheries Innovation for sustainable SHared INTerchannEL resources
#> 67                                           Fish migration from source to Waddensea
#> 68                                                              FISHOWF_Array_France
#> 69                                                               Friesland 2011-2012
#> 70                                     Galician (NW Spain) acoustic Tracking Network
#> 71                                                         GIBRALTAR_STRAITS_CURTAIN
#> 72                                                                    Gudena_network
#> 73                                                                   IG II Fish Pass
#> 74                                                                         Iller_VPS
#> 75                                         Immature bass acoustic stock surveillance
#> 76                                                                     Inforbiomares
#> 77                                                                   INTERREG_CANAPE
#> 78                                                      Jersey Coastal Fish Tracking
#> 79                                                               JSATS-PalmaBay-2019
#> 80                                                    Kalmar Baltic Tracking Network
#> 81                                                                    Kornwerderzand
#> 82                                                                 Lake_Anundsjo_VPS
#> 83                        Land-use change and human activity on anadromous salmonids
#> 84                                              Large_Migrants_Haringvliet_2023-2025
#> 85                                                                       Lauwersmeer
#> 86                                                                     Leopoldkanaal
#> 87                                                                         Lifewatch
#> 88                                                                          Limfjord
#> 89                                              Ljusne Bay Salmon Spawning Migration
#> 90  Long-term array MOnitoring BEhaviour in Illas Atlánticas National Park, NW Spain
#> 91                                                                           MacFish
#> 92                                                           Marble vs rainbow trout
#> 93                                                             MariagerFjord_network
#> 94                                                                        MBA_Massmo
#> 95                                                                       MBA_Wavehub
#> 96                                                                      MBA_Whitsand
#> 97                                                                            MEPNSW
#> 98                                                           Middle_Danube_sturgeons
#> 99    Migration of pikeperch within the Lake Peipus complex (Estonia) from 2021-2023
#> 100                                                                      MIGRATOEBRE
#> 101                                                                      Mobula Imar
#> 102                            Monitoring Gibraltral Strait Using Acoustic Telemetry
#> 103                                                                 Moray Firth 2016
#> 104      Movement and habitat connectivity of Atlantic cod in a coastal fjord system
#> 105     Movement, Range Size and Habitat Utilization of magafish in Turkey 2019-2021
#> 106   Movements of cod, plaice, flounder and turbot in the western German Baltic Sea
#> 107                                                                         mrc_vliz
#> 108                                                                         NARVAEEL
#> 109                                                         Noordzeekanaal 2017-2018
#> 110                                                          Northern_Norfolk_Broads
#> 111                                                                 North_sea_wrecks
#> 112                                                                    NTNU-Gaulosen
#> 113                                                       NTNU River Nidelva Project
#> 114   Observing the effect of group dynamics on individual behavior in basking shark
#> 115                                                                       Orsted Cod
#> 116                                                                  OTN-Hemnfjorden
#> 117                                                          OTN Loan - Beiarfjorden
#> 118                                                             OTN-Skjerstadfjorden
#> 119                                                                 OTN-Tosenfjorden
#> 120                                                           OTN VR2W Loan - Norway
#> 121                                                                    paintedcomber
#> 122                                         PAM Harbour porpoise Borssele, 2019-2024
#> 123                                                                             PC4C
#> 124                                                                PhD Barbara Koeck
#> 125                                                                PhD Jeremy Pastor
#> 126                                                                        PhD_Nolan
#> 127                                                                         PhysFish
#> 128                       Pilot permanent deployment acoustic array Gibraltar Strait
#> 129                          Predators and prey around renewable energy developments
#> 130                                                                  PTN/PROTECT2012
#> 131                                                                  PTN/PROTECT2013
#> 132                                                           PTN-Silver eel-Mondego
#> 133                                                          Purewind fish detectors
#> 134  Redes de Sensores submarinos autonomos para la medida de indicadores biologicos
#> 135                                                                         Reelease
#> 136                 Research on the impact of management actions on beaver behaviour
#> 137                                                                           ResMed
#> 138                                                River_lamprey_River Trent_2020/21
#> 139                                             River_Lamprey_Yorkshire Ouse_2018-20
#> 140                                             River Usk receiver network 2021-2024
#> 141                                                                rt2020_zeeschelde
#> 142                                                             Ruim_baan_voor_vis_2
#> 143                                                                      Running_eel
#> 144                                                                       Saeftinghe
#> 145                                                                          SAMARCH
#> 146                                                                       SARTELARAM
#> 147                                                                         SARTELTG
#> 148                                                                    SARTELZINGARO
#> 149                                                                         Scytrack
#> 150                                                                       SeaMonitor
#> 151           SeaMonitor - Costs and Pathways of Marine Migration in Atlantic salmon
#> 152                                                        SeaMonitor - GMIT Project
#> 153            SeaMonitor - Migration route of Atlantic salmon in north west Ireland
#> 154                                                    SeaMonitor - River Bush study
#> 155                                                        sea trout_2018-9_gullmarn
#> 156 SEAWave: small cetacean presence at the EMEC Orkney Billia Croo wave energy site
#> 157                                                                   Semp Lithuania
#> 158                                                                 Siganid_East_Med
#> 159                                                               Siganid_Gulf_Aqaba
#> 160              Silver Eel downstream migration in the Walloon Meuse - 2021 to 2024
#> 161                                                 Skagerrak_temeletry_array_Norway
#> 162                                                                        Skye 2017
#> 163                                                             SMOLTRACK-I-Engeland
#> 164                                                            SMOLTRACK-II-Engeland
#> 165                                                      Southern Norway brown trout
#> 166         Spatial Ecology of spurdog (Squalus acanthias) in a Scottish loch system
#> 167 Spawning migration of pike in MatsaluBay and RiverKasari (Baltic Sea, 2018-2020)
#> 168  Spawning migration of pike in Teorehe-Sauemeri floodplain (BalticSea 2017-2019)
#> 169                                                                        SPAWNSEIS
#> 170                                                                S_rivulatus_Aqaba
#> 171                                                S_rivulatus_Eastern_Mediterranean
#> 172                                                                      STRAITS_PAM
#> 173                                                         STRAITS_Portugal_network
#> 174                                                                      STR_MOV_ANG
#> 175                                                  Sud_le_Impuls_voor_Diadrome_Vis
#> 176                                                    Sud_le_Innovatieve_monitoring
#> 177                                          survival_movement_elasmobranch_Adriatic
#> 178                                                                  SVNL-FISH-WATCH
#> 179 Swansea Uni acoustic receiver array: Bristol Channel, Swansea Bay and River Tawe
#> 180                                                            Swedish eel migration
#> 181                                                                     SWIMWAY_2021
#> 182                                                                   Swimway Vechte
#> 183                                  Temporary acoustic network in the Sines harbour
#> 184    Temporary acoustic network in the SW Alentejo and Vicentine Coast Marine Park
#> 185                                                                             TeRi
#> 186                                                                       Test VR2AR
#> 187                                                     The Bergen Telemetry Network
#> 188                                                                     Thorntonbank
#> 189                                                   Turkish Straits Detection Data
#> 190       Understanding the movement of dolphinfish in the Mediterranean FAD fishery
#> 191    Underwater acoustic data from the Belgian ARMS-MBON hard-substrate monitoring
#> 192                                                             Unlocking the Severn
#> 193                  Unravelling the migrations of the invasive blue crab in Algarve
#> 194                                                                 VFM_GulfOfBotnia
#> 195                                                                    VFM_Hjalmaren
#> 196                                                                      VFM_Malaren
#> 197                                                                       VFM_Siljan
#> 198                                                                      VFM_Vattern
#> 199                                                              VFM_WestCoastCostal
#> 200                                                           Vissen_voor_verbinding
#> 201                                           VMT Loan - SeaMonitor Marine Institute
#> 202                                                                  Westerschelde 1
#> 203                                                                  Westerschelde 2
#> 204                                                                  Westerschelde 3
#> 205                                      Yellow eel movement behaviour in the Azores
#> 206                                                                       Zeeschelde
#> 207                                                                 2010_phd_reubens
#> 208                                                            2010_phd_reubens_sync
#> 209                                                                  2011 Rivierprik
#> 210                                                               2012 Leopoldkanaal
#> 211                                                                2013 Albertkanaal
#> 212                                                                       2014 Demer
#> 213                                                        2015_Albertkanaal_VPS_Ham
#> 214                                                                       2015 Dijle
#> 215                                                                     2015_Homarus
#> 216                                                            2015 PhD Verhelst_cod
#> 217                                                            2015 PhD Verhelst_eel
#> 218                                                        2016_Albertkanaal_VPS_Ham
#> 219                                                               2016 PhD Vergeynst
#> 220                                                        2017_Albertkanaal_VPS_Ham
#> 221                                                         2021_Sturgeons_Danube_BG
#> 222                                                    2022_Albertkanaal_VPS_Hasselt
#> 223   Acoustic telemetry data for Atlantic cod at an oil and gas platform in Denmark
#> 224                                                                       ARAISOLA02
#> 225                                                                       ARAISOLA03
#> 226                                                                              Ars
#> 227                                                                         Artevigo
#> 228                                                        Balearic Tracking Network
#> 229                  Balearic Tracking Network - Spanish Deep Water Receiver Network
#> 230      Comparative study of marble and rainbow trout behaviour in Kanomljica River
#> 231                                                                    DAbecasis PhD
#> 232                                                                 Danube_Sturgeons
#> 233                                                         Delfzijl Drie Delfzijlen
#> 234   Dusky grouper´s resettlement - Acoustic telemetry in Maro-Cerro Gordo´s cliffs
#> 235                                                                             EMMN
#> 236                                                                             ESGL
#> 237                                                              ETN_project_group_4
#> 238                                                              ETN_Workshop_group2
#> 239   Evaluation of the effectiveness of the Balearic Islands Marine Reserve Network
#> 240                                                                        Hunze_eel
#> 241                                                              JSATS-PalmaBay-2019
#> 242                                                                         Limfjord
#> 243                                                          Marble vs rainbow trout
#> 244                                                                       MBA_Massmo
#> 245                                                                      MBA_Wavehub
#> 246                                                                     MBA_Whitsand
#> 247                                                          Middle_Danube_sturgeons
#> 248                                                                      Mobula Imar
#> 249                            Monitoring Gibraltral Strait Using Acoustic Telemetry
#> 250      Movement and habitat connectivity of Atlantic cod in a coastal fjord system
#> 251                                                   North Aegean Anguilla Tracking
#> 252                                                                    NTNU-Gaulosen
#> 253                                                       NTNU River Nidelva Project
#> 254                                                                       Orsted Cod
#> 255                                                                  OTN-Hemnfjorden
#> 256                                                                 OTN-Tosenfjorden
#> 257                                                           OTN VR2W Loan - Norway
#> 258                                                                PhD Barbara Koeck
#> 259                                                                PhD Jeremy Pastor
#> 260                                                                     PhD_Marrocco
#> 261                                                                         PhysFish
#> 262                                                                       SARTELARAM
#> 263                                                                         SARTELTG
#> 264                                                                    SARTELZINGARO
#> 265                                                                         Scytrack
#> 266                                                        sea trout_2018-9_gullmarn
#> 267                                                                   Semp Lithuania
#> 268                                                                 Siganid_East_Med
#> 269                                                               Siganid_Gulf_Aqaba
#> 270              Silver Eel downstream migration in the Walloon Meuse - 2021 to 2024
#> 271         Spatial Ecology of spurdog (Squalus acanthias) in a Scottish loch system
#> 272                                                S_rivulatus_Eastern_Mediterranean
#> 273                                                          STRAITS_Portugal_animal
#> 274                                                                             TeRi
#> 275       Understanding the movement of dolphinfish in the Mediterranean FAD fishery
#> 276                                                                    VPS Duurswold
#> 277                                                                  VPS Garmerwolde
#>        type
#> 1   network
#> 2   network
#> 3   network
#> 4   network
#> 5   network
#> 6   network
#> 7   network
#> 8   network
#> 9   network
#> 10  network
#> 11  network
#> 12  network
#> 13  network
#> 14  network
#> 15  network
#> 16  network
#> 17  network
#> 18  network
#> 19  network
#> 20  network
#> 21  network
#> 22  network
#> 23  network
#> 24  network
#> 25  network
#> 26  network
#> 27  network
#> 28  network
#> 29  network
#> 30  network
#> 31  network
#> 32  network
#> 33  network
#> 34  network
#> 35  network
#> 36  network
#> 37  network
#> 38  network
#> 39  network
#> 40  network
#> 41  network
#> 42  network
#> 43  network
#> 44  network
#> 45  network
#> 46  network
#> 47  network
#> 48  network
#> 49  network
#> 50  network
#> 51  network
#> 52  network
#> 53  network
#> 54  network
#> 55  network
#> 56  network
#> 57  network
#> 58  network
#> 59  network
#> 60  network
#> 61  network
#> 62  network
#> 63  network
#> 64  network
#> 65  network
#> 66  network
#> 67  network
#> 68  network
#> 69  network
#> 70  network
#> 71  network
#> 72  network
#> 73  network
#> 74  network
#> 75  network
#> 76  network
#> 77  network
#> 78  network
#> 79  network
#> 80  network
#> 81  network
#> 82  network
#> 83  network
#> 84  network
#> 85  network
#> 86  network
#> 87  network
#> 88  network
#> 89  network
#> 90  network
#> 91  network
#> 92  network
#> 93  network
#> 94  network
#> 95  network
#> 96  network
#> 97  network
#> 98  network
#> 99  network
#> 100 network
#> 101 network
#> 102 network
#> 103 network
#> 104 network
#> 105 network
#> 106 network
#> 107 network
#> 108 network
#> 109 network
#> 110 network
#> 111 network
#> 112 network
#> 113 network
#> 114 network
#> 115 network
#> 116 network
#> 117 network
#> 118 network
#> 119 network
#> 120 network
#> 121 network
#> 122 network
#> 123 network
#> 124 network
#> 125 network
#> 126 network
#> 127 network
#> 128 network
#> 129 network
#> 130 network
#> 131 network
#> 132 network
#> 133 network
#> 134 network
#> 135 network
#> 136 network
#> 137 network
#> 138 network
#> 139 network
#> 140 network
#> 141 network
#> 142 network
#> 143 network
#> 144 network
#> 145 network
#> 146 network
#> 147 network
#> 148 network
#> 149 network
#> 150 network
#> 151 network
#> 152 network
#> 153 network
#> 154 network
#> 155 network
#> 156 network
#> 157 network
#> 158 network
#> 159 network
#> 160 network
#> 161 network
#> 162 network
#> 163 network
#> 164 network
#> 165 network
#> 166 network
#> 167 network
#> 168 network
#> 169 network
#> 170 network
#> 171 network
#> 172 network
#> 173 network
#> 174 network
#> 175 network
#> 176 network
#> 177 network
#> 178 network
#> 179 network
#> 180 network
#> 181 network
#> 182 network
#> 183 network
#> 184 network
#> 185 network
#> 186 network
#> 187 network
#> 188 network
#> 189 network
#> 190 network
#> 191 network
#> 192 network
#> 193 network
#> 194 network
#> 195 network
#> 196 network
#> 197 network
#> 198 network
#> 199 network
#> 200 network
#> 201 network
#> 202 network
#> 203 network
#> 204 network
#> 205 network
#> 206 network
#> 207  animal
#> 208  animal
#> 209  animal
#> 210  animal
#> 211  animal
#> 212  animal
#> 213  animal
#> 214  animal
#> 215  animal
#> 216  animal
#> 217  animal
#> 218  animal
#> 219  animal
#> 220  animal
#> 221  animal
#> 222  animal
#> 223  animal
#> 224  animal
#> 225  animal
#> 226  animal
#> 227  animal
#> 228  animal
#> 229  animal
#> 230  animal
#> 231  animal
#> 232  animal
#> 233  animal
#> 234  animal
#> 235  animal
#> 236  animal
#> 237  animal
#> 238  animal
#> 239  animal
#> 240  animal
#> 241  animal
#> 242  animal
#> 243  animal
#> 244  animal
#> 245  animal
#> 246  animal
#> 247  animal
#> 248  animal
#> 249  animal
#> 250  animal
#> 251  animal
#> 252  animal
#> 253  animal
#> 254  animal
#> 255  animal
#> 256  animal
#> 257  animal
#> 258  animal
#> 259  animal
#> 260  animal
#> 261  animal
#> 262  animal
#> 263  animal
#> 264  animal
#> 265  animal
#> 266  animal
#> 267  animal
#> 268  animal
#> 269  animal
#> 270  animal
#> 271  animal
#> 272  animal
#> 273  animal
#> 274  animal
#> 275  animal
#> 276  animal
#> 277  animal