Retrieve an overview of the ETN data.

Retrieves the name of network/animals? To get an account, register via the Lifewatch RShiny registration webpage.

listEtnProjects(usr = NULL, pwd = NULL, ...)

Arguments

usr

Username to connect to ETN database

pwd

Password to connect to ETN database

...

Reserved for internal use.

Value

Dataframe with name and type of networks.

Examples

listEtnProjects()
#> - Query mode: Post request to OpenCPU server
#> No encoding supplied: defaulting to UTF-8.
#>                                                                                 name
#> 1                                                                        2004_Gudena
#> 2                                                                         2011_Loire
#> 3                                                                        2011_Warnow
#> 4                                                                         2013_Foyle
#> 5                                                                          2013_Maas
#> 6                                                                          2014_Nene
#> 7                                                           2015_PhD_Gutmann_Roberts
#> 8                                                               2016_Diaccia_Botrona
#> 9                                                                        2017_Fremur
#> 10                                                                    2019_Grotenete
#> 11                                                                   2020_PhD_Winter
#> 12                                                                     Aberdeen 2017
#> 13                                      Acoustic Ecology of Pelagic Fish Communities
#> 14                                                                      Albertkanaal
#> 15                                                              Albertkanaal_VPS_Ham
#> 16                                                                        ARAISOLA02
#> 17                                                                        ARAISOLA03
#> 18                                                                               Ars
#> 19                                                                          Artevigo
#> 20                                                 Azorean acoustic receiver network
#> 21                                                         Balearic Tracking Network
#> 22                   Balearic Tracking Network - Spanish Deep Water Receiver Network
#> 23                                                            Basking_shark_Hebrides
#> 24                                                                            BECORV
#> 25                           Belwind temporary acoustic receiver network 2020 - 2021
#> 26                                                                       Boddenhecht
#> 27                                                                      Bovenschelde
#> 28                                                                              BPNS
#> 29                                                           Brasem Ijsselmeergebied
#> 30                                               Brown trout in the Falkland Islands
#> 31                                                      Canaries Angel Shark Project
#> 32                                                          Catalan Tracking Network
#> 33                          Connectivity of coastal fish in the French Mediterranean
#> 34                                                                        Conon 2017
#> 35                                                                       cpodnetwork
#> 36                                                                     DAbecasis PhD
#> 37                                                                     DAK 2018-2019
#> 38                                                          Dalalven Smolt Migration
#> 39                                                           Danish Belts and Sounds
#> 40                                                        Danish_southeastern_strait
#> 41                                                                  Danube_Sturgeons
#> 42                                                                Delfzijl 2014-2015
#> 43                                                                             Demer
#> 44                                                                      Deveron 2016
#> 45                                                                             Dijle
#> 46                                                                         Dijle_VPS
#> 47                                                                        DTU-Skjern
#> 48                                                  Eel migration from source to sea
#> 49                                                                      Eemskanaal I
#> 50                                                                     Eemskanaal II
#> 51                                                                    Eemskanaal III
#> 52                                                                              EMMN
#> 53                                                                              ESGL
#> 54                                                               ETN_project_group_1
#> 55                                                               ETN_project_group_4
#> 56                                                               ETN_Workshop_group2
#> 57                                                                 FarmTrack_Spain22
#> 58       Fine scale tracking of silver eel migrating along a fish detterrence system
#> 59                Fisheries Innovation for sustainable SHared INTerchannEL resources
#> 60                                           Fish migration from source to Waddensea
#> 61                                                              FISHOWF_Array_France
#> 62                                                               Friesland 2011-2012
#> 63                                     Galician (NW Spain) acoustic Tracking Network
#> 64                                                                    Gudena_network
#> 65                                                                   IG II Fish Pass
#> 66                                                                         Iller_VPS
#> 67                                         Immature bass acoustic stock surveillance
#> 68                                                                     Inforbiomares
#> 69                                                                   INTERREG_CANAPE
#> 70                                                      Jersey Coastal Fish Tracking
#> 71                                                               JSATS-PalmaBay-2019
#> 72                                                    Kalmar Baltic Tracking Network
#> 73                                                                    Kornwerderzand
#> 74                                                                 Lake_Anundsjo_VPS
#> 75                        Land-use change and human activity on anadromous salmonids
#> 76                                                                       Lauwersmeer
#> 77                                                                     Leopoldkanaal
#> 78                                                                         Lifewatch
#> 79                                                                          Limfjord
#> 80                                              Ljusne Bay Salmon Spawning Migration
#> 81  Long-term array MOnitoring BEhaviour in Illas Atlánticas National Park, NW Spain
#> 82                                                                           MacFish
#> 83                                                           Marble vs rainbow trout
#> 84                                                             MariagerFjord_network
#> 85                                                                        MBA_Massmo
#> 86                                                                       MBA_Wavehub
#> 87                                                                      MBA_Whitsand
#> 88                                                                            MEPNSW
#> 89                                                           Middle_Danube_sturgeons
#> 90    Migration of pikeperch within the Lake Peipus complex (Estonia) from 2021-2023
#> 91                                                                       MIGRATOEBRE
#> 92                                                                       Mobula Imar
#> 93                                                                  Moray Firth 2016
#> 94       Movement and habitat connectivity of Atlantic cod in a coastal fjord system
#> 95      Movement, Range Size and Habitat Utilization of magafish in Turkey 2019-2021
#> 96                                                                          NARVAEEL
#> 97                                                          Noordzeekanaal 2017-2018
#> 98                                                           Northern_Norfolk_Broads
#> 99                                                                  North_sea_wrecks
#> 100                                                                    NTNU-Gaulosen
#> 101                                                       NTNU River Nidelva Project
#> 102   Observing the effect of group dynamics on individual behavior in basking shark
#> 103                                                                       Orsted Cod
#> 104                                                                  OTN-Hemnfjorden
#> 105                                                          OTN Loan - Beiarfjorden
#> 106                                                                 OTN-Tosenfjorden
#> 107                                                           OTN VR2W Loan - Norway
#> 108                                                                    paintedcomber
#> 109                                                                             PC4C
#> 110                                                                PhD Barbara Koeck
#> 111                                                                PhD Jeremy Pastor
#> 112                                                                        PhD_Nolan
#> 113                                                                         PhysFish
#> 114                       Pilot permanent deployment acoustic array Gibraltar Strait
#> 115                                                                  PTN/PROTECT2012
#> 116                                                                  PTN/PROTECT2013
#> 117                                                           PTN-Silver eel-Mondego
#> 118  Redes de Sensores submarinos autonomos para la medida de indicadores biologicos
#> 119                                                                         Reelease
#> 120                 Research on the impact of management actions on beaver behaviour
#> 121                                                                           ResMed
#> 122                                                River_lamprey_River Trent_2020/21
#> 123                                             River_Lamprey_Yorkshire Ouse_2018-20
#> 124                                             River Usk receiver network 2021-2024
#> 125                                                             Ruim_baan_voor_vis_2
#> 126                                                                      Running_eel
#> 127                                                                       Saeftinghe
#> 128                                                                          SAMARCH
#> 129                                                                       SARTELARAM
#> 130                                                                         SARTELTG
#> 131                                                                    SARTELZINGARO
#> 132                                                                         Scytrack
#> 133                                                                       SeaMonitor
#> 134           SeaMonitor - Costs and Pathways of Marine Migration in Atlantic salmon
#> 135                                                        SeaMonitor - GMIT Project
#> 136            SeaMonitor - Migration route of Atlantic salmon in north west Ireland
#> 137                                                    SeaMonitor - River Bush study
#> 138                                                        sea trout_2018-9_gullmarn
#> 139                                                                   Semp Lithuania
#> 140                                                                 Siganid_East_Med
#> 141                                                               Siganid_Gulf_Aqaba
#> 142              Silver Eel downstream migration in the Walloon Meuse - 2021 to 2024
#> 143                                                                        Skye 2017
#> 144                                                             SMOLTRACK-I-Engeland
#> 145                                                            SMOLTRACK-II-Engeland
#> 146                                                      Southern Norway brown trout
#> 147         Spatial Ecology of spurdog (Squalus acanthias) in a Scottish loch system
#> 148 Spawning migration of pike in MatsaluBay and RiverKasari (Baltic Sea, 2018-2020)
#> 149  Spawning migration of pike in Teorehe-Sauemeri floodplain (BalticSea 2017-2019)
#> 150                                                                        SPAWNSEIS
#> 151                                                                S_rivulatus_Aqaba
#> 152                                                S_rivulatus_Eastern_Mediterranean
#> 153                                                  Sud_le_Impuls_voor_Diadrome_Vis
#> 154                                                    Sud_le_Innovatieve_monitoring
#> 155                                          survival_movement_elasmobranch_Adriatic
#> 156                                                                  SVNL-FISH-WATCH
#> 157 Swansea Uni acoustic receiver array: Bristol Channel, Swansea Bay and River Tawe
#> 158                                                            Swedish eel migration
#> 159                                                                     SWIMWAY_2021
#> 160                                                                   Swimway Vechte
#> 161                                  Temporary acoustic network in the Sines harbour
#> 162                                                                             TeRi
#> 163                                                                       Test VR2AR
#> 164                                                     The Bergen Telemetry Network
#> 165                                                                     Thorntonbank
#> 166       Understanding the movement of dolphinfish in the Mediterranean FAD fishery
#> 167                                                             Unlocking the Severn
#> 168                  Unravelling the migrations of the invasive blue crab in Algarve
#> 169                                                                 VFM_GulfOfBotnia
#> 170                                                                    VFM_Hjalmaren
#> 171                                                                      VFM_Malaren
#> 172                                                                       VFM_Siljan
#> 173                                                                      VFM_Vattern
#> 174                                                              VFM_WestCoastCostal
#> 175                                                           Vissen_voor_verbinding
#> 176                                           VMT Loan - SeaMonitor Marine Institute
#> 177                                                                  Westerschelde 1
#> 178                                                                  Westerschelde 2
#> 179                                                                  Westerschelde 3
#> 180                                      Yellow eel movement behaviour in the Azores
#> 181                                                                       Zeeschelde
#> 182                                                                 2010_phd_reubens
#> 183                                                            2010_phd_reubens_sync
#> 184                                                                  2011 Rivierprik
#> 185                                                               2012 Leopoldkanaal
#> 186                                                                2013 Albertkanaal
#> 187                                                                       2014 Demer
#> 188                                                        2015_Albertkanaal_VPS_Ham
#> 189                                                                       2015 Dijle
#> 190                                                                     2015_Homarus
#> 191                                                            2015 PhD Verhelst_cod
#> 192                                                            2015 PhD Verhelst_eel
#> 193                                                        2016_Albertkanaal_VPS_Ham
#> 194                                                               2016 PhD Vergeynst
#> 195                                                        2017_Albertkanaal_VPS_Ham
#> 196                                                         2021_Sturgeons_Danube_BG
#> 197                                                                       ARAISOLA02
#> 198                                                                       ARAISOLA03
#> 199                                                                              Ars
#> 200                                                                         Artevigo
#> 201                                                        Balearic Tracking Network
#> 202                  Balearic Tracking Network - Spanish Deep Water Receiver Network
#> 203      Comparative study of marble and rainbow trout behaviour in Kanomljica River
#> 204                                                                    DAbecasis PhD
#> 205                                                                 Danube_Sturgeons
#> 206                                                         Delfzijl Drie Delfzijlen
#> 207                                                                             EMMN
#> 208                                                                             ESGL
#> 209                                                              ETN_project_group_4
#> 210                                                              ETN_Workshop_group2
#> 211      Fine scale tracking of silver eel migrating along a fish detterrence system
#> 212                                                                        Hunze_eel
#> 213                                                              JSATS-PalmaBay-2019
#> 214                                             Large_Migrants_Haringvliet_2023-2025
#> 215                                                                         Limfjord
#> 216                                                          Marble vs rainbow trout
#> 217                                                                       MBA_Massmo
#> 218                                                                      MBA_Wavehub
#> 219                                                                     MBA_Whitsand
#> 220                                                          Middle_Danube_sturgeons
#> 221                                                                      Mobula Imar
#> 222      Movement and habitat connectivity of Atlantic cod in a coastal fjord system
#> 223                                                   North Aegean Anguilla Tracking
#> 224                                                                    NTNU-Gaulosen
#> 225                                                       NTNU River Nidelva Project
#> 226                                                                       Orsted Cod
#> 227                                                                  OTN-Hemnfjorden
#> 228                                                                 OTN-Tosenfjorden
#> 229                                                           OTN VR2W Loan - Norway
#> 230                                                                PhD Barbara Koeck
#> 231                                                                PhD Jeremy Pastor
#> 232                                                                     PhD_Marrocco
#> 233                                                                         PhysFish
#> 234                                                                       SARTELARAM
#> 235                                                                         SARTELTG
#> 236                                                                    SARTELZINGARO
#> 237                                                                         Scytrack
#> 238                                                        sea trout_2018-9_gullmarn
#> 239                                                                   Semp Lithuania
#> 240                                                                 Siganid_East_Med
#> 241                                                               Siganid_Gulf_Aqaba
#> 242              Silver Eel downstream migration in the Walloon Meuse - 2021 to 2024
#> 243         Spatial Ecology of spurdog (Squalus acanthias) in a Scottish loch system
#> 244                                                S_rivulatus_Eastern_Mediterranean
#> 245                                                                             TeRi
#> 246       Understanding the movement of dolphinfish in the Mediterranean FAD fishery
#> 247                                                                    VPS Duurswold
#> 248                                                                  VPS Garmerwolde
#>        type
#> 1   network
#> 2   network
#> 3   network
#> 4   network
#> 5   network
#> 6   network
#> 7   network
#> 8   network
#> 9   network
#> 10  network
#> 11  network
#> 12  network
#> 13  network
#> 14  network
#> 15  network
#> 16  network
#> 17  network
#> 18  network
#> 19  network
#> 20  network
#> 21  network
#> 22  network
#> 23  network
#> 24  network
#> 25  network
#> 26  network
#> 27  network
#> 28  network
#> 29  network
#> 30  network
#> 31  network
#> 32  network
#> 33  network
#> 34  network
#> 35  network
#> 36  network
#> 37  network
#> 38  network
#> 39  network
#> 40  network
#> 41  network
#> 42  network
#> 43  network
#> 44  network
#> 45  network
#> 46  network
#> 47  network
#> 48  network
#> 49  network
#> 50  network
#> 51  network
#> 52  network
#> 53  network
#> 54  network
#> 55  network
#> 56  network
#> 57  network
#> 58  network
#> 59  network
#> 60  network
#> 61  network
#> 62  network
#> 63  network
#> 64  network
#> 65  network
#> 66  network
#> 67  network
#> 68  network
#> 69  network
#> 70  network
#> 71  network
#> 72  network
#> 73  network
#> 74  network
#> 75  network
#> 76  network
#> 77  network
#> 78  network
#> 79  network
#> 80  network
#> 81  network
#> 82  network
#> 83  network
#> 84  network
#> 85  network
#> 86  network
#> 87  network
#> 88  network
#> 89  network
#> 90  network
#> 91  network
#> 92  network
#> 93  network
#> 94  network
#> 95  network
#> 96  network
#> 97  network
#> 98  network
#> 99  network
#> 100 network
#> 101 network
#> 102 network
#> 103 network
#> 104 network
#> 105 network
#> 106 network
#> 107 network
#> 108 network
#> 109 network
#> 110 network
#> 111 network
#> 112 network
#> 113 network
#> 114 network
#> 115 network
#> 116 network
#> 117 network
#> 118 network
#> 119 network
#> 120 network
#> 121 network
#> 122 network
#> 123 network
#> 124 network
#> 125 network
#> 126 network
#> 127 network
#> 128 network
#> 129 network
#> 130 network
#> 131 network
#> 132 network
#> 133 network
#> 134 network
#> 135 network
#> 136 network
#> 137 network
#> 138 network
#> 139 network
#> 140 network
#> 141 network
#> 142 network
#> 143 network
#> 144 network
#> 145 network
#> 146 network
#> 147 network
#> 148 network
#> 149 network
#> 150 network
#> 151 network
#> 152 network
#> 153 network
#> 154 network
#> 155 network
#> 156 network
#> 157 network
#> 158 network
#> 159 network
#> 160 network
#> 161 network
#> 162 network
#> 163 network
#> 164 network
#> 165 network
#> 166 network
#> 167 network
#> 168 network
#> 169 network
#> 170 network
#> 171 network
#> 172 network
#> 173 network
#> 174 network
#> 175 network
#> 176 network
#> 177 network
#> 178 network
#> 179 network
#> 180 network
#> 181 network
#> 182  animal
#> 183  animal
#> 184  animal
#> 185  animal
#> 186  animal
#> 187  animal
#> 188  animal
#> 189  animal
#> 190  animal
#> 191  animal
#> 192  animal
#> 193  animal
#> 194  animal
#> 195  animal
#> 196  animal
#> 197  animal
#> 198  animal
#> 199  animal
#> 200  animal
#> 201  animal
#> 202  animal
#> 203  animal
#> 204  animal
#> 205  animal
#> 206  animal
#> 207  animal
#> 208  animal
#> 209  animal
#> 210  animal
#> 211  animal
#> 212  animal
#> 213  animal
#> 214  animal
#> 215  animal
#> 216  animal
#> 217  animal
#> 218  animal
#> 219  animal
#> 220  animal
#> 221  animal
#> 222  animal
#> 223  animal
#> 224  animal
#> 225  animal
#> 226  animal
#> 227  animal
#> 228  animal
#> 229  animal
#> 230  animal
#> 231  animal
#> 232  animal
#> 233  animal
#> 234  animal
#> 235  animal
#> 236  animal
#> 237  animal
#> 238  animal
#> 239  animal
#> 240  animal
#> 241  animal
#> 242  animal
#> 243  animal
#> 244  animal
#> 245  animal
#> 246  animal
#> 247  animal
#> 248  animal