Retrieve an overview of the ETN data.

Retrieves the name of network/animals? To get an account, register via the Lifewatch RShiny registration webpage.

listEtnProjects(usr = NULL, pwd = NULL)

Arguments

usr
Username to connect to ETN database

pwd
Password to connect to ETN database

Value

Dataframe with name and type of networks.

Examples

listEtnProjects()

#> No encoding supplied: defaulting to UTF-8.

#>                                                                                 name
#> 1                                                                        2004_Gudena
#> 2                                                                         2011_Loire
#> 3                                                                        2011_Warnow
#> 4                                                                         2013_Foyle
#> 5                                                                          2013_Maas
#> 6                                                                          2014_Nene
#> 7                                                           2015_PhD_Gutmann_Roberts
#> 8                                                               2016_Diaccia_Botrona
#> 9                                                                        2017_Fremur
#> 10                                                                    2019_Grotenete
#> 11                                                                   2020_PhD_Winter
#> 12                                                                     Aberdeen 2017
#> 13                                      Acoustic Ecology of Pelagic Fish Communities
#> 14                                                                      Albertkanaal
#> 15                                                              Albertkanaal_VPS_Ham
#> 16                                                                        ARAISOLA02
#> 17                                                                        ARAISOLA03
#> 18                                                                               Ars
#> 19                                                                          Artevigo
#> 20                                                 Azorean acoustic receiver network
#> 21                                                                            BECORV
#> 22                           Belwind temporary acoustic receiver network 2020 - 2021
#> 23                                                                      Bovenschelde
#> 24                                                                              BPNS
#> 25                                                           Brasem Ijsselmeergebied
#> 26                          Connectivity of coastal fish in the French Mediterranean
#> 27                                                                        Conon 2017
#> 28                                                                       cpodnetwork
#> 29                                                                     DAbecasis PhD
#> 30                                                                     DAK 2018-2019
#> 31                                                          Dalalven Smolt Migration
#> 32                                                                  Danube_Sturgeons
#> 33                                                                Delfzijl 2014-2015
#> 34                                                                             Demer
#> 35                                                                      Deveron 2016
#> 36                                                                             Dijle
#> 37                                                                        DTU-Skjern
#> 38                                                  Eel migration from source to sea
#> 39                                                                      Eemskanaal I
#> 40                                                                     Eemskanaal II
#> 41                                                                    Eemskanaal III
#> 42                                                                              EMMN
#> 43                                                                              ESGL
#> 44                Fisheries Innovation for sustainable SHared INTerchannEL resources
#> 45                                           Fish migration from source to Waddensea
#> 46                                                               Friesland 2011-2012
#> 47                                     Galician (NW Spain) acoustic Tracking Network
#> 48                                                                   IG II Fish Pass
#> 49                                                                             Iller
#> 50                                                                     Inforbiomares
#> 51                                                                   INTERREG_CANAPE
#> 52                                                      Jersey Coastal Fish Tracking
#> 53                                                               JSATS-PalmaBay-2019
#> 54                                                    Kalmar Baltic Tracking Network
#> 55                                                                    Kornwerderzand
#> 56                                                                       Lauwersmeer
#> 57                                                                     Leopoldkanaal
#> 58                                                                         Lifewatch
#> 59                                                                          Limfjord
#> 60                                              Ljusne Bay Salmon Spawning Migration
#> 61  Long-term array MOnitoring BEhaviour in Illas Atlánticas National Park, NW Spain
#> 62                                                                           MacFish
#> 63                                                           Marble vs rainbow trout
#> 64                                                                        MBA_Massmo
#> 65                                                                       MBA_Wavehub
#> 66                                                                      MBA_Whitsand
#> 67                                                                            MEPNSW
#> 68                                                           Middle_Danube_sturgeons
#> 69                                                                       MIGRATOEBRE
#> 70                                                                       Mobula Imar
#> 71                                                                  Moray Firth 2016
#> 72      Movement, Range Size and Habitat Utilization of magafish in Turkey 2019-2021
#> 73                                                                          NARVAEEL
#> 74                                                          Noordzeekanaal 2017-2018
#> 75                                                           Northern_Norfolk_Broads
#> 76                                                                  North_sea_wrecks
#> 77                                                                     NTNU-Gaulosen
#> 78                                                                        Orsted Cod
#> 79                                                                   OTN-Hemnfjorden
#> 80                                                                  OTN-Tosenfjorden
#> 81                                                                     paintedcomber
#> 82                                                                              PC4C
#> 83                                                                 PhD Barbara Koeck
#> 84                                                                 PhD Jeremy Pastor
#> 85                                                                         PhD_Nolan
#> 86                                                                          PhysFish
#> 87                        Pilot permanent deployment acoustic array Gibraltar Strait
#> 88                                                                   PTN/PROTECT2012
#> 89                                                                   PTN/PROTECT2013
#> 90                                                            PTN-Silver eel-Mondego
#> 91   Redes de Sensores submarinos autonomos para la medida de indicadores biologicos
#> 92                                                                          Reelease
#> 93                                                                            ResMed
#> 94                                              River Usk receiver network 2021-2024
#> 95                                                              Ruim_baan_voor_vis_2
#> 96                                                                       Running_eel
#> 97                                                                        Saeftinghe
#> 98                                                                           SAMARCH
#> 99                                                                        SARTELARAM
#> 100                                                                         SARTELTG
#> 101                                                                    SARTELZINGARO
#> 102                                                                         Scytrack
#> 103                                                                       SeaMonitor
#> 104                                                                   Semp Lithuania
#> 105                                                                 Siganid_East_Med
#> 106                                                               Siganid_Gulf_Aqaba
#> 107              Silver Eel downstream migration in the Walloon Meuse - 2021 to 2024
#> 108                                                                        Skye 2017
#> 109                                                             SMOLTRACK-I-Engeland
#> 110                                                            SMOLTRACK-II-Engeland
#> 111         Spatial Ecology of spurdog (Squalus acanthias) in a Scottish loch system
#> 112                                                                        SPAWNSEIS
#> 113                                                                S_rivulatus_Aqaba
#> 114                                                S_rivulatus_Eastern_Mediterranean
#> 115                                                  Sud_le_Impuls_voor_Diadrome_Vis
#> 116                                                    Sud_le_Innovatieve_monitoring
#> 117                                                                  SVNL-FISH-WATCH
#> 118                                                                   Swimway Vechte
#> 119                                                                             TeRi
#> 120                                                                     Thorntonbank
#> 121                                                             Unlocking the Severn
#> 122                                                           Vissen_voor_verbinding
#> 123                                                                  Westerschelde 1
#> 124                                                                  Westerschelde 2
#> 125                                                                  Westerschelde 3
#> 126                                      Yellow eel movement behaviour in the Azores
#> 127                                                                       Zeeschelde
#> 128                                                                 2010_phd_reubens
#> 129                                                            2010_phd_reubens_sync
#> 130                                                                  2011 Rivierprik
#> 131                                                               2012 Leopoldkanaal
#> 132                                                                2013 Albertkanaal
#> 133                                                                       2014 Demer
#> 134                                                        2015_Albertkanaal_VPS_Ham
#> 135                                                                       2015 Dijle
#> 136                                                                     2015_Homarus
#> 137                                                            2015 PhD Verhelst_cod
#> 138                                                            2015 PhD Verhelst_eel
#> 139                                                        2016_Albertkanaal_VPS_Ham
#> 140                                                               2016 PhD Vergeynst
#> 141                                                        2017_Albertkanaal_VPS_Ham
#> 142                                                                       ARAISOLA02
#> 143                                                                       ARAISOLA03
#> 144                                                                              Ars
#> 145                                                                         Artevigo
#> 146      Comparative study of marble and rainbow trout behaviour in Kanomljica River
#> 147                                                                    DAbecasis PhD
#> 148                                                                 Danube_Sturgeons
#> 149                                                                             EMMN
#> 150                                                                             ESGL
#> 151                                                              JSATS-PalmaBay-2019
#> 152                                                                         Limfjord
#> 153                                                          Marble vs rainbow trout
#> 154                                                                       MBA_Massmo
#> 155                                                                      MBA_Wavehub
#> 156                                                                     MBA_Whitsand
#> 157                                                          Middle_Danube_sturgeons
#> 158                                                                      Mobula Imar
#> 159                                                   North Aegean Anguilla Tracking
#> 160                                                                    NTNU-Gaulosen
#> 161                                                           Ocean Tracking Network
#> 162                                                                       Orsted Cod
#> 163                                                                  OTN-Hemnfjorden
#> 164                                                                 OTN-Tosenfjorden
#> 165                                                                PhD Barbara Koeck
#> 166                                                                PhD Jeremy Pastor
#> 167                                                                         PhysFish
#> 168                                                                       SARTELARAM
#> 169                                                                         SARTELTG
#> 170                                                                    SARTELZINGARO
#> 171                                                                         Scytrack
#> 172                                                                   Semp Lithuania
#> 173                                                               Siganid_Gulf_Aqaba
#> 174              Silver Eel downstream migration in the Walloon Meuse - 2021 to 2024
#> 175         Spatial Ecology of spurdog (Squalus acanthias) in a Scottish loch system
#> 176                                                S_rivulatus_Eastern_Mediterranean
#> 177                                                                             TeRi
#>        type
#> 1   network
#> 2   network
#> 3   network
#> 4   network
#> 5   network
#> 6   network
#> 7   network
#> 8   network
#> 9   network
#> 10  network
#> 11  network
#> 12  network
#> 13  network
#> 14  network
#> 15  network
#> 16  network
#> 17  network
#> 18  network
#> 19  network
#> 20  network
#> 21  network
#> 22  network
#> 23  network
#> 24  network
#> 25  network
#> 26  network
#> 27  network
#> 28  network
#> 29  network
#> 30  network
#> 31  network
#> 32  network
#> 33  network
#> 34  network
#> 35  network
#> 36  network
#> 37  network
#> 38  network
#> 39  network
#> 40  network
#> 41  network
#> 42  network
#> 43  network
#> 44  network
#> 45  network
#> 46  network
#> 47  network
#> 48  network
#> 49  network
#> 50  network
#> 51  network
#> 52  network
#> 53  network
#> 54  network
#> 55  network
#> 56  network
#> 57  network
#> 58  network
#> 59  network
#> 60  network
#> 61  network
#> 62  network
#> 63  network
#> 64  network
#> 65  network
#> 66  network
#> 67  network
#> 68  network
#> 69  network
#> 70  network
#> 71  network
#> 72  network
#> 73  network
#> 74  network
#> 75  network
#> 76  network
#> 77  network
#> 78  network
#> 79  network
#> 80  network
#> 81  network
#> 82  network
#> 83  network
#> 84  network
#> 85  network
#> 86  network
#> 87  network
#> 88  network
#> 89  network
#> 90  network
#> 91  network
#> 92  network
#> 93  network
#> 94  network
#> 95  network
#> 96  network
#> 97  network
#> 98  network
#> 99  network
#> 100 network
#> 101 network
#> 102 network
#> 103 network
#> 104 network
#> 105 network
#> 106 network
#> 107 network
#> 108 network
#> 109 network
#> 110 network
#> 111 network
#> 112 network
#> 113 network
#> 114 network
#> 115 network
#> 116 network
#> 117 network
#> 118 network
#> 119 network
#> 120 network
#> 121 network
#> 122 network
#> 123 network
#> 124 network
#> 125 network
#> 126 network
#> 127 network
#> 128  animal
#> 129  animal
#> 130  animal
#> 131  animal
#> 132  animal
#> 133  animal
#> 134  animal
#> 135  animal
#> 136  animal
#> 137  animal
#> 138  animal
#> 139  animal
#> 140  animal
#> 141  animal
#> 142  animal
#> 143  animal
#> 144  animal
#> 145  animal
#> 146  animal
#> 147  animal
#> 148  animal
#> 149  animal
#> 150  animal
#> 151  animal
#> 152  animal
#> 153  animal
#> 154  animal
#> 155  animal
#> 156  animal
#> 157  animal
#> 158  animal
#> 159  animal
#> 160  animal
#> 161  animal
#> 162  animal
#> 163  animal
#> 164  animal
#> 165  animal
#> 166  animal
#> 167  animal
#> 168  animal
#> 169  animal
#> 170  animal
#> 171  animal
#> 172  animal
#> 173  animal
#> 174  animal
#> 175  animal
#> 176  animal
#> 177  animal