Retrieve an overview of the ETN data.

Retrieves the name of network/animals? To get an account, register via the Lifewatch RShiny registration webpage.

listEtnProjects(usr = NULL, pwd = NULL)

Arguments

usr
Username to connect to ETN database

pwd
Password to connect to ETN database

Value

Dataframe with name and type of networks.

Examples

listEtnProjects()

#> No encoding supplied: defaulting to UTF-8.

#>                                                                                name
#> 1                                                                       2004_Gudena
#> 2                                                                        2011_Loire
#> 3                                                                       2011_Warnow
#> 4                                                                        2013_Foyle
#> 5                                                                         2013_Maas
#> 6                                                                         2014_Nene
#> 7                                                          2015_PhD_Gutmann_Roberts
#> 8                                                              2016_Diaccia_Botrona
#> 9                                                                       2017_Fremur
#> 10                                                                   2019_Grotenete
#> 11                                                                  2020_PhD_Winter
#> 12                                                                    Aberdeen 2017
#> 13                                     Acoustic Ecology of Pelagic Fish Communities
#> 14                                                                     Albertkanaal
#> 15                                                             Albertkanaal_VPS_Ham
#> 16                                                                        Apelafico
#> 17                                                                       ARAISOLA02
#> 18                                                                       ARAISOLA03
#> 19                                                                              Ars
#> 20                                                                         Artevigo
#> 21                                                Azorean acoustic receiver network
#> 22                                                                           BECORV
#> 23                          Belwind temporary acoustic receiver network 2020 - 2021
#> 24                                                                     Bovenschelde
#> 25                                                                             BPNS
#> 26                                                          Brasem Ijsselmeergebied
#> 27                         Connectivity of coastal fish in the French Mediterranean
#> 28                                                                       Conon 2017
#> 29                                                                      cpodnetwork
#> 30                                                                    DAbecasis PhD
#> 31                                                                    DAK 2018-2019
#> 32                                                         Dalalven Smolt Migration
#> 33                                                                 Danube_Sturgeons
#> 34                                                               Delfzijl 2014-2015
#> 35                                                                            Demer
#> 36                                                                     Deveron 2016
#> 37                                                                            Dijle
#> 38                                                                       DTU-Skjern
#> 39                                                 Eel migration from source to sea
#> 40                                                                     Eemskanaal I
#> 41                                                                    Eemskanaal II
#> 42                                                                   Eemskanaal III
#> 43                                                                             EMMN
#> 44                                                                             ESGL
#> 45                                          Fish migration from source to Waddensea
#> 46                                                              Friesland 2011-2012
#> 47                                                                  IG II Fish Pass
#> 48                                                                            Iller
#> 49                                                                    Inforbiomares
#> 50                                                                  INTERREG_CANAPE
#> 51                                                     Jersey Coastal Fish Tracking
#> 52                                                              JSATS-PalmaBay-2019
#> 53                                                   Kalmar Baltic Tracking Network
#> 54                                                                   Kornwerderzand
#> 55                                                                      Lauwersmeer
#> 56                                                                    Leopoldkanaal
#> 57                                                                        Lifewatch
#> 58                                                                         Limfjord
#> 59                                             Ljusne Bay Salmon Spawning Migration
#> 60                                                                          MacFish
#> 61                                                          Marble vs rainbow trout
#> 62                                                                       MBA_Massmo
#> 63                                                                      MBA_Wavehub
#> 64                                                                     MBA_Whitsand
#> 65                                                                           MEPNSW
#> 66                                                          Middle_Danube_sturgeons
#> 67                                                                      MIGRATOEBRE
#> 68                                                                           MOBEIA
#> 69                                                                      Mobula Imar
#> 70                                                                 Moray Firth 2016
#> 71                                                                         NARVAEEL
#> 72                                                         Noordzeekanaal 2017-2018
#> 73                                                          Northern_Norfolk_Broads
#> 74                                                                 North_sea_wrecks
#> 75                                                                    NTNU-Gaulosen
#> 76                                                                       Orsted Cod
#> 77                                                                  OTN-Hemnfjorden
#> 78                                                                 OTN-Tosenfjorden
#> 79                                                                    paintedcomber
#> 80                                                                             PC4C
#> 81                                                                PhD Barbara Koeck
#> 82                                                                PhD Jeremy Pastor
#> 83                                                                        PhD_Nolan
#> 84                                                                         PhysFish
#> 85                       Pilot permanent deployment acoustic array Gibraltar Strait
#> 86                                                                  PTN/PROTECT2012
#> 87                                                                  PTN/PROTECT2013
#> 88                                                           PTN-Silver eel-Mondego
#> 89  Redes de Sensores submarinos autonomos para la medida de indicadores biologicos
#> 90                                                                         Reelease
#> 91                                                                           ResMed
#> 92                                             River Usk receiver network 2021-2024
#> 93                                                             Ruim_baan_voor_vis_2
#> 94                                                                      Running_eel
#> 95                                                                       Saeftinghe
#> 96                                                                          SAMARCH
#> 97                                                                       SARTELARAM
#> 98                                                                         SARTELTG
#> 99                                                                    SARTELZINGARO
#> 100                                                                        Scytrack
#> 101                                                                      SeaMonitor
#> 102                                                                  Semp Lithuania
#> 103                                                                Siganid_East_Med
#> 104                                                              Siganid_Gulf_Aqaba
#> 105                                                                       Skye 2017
#> 106                                                            SMOLTRACK-I-Engeland
#> 107                                                           SMOLTRACK-II-Engeland
#> 108        Spatial Ecology of spurdog (Squalus acanthias) in a Scottish loch system
#> 109                                                                       SPAWNSEIS
#> 110                                                 Sud_le_Impuls_voor_Diadrome_Vis
#> 111                                                   Sud_le_Innovatieve_monitoring
#> 112                                                                 SVNL-FISH-WATCH
#> 113                                                                  Swimway Vechte
#> 114                                                                            TeRi
#> 115                                                                    Thorntonbank
#> 116                                                            Unlocking the Severn
#> 117                                                          Vissen_voor_verbinding
#> 118                                                                 Westerschelde 1
#> 119                                                                 Westerschelde 2
#> 120                                                                 Westerschelde 3
#> 121                                                                      Zeeschelde
#> 122                                                                2010_phd_reubens
#> 123                                                           2010_phd_reubens_sync
#> 124                                                                 2011 Rivierprik
#> 125                                                              2012 Leopoldkanaal
#> 126                                                               2013 Albertkanaal
#> 127                                                                      2014 Demer
#> 128                                                       2015_Albertkanaal_VPS_Ham
#> 129                                                                      2015 Dijle
#> 130                                                                    2015_Homarus
#> 131                                                           2015 PhD Verhelst_cod
#> 132                                                           2015 PhD Verhelst_eel
#> 133                                                       2016_Albertkanaal_VPS_Ham
#> 134                                                              2016 PhD Vergeynst
#> 135                                                       2017_Albertkanaal_VPS_Ham
#> 136                                                                      ARAISOLA02
#> 137                                                                      ARAISOLA03
#> 138                                                                             Ars
#> 139                                                                        Artevigo
#> 140     Comparative study of marble and rainbow trout behaviour in Kanomljica River
#> 141                                                                   DAbecasis PhD
#> 142                                                                Danube_Sturgeons
#> 143                                                                            EMMN
#> 144                                                                            ESGL
#> 145                                                             JSATS-PalmaBay-2019
#> 146                                                                        Limfjord
#> 147                                                         Marble vs rainbow trout
#> 148                                                                      MBA_Massmo
#> 149                                                                     MBA_Wavehub
#> 150                                                                    MBA_Whitsand
#> 151                                                         Middle_Danube_sturgeons
#> 152                                                                     Mobula Imar
#> 153                                                  North Aegean Anguilla Tracking
#> 154                                                                   NTNU-Gaulosen
#> 155                                                          Ocean Tracking Network
#> 156                                                                      Orsted Cod
#> 157                                                                 OTN-Hemnfjorden
#> 158                                                                OTN-Tosenfjorden
#> 159                                                               PhD Barbara Koeck
#> 160                                                               PhD Jeremy Pastor
#> 161                                                                        PhysFish
#> 162                                                                      SARTELARAM
#> 163                                                                        SARTELTG
#> 164                                                                   SARTELZINGARO
#> 165                                                                        Scytrack
#> 166                                                                  Semp Lithuania
#> 167                                                              Siganid_Gulf_Aqaba
#> 168        Spatial Ecology of spurdog (Squalus acanthias) in a Scottish loch system
#> 169                                                                            TeRi
#>        type
#> 1   network
#> 2   network
#> 3   network
#> 4   network
#> 5   network
#> 6   network
#> 7   network
#> 8   network
#> 9   network
#> 10  network
#> 11  network
#> 12  network
#> 13  network
#> 14  network
#> 15  network
#> 16  network
#> 17  network
#> 18  network
#> 19  network
#> 20  network
#> 21  network
#> 22  network
#> 23  network
#> 24  network
#> 25  network
#> 26  network
#> 27  network
#> 28  network
#> 29  network
#> 30  network
#> 31  network
#> 32  network
#> 33  network
#> 34  network
#> 35  network
#> 36  network
#> 37  network
#> 38  network
#> 39  network
#> 40  network
#> 41  network
#> 42  network
#> 43  network
#> 44  network
#> 45  network
#> 46  network
#> 47  network
#> 48  network
#> 49  network
#> 50  network
#> 51  network
#> 52  network
#> 53  network
#> 54  network
#> 55  network
#> 56  network
#> 57  network
#> 58  network
#> 59  network
#> 60  network
#> 61  network
#> 62  network
#> 63  network
#> 64  network
#> 65  network
#> 66  network
#> 67  network
#> 68  network
#> 69  network
#> 70  network
#> 71  network
#> 72  network
#> 73  network
#> 74  network
#> 75  network
#> 76  network
#> 77  network
#> 78  network
#> 79  network
#> 80  network
#> 81  network
#> 82  network
#> 83  network
#> 84  network
#> 85  network
#> 86  network
#> 87  network
#> 88  network
#> 89  network
#> 90  network
#> 91  network
#> 92  network
#> 93  network
#> 94  network
#> 95  network
#> 96  network
#> 97  network
#> 98  network
#> 99  network
#> 100 network
#> 101 network
#> 102 network
#> 103 network
#> 104 network
#> 105 network
#> 106 network
#> 107 network
#> 108 network
#> 109 network
#> 110 network
#> 111 network
#> 112 network
#> 113 network
#> 114 network
#> 115 network
#> 116 network
#> 117 network
#> 118 network
#> 119 network
#> 120 network
#> 121 network
#> 122  animal
#> 123  animal
#> 124  animal
#> 125  animal
#> 126  animal
#> 127  animal
#> 128  animal
#> 129  animal
#> 130  animal
#> 131  animal
#> 132  animal
#> 133  animal
#> 134  animal
#> 135  animal
#> 136  animal
#> 137  animal
#> 138  animal
#> 139  animal
#> 140  animal
#> 141  animal
#> 142  animal
#> 143  animal
#> 144  animal
#> 145  animal
#> 146  animal
#> 147  animal
#> 148  animal
#> 149  animal
#> 150  animal
#> 151  animal
#> 152  animal
#> 153  animal
#> 154  animal
#> 155  animal
#> 156  animal
#> 157  animal
#> 158  animal
#> 159  animal
#> 160  animal
#> 161  animal
#> 162  animal
#> 163  animal
#> 164  animal
#> 165  animal
#> 166  animal
#> 167  animal
#> 168  animal
#> 169  animal