Retrieve an overview of the ETN data.

Retrieves the name of network/animals? To get an account, register via the Lifewatch RShiny registration webpage.

listEtnProjects(usr = NULL, pwd = NULL)

Arguments

usr

Username to connect to ETN database

pwd

Password to connect to ETN database

Value

Dataframe with name and type of networks.

Examples

listEtnProjects()
#> - Query mode: Post request to OpenCPU server
#> No encoding supplied: defaulting to UTF-8.
#>                                                                                 name
#> 1                                                                        2004_Gudena
#> 2                                                                         2011_Loire
#> 3                                                                        2011_Warnow
#> 4                                                                         2013_Foyle
#> 5                                                                          2013_Maas
#> 6                                                                          2014_Nene
#> 7                                                           2015_PhD_Gutmann_Roberts
#> 8                                                               2016_Diaccia_Botrona
#> 9                                                                        2017_Fremur
#> 10                                                                    2019_Grotenete
#> 11                                                                   2020_PhD_Winter
#> 12                                                                     Aberdeen 2017
#> 13                                      Acoustic Ecology of Pelagic Fish Communities
#> 14                                                                      Albertkanaal
#> 15                                                              Albertkanaal_VPS_Ham
#> 16                                                                        ARAISOLA02
#> 17                                                                        ARAISOLA03
#> 18                                                                               Ars
#> 19                                                                          Artevigo
#> 20                                                 Azorean acoustic receiver network
#> 21                                                         Balearic Tracking Network
#> 22                   Balearic Tracking Network - Spanish Deep Water Receiver Network
#> 23                                                                            BECORV
#> 24                           Belwind temporary acoustic receiver network 2020 - 2021
#> 25                                                                      Bovenschelde
#> 26                                                                              BPNS
#> 27                                                           Brasem Ijsselmeergebied
#> 28                                               Brown trout in the Falkland Islands
#> 29                                                      Canaries Angel Shark Project
#> 30                          Connectivity of coastal fish in the French Mediterranean
#> 31                                                                        Conon 2017
#> 32                                                                       cpodnetwork
#> 33                                                                     DAbecasis PhD
#> 34                                                                     DAK 2018-2019
#> 35                                                          Dalalven Smolt Migration
#> 36                                                           Danish Belts and Sounds
#> 37                                                                  Danube_Sturgeons
#> 38                                                                Delfzijl 2014-2015
#> 39                                                                             Demer
#> 40                                                                      Deveron 2016
#> 41                                                                             Dijle
#> 42                                                                        DTU-Skjern
#> 43                                                  Eel migration from source to sea
#> 44                                                                      Eemskanaal I
#> 45                                                                     Eemskanaal II
#> 46                                                                    Eemskanaal III
#> 47                                                                              EMMN
#> 48                                                                              ESGL
#> 49                Fisheries Innovation for sustainable SHared INTerchannEL resources
#> 50                                           Fish migration from source to Waddensea
#> 51                                                               Friesland 2011-2012
#> 52                                     Galician (NW Spain) acoustic Tracking Network
#> 53                                                                   IG II Fish Pass
#> 54                                                                             Iller
#> 55                                                                     Inforbiomares
#> 56                                                                   INTERREG_CANAPE
#> 57                                                      Jersey Coastal Fish Tracking
#> 58                                                               JSATS-PalmaBay-2019
#> 59                                                    Kalmar Baltic Tracking Network
#> 60                                                                    Kornwerderzand
#> 61                        Land-use change and human activity on anadromous salmonids
#> 62                                                                       Lauwersmeer
#> 63                                                                     Leopoldkanaal
#> 64                                                                         Lifewatch
#> 65                                                                          Limfjord
#> 66                                              Ljusne Bay Salmon Spawning Migration
#> 67  Long-term array MOnitoring BEhaviour in Illas Atlánticas National Park, NW Spain
#> 68                                                                           MacFish
#> 69                                                           Marble vs rainbow trout
#> 70                                                                        MBA_Massmo
#> 71                                                                       MBA_Wavehub
#> 72                                                                      MBA_Whitsand
#> 73                                                                            MEPNSW
#> 74                                                           Middle_Danube_sturgeons
#> 75    Migration of pikeperch within the Lake Peipus complex (Estonia) from 2021-2023
#> 76                                                                       MIGRATOEBRE
#> 77                                                                       Mobula Imar
#> 78                                                                  Moray Firth 2016
#> 79      Movement, Range Size and Habitat Utilization of magafish in Turkey 2019-2021
#> 80                                                                          NARVAEEL
#> 81                                                          Noordzeekanaal 2017-2018
#> 82                                                           Northern_Norfolk_Broads
#> 83                                                                  North_sea_wrecks
#> 84                                                                     NTNU-Gaulosen
#> 85                                                        NTNU River Nidelva Project
#> 86    Observing the effect of group dynamics on individual behavior in basking shark
#> 87                                                                        Orsted Cod
#> 88                                                                   OTN-Hemnfjorden
#> 89                                                           OTN Loan - Beiarfjorden
#> 90                                                                  OTN-Tosenfjorden
#> 91                                                            OTN VR2W Loan - Norway
#> 92                                                                     paintedcomber
#> 93                                                                              PC4C
#> 94                                                                 PhD Barbara Koeck
#> 95                                                                 PhD Jeremy Pastor
#> 96                                                                         PhD_Nolan
#> 97                                                                          PhysFish
#> 98                        Pilot permanent deployment acoustic array Gibraltar Strait
#> 99                                                                   PTN/PROTECT2012
#> 100                                                                  PTN/PROTECT2013
#> 101                                                           PTN-Silver eel-Mondego
#> 102  Redes de Sensores submarinos autonomos para la medida de indicadores biologicos
#> 103                                                                         Reelease
#> 104                                                                           ResMed
#> 105                                             River Usk receiver network 2021-2024
#> 106                                                             Ruim_baan_voor_vis_2
#> 107                                                                      Running_eel
#> 108                                                                       Saeftinghe
#> 109                                                                          SAMARCH
#> 110                                                                       SARTELARAM
#> 111                                                                         SARTELTG
#> 112                                                                    SARTELZINGARO
#> 113                                                                         Scytrack
#> 114                                                                       SeaMonitor
#> 115            SeaMonitor: Costs and Pathways of Marine Migration in Atlantic salmon
#> 116                                                                   Semp Lithuania
#> 117                                                                 Siganid_East_Med
#> 118                                                               Siganid_Gulf_Aqaba
#> 119              Silver Eel downstream migration in the Walloon Meuse - 2021 to 2024
#> 120                                                                        Skye 2017
#> 121                                                             SMOLTRACK-I-Engeland
#> 122                                                            SMOLTRACK-II-Engeland
#> 123                                                      Southern Norway brown trout
#> 124         Spatial Ecology of spurdog (Squalus acanthias) in a Scottish loch system
#> 125 Spawning migration of pike in MatsaluBay and RiverKasari (Baltic Sea, 2018-2020)
#> 126  Spawning migration of pike in Teorehe-Sauemeri floodplain (BalticSea 2017-2019)
#> 127                                                                        SPAWNSEIS
#> 128                                                                S_rivulatus_Aqaba
#> 129                                                S_rivulatus_Eastern_Mediterranean
#> 130                                                  Sud_le_Impuls_voor_Diadrome_Vis
#> 131                                                    Sud_le_Innovatieve_monitoring
#> 132                                                                  SVNL-FISH-WATCH
#> 133 Swansea Uni acoustic receiver array: Bristol Channel, Swansea Bay and River Tawe
#> 134                                                                   Swimway Vechte
#> 135                                                                             TeRi
#> 136                                                     The Bergen Telemetry Network
#> 137                                                                     Thorntonbank
#> 138                                                             Unlocking the Severn
#> 139                                                           Vissen_voor_verbinding
#> 140                                           VMT Loan - SeaMonitor Marine Institute
#> 141                                                                  Westerschelde 1
#> 142                                                                  Westerschelde 2
#> 143                                                                  Westerschelde 3
#> 144                                      Yellow eel movement behaviour in the Azores
#> 145                                                                       Zeeschelde
#> 146                                                                 2010_phd_reubens
#> 147                                                            2010_phd_reubens_sync
#> 148                                                                  2011 Rivierprik
#> 149                                                               2012 Leopoldkanaal
#> 150                                                                2013 Albertkanaal
#> 151                                                                       2014 Demer
#> 152                                                        2015_Albertkanaal_VPS_Ham
#> 153                                                                       2015 Dijle
#> 154                                                                     2015_Homarus
#> 155                                                            2015 PhD Verhelst_cod
#> 156                                                            2015 PhD Verhelst_eel
#> 157                                                        2016_Albertkanaal_VPS_Ham
#> 158                                                               2016 PhD Vergeynst
#> 159                                                        2017_Albertkanaal_VPS_Ham
#> 160                                                                       ARAISOLA02
#> 161                                                                       ARAISOLA03
#> 162                                                                              Ars
#> 163                                                                         Artevigo
#> 164                                                        Balearic Tracking Network
#> 165                  Balearic Tracking Network - Spanish Deep Water Receiver Network
#> 166      Comparative study of marble and rainbow trout behaviour in Kanomljica River
#> 167                                                                    DAbecasis PhD
#> 168                                                                 Danube_Sturgeons
#> 169                                                                             EMMN
#> 170                                                                             ESGL
#> 171                                                              JSATS-PalmaBay-2019
#> 172                                                                         Limfjord
#> 173                                                          Marble vs rainbow trout
#> 174                                                                       MBA_Massmo
#> 175                                                                      MBA_Wavehub
#> 176                                                                     MBA_Whitsand
#> 177                                                          Middle_Danube_sturgeons
#> 178                                                                      Mobula Imar
#> 179                                                   North Aegean Anguilla Tracking
#> 180                                                                    NTNU-Gaulosen
#> 181                                                       NTNU River Nidelva Project
#> 182                                                           Ocean Tracking Network
#> 183                                                                       Orsted Cod
#> 184                                                                  OTN-Hemnfjorden
#> 185                                                                 OTN-Tosenfjorden
#> 186                                                           OTN VR2W Loan - Norway
#> 187                                                                PhD Barbara Koeck
#> 188                                                                PhD Jeremy Pastor
#> 189                                                                     PhD_Marrocco
#> 190                                                                         PhysFish
#> 191                                                                       SARTELARAM
#> 192                                                                         SARTELTG
#> 193                                                                    SARTELZINGARO
#> 194                                                                         Scytrack
#> 195                                                                   Semp Lithuania
#> 196                                                               Siganid_Gulf_Aqaba
#> 197              Silver Eel downstream migration in the Walloon Meuse - 2021 to 2024
#> 198         Spatial Ecology of spurdog (Squalus acanthias) in a Scottish loch system
#> 199                                                S_rivulatus_Eastern_Mediterranean
#> 200                                                                             TeRi
#>        type
#> 1   network
#> 2   network
#> 3   network
#> 4   network
#> 5   network
#> 6   network
#> 7   network
#> 8   network
#> 9   network
#> 10  network
#> 11  network
#> 12  network
#> 13  network
#> 14  network
#> 15  network
#> 16  network
#> 17  network
#> 18  network
#> 19  network
#> 20  network
#> 21  network
#> 22  network
#> 23  network
#> 24  network
#> 25  network
#> 26  network
#> 27  network
#> 28  network
#> 29  network
#> 30  network
#> 31  network
#> 32  network
#> 33  network
#> 34  network
#> 35  network
#> 36  network
#> 37  network
#> 38  network
#> 39  network
#> 40  network
#> 41  network
#> 42  network
#> 43  network
#> 44  network
#> 45  network
#> 46  network
#> 47  network
#> 48  network
#> 49  network
#> 50  network
#> 51  network
#> 52  network
#> 53  network
#> 54  network
#> 55  network
#> 56  network
#> 57  network
#> 58  network
#> 59  network
#> 60  network
#> 61  network
#> 62  network
#> 63  network
#> 64  network
#> 65  network
#> 66  network
#> 67  network
#> 68  network
#> 69  network
#> 70  network
#> 71  network
#> 72  network
#> 73  network
#> 74  network
#> 75  network
#> 76  network
#> 77  network
#> 78  network
#> 79  network
#> 80  network
#> 81  network
#> 82  network
#> 83  network
#> 84  network
#> 85  network
#> 86  network
#> 87  network
#> 88  network
#> 89  network
#> 90  network
#> 91  network
#> 92  network
#> 93  network
#> 94  network
#> 95  network
#> 96  network
#> 97  network
#> 98  network
#> 99  network
#> 100 network
#> 101 network
#> 102 network
#> 103 network
#> 104 network
#> 105 network
#> 106 network
#> 107 network
#> 108 network
#> 109 network
#> 110 network
#> 111 network
#> 112 network
#> 113 network
#> 114 network
#> 115 network
#> 116 network
#> 117 network
#> 118 network
#> 119 network
#> 120 network
#> 121 network
#> 122 network
#> 123 network
#> 124 network
#> 125 network
#> 126 network
#> 127 network
#> 128 network
#> 129 network
#> 130 network
#> 131 network
#> 132 network
#> 133 network
#> 134 network
#> 135 network
#> 136 network
#> 137 network
#> 138 network
#> 139 network
#> 140 network
#> 141 network
#> 142 network
#> 143 network
#> 144 network
#> 145 network
#> 146  animal
#> 147  animal
#> 148  animal
#> 149  animal
#> 150  animal
#> 151  animal
#> 152  animal
#> 153  animal
#> 154  animal
#> 155  animal
#> 156  animal
#> 157  animal
#> 158  animal
#> 159  animal
#> 160  animal
#> 161  animal
#> 162  animal
#> 163  animal
#> 164  animal
#> 165  animal
#> 166  animal
#> 167  animal
#> 168  animal
#> 169  animal
#> 170  animal
#> 171  animal
#> 172  animal
#> 173  animal
#> 174  animal
#> 175  animal
#> 176  animal
#> 177  animal
#> 178  animal
#> 179  animal
#> 180  animal
#> 181  animal
#> 182  animal
#> 183  animal
#> 184  animal
#> 185  animal
#> 186  animal
#> 187  animal
#> 188  animal
#> 189  animal
#> 190  animal
#> 191  animal
#> 192  animal
#> 193  animal
#> 194  animal
#> 195  animal
#> 196  animal
#> 197  animal
#> 198  animal
#> 199  animal
#> 200  animal